Internal API

PowerSystems._pti_dtypes — Constant

lookup array of data types for PTI file sections given by field_name, as enumerated by PSS/e Program Operation Manual.

PowerSystems._pti_sections — Constant

A list of data file sections in the order that they appear in a PTI v33 file

PowerSystems.pm_component_status — Constant

maps component types to status parameters

PowerSystems.pm_component_status_inactive — Constant

maps component types to inactive status values

PowerSystems.PowerSystemTableData — Method

PowerSystemTableData(directory::AbstractString, base_power::Float64, user_descriptor_file::AbstractString; descriptor_file, generator_mapping_file, timeseries_metadata_file) -> PowerSystems.PowerSystemTableData

Reads in all the data stored in csv files The general format for data is folder: gen.csv branch.csv bus.csv .. load.csv

Arguments

directory::AbstractString: directory containing CSV files

base_power::Float64: base power for System

user_descriptor_file::AbstractString: customized input descriptor file

descriptor_file=POWER_SYSTEM_DESCRIPTOR_FILE: PowerSystems descriptor file

generator_mapping_file=GENERATOR_MAPPING_FILE: generator mapping configuration file

PowerSystems._FieldInfo — Type

Stores user-customized information for required dataframe columns.

Base.convert — Method

convert(_::Type{NamedTuple{(:min, :max),Tuple{Float64,Float64}}}, input::Tuple{Float64,Float64}) -> NamedTuple{(:min, :max),Tuple{Float64,Float64}}

Convert Tuple to Min Max Named Tuple

Base.convert — Method

convert(_::Type{NamedTuple{(:up, :down),Tuple{Float64,Float64}}}, input::Tuple{Float64,Float64}) -> NamedTuple{(:up, :down),Tuple{Float64,Float64}}

Convert Tuple to Up Down Named Tuple

InfrastructureSystems.convert_type — Method

convert_type(_::Type{T<:Device}, data::NamedTuple, component_cache::Dict) -> DynamicBranch

Creates a Device object by decoding the data that was in JSON. This data stores the values for buses and services as UUIDs, so this will lookup each in component_cache.

PowerSystems._add_dcline_costs! — Method

_add_dcline_costs!(data::Dict{String,Any})

adds dcline costs, if gen costs exist

PowerSystems._add_line_delimiter — Method

_add_line_delimiter(mp_line::AbstractString, start_char::Any, end_char::Any) -> AbstractString

PowerSystems._apply_func! — Method

_apply_func!(data::Dict{String,#s857} where #s857, key::String, func::Any) -> Any

PowerSystems._biggest_generator — Method

_biggest_generator(gens::Any) -> Any

find the largest active generator in the network

PowerSystems._bold — Method

_bold(s::String) -> String

Makes a string bold in the terminal

PowerSystems._calc_branch_flow_ac — Method

_calc_branch_flow_ac(data::Dict{String,#s849} where #s849) -> Dict{String,Any}

helper function for calcbranchflow_ac

PowerSystems._calc_branch_flow_dc — Method

_calc_branch_flow_dc(data::Dict{String,#s850} where #s850) -> Dict{String,Any}

helper function for calcbranchflow_dc

PowerSystems._calc_comp_lines — Method

_calc_comp_lines(component::Dict{String,#s855} where #s855) -> Array{Any,1}

compute lines in m and b from from pwl cost models

PowerSystems._calc_max_cost_index — Method

_calc_max_cost_index(data::Dict{String,#s853} where #s853) -> Any

PowerSystems._calc_power_balance — Method

_calc_power_balance(data::Dict{String,#s837} where #s837) -> Dict{String,Any}

helper function for calcpowerbalance

PowerSystems._check_conductors — Method

_check_conductors(data::Dict{String,#s857} where #s857)

PowerSystems._check_keys — Method

_check_keys(data::Any, keys::Any)

PowerSystems._compare_numbers — Method

_compare_numbers(v1::Any, v2::Any) -> Bool

tests if two numbers are equal, up to floating point precision

PowerSystems._correct_cost_function! — Method

_correct_cost_function!(id::Any, comp::Any, type_name::Any) -> Bool

PowerSystems._correct_reference_buses! — Method

_correct_reference_buses!(data::Dict{String,#s846} where #s846)

PowerSystems._create_starbus_from_transformer — Method

_create_starbus_from_transformer(pm_data::Dict, transformer::Dict) -> Dict{String,Any}

create_starbus(pm_data, transformer)

Creates a starbus from a given three-winding transformer. "sourceid" is given by `["busi", "name", "I", "J", "K", "CKT"]` where "bus_i" and "name" are the modified names for the starbus, and "I", "J", "K" and "CKT" come from the originating transformer, in the PSS(R)E transformer specification.

PowerSystems._dfs — Method

_dfs(i::Any, neighbors::Any, component_lookup::Any, touched::Any)

performs DFS on a graph

PowerSystems._export_cost_data — Method

_export_cost_data(io::IO, components::Dict{Int64,Dict}, prefix::String)

Export cost data

PowerSystems._export_extra_data — Function

_export_extra_data(io::IO, data::Dict{String,#s843} where #s843, component::Any)
_export_extra_data(io::IO, data::Dict{String,#s842} where #s842, component::Any, excluded_fields::Any; postfix)

Export fields of a component type

PowerSystems._extract_matlab_assignment — Method

_extract_matlab_assignment(string::AbstractString) -> Tuple{SubString,SubString}

breaks up matlab strings of the form 'name = value;'

PowerSystems._find_max_bus_id — Method

_find_max_bus_id(pm_data::Dict) -> Int64

_find_max_bus_id(pm_data)

Returns the maximum bus id in pm_data

PowerSystems._float2string — Method

_float2string(v::AbstractFloat, float_precision::Int64) -> String

converts a float value into a string of fixed precision

sprintf would do the job but this work around is needed because sprintf cannot take format strings during runtime

PowerSystems._get_bus_value — Method

_get_bus_value(bus_i::Any, field::Any, pm_data::Any) -> Any

_get_bus_value(bus_i, field, pm_data)

Returns the value of field of bus_i from the PowerModels data. Requires "bus" Dict to already be populated.

PowerSystems._get_contributing_devices — Method

_get_contributing_devices(sys::System, service::T<:Service) -> ServiceContributingDevices

Returns a ServiceContributingDevices object.

PowerSystems._get_default — Function

_get_default(dict::Any, key::Any) -> Any
_get_default(dict::Any, key::Any, default::Any) -> Any

Get a default value for dict entry

PowerSystems._get_line_elements — Method

_get_line_elements(line::AbstractString) -> Tuple{Array{_A,1} where _A,Union{String, SubString}}

_get_line_elements(line)

Internal function. Uses regular expressions to extract all separate data elements from a line of a PTI file and populate them into an Array{String}. Comments, typically indicated at the end of a line with a '/' character, are also extracted separately, and Array{Array{String}, String} is returned.

PowerSystems._get_load_data — Method

_get_load_data(sys::System, b::Bus) -> Tuple{Any,Any}

Obtain total load on bus b

PowerSystems._get_pm_dict_name — Method

_get_pm_dict_name(device_dict::Any) -> Any

Internal component name retreval from pm2ps_dict

PowerSystems._grey — Method

_grey(s::String) -> String

Makes a string grey in the terminal, does not seem to work well on Windows terminals more info can be found at https://en.wikipedia.org/wiki/ANSIescapecode

PowerSystems._import_remaining! — Method

_import_remaining!(data_out::Dict, data_in::Dict, import_all::Bool; exclude)

Imports remaining keys from data_in into data_out, excluding keys in exclude

PowerSystems._init_bus! — Method

_init_bus!(bus::Dict{String,Any}, id::Int64) -> Int64

_init_bus!(bus, id)

Initializes a bus of id id with default values given in the PSS(R)E specification.

PowerSystems._make_mixed_units! — Method

_make_mixed_units!(data::Dict{String,#s842} where #s842, mva_base::Real)

PowerSystems._make_per_unit! — Method

_make_per_unit!(data::Dict{String,#s842} where #s842, mva_base::Real)

PowerSystems._make_shaft — Method

_make_shaft(param_map::Any, val::Any) -> Any

Create a SingleMass shaft struct directly using the parameter mapping.

PowerSystems._matpower_to_powermodels! — Method

_matpower_to_powermodels!(mp_data::Dict{String,#s147} where #s147) -> Dict{String,#s147} where #s147

Converts a Matpower dict into a PowerModels dict

PowerSystems._merge_bus_name_data! — Method

_merge_bus_name_data!(data::Dict{String,Any}) -> Union{Nothing, Dict{String,Any}}

merges bus name data into buses, if names exist

PowerSystems._merge_generator_cost_data! — Method

_merge_generator_cost_data!(data::Dict{String,Any}) -> Union{Nothing, Dict{String,Any}}

merges generator cost functions into generator data, if costs exist

PowerSystems._merge_generic_data! — Method

_merge_generic_data!(data::Dict{String,Any})

merges Matpower tables based on the table extension syntax

PowerSystems._mp2pm_branch! — Method

_mp2pm_branch!(data::Dict{String,Any})

sets all branch transformer taps to 1.0, to simplify branch models

PowerSystems._mp2pm_dcline! — Method

_mp2pm_dcline!(data::Dict{String,Any})

adds pmin and pmax values at to and from buses

PowerSystems._mp_cost_data — Method

_mp_cost_data(cost_row::Any) -> Dict{String,Any}

PowerSystems._parse_dyr_components — Method

_parse_dyr_components(dyr_file::AbstractString) -> Dict{Int64,Any}

Parse a .dyr file directly from its name by constructing its dictionary of dictionaries.

PowerSystems._parse_dyr_components — Method

_parse_dyr_components(data::Dict) -> Dict{Int64,Any}

Parse dictionary of dictionaries of data (from _parse_dyr_file) into a dictionary of struct components. The function receives the parsed dictionary and construct a dictionary indexed by bus, that containts a dictionary with each dynamic generator components (indexed via its id).

Each dictionary indexed by id contains a vector with 5 of its components:

Machine

Shaft

AVR

TurbineGov

PSS

PowerSystems._parse_dyr_file — Method

_parse_dyr_file(file::AbstractString) -> Dict{Int64,Dict}

Parse .dyr file into a dictionary indexed by bus number. Each bus number key has a dictionary indexed by component type and id.

PowerSystems._parse_elements — Method

_parse_elements(elements::Array, dtypes::Array, defaults::Dict, section::AbstractString) -> Dict{String,Any}

This is an experimental method for parsing elements and setting defaults at the same time. It is not currently working but would reduce memory allocations if implemented correctly.

PowerSystems._parse_line_element! — Method

_parse_line_element!(data::Dict, elements::Array, section::AbstractString)

_parse_line_element!(data, elements, section)

Internal function. Parses a single "line" of data elements from a PTI file, as given by elements which is an array of the line, typically split at ,. Elements are parsed into data types given by section and saved into data::Dict.

PowerSystems._parse_matlab_cells — Method

_parse_matlab_cells(lines::Any, index::Any) -> Dict{_A,_B} where _B where _A

PowerSystems._parse_matlab_data — Method

_parse_matlab_data(lines::Any, index::Any, start_char::Any, end_char::Any) -> Dict{_A,_B} where _B where _A

PowerSystems._parse_matlab_matrix — Method

_parse_matlab_matrix(lines::Any, index::Any) -> Dict{_A,_B} where _B where _A

PowerSystems._parse_matpower_string — Method

_parse_matpower_string(data_string::String) -> Dict{String,Any}

PowerSystems._parse_pti_data — Method

_parse_pti_data(data_io::IO) -> Dict{String,Array{Dict,N} where N}

_parse_pti_data(data_string, sections)

Internal function. Parse a PTI raw file into a Dict, given the data_string of the file and a list of the sections in the PTI file (typically given by default by get_pti_sections().

PowerSystems._populate_args — Method

_populate_args(param_map::Any, val::Any) -> Array{Any,1}

Populate arguments in a vector for each dynamic component (except Shafts). Returns a vector with the parameter values of the argument of each component.

PowerSystems._populate_defaults! — Method

_populate_defaults!(data::Dict)

_populate_defaults!(pti_data)

Internal function. Populates empty fields with PSS(R)E PTI v33 default values

PowerSystems._propagate_topology_status! — Method

_propagate_topology_status!(data::Dict{String,#s142} where #s142)

PowerSystems._psse2pm_branch! — Method

_psse2pm_branch!(pm_data::Dict, pti_data::Dict, import_all::Bool)

_psse2pm_branch!(pm_data, pti_data)

Parses PSS(R)E-style Branch data into a PowerModels-style Dict. "source_id" is given by ["I", "J", "CKT"] in PSS(R)E Branch specification.

PowerSystems._psse2pm_bus! — Method

_psse2pm_bus!(pm_data::Dict, pti_data::Dict, import_all::Bool)

_psse2pm_bus!(pm_data, pti_data)

Parses PSS(R)E-style Bus data into a PowerModels-style Dict. "source_id" is given by ["I", "NAME"] in PSS(R)E Bus specification.

PowerSystems._psse2pm_dcline! — Method

_psse2pm_dcline!(pm_data::Dict, pti_data::Dict, import_all::Bool)

_psse2pm_dcline!(pm_data, pti_data)

Parses PSS(R)E-style Two-Terminal and VSC DC Lines data into a PowerModels compatible Dict structure by first converting them to a simple DC Line Model. For Two-Terminal DC lines, "sourceid" is given by ["IPR", "IPI", "NAME"] in the PSS(R)E Two-Terminal DC specification. For Voltage Source Converters, "sourceid" is given by ["IBUS1", "IBUS2", "NAME"], where "IBUS1" is "IBUS" of the first converter bus, and "IBUS2" is the "IBUS" of the second converter bus, in the PSS(R)E Voltage Source Converter specification.

PowerSystems._psse2pm_generator! — Method

_psse2pm_generator!(pm_data::Dict, pti_data::Dict, import_all::Bool)

_psse2pm_generator!(pm_data, pti_data)

Parses PSS(R)E-style Generator data in a PowerModels-style Dict. "source_id" is given by ["I", "ID"] in PSS(R)E Generator specification.

PowerSystems._psse2pm_load! — Method

_psse2pm_load!(pm_data::Dict, pti_data::Dict, import_all::Bool)

_psse2pm_load!(pm_data, pti_data)

Parses PSS(R)E-style Load data into a PowerModels-style Dict. "source_id" is given by ["I", "ID"] in the PSS(R)E Load specification.

PowerSystems._psse2pm_shunt! — Method

_psse2pm_shunt!(pm_data::Dict, pti_data::Dict, import_all::Bool)

_psse2pm_shunt!(pm_data, pti_data)

Parses PSS(R)E-style Fixed and Switched Shunt data into a PowerModels-style Dict. "source_id" is given by ["I", "ID"] for Fixed Shunts, and ["I", "SWREM"] for Switched Shunts, as given by the PSS(R)E Fixed and Switched Shunts specifications.

PowerSystems._psse2pm_transformer! — Method

_psse2pm_transformer!(pm_data::Dict, pti_data::Dict, import_all::Bool)

_psse2pm_transformer!(pm_data, pti_data)

Parses PSS(R)E-style Transformer data into a PowerModels-style Dict. "source_id" is given by ["I", "J", "K", "CKT", "winding"], where "winding" is 0 if transformer is two-winding, and 1, 2, or 3 for three-winding, and the remaining keys are defined in the PSS(R)E Transformer specification.

PowerSystems._pti_to_powermodels! — Method

_pti_to_powermodels!(pti_data::Dict; import_all, validate) -> Dict{String,Any}

_pti_to_powermodels!(pti_data)

Converts PSS(R)E-style data parsed from a PTI raw file, passed by pti_data into a format suitable for use internally in PowerModels. Imports all remaining data from the PTI file if import_all is true (Default: false).

PowerSystems._read_data_row — Method

_read_data_row(data::PowerSystems.PowerSystemTableData, row::Any, field_infos::Any; na_to_nothing) -> NamedTuple

Reads values from dataframe row and performs necessary conversions.

PowerSystems._remove_aggregration_topology! — Method

_remove_aggregration_topology!(bus::Bus, _::LoadZone)

Remove the aggregation topology in a Bus

PowerSystems._remove_pwl_cost_duplicates! — Function

_remove_pwl_cost_duplicates!(id::Any, comp::Any, type_name::Any) -> Bool
_remove_pwl_cost_duplicates!(id::Any, comp::Any, type_name::Any, tolerance::Any) -> Bool

checks that each point in the a pwl function is unqiue, simplifies the function if duplicates appear

PowerSystems._remove_service! — Method

_remove_service!(device::Device, service::Service) -> Bool

Remove service from device if it is attached.

PowerSystems._rescale_cost_model! — Method

_rescale_cost_model!(comp::Dict{String,#s854} where #s854, scale::Real)

PowerSystems._select_largest_component! — Method

_select_largest_component!(data::Dict{String,#s855} where #s855) -> Any

PowerSystems._simplify_pwl_cost! — Function

_simplify_pwl_cost!(id::Any, comp::Any, type_name::Any) -> Bool
_simplify_pwl_cost!(id::Any, comp::Any, type_name::Any, tolerance::Any) -> Bool

checks the slope of each segment in a pwl function, simplifies the function if the slope changes is below a tolerance

PowerSystems._split_loads_shunts! — Method

_split_loads_shunts!(data::Dict{String,Any})

_split_loads_shunts!(data)

Seperates Loads and Shunts in data under separate "load" and "shunt" keys in the PowerModels data format. Includes references to originating bus via "loadbus" and "shuntbus" keys, respectively.

PowerSystems._standardize_cost_terms! — Method

_standardize_cost_terms!(components::Dict{String,#s854} where #s854, comp_order::Int64, cost_comp_name::String) -> Set{Int64}

ensures all polynomial costs functions have at exactly comp_order terms

PowerSystems._type_array — Method

_type_array(string_array::Array{T<:AbstractString,1}) -> Any

Attempts to determine the type of an array of strings extracted from a matlab file

PowerSystems._type_value — Method

_type_value(value_string::AbstractString) -> Union{Float64, Int64, SubString}

Attempts to determine the type of a string extracted from a matlab file

PowerSystems._update_branch_flow! — Method

_update_branch_flow!(sys::System)

Updates the flow on the branches

PowerSystems._update_data! — Method

_update_data!(data::Dict{String,#s854} where #s854, new_data::Dict{String,#s853} where #s853)

recursive call of updatedata

PowerSystems._value2string — Method

_value2string(v::Any, float_precision::Int64) -> Any

converts any value to a string, summarizes arrays and dicts

PowerSystems._write_pf_sol! — Method

_write_pf_sol!(sys::System, nl_result::Any)

Updates system voltages and powers with power flow results

PowerSystems._write_results — Method

_write_results(sys::System, nl_result::Any) -> Dict{String,DataFrames.DataFrame}

Return power flow results in dictionary of dataframes.

PowerSystems.adequacy_check — Method

adequacy_check(sys::System) -> Union{Nothing, Bool}

adequacy_check(sys::System)

Checks the system for sum(generator ratings) >= sum(load ratings).

Arguments

sys::System: system

PowerSystems.arrays_to_dicts! — Method

arrays_to_dicts!(data::Dict{String,#s853} where #s853)

turns top level arrays into dicts

PowerSystems.branch_csv_parser! — Method

branch_csv_parser!(sys::System, data::PowerSystems.PowerSystemTableData)

Add branches to the System from the raw data.

PowerSystems.bus_csv_parser! — Method

bus_csv_parser!(sys::System, data::PowerSystems.PowerSystemTableData)

Add buses and areas to the System from the raw data.

PowerSystems.bus_gen_lookup — Method

bus_gen_lookup(gen_data::Dict{String,#s852} where #s852, bus_data::Dict{String,#s851} where #s851) -> Dict{_A,_B} where _B where _A

builds a lookup list of what generators are connected to a given bus

PowerSystems.bus_load_lookup — Method

bus_load_lookup(load_data::Dict{String,#s852} where #s852, bus_data::Dict{String,#s851} where #s851) -> Dict{_A,_B} where _B where _A

builds a lookup list of what loads are connected to a given bus

PowerSystems.bus_shunt_lookup — Method

bus_shunt_lookup(shunt_data::Dict{String,#s852} where #s852, bus_data::Dict{String,#s851} where #s851) -> Dict{_A,_B} where _B where _A

builds a lookup list of what shunts are connected to a given bus

PowerSystems.bus_storage_lookup — Method

bus_storage_lookup(storage_data::Dict{String,#s852} where #s852, bus_data::Dict{String,#s851} where #s851) -> Dict{_A,_B} where _B where _A

builds a lookup list of what storage is connected to a given bus

PowerSystems.calc_branch_flow_ac — Method

calc_branch_flow_ac(data::Dict{String,#s855} where #s855) -> Dict{String,Any}

assumes a vaild ac solution is included in the data and computes the branch flow values

PowerSystems.calc_branch_flow_dc — Method

calc_branch_flow_dc(data::Dict{String,#s855} where #s855) -> Dict{String,Any}

assumes a vaild dc solution is included in the data and computes the branch flow values

PowerSystems.calc_branch_t — Method

calc_branch_t(branch::Dict{String,#s857} where #s857) -> Tuple{Any,Any}

PowerSystems.calc_branch_y — Method

calc_branch_y(branch::Dict{String,#s857} where #s857) -> Tuple{Any,Any}

PowerSystems.calc_connected_components — Method

calc_connected_components(data::Dict{String,#s849} where #s849; edges) -> Set{Set{Int64}}

computes the connected components of the network graph returns a set of sets of bus ids, each set is a connected component

PowerSystems.calc_cost_pwl_lines — Method

calc_cost_pwl_lines(comp_dict::Dict) -> Dict{Any,Any}

compute lines in m and b from from pwl cost models data is a list of components.

Can be run on data or ref data structures

PowerSystems.calc_dcline_cost — Method

calc_dcline_cost(data::Dict{String,#s853} where #s853) -> Any

computes the dcline cost from given network data

PowerSystems.calc_gen_cost — Method

calc_gen_cost(data::Dict{String,#s853} where #s853) -> Any

computes the generator cost from given network data

PowerSystems.calc_max_cost_index — Method

calc_max_cost_index(data::Dict{String,Any}) -> Any

PowerSystems.calc_power_balance — Method

calc_power_balance(data::Dict{String,#s855} where #s855) -> Dict{String,Any}

assumes a vaild solution is included in the data and computes the power balance at each bus

PowerSystems.calc_theta_delta_bounds — Method

calc_theta_delta_bounds(data::Dict{String,#s856} where #s856) -> Tuple{Real,Real}

PowerSystems.check_ascending_order — Method

check_ascending_order(array::Array{Int64,N} where N, name::AbstractString)

Throws DataFormatError if the array is not in ascending order.

PowerSystems.check_branch_loops — Method

check_branch_loops(data::Dict{String,#s855} where #s855)

checks that all branches connect two distinct buses

PowerSystems.check_component_addition — Method

check_component_addition(sys::System, component::Component)

Throws ArgumentError if a PowerSystems rule blocks addition to the system.

This method is tied with handlecomponentaddition!. If the methods are re-implemented for a subtype then whatever is added in handlecomponentaddition! must be checked here.

PowerSystems.check_component_removal — Method

check_component_removal(sys::System, component::Component)

Throws ArgumentError if a PowerSystems rule blocks removal from the system.

PowerSystems.check_component_removal — Method

Throws ArgumentError if a PowerSystems rule blocks removal from the system.

PowerSystems.check_conductors — Method

check_conductors(data::Dict{String,#s855} where #s855)

PowerSystems.check_connectivity — Method

check_connectivity(data::Dict{String,#s841} where #s841)

checks that all buses are unique and other components link to valid buses

PowerSystems.check_reference_bus — Method

check_reference_bus(data::Dict{String,#s854} where #s854)

checks that contains at least one refrence bus

PowerSystems.check_reference_buses — Method

check_reference_buses(data::Dict{String,Any})

checks that each connected components has a reference bus, if not, adds one

PowerSystems.check_status — Method

check_status(data::Dict{String,#s841} where #s841)

checks that active components are not connected to inactive buses, otherwise prints warnings

PowerSystems.check_storage_parameters — Method

check_storage_parameters(data::Dict{String,Any})

checks that each storage unit has a reasonable parameters

PowerSystems.check_switch_parameters — Method

check_switch_parameters(data::Dict{String,#s855} where #s855)

checks that each switch has a reasonable parameters

PowerSystems.check_type — Method

check_type(typ::Any, value::Any) -> Any

Checks if the given value is of a given type, if not tries to make it that type

PowerSystems.check_voltage_setpoints — Method

check_voltage_setpoints(data::Dict{String,#s852} where #s852)

throws warnings if generator and dc line voltage setpoints are not consistent with the bus voltage setpoint

PowerSystems.compare_dict — Method

compare_dict(d1::Any, d2::Any) -> Bool

tests if two dicts are equal, up to floating point precision

PowerSystems.copy_forecasts! — Function

copy_forecasts!(src::InfrastructureSystems.InfrastructureSystemsType, dst::InfrastructureSystems.InfrastructureSystemsType)
copy_forecasts!(src::InfrastructureSystems.InfrastructureSystemsType, dst::InfrastructureSystems.InfrastructureSystemsType, label_mapping::Union{Nothing, Dict{String,String}})

Efficiently add all forecasts in one component to another by copying the underlying references.

Arguments

src::InfrastructureSystemsType: Source component

dst::InfrastructureSystemsType: Destination component

label_mapping::Dict = nothing: Optionally map src labels to different dst labels. If provided and src has a forecast with a label not present in labelmapping, that forecast will not copied. If labelmapping is nothing then all forecasts will be copied with src's labels.

PowerSystems.correct_branch_directions! — Method

correct_branch_directions!(data::Dict{String,#s855} where #s855) -> Set{Int64}

checks that all parallel branches have the same orientation

PowerSystems.correct_bus_types! — Method

correct_bus_types!(data::Dict{String,#s851} where #s851) -> Set{Int64}

checks bus types are consistent with generator connections, if not, fixes them

PowerSystems.correct_component_refrence_bus! — Method

correct_component_refrence_bus!(component_bus_ids::Any, bus_lookup::Any, component_gens::Any)

checks that a connected component has a reference bus, if not, adds one

PowerSystems.correct_cost_functions! — Method

correct_cost_functions!(data::Dict{String,#s853} where #s853) -> Tuple{Set{Int64},Set{Int64}}

throws warnings if cost functions are malformed

PowerSystems.correct_current_limits! — Method

correct_current_limits!(data::Dict{String,#s855} where #s855) -> Set{Int64}

checks that each branch has a reasonable current rating-a, if not computes one

PowerSystems.correct_dcline_limits! — Method

correct_dcline_limits!(data::Dict{String,Any}) -> Set{Int64}

checks that parameters for dc lines are reasonable

PowerSystems.correct_network_data! — Method

correct_network_data!(data::Dict{String,#s856} where #s856) -> Dict{String,Dict{Symbol,Set{Int64}}}

Runs various data quality checks on a PowerModels data dictionary. Applies modifications in some cases. Reports modified component ids.

PowerSystems.correct_thermal_limits! — Method

correct_thermal_limits!(data::Dict{String,#s855} where #s855) -> Set{Int64}

checks that each branch has a reasonable thermal rating-a, if not computes one

PowerSystems.correct_transformer_parameters! — Method

correct_transformer_parameters!(data::Dict{String,#s854} where #s854) -> Set{Int64}

checks that each branch has a reasonable transformer parameters

this is important because setting tap == 0.0 leads to NaN computations, which are hard to debug

PowerSystems.correct_voltage_angle_differences! — Function

correct_voltage_angle_differences!(data::Dict{String,#s853} where #s853) -> Set{Int64}
correct_voltage_angle_differences!(data::Dict{String,#s852} where #s852, default_pad::Any) -> Set{Int64}

checks that voltage angle differences are within 90 deg., if not tightens

PowerSystems.dc_branch_csv_parser! — Method

dc_branch_csv_parser!(sys::System, data::PowerSystems.PowerSystemTableData)

Add DC branches to the System from raw data.

PowerSystems.encode_for_json — Method

Encode composed buses, injectors, and services as UUIDs.

PowerSystems.export_matpower — Method

export_matpower(data::Dict{String,Any}) -> String

Export power network data in the matpower format

PowerSystems.export_matpower — Method

export_matpower(io::IO, data::Dict{String,Any})

Export power network data in the matpower format

PowerSystems.flow_func — Method

flow_func(b::ACBranch, V_from::Complex{Float64}, V_to::Complex{Float64})

Calculates the From - To complex power flow using external data of voltages of branch of type Line

PowerSystems.flow_func — Method

flow_func(b::TapTransformer, V_from::Complex{Float64}, V_to::Complex{Float64}) -> Tuple{Float64,Float64}

Calculates the From - To complex power flow using external data of voltages of branch of type TapTransformer

PowerSystems.flow_func — Method

flow_func(b::Transformer2W, V_from::Complex{Float64}, V_to::Complex{Float64}) -> Tuple{Float64,Float64}

Calculates the From - To complex power flow using external data of voltages of branch of type Transformer2W

PowerSystems.flow_val — Method

flow_val(b::ACBranch)

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Line

PowerSystems.flow_val — Method

flow_val(b::TapTransformer) -> Complex{Float64}

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type TapTransformer

PowerSystems.flow_val — Method

flow_val(b::Transformer2W) -> Complex{Float64}

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Transformer2W

PowerSystems.gen_csv_parser! — Method

gen_csv_parser!(sys::System, data::PowerSystems.PowerSystemTableData)

Add generators to the System from the raw data.

PowerSystems.get_avr_saturation — Method

get_avr_saturation(E::Tuple{Float64,Float64}, Se::Tuple{Float64,Float64}) -> Tuple{Float64,Float64}

Obtain coefficients (A, B) of the function Se(x) = B(x - A)^2/x for Se(E1) = B(E1 - A)^2/E1 and Se(E2) = B(E2 - A)^2/2 and uses the negative solution of the quadratic equation

PowerSystems.get_branch — Method

get_branch(value::DynamicBranch) -> ACBranch

Get branch

PowerSystems.get_dataframe — Method

get_dataframe(data::PowerSystems.PowerSystemTableData, category::PowerSystems.InputCategory) -> DataFrames.DataFrame

Return the dataframe for the category.

PowerSystems.get_enum_value — Method

get_enum_value(enum::Any, value::String) -> Any

Get the enum value for the string. Case insensitive.

PowerSystems.get_exponential_saturation — Method

get_exponential_saturation(Se::Tuple{Float64,Float64}) -> Tuple{Float64,Float64}

Obtain coefficients (A, B) of the function Se = Bx^A for Se(1.2) = B(1.2)^A and Se(1.0) = B(1.0)^A as: B = Se(1.0) and hence (1.2)^A = Se(1.2)/B -> A = log(Se(1.2)/B) / log(1.2)

PowerSystems.get_forecast_keys — Method

get_forecast_keys(component::Component) -> Any

Return an iterable of NamedTuple keys for forecasts stored for this component.

PowerSystems.get_generator_mapping — Function

get_generator_mapping() -> Dict{NamedTuple,DataType}
get_generator_mapping(filename::Any) -> Dict{NamedTuple,DataType}

Return a dict where keys are a tuple of input parameters (fuel, unit_type) and values are generator types.

PowerSystems.get_generator_type — Method

get_generator_type(fuel::Any, unit_type::Any, mappings::Dict{NamedTuple,DataType}) -> Union{Nothing, DataType}

Return the PowerSystems generator type for this fuel and unit_type.

PowerSystems.get_quadratic_saturation — Method

get_quadratic_saturation(Se::Tuple{Float64,Float64}) -> Tuple{Float64,Float64}

Obtain coefficients (A, B) of the function Se = B(x - A)^2 for Se(1.2) = B(1.2 - A)^2 and Se(1.0) = B(1.0 - A)^2 as: Se(1.0) = Se(1.2)/(1.2 - A)^2 * (1.0 - A)^2 that yields (1.2 - A)^2 Se(1.0) = Se(1.2) * (1.0 - A)^2 or expanding: (Se(1.2) - Se(1.0)) A^2 + (2.4 Se(1.0) - 2 Se(1.2)) A + (Se(1.2) - 1.44 Se(1.0)) = 0 and uses the negative solution of the quadratic equation

PowerSystems.get_series_susceptance — Method

get_series_susceptance(b::ACBranch) -> Float64

Calculate the admittance of AC branches

PowerSystems.get_static_injector — Method

Return the static injector stored with a dynamic injector.

PowerSystems.get_user_field — Method

get_user_field(data::PowerSystems.PowerSystemTableData, category::PowerSystems.InputCategory, field::AbstractString) -> Any

Return the custom name stored in the user descriptor file.

Throws DataFormatError if a required value is not found in the file.

PowerSystems.get_user_fields — Method

get_user_fields(data::PowerSystems.PowerSystemTableData, category::PowerSystems.InputCategory) -> Any

Return a vector of user-defined fields for the category.

PowerSystems.handle_component_addition! — Method

handle_component_addition!(sys::System, component::Component)

Refer to docstring for checkcomponentaddition!

PowerSystems.handle_component_removal! — Method

handle_component_removal!(sys::System, bus::Bus) -> Int64

Throws ArgumentError if the bus number is not stored in the system.

PowerSystems.im_replicate — Method

im_replicate(sn_data::Dict{String,#s855} where #s855, count::Int64, global_keys::Set{String}) -> Dict{String,Any}

Transforms a single network into a multinetwork with several deepcopies of the original network

PowerSystems.is_attached — Method

is_attached(component::T<:Component, sys::System) -> Bool

Return true if the component is attached to the system.

PowerSystems.isafield — Method

isafield(component::Component, field::Symbol) -> Any

Checks if a PowerSystemDevice has a field or subfield name.

PowerSystems.ismultinetwork — Method

ismultinetwork(data::Dict{String,#s857} where #s857) -> Any

checks if a given network data is a multinetwork

PowerSystems.iterate_rows — Method

iterate_rows(data::PowerSystems.PowerSystemTableData, category::Any; na_to_nothing) -> Channel{Any}

Return a NamedTuple of parameters from the descriptor file for each row of a dataframe, making type conversions as necessary.

Refer to the PowerSystems descriptor file for field names that will be created.

PowerSystems.load_csv_parser! — Method

load_csv_parser!(sys::System, data::PowerSystems.PowerSystemTableData)

load_csv_parser!(sys::System, data::PowerSystemTableData)

Add loads to the System from the raw load data.

PowerSystems.loadzone_csv_parser! — Method

loadzone_csv_parser!(sys::System, data::PowerSystems.PowerSystemTableData)

loadzone_csv_parser!(sys::System, data::PowerSystemTableData)

Add branches to the System from the raw data.

PowerSystems.make_bus — Method

make_bus(bus_dict::Dict{String,Any}) -> Bus

Creates a PowerSystems.Bus from a PowerSystems bus dictionary

PowerSystems.make_generator — Method

make_generator(data::PowerSystems.PowerSystemTableData, gen::Any, cost_colnames::Any, bus::Any) -> Any

Creates a generator of any type.

PowerSystems.make_mixed_units! — Method

make_mixed_units!(data::Dict{String,#s855} where #s855)

Transforms network data into mixed-units (inverse of per-unit)

PowerSystems.make_per_unit! — Method

make_per_unit!(data::Dict{String,#s855} where #s855)

Transforms network data into per-unit

PowerSystems.make_thermal_gen — Method

make_thermal_gen(gen_name::AbstractString, d::Dict, bus::Bus, sys_mbase::Number) -> ThermalStandard

The polynomial term follows the convention that for an n-degree polynomial, at least n + 1 components are needed. c(p) = cn*p^n+...+c1p+c0 co is stored in the field in of the Econ Struct

PowerSystems.move_genfuel_and_gentype! — Method

move_genfuel_and_gentype!(data::Dict{String,Any})

Move gentype and genfuel fields to be subfields of gen

PowerSystems.parse_matpower — Method

parse_matpower(io::IO; validate) -> Dict{String,Any}

Parses the matpwer data from either a filename or an IO object

PowerSystems.parse_psse — Method

parse_psse(io::IO; kwargs...) -> Dict{String,Any}

Parses directly from iostream

PowerSystems.parse_psse — Method

parse_psse(filename::String; kwargs...) -> Dict{String,Any}

Parses directly from file

PowerSystems.parse_pti — Method

parse_pti(io::IO) -> Dict{String,Array{Dict,N} where N}

parse_pti(io::IO)

Reads PTI data in io::IO, returning a Dict of the data parsed into the proper types.

PowerSystems.parse_pti — Method

parse_pti(filename::String) -> Dict{String,Array{Dict,N} where N}

parse_pti(filename::String)

Open PTI raw file given by filename, returning a Dict of the data parsed into the proper types.

PowerSystems.propagate_topology_status! — Method

propagate_topology_status!(data::Dict{String,#s855} where #s855)

finds active network buses and branches that are not necessary for the computation and sets their status to off.

Works on a PowerModels data dict, so that a it can be used without a GenericPowerModel object

Warning: this implementation has quadratic complexity, in the worst case

PowerSystems.read_gen! — Method

read_gen!(sys::System, data::Any, bus_number_to_bus::Dict{Int64,Bus}; kwargs...)

Transfer generators to ps_dict according to their classification

PowerSystems.replicate — Method

replicate(sn_data::Dict{String,#s854} where #s854, count::Int64; global_keys) -> Dict{String,Any}

Turns in given single network data in multinetwork data with a count replicate of the given network. Note that this function performs a deepcopy of the network data. Significant multinetwork space savings can often be achieved by building application specific methods of building multinetwork with minimal data replication.

PowerSystems.row_to_dict — Method

row_to_dict(row_data::Any, columns::Any) -> Dict{String,Any}

takes a row from a matrix and assigns the values names

PowerSystems.row_to_typed_dict — Method

row_to_typed_dict(row_data::Any, columns::Any) -> Dict{String,Any}

takes a row from a matrix and assigns the values names and types

PowerSystems.select_largest_component — Method

select_largest_component(data::Dict{String,Any}) -> Any

determines the largest connected component of the network and turns everything else off

PowerSystems.services_csv_parser! — Method

services_csv_parser!(sys::System, data::PowerSystems.PowerSystemTableData)

Add services to the System from the raw data.

PowerSystems.set_branch! — Method

set_branch!(value::DynamicBranch, val::ACBranch) -> ACBranch

Set branch

PowerSystems.simplify_cost_terms! — Method

simplify_cost_terms!(data::Dict{String,#s849} where #s849) -> Tuple{Set{Int64},Set{Int64}}

trims zeros from higher order cost terms

PowerSystems.split_line — Method

split_line(mp_line::AbstractString) -> Array{_A,1} where _A

PowerSystems.standardize_cost_terms! — Method

standardize_cost_terms!(data::Dict{String,#s844} where #s844; order)

ensures all polynomial costs functions have the same number of terms

PowerSystems.throw_if_not_attached — Method

throw_if_not_attached(component::Component, sys::System)

Throws ArgumentError if the component is not attached to the system.

PowerSystems.total_capacity_rating — Method

total_capacity_rating(sys::System) -> Any

total_capacity_rating(sys::System)

Sum of system generator and storage ratings.

Arguments

sys::System: system

PowerSystems.total_load_rating — Method

total_load_rating(sys::System) -> Any

total_load_rating(sys::System)

Checks the system for sum(generator ratings) >= sum(load ratings).

Arguments

sys::System: system

PowerSystems.validate_struct — Method

validate_struct(sys::System, value::PowerSystemType) -> Bool

Validate an instance of a PowerSystemType against System data. Returns true if the instance is valid.

Users implementing this function for custom types should consider implementing InfrastructureSystems.validate_struct instead if the validation logic only requires data contained within the instance.