Public API Reference

Abstract Types

PowerSystems.Component — Type

Supertype for all PowerSystems components. All subtypes must include a InfrastructureSystemsInternal member. Subtypes should call InfrastructureSystemsInternal() by default, but also must provide a constructor that allows existing values to be deserialized.

PowerSystems.Device — Type

Supertype for "devices" (bus, line, etc.)

PowerSystems.Generator — Type

Supertype for all generation technologies

PowerSystems.HydroGen — Type

Supertype for all Hydropower generation technologies

PowerSystems.RenewableGen — Type

Supertype for all renewable generation technologies Requires the implementation of get_ratingand get_power_factor methods

PowerSystems.ThermalGen — Type

Supertype for all Thermal generation technologies

PowerSystems.AggregationTopology — Type

Represents a geographical region of system components.

All subtypes must implement the method get_aggregation_topology_accessor.

PowerSystems.Topology — Type

Abstract type to represent the structure and interconnectedness of the system

PowerSystems.DynamicComponent — Type

Abstract type for all components used to compose a DynamicInjection device

PowerSystems.DynamicInjection — Type

Abstract type for all dynamic injection types

PowerSystems.StaticInjection — Type

Abstract type of Devices that inject current/power

System

PowerSystems.System — Type

System

A power system defined by fields for base_power, components, and time series.

System(base_power)
System(base_power, buses, components...)
System(base_power, buses, generators, loads, branches, storage, services; kwargs...)
System(base_power, buses, generators, loads; kwargs...)
System(file; kwargs...)
System(; buses, generators, loads, branches, storage, base_power, services, kwargs...)
System(; kwargs...)

Arguments

base_power::Float64: the base power value for the system

buses::Vector{Bus}: an array of buses

components...: Each element must be an iterable containing subtypes of Component.

Keyword arguments

ext::Dict: Contains user-defined parameters. Should only contain standard types.

runchecks::Bool: Run available checks on input fields and when add_component! is called. Throws InvalidRange if an error is found.

time_series_in_memory::Bool=false: Store time series data in memory instead of HDF5.

config_path::String: specify path to validation config file

PowerSystems.System — Method

System(file_path::AbstractString; kwargs...) -> Any

Constructs a System from a file path ending with .m, .RAW, or .json

PowerSystems.System — Method

System(data::Any, base_power::Any; internal, kwargs...) -> System

Construct a System from InfrastructureSystems.SystemData

PowerSystems.System — Method

System(base_power::Any; kwargs...) -> System

Construct an empty System. Useful for building a System while parsing raw data.

PowerSystems.System — Method

System(base_power::Float64, buses::Array{Bus,1}, components::Vararg{Any,N} where N; kwargs...) -> System

System constructor when components are constructed externally.

PowerSystems.System — Method

System(::Nothing; buses, generators, loads, branches, storage, base_power, services, kwargs...) -> System

Constructs a non-functional System for demo purposes.

InfrastructureSystems.get_time_series_multiple — Function

get_time_series_multiple(sys::System) -> Channel{Any}
get_time_series_multiple(sys::System, filter_func::Any; type, name) -> Channel{Any}

Return an iterator of time series in order of initial time.

Note that passing a filter function can be much slower than the other filtering parameters because it reads time series data from media.

Call collect on the result to get an array.

Arguments

data::SystemData: system

filter_func = nothing: Only return time series for which this returns true.

type = nothing: Only return time series with this type.

name = nothing: Only return time series matching this value.

Examples

for time_series in get_time_series_multiple(sys)
    @show time_series
end

ts = collect(get_time_series_multiple(sys; type = SingleTimeSeries))

InfrastructureSystems.to_json — Method

to_json(sys::System, filename::AbstractString; force)

Serializes a system to a JSON string.

PowerSystems.add_component! — Method

add_component!(sys::System, dyn_injector::DynamicInjection, static_injector::StaticInjection; kwargs...)

Add a dynamic injector to the system.

Throws ArgumentError if the name does not match the staticinjector name. Throws ArgumentError if the staticinjector is not attached to the system.

All rules for the generic add_component! method also apply.

PowerSystems.add_component! — Method

Add a component to the system.

Throws ArgumentError if the component's name is already stored for its concrete type. Throws ArgumentError if any Component-specific rule is violated. Throws InvalidRange if any of the component's field values are outside of defined valid range.

Examples

sys = System(100.0)

# Add a single component.
add_component!(sys, bus)

# Add many at once.
buses = [bus1, bus2, bus3]
generators = [gen1, gen2, gen3]
foreach(x -> add_component!(sys, x), Iterators.flatten((buses, generators)))

PowerSystems.add_service! — Method

add_service!(device::Device, service::Service, sys::System)

Similar to add_service! but for Service and Device already stored in the system. Performs validation checks on the device and the system

Arguments

device::Device: Device

service::Service: Service

sys::System: system

PowerSystems.add_service! — Method

add_service!(sys::System, service::Service, contributing_devices::Any; kwargs...)

Similar to add_component! but for services.

Arguments

sys::System: system

service::Service: service to add

contributing_devices: Must be an iterable of type Device

PowerSystems.add_service! — Method

add_service!(sys::System, service::Service, contributing_device::Device; kwargs...)

Similar to add_component! but for services.

Arguments

sys::System: system

service::Service: service to add

contributing_device::Device: Valid Device

PowerSystems.add_service! — Method

add_service!(sys::System, service::StaticReserveGroup, contributing_services::Array{var"#s263",1} where var"#s263"<:Service; kwargs...)

Similar to add_component! but for StaticReserveGroup.

Arguments

sys::System: system

service::StaticReserveGroup: service to add

contributing_services: contributing services to the group

PowerSystems.add_service! — Method

add_service!(sys::System, service::StaticReserveGroup; kwargs...)

Similar to add_component! but for StaticReserveGroup.

Arguments

sys::System: system

service::StaticReserveGroup: service to add

PowerSystems.add_time_series! — Method

add_time_series!(sys::System, metadata_file::AbstractString; resolution)

Add time series data from a metadata file or metadata descriptors.

Arguments

sys::System: system

metadata_file::AbstractString: metadata file for timeseries that includes an array of IS.TimeSeriesFileMetadata instances or a vector.

resolution::DateTime.Period=nothing: skip time series that don't match this resolution.

PowerSystems.add_time_series! — Method

add_time_series!(sys::System, components::Any, time_series::TimeSeriesData)

Add the same time series data to multiple components.

This is significantly more efficent than calling add_time_series! for each component individually with the same data because in this case, only one time series array is stored.

Throws ArgumentError if a component is not stored in the system.

PowerSystems.add_time_series! — Method

add_time_series!(sys::System, file_metadata::Array{InfrastructureSystems.TimeSeriesFileMetadata,1}; resolution)

Add time series data from a metadata file or metadata descriptors.

Arguments

sys::System: system

timeseries_metadata::Vector{IS.TimeSeriesFileMetadata}: metadata for timeseries

resolution::DateTime.Period=nothing: skip time series that don't match this resolution.

PowerSystems.add_time_series! — Method

add_time_series!(sys::System, component::Component, time_series::TimeSeriesData)

Add time series data to a component.

Throws ArgumentError if the component is not stored in the system.

PowerSystems.check_time_series_consistency — Method

check_time_series_consistency(sys::System) -> Any

Throws DataFormatError if time series data have inconsistent parameters.

PowerSystems.clear_components! — Method

clear_components!(sys::System)

Remove all components from the system.

PowerSystems.clear_ext! — Method

clear_ext!(sys::System)

Clear any value stored in ext.

PowerSystems.clear_time_series! — Method

clear_time_series!(sys::System)

Remove all time series data from the system.

PowerSystems.convert_component! — Method

convert_component!(linetype::Type{Line}, line::MonitoredLine, sys::System; kwargs...)

Converts a MonitoredLine component to a Line component and replaces the original in the system

PowerSystems.convert_component! — Method

convert_component!(linetype::Type{MonitoredLine}, line::Line, sys::System; kwargs...)

Converts a Line component to a MonitoredLine component and replaces the original in the system

PowerSystems.copy_time_series! — Method

copy_time_series!(dst::Component, src::Component; name_mapping, scaling_factor_multiplier_mapping)

Efficiently add all time series data in one component to another by copying the underlying references.

Arguments

dst::Component: Destination component

src::Component: Source component

name_mapping::Dict = nothing: Optionally map src names to different dst names. If provided and src has a time series with a name not present in namemapping, that time series will not copied. If namemapping is nothing then all time series will be copied with src's names.

scaling_factor_multiplier_mapping::Dict = nothing: Optionally map src scalingfactormultipliers to dst scalingfactormultipliers. Same behaviors as name_mapping.

PowerSystems.get_aggregation_topology_mapping — Method

get_aggregation_topology_mapping(_::Type{T<:AggregationTopology}, sys::System) -> Dict{String,Array{Bus,1}}

Return a mapping of AggregationTopology name to vector of buses within it.

PowerSystems.get_base_power — Method

get_base_power(sys::System) -> Float64

Return the system's base power.

PowerSystems.get_bus — Method

get_bus(sys::System, bus_number::Int64) -> Any

Return bus with bus_number.

PowerSystems.get_bus — Method

get_bus(sys::System, name::String) -> Union{Nothing, Bus}

Return bus with name.

PowerSystems.get_bus_numbers — Method

get_bus_numbers(sys::System) -> Array{Int64,1}

Return a sorted vector of bus numbers in the system.

PowerSystems.get_buses — Method

get_buses(sys::System, aggregator::AggregationTopology) -> Array{Bus,1}

Return a vector of buses contained within the AggregationTopology.

PowerSystems.get_buses — Method

get_buses(sys::System, bus_numbers::Set{Int64}) -> Array{Bus,1}

Return all buses values with bus_numbers.

PowerSystems.get_component — Method

get_component(_::Type{T<:Component}, sys::System, name::AbstractString) -> Union{Nothing, T<:Component}

Get the component of type T with name. Returns nothing if no component matches. If T is an abstract type then the names of components across all subtypes of T must be unique.

See get_components_by_name for abstract types with non-unique names across subtypes.

Throws ArgumentError if T is not a concrete type and there is more than one component with requested name

PowerSystems.get_components — Method

get_components(_::Type{T<:Component}, sys::System) -> InfrastructureSystems.FlattenIteratorWrapper{_A} where _A

Returns an iterator of components. T can be concrete or abstract. Call collect on the result if an array is desired.

Examples

iter = PowerSystems.get_components(ThermalStandard, sys)
iter = PowerSystems.get_components(Generator, sys)
iter = PowerSystems.get_components(Generator, sys, x->(PowerSystems.get_available(x)))
generators = collect(PowerSystems.get_components(Generator, sys))

See also: iterate_components

PowerSystems.get_components_by_name — Method

Get the components of abstract type T with name. Note that PowerSystems enforces unique names on each concrete type but not across concrete types.

See get_component if the concrete type is known.

Throws ArgumentError if T is not an abstract type.

PowerSystems.get_components_in_aggregation_topology — Method

get_components_in_aggregation_topology(_::Type{T<:StaticInjection}, sys::System, aggregator::AggregationTopology) -> Array{_A,1} where _A

Return a vector of components with buses in the AggregationTopology.

PowerSystems.get_contributing_device_mapping — Method

get_contributing_device_mapping(sys::System) -> Dict{NamedTuple{(:type, :name),Tuple{DataType,String}},ServiceContributingDevices}

Return an instance of ServiceContributingDevicesMapping.

PowerSystems.get_contributing_devices — Method

Return a vector of devices contributing to the service.

PowerSystems.get_ext — Method

get_ext(sys::System) -> Union{Nothing, Dict{String,Any}}

Return a user-modifiable dictionary to store extra information.

PowerSystems.get_forecast_horizon — Method

get_forecast_horizon(sys::System) -> Int64

Return the horizon for all forecasts.

PowerSystems.get_forecast_initial_times — Method

get_forecast_initial_times(sys::System) -> Union{Array{Any,1}, StepRange{Dates.DateTime,_A} where _A}

Return the initial times for all forecasts.

PowerSystems.get_forecast_initial_timestamp — Method

get_forecast_initial_timestamp(sys::System) -> Dates.DateTime

Return the initial_timestamp for all forecasts.

PowerSystems.get_forecast_interval — Method

get_forecast_interval(sys::System) -> Dates.Period

Return the interval for all forecasts.

PowerSystems.get_forecast_total_period — Method

get_forecast_total_period(sys::System) -> Any

Return the total period covered by all forecasts.

PowerSystems.get_forecast_window_count — Method

get_forecast_window_count(sys::System) -> Int64

Return the window count for all forecasts.

PowerSystems.get_frequency — Method

get_frequency(sys::System) -> Float64

Return the system's frequency.

PowerSystems.get_time_series_resolution — Method

get_time_series_resolution(sys::System) -> Dates.Period

Return the resolution for all time series.

PowerSystems.iterate_components — Method

iterate_components(sys::System) -> Channel{Any}

Iterates over all components.

Examples

for component in iterate_components(sys)
    @show component
end

See also: get_components

PowerSystems.remove_component! — Method

remove_component!(sys::System, component::T<:Component)

Remove a component from the system by its value.

Throws ArgumentError if the component is not stored.

PowerSystems.remove_component! — Method

Remove a component from the system by its name.

Throws ArgumentError if the component is not stored.

PowerSystems.remove_components! — Method

Remove all components of type T from the system.

Throws ArgumentError if the type is not stored.

PowerSystems.remove_time_series! — Method

remove_time_series!(sys::System, _::Type{T<:TimeSeriesData}, component::Component, name::String)

Remove the time series data for a component.

PowerSystems.set_contributing_services! — Method

set_contributing_services!(sys::System, service::StaticReserveGroup, val::Array{var"#s265",1} where var"#s265"<:Service) -> Array{var"#s265",1} where var"#s265"<:Service

Set StaticReserveGroup contributing_services with check

PowerSystems.set_units_base_system! — Method

set_units_base_system!(system::System, settings::String)

Sets the units base for the getter functions on the devices. It modifies the behavior of all getter functions

PowerSystems.transform_single_time_series! — Method

transform_single_time_series!(sys::System)

Transform all instances of SingleTimeSeries to DeterministicSingleTimeSeries.

PowerSystems.validate_time_series_consistency — Method

validate_time_series_consistency(sys::System) -> Any

Return true if all time series data have consistent parameters.

Component Methods

PowerSystems.get_base_power — Method

get_base_power(c::Component) -> Float64

Default behavior of a component. If there is no base_power field, assume is in the system's base power.

PowerSystems.get_max_active_power — Method

get_max_active_power(d::T<:RenewableGen) -> Any

Return the max active power for the Renewable Generation calculated as the rating * power_factor

PowerSystems.get_max_reactive_power — Method

Return the max reactive power for the Renewable Generation calculated as the rating * power_factor

PowerSystems.get_aggregation_topology_accessor — Method

get_aggregation_topology_accessor(_::Type{T<:AggregationTopology}) -> typeof(get_area)

Return the method to be called on a Bus to get its AggregationTopology value for this type.

Dynamic Modeling Methods

PowerSystems.get_dynamic_components — Method

get_dynamic_components(device::T<:DynamicInjection) -> Base.Generator{_A,_B} where _B where _A

Return all the dynamic components of a DynamicInjection device

Network Matrices

PowerSystems.Ybus — Type

Nodal admittance matrix (Ybus) is an N x N matrix describing a power system with N buses. It represents the nodal admittance of the buses in a power system.

The Ybus Struct is indexed using the Bus Numbers, no need for them to be sequential

PowerSystems.Ybus — Method

Builds a Ybus from a collection of buses and branches. The return is a Ybus Array indexed with the bus numbers and the branch names.

Keyword arguments

check_connectivity::Bool: Checks connectivity of the network using Goderya's algorithm

PowerSystems.PTDF — Type

Power Transfer Distribution Factors (PTDF) indicate the incremental change in real power that occurs on transmission lines due to real power injections changes at the buses.

The PTDF struct is indexed using the Bus numbers and branch names

PowerSystems.PTDF — Type

Builds the PTDF matrix from a group of branches and nodes. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerSystems.PTDF — Type

Builds the PTDF matrix from a system. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerSystems.LODF — Type

Line Outage Distribution Factors (LODFs) are a sensitivity measure of how a change in a line’s flow affects the flows on other lines in the system.

PowerSystems.LODF — Type

Builds the LODF matrix from a group of branches and nodes. The return is a LOLDF array indexed with the branch name.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerSystems.LODF — Type

Builds the LODF matrix from a system. The return is a LOLDF array indexed with the branch name.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

Power Flow

PowerSystems.solve_powerflow! — Method

solve_powerflow!(system::System; finite_diff, kwargs...) -> Bool

Solves a the power flow into the system and writes the solution into the relevant structs. Updates generators active and reactive power setpoints and branches active and reactive power flows (calculated in the From - To direction) (see flow_val)

Supports solving using Finite Differences Method (instead of using analytic Jacobian) by setting finite_diff = true. Supports passing NLsolve kwargs in the args. By default shows the solver trace.

Arguments available for nlsolve:

method : See NLSolve.jl documentation for available solvers

xtol: norm difference in x between two successive iterates under which convergence is declared. Default: 0.0.

ftol: infinite norm of residuals under which convergence is declared. Default: 1e-8.

iterations: maximum number of iterations. Default: 1_000.

store_trace: should a trace of the optimization algorithm's state be stored? Default: false.

show_trace: should a trace of the optimization algorithm's state be shown on STDOUT? Default: false.

extended_trace: should additifonal algorithm internals be added to the state trace? Default: false.

Examples

solve_powerflow!(sys)
# Passing NLsolve arguments
solve_powerflow!(sys, method = :newton)
# Using Finite Differences
solve_powerflow!(sys, finite_diff = true)

PowerSystems.solve_powerflow — Method

solve_powerflow(system::System; finite_diff, kwargs...) -> Union{Bool, Dict{String,DataFrames.DataFrame}}

Similar to solve_powerflow!(sys) but does not update the system struct with results. Returns the results in a dictionary of dataframes.

Examples

res = solve_powerflow(sys)
# Passing NLsolve arguments
res = solve_powerflow(sys, method = :newton)
# Using Finite Differences
res = solve_powerflow(sys, finite_diff = true)