Variable Cost Curves

PowerSystems.CostCurve — Type

@kwdef struct CostCurve{T <: ValueCurve} <: ProductionVariableCostCurve{T}
    value_curve::T
    power_units::UnitSystem = UnitSystem.NATURAL_UNITS
    vom_cost::LinearCurve = LinearCurve(0.0)
end

Direct representation of the variable operation cost of a power plant in currency. Composed of a ValueCurve that may represent input-output, incremental, or average rate data. The default units for the x-axis are megawatts and can be specified with power_units.

Arguments

value_curve::T: The underlying ValueCurve representation of this ProductionVariableCostCurve

power_units::UnitSystem = UnitSystem.NATURAL_UNITS: The units for the x-axis of the curve; defaults to natural units (megawatts)

vom_cost::LinearCurve = LinearCurve(0.0): Additional proportional Variable Operation and Maintenance cost in currency/(power_unit h)

PowerSystems.FuelCurve — Type

@kwdef struct FuelCurve{T <: ValueCurve} <: ProductionVariableCostCurve{T}
    value_curve::T
    power_units::UnitSystem = UnitSystem.NATURAL_UNITS
    fuel_cost::Union{Float64, TimeSeriesKey}
    vom_cost::LinearCurve = LinearCurve(0.0)
end

Representation of the variable operation cost of a power plant in terms of fuel (MBTU, liters, m^3, etc.), coupled with a conversion factor between fuel and currency. Composed of a ValueCurve that may represent input-output, incremental, or average rate data. The default units for the x-axis are megawatts and can be specified with power_units.

Arguments

value_curve::T: The underlying ValueCurve representation of this ProductionVariableCostCurve

power_units::UnitSystem = UnitSystem.NATURAL_UNITS: The units for the x-axis of the curve; defaults to natural units (megawatts)

fuel_cost::Union{Float64, TimeSeriesKey}: Either a fixed value for fuel cost or the key to a fuel cost time series

vom_cost::LinearCurve = LinearCurve(0.0): Additional proportional Variable Operation and Maintenance cost in currency/(power_unit h)

ValueCurves

PowerSystems.LinearCurve — Type

LinearCurve(proportional_term::Float64, constant_term::Float64)

A linear input-output curve, representing a constant marginal rate. May have zero no-load cost (i.e., constant average rate) or not.

Arguments

proportional_term::Float64: marginal rate

constant_term::Float64: optional, cost at zero production, defaults to 0.0

PowerSystems.QuadraticCurve — Type

QuadraticCurve(quadratic_term::Float64, proportional_term::Float64, constant_term::Float64)

A quadratic input-output curve, may have nonzero no-load cost.

Arguments

quadratic_term::Float64: quadratic term of the curve

proportional_term::Float64: proportional term of the curve

constant_term::Float64: constant term of the curve

PowerSystems.PiecewisePointCurve — Type

PiecewisePointCurve(points::Vector{Tuple{Float64, Float64}})

A piecewise linear curve specified by cost values at production points.

Arguments

points::Vector{Tuple{Float64, Float64}} or similar: vector of (production, cost) pairs

PowerSystems.PiecewiseIncrementalCurve — Type

PiecewiseIncrementalCurve(initial_input::Float64, x_coords::Vector{Float64}, slopes::Vector{Float64})

A piecewise linear curve specified by marginal rates (slopes) between production points. May have nonzero initial value.

Arguments

initial_input::Float64: cost at minimum production point

x_coords::Vector{Float64}: vector of n production points

slopes::Vector{Float64}: vector of n-1 marginal rates/slopes of the curve segments between the points

PowerSystems.PiecewiseAverageCurve — Type

PiecewiseAverageCurve(initial_input::Float64, x_coords::Vector{Float64}, slopes::Vector{Float64})

A piecewise linear curve specified by average rates between production points. May have nonzero initial value.

Arguments

initial_input::Float64: cost at minimum production point

x_coords::Vector{Float64}: vector of n production points

slopes::Vector{Float64}: vector of n-1 average rates/slopes of the curve segments between the points