Operating Costs

PowerSystems.ThermalGenerationCost — Type

mutable struct ThermalGenerationCost <: OperationalCost
    variable::ProductionVariableCostCurve
    fixed::Float64
    start_up::NamedTuple{(:hot, :warm, :cold), NTuple{3, Float64}}
    shut_down::Float64
end

Data Structure Operational Cost Data which includes fixed, variable cost, multiple start up cost and stop costs.

Arguments

variable::ProductionVariableCostCurve: Production Variable Cost. Can take fuel curves or cost curve represenations

fixed::Union{Nothing, Float64}: Fixed cost of keeping the unit online. For some cost represenations this field can be duplicative.

start_up::Union{NamedTuple{(:hot, :warm, :cold), NTuple{3, Float64}}, Float64}: start-up cost can take linear of multi-stage costs

shut_down::Float64: cost of turning the unit offline.

PowerSystems.HydroGenerationCost — Type

mutable struct HydroGenerationCost <: OperationalCost
    variable::ProductionVariableCostCurve
    fixed::Float64
end

Data Structure for the Operational Cost of Hydro Power Plants which includes fixed and variable cost. Variable Costs can be used to represent the cost of curtailment if negative values are used or the opportunity cost of water if the costs are positive. It also supports fuel curves to model specific water intake.

Arguments

variable::ProductionVariableCostCurve: Production variable cost represented by a FuelCurve, where the fuel is water, or a CostCurve in currency.

fixed::Union{Nothing, Float64}: Fixed cost of keeping the unit online. For some cost represenations this field can be duplicative.

PowerSystems.RenewableGenerationCost — Type

mutable struct RenewableGenerationCost <: OperationalCost
    variable::CostCurve
    curtailment_cost::CostCurve
end

Data Structure for the Operational Cost of Renewable Power Plants which includes the variable cost of energy (like a PPA) and the cost of curtailing power. For example, curtailment Costs can be used to represent the loss of tax incentives.

Arguments

variable::CostCurve: Variable cost.

curtailment_cost::CostCurve: Input/output cost of curtailing power.

PowerSystems.StorageCost — Type

mutable struct StorageCost <: OperationalCost
    charge_variable_cost::CostCurve
    discharge_variable_cost::CostCurve
    fixed::Float64
    start_up::Union{STORAGE_OPERATION_MODES, Float64}
    shut_down::Float64
    energy_shortage_cost::Float64
    energy_surplus_cost::Float64
end

Data Structure for Operational Cost Data like variable cost and start - stop costs and energy storage cost. This data structure is not intended to represent market storage systems market operations like the submission of buy/sell bids.

Arguments

charge_variable_cost::CostCurve: variable cost represented as a CostCurve

discharge_variable_cost::CostCurve: variable cost represented as a CostCurve

fixed::Float64: fixed cost of operating the storage system

start_up::Union{STORAGE_OPERATION_MODES, Float64}: start-up cost, validation range:, action if invalid

shut_down::Float64: shut-down cost, validation range: action if invalid

energy_shortage_cost::Float64: Cost incurred by the model for being short of the energy target.

energy_surplus_cost::Float64: Cost incurred by the model for surplus energy stored.

PowerSystems.LoadCost — Type

mutable struct LoadCost <: OperationalCost
    "variable cost"
    variable::CostCurve
    "fixed cost"
    fixed::Float64
end

Data structure for the operational cost of loads (e.g., InterruptiblePowerLoad), including fixed and variable cost components.

Arguments

variable::CostCurve: Variable cost.

fixed::Union{Nothing, Float64}: Fixed cost of keeping the unit online. For some cost represenations this field can be duplicative.

PowerSystems.MarketBidCost — Type

mutable struct MarketBidCost <: OperationalCost
    no_load_cost::Float64
    start_up::NamedTuple{(:hot, :warm, :cold), NTuple{3, Float64}}
    shut_down::Float64
    incremental_offer_curves::Union{Nothing, TimeSeriesKey, PiecewiseLinearData}  # TODO update docs given struct
    decremental_offer_curves::Union{Nothing, TimeSeriesKey, PiecewiseLinearData}
    ancillary_service_offers::Vector{Service}
end

Data Structure Operational Cost to reflect market bids of energy and ancilliary services for any asset. Compatible with most US Market bidding mechanisms that support demand and generation side.

Arguments

no_load_cost::Float64: no load cost

start_up::NamedTuple{(:hot, :warm, :cold), NTuple{3, Float64}}: start-up cost at different stages of the thermal cycle. Warm is also refered as intermediate in some markets

shut_down::Float64: shut-down cost, validation range: (0, nothing), action if invalid: warn

incremental_offer_curves::Union{Nothing, TimeSeriesKey, PiecewiseLinearData}: Sell Offer Curves data, can be a time series or a fixed PiecewiseLinearData

decremental_offer_curves::Union{Nothing, TimeSeriesKey, PiecewiseLinearData}: Buy Offer Curves data, can be a time series or a fixed PiecewiseLinearData

ancillary_service_offers::Vector{Service}: Bids for the ancillary services