Dynamic Branch

PowerSystems.DynamicBranch — Type

Extends the branch type to add the information required for dynamic modeling of branches. Includes the fields for the states and the number of states

Arguments

branch::ACBranch

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::DynamicBranch)

Get DynamicBranch activepowerflow.

PowerSystems.get_angle_limits — Method

get_angle_limits(
    value::DynamicBranch
) -> @NamedTuple{min::Float64, max::Float64}

Get DynamicBranch angle_limits.

PowerSystems.get_arc — Method

get_arc(value::DynamicBranch) -> Any

Get DynamicBranch arc.

PowerSystems.get_available — Method

get_available(value::DynamicBranch) -> Any

Get DynamicBranch available.

PowerSystems.get_b — Method

get_b(
    value::DynamicBranch
) -> @NamedTuple{from::Float64, to::Float64}

Get DynamicBranch b.

PowerSystems.get_ext — Method

get_ext(value::DynamicBranch) -> Any

Get DynamicBranch ext.

PowerSystems.get_n_states — Method

get_n_states(value::DynamicBranch) -> Int64

Get n_states

PowerSystems.get_r — Method

get_r(value::DynamicBranch) -> Any

Get DynamicBranch r.

PowerSystems.get_rating — Method

get_rating(value::DynamicBranch) -> Any

Get DynamicBranch rating.

PowerSystems.get_reactive_power_flow — Method

get_reactive_power_flow(value::DynamicBranch)

Get DynamicBranch reactivepowerflow.

PowerSystems.get_services — Method

get_services(value::DynamicBranch) -> Any

Get DynamicBranch services.

PowerSystems.get_states — Method

get_states(value::DynamicBranch) -> Vector{Symbol}

Get states

PowerSystems.get_x — Method

get_x(value::DynamicBranch) -> Any

Get DynamicBranch x.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(
    value::DynamicBranch,
    val::Float64
) -> Any

Set DynamicBranch activepowerflow.

PowerSystems.set_angle_limits! — Method

set_angle_limits!(
    value::DynamicBranch,
    val::@NamedTuple{min::Float64, max::Float64}
) -> @NamedTuple{min::Float64, max::Float64}

Set DynamicBranch angle_limits.

PowerSystems.set_arc! — Method

set_arc!(value::DynamicBranch, val::Arc) -> Any

Set DynamicBranch arc.

PowerSystems.set_available! — Method

set_available!(value::DynamicBranch, val::Bool) -> Any

Set DynamicBranch available.

PowerSystems.set_b! — Method

set_b!(value::DynamicBranch, val) -> Any

Set DynamicBranch b.

PowerSystems.set_ext! — Method

set_ext!(
    value::DynamicBranch,
    val::Dict{String, Any}
) -> Any

Set DynamicBranch ext.

PowerSystems.set_n_states! — Method

set_n_states!(value::DynamicBranch, val::Int64) -> Int64

Set n_states

PowerSystems.set_r! — Method

set_r!(value::DynamicBranch, val::Float64) -> Any

Set DynamicBranch r.

PowerSystems.set_rating! — Method

set_rating!(value::DynamicBranch, val::Float64) -> Any

Set DynamicBranch rating.

PowerSystems.set_reactive_power_flow! — Method

set_reactive_power_flow!(
    value::DynamicBranch,
    val::Float64
) -> Any

Set DynamicBranch reactivepowerflow.

PowerSystems.set_services! — Method

set_services!(
    value::DynamicBranch,
    val::Vector{Service}
) -> Any

Set DynamicBranch services.

PowerSystems.set_states! — Method

set_states!(
    value::DynamicBranch,
    val::Vector{Symbol}
) -> Vector{Symbol}

Set states

PowerSystems.set_x! — Method

set_x!(value::DynamicBranch, val::Float64) -> Any

Set DynamicBranch x.