ACBus

PowerSystems.ACBus — Type

mutable struct ACBus <: Bus
    number::Int
    name::String
    bustype::Union{Nothing, ACBusTypes}
    angle::Union{Nothing, Float64}
    magnitude::Union{Nothing, Float64}
    voltage_limits::Union{Nothing, MinMax}
    base_voltage::Union{Nothing, Float64}
    area::Union{Nothing, Area}
    load_zone::Union{Nothing, LoadZone}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

An AC bus

Arguments

number::Int: A unique bus identification number (positive integer)

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

bustype::Union{Nothing, ACBusTypes}: Used to describe the connectivity and behavior of this bus. Options are listed here.

angle::Union{Nothing, Float64}: angle of the bus in radians, validation range: (-1.571, 1.571)

magnitude::Union{Nothing, Float64}: voltage as a multiple of base_voltage, validation range: voltage_limits

voltage_limits::Union{Nothing, MinMax}: limits on the voltage variation as multiples of base_voltage

base_voltage::Union{Nothing, Float64}: the base voltage in kV, validation range: (0, nothing)

area::Union{Nothing, Area}: (default: nothing) the area containing the bus

load_zone::Union{Nothing, LoadZone}: (default: nothing) the load zone containing the bus

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::ACBus) -> String

Get ACBus name.

PowerSystems.get_angle — Method

get_angle(value::ACBus) -> Union{Nothing, Float64}

Get ACBus angle.

PowerSystems.get_area — Method

get_area(value::ACBus) -> Union{Nothing, Area}

Get ACBus area.

PowerSystems.get_base_voltage — Method

get_base_voltage(value::ACBus) -> Union{Nothing, Float64}

Get ACBus base_voltage.

PowerSystems.get_bustype — Method

get_bustype(value::ACBus) -> Union{Nothing, ACBusTypes}

Get ACBus bustype.

PowerSystems.get_ext — Method

get_ext(value::ACBus) -> Dict{String, Any}

Get ACBus ext.

PowerSystems.get_load_zone — Method

get_load_zone(value::ACBus) -> Union{Nothing, LoadZone}

Get ACBus load_zone.

PowerSystems.get_magnitude — Method

get_magnitude(value::ACBus) -> Union{Nothing, Float64}

Get ACBus magnitude.

PowerSystems.get_number — Method

get_number(value::ACBus) -> Int64

Get ACBus number.

PowerSystems.get_voltage_limits — Method

get_voltage_limits(
    value::ACBus
) -> Union{Nothing, @NamedTuple{min::Float64, max::Float64}}

Get ACBus voltage_limits.

PowerSystems.set_angle! — Method

set_angle!(value::ACBus, val) -> Any

Set ACBus angle.

PowerSystems.set_area! — Method

set_area!(value::ACBus, val) -> Area

Set ACBus area.

PowerSystems.set_base_voltage! — Method

set_base_voltage!(value::ACBus, val) -> Any

Set ACBus base_voltage.

PowerSystems.set_bustype! — Method

set_bustype!(value::ACBus, val) -> Any

Set ACBus bustype.

PowerSystems.set_ext! — Method

set_ext!(value::ACBus, val) -> Any

Set ACBus ext.

PowerSystems.set_load_zone! — Method

set_load_zone!(value::ACBus, val) -> LoadZone

Set ACBus load_zone.

PowerSystems.set_magnitude! — Method

set_magnitude!(value::ACBus, val) -> Any

Set ACBus magnitude.

PowerSystems.set_number! — Method

set_number!(value::ACBus, val) -> Any

Set ACBus number.

PowerSystems.set_voltage_limits! — Method

set_voltage_limits!(value::ACBus, val) -> Any

Set ACBus voltage_limits.