AGC

PowerSystems.AGC — Type

mutable struct AGC <: Service
    name::String
    available::Bool
    bias::Float64
    K_p::Float64
    K_i::Float64
    K_d::Float64
    delta_t::Float64
    area::Union{Nothing, Area}
    initial_ace::Float64
    reserves::Vector{Reserve}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

Automatic generation control (AGC) for the system or a certain Area within the system.

This model uses a proportional–integral–derivative (PID) control to simulate a "smooth" response of the AGC to the area control error (ACE). Refer to "AGC Simulation Model for Large Renewable Energy Penetration Studies."

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

bias::Float64: Area frequency bias in MW/Hz

K_p::Float64: PID Proportional Constant

K_i::Float64: PID Integral Constant

K_d::Float64: PID Derivative Constant

delta_t::Float64: PID Discretization period [Seconds]

area::Union{Nothing, Area}: (default: nothing) the area controlled by the AGC

initial_ace::Float64: (default: 0.0) Initial condition for ACE

reserves::Vector{Reserve}: (default: Device[]) Reserves that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::AGC) -> Bool

Get AGC available.

InfrastructureSystems.get_name — Method

get_name(value::AGC) -> String

Get AGC name.

InfrastructureSystems.set_available! — Method

set_available!(value::AGC, val) -> Any

Set AGC available.

PowerSystems.get_K_d — Method

get_K_d(value::AGC) -> Float64

Get AGC K_d.

PowerSystems.get_K_i — Method

get_K_i(value::AGC) -> Float64

Get AGC K_i.

PowerSystems.get_K_p — Method

get_K_p(value::AGC) -> Float64

Get AGC K_p.

PowerSystems.get_area — Method

get_area(value::AGC) -> Union{Nothing, Area}

Get AGC area.

PowerSystems.get_bias — Method

get_bias(value::AGC) -> Float64

Get AGC bias.

PowerSystems.get_delta_t — Method

get_delta_t(value::AGC) -> Float64

Get AGC delta_t.

PowerSystems.get_ext — Method

get_ext(value::AGC) -> Dict{String, Any}

Get AGC ext.

PowerSystems.get_initial_ace — Method

get_initial_ace(value::AGC) -> Float64

Get AGC initial_ace.

PowerSystems.get_reserves — Method

get_reserves(value::AGC) -> Vector{Reserve}

Get AGC reserves.

PowerSystems.set_K_d! — Method

set_K_d!(value::AGC, val) -> Any

Set AGC K_d.

PowerSystems.set_K_i! — Method

set_K_i!(value::AGC, val) -> Any

Set AGC K_i.

PowerSystems.set_K_p! — Method

set_K_p!(value::AGC, val) -> Any

Set AGC K_p.

PowerSystems.set_area! — Method

set_area!(value::AGC, val) -> Any

Set AGC area.

PowerSystems.set_bias! — Method

set_bias!(value::AGC, val) -> Any

Set AGC bias.

PowerSystems.set_delta_t! — Method

set_delta_t!(value::AGC, val) -> Any

Set AGC delta_t.

PowerSystems.set_ext! — Method

set_ext!(value::AGC, val) -> Any

Set AGC ext.

PowerSystems.set_initial_ace! — Method

set_initial_ace!(value::AGC, val) -> Any

Set AGC initial_ace.

PowerSystems.set_reserves! — Method

set_reserves!(value::AGC, val) -> Any

Set AGC reserves.