AVR

AC1A

PowerSystems.AC1A — Type

mutable struct AC1A <: AVR
    Tr::Float64
    Tb::Float64
    Tc::Float64
    Ka::Float64
    Ta::Float64
    Va_lim::Tuple{Float64, Float64}
    Te::Float64
    Kf::Float64
    Tf::Float64
    Kc::Float64
    Kd::Float64
    Ke::Float64
    E_sat::Tuple{Float64, Float64}
    Se::Tuple{Float64, Float64}
    Vr_lim::Tuple{Float64, Float64}
    V_ref::Float64
    saturation_coeffs::Tuple{Float64, Float64}
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    states_types::Vector{StateTypes.StateType}
    internal::InfrastructureSystemsInternal
end

This excitation systems consists of an alternator main exciter feeding its output via non-controlled rectifiers. The exciter does not employ self-excitation, and the voltage regulator power is taken from a source that is not affected by external transients. Parameters of IEEE Std 421.5 Type AC1A Excitacion System. ESAC1A in PSSE and PSLF

Arguments

Tr::Float64: Regulator input filter time constant in s, validation range: (0, nothing)

Tb::Float64: Regulator denominator (lag) time constant in s, validation range: (0, nothing)

Tc::Float64: Regulator numerator (lead) time constant in s, validation range: (0, nothing)

Ka::Float64: Regulator output gain, validation range: (0, nothing)

Ta::Float64: Regulator output time constant in s, validation range: (0, nothing)

Va_lim::Tuple{Float64, Float64}: Limits for regulator output (Vamin, Vamax)

Te::Float64: Exciter field time constant in s, validation range: ("eps()", nothing)

Kf::Float64: Rate feedback excitation system stabilizer gain, validation range: (0, nothing)

Tf::Float64: Rate feedback time constant, validation range: ("eps()", nothing)

Kc::Float64: Rectifier loading factor proportional to commutating reactance, validation range: (0, nothing)

Kd::Float64: Demagnetizing factor, function of exciter alternator reactances, validation range: (0, nothing)

Ke::Float64: Exciter field proportional constant, validation range: (0, nothing)

E_sat::Tuple{Float64, Float64}: Exciter output voltage for saturation factor: (E1, E2)

Se::Tuple{Float64, Float64}: Exciter saturation factor at exciter output voltage: (Se(E1), Se(E2))

Vr_lim::Tuple{Float64, Float64}: Limits for exciter field voltage: (Vrmin, Vrmax)

V_ref::Float64: Reference Voltage Set-point, validation range: (0, nothing)

saturation_coeffs::Tuple{Float64, Float64}: Coefficients (A,B) of the function: Se(x) = B(x - A)^2/x

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

Vm: Sensed terminal voltage,
Vr1: Lead-lag state,
Vr2: Regulator output state,
Ve: Integrator output state,
Vr3: Feedback output state

n_states::Int64: AC1A has 5 states

states_types::Vector{StateTypes.StateType}: AC1A has 5 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_E_sat — Method

get_E_sat(value::AC1A) -> Tuple{Float64,Float64}

Get AC1A E_sat.

PowerSystems.get_Ka — Method

get_Ka(value::AC1A) -> Float64

Get AC1A Ka.

PowerSystems.get_Kc — Method

get_Kc(value::AC1A) -> Float64

Get AC1A Kc.

PowerSystems.get_Kd — Method

get_Kd(value::AC1A) -> Float64

Get AC1A Kd.

PowerSystems.get_Ke — Method

get_Ke(value::AC1A) -> Float64

Get AC1A Ke.

PowerSystems.get_Kf — Method

get_Kf(value::AC1A) -> Float64

Get AC1A Kf.

PowerSystems.get_Se — Method

get_Se(value::AC1A) -> Tuple{Float64,Float64}

Get AC1A Se.

PowerSystems.get_Ta — Method

get_Ta(value::AC1A) -> Float64

Get AC1A Ta.

PowerSystems.get_Tb — Method

get_Tb(value::AC1A) -> Float64

Get AC1A Tb.

PowerSystems.get_Tc — Method

get_Tc(value::AC1A) -> Float64

Get AC1A Tc.

PowerSystems.get_Te — Method

get_Te(value::AC1A) -> Float64

Get AC1A Te.

PowerSystems.get_Tf — Method

get_Tf(value::AC1A) -> Float64

Get AC1A Tf.

PowerSystems.get_Tr — Method

get_Tr(value::AC1A) -> Float64

Get AC1A Tr.

PowerSystems.get_V_ref — Method

get_V_ref(value::AC1A) -> Float64

Get AC1A V_ref.

PowerSystems.get_Va_lim — Method

get_Va_lim(value::AC1A) -> Tuple{Float64,Float64}

Get AC1A Va_lim.

PowerSystems.get_Vr_lim — Method

get_Vr_lim(value::AC1A) -> Tuple{Float64,Float64}

Get AC1A Vr_lim.

PowerSystems.get_ext — Method

get_ext(value::AC1A) -> Dict{String,Any}

Get AC1A ext.

PowerSystems.get_internal — Method

get_internal(value::AC1A) -> InfrastructureSystems.InfrastructureSystemsInternal

Get AC1A internal.

PowerSystems.get_n_states — Method

get_n_states(value::AC1A) -> Int64

Get AC1A n_states.

PowerSystems.get_saturation_coeffs — Method

get_saturation_coeffs(value::AC1A) -> Tuple{Float64,Float64}

Get AC1A saturation_coeffs.

PowerSystems.get_states — Method

get_states(value::AC1A) -> Array{Symbol,1}

Get AC1A states.

PowerSystems.get_states_types — Method

get_states_types(value::AC1A) -> Array{PowerSystems.StateTypes.StateType,1}

Get AC1A states_types.

PowerSystems.set_E_sat! — Method

set_E_sat!(value::AC1A, val::Any) -> Any

Set AC1A E_sat.

PowerSystems.set_Ka! — Method

set_Ka!(value::AC1A, val::Any) -> Any

Set AC1A Ka.

PowerSystems.set_Kc! — Method

set_Kc!(value::AC1A, val::Any) -> Any

Set AC1A Kc.

PowerSystems.set_Kd! — Method

set_Kd!(value::AC1A, val::Any) -> Any

Set AC1A Kd.

PowerSystems.set_Ke! — Method

set_Ke!(value::AC1A, val::Any) -> Any

Set AC1A Ke.

PowerSystems.set_Kf! — Method

set_Kf!(value::AC1A, val::Any) -> Any

Set AC1A Kf.

PowerSystems.set_Se! — Method

set_Se!(value::AC1A, val::Any) -> Any

Set AC1A Se.

PowerSystems.set_Ta! — Method

set_Ta!(value::AC1A, val::Any) -> Any

Set AC1A Ta.

PowerSystems.set_Tb! — Method

set_Tb!(value::AC1A, val::Any) -> Any

Set AC1A Tb.

PowerSystems.set_Tc! — Method

set_Tc!(value::AC1A, val::Any) -> Any

Set AC1A Tc.

PowerSystems.set_Te! — Method

set_Te!(value::AC1A, val::Any) -> Any

Set AC1A Te.

PowerSystems.set_Tf! — Method

set_Tf!(value::AC1A, val::Any) -> Any

Set AC1A Tf.

PowerSystems.set_Tr! — Method

set_Tr!(value::AC1A, val::Any) -> Any

Set AC1A Tr.

PowerSystems.set_V_ref! — Method

set_V_ref!(value::AC1A, val::Any) -> Any

Set AC1A V_ref.

PowerSystems.set_Va_lim! — Method

set_Va_lim!(value::AC1A, val::Any) -> Any

Set AC1A Va_lim.

PowerSystems.set_Vr_lim! — Method

set_Vr_lim!(value::AC1A, val::Any) -> Any

Set AC1A Vr_lim.

PowerSystems.set_ext! — Method

set_ext!(value::AC1A, val::Any) -> Any

Set AC1A ext.

PowerSystems.set_internal! — Method

set_internal!(value::AC1A, val::Any) -> Any

Set AC1A internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::AC1A, val::Any) -> Any

Set AC1A n_states.

PowerSystems.set_saturation_coeffs! — Method

set_saturation_coeffs!(value::AC1A, val::Any) -> Any

Set AC1A saturation_coeffs.

PowerSystems.set_states! — Method

set_states!(value::AC1A, val::Any) -> Any

Set AC1A states.

PowerSystems.set_states_types! — Method

set_states_types!(value::AC1A, val::Any) -> Any

Set AC1A states_types.

AVRFixed

PowerSystems.AVRFixed — Type

mutable struct AVRFixed <: AVR
    Vf::Float64
    V_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    states_types::Vector{StateTypes.StateType}
    internal::InfrastructureSystemsInternal
end

Parameters of a AVR that returns a fixed voltage to the rotor winding

Arguments

Vf::Float64: Fixed voltage field applied to the rotor winding, validation range: (0, nothing)

V_ref::Float64: Reference Voltage Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: Fixed AVR has no states

n_states::Int64: Fixed AVR has no states

states_types::Vector{StateTypes.StateType}: Fixed AVR has no states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_V_ref — Method

get_V_ref(value::AVRFixed) -> Float64

Get AVRFixed V_ref.

PowerSystems.get_Vf — Method

get_Vf(value::AVRFixed) -> Float64

Get AVRFixed Vf.

PowerSystems.get_ext — Method

get_ext(value::AVRFixed) -> Dict{String,Any}

Get AVRFixed ext.

PowerSystems.get_internal — Method

get_internal(value::AVRFixed) -> InfrastructureSystems.InfrastructureSystemsInternal

Get AVRFixed internal.

PowerSystems.get_n_states — Method

get_n_states(value::AVRFixed) -> Int64

Get AVRFixed n_states.

PowerSystems.get_states — Method

get_states(value::AVRFixed) -> Array{Symbol,1}

Get AVRFixed states.

PowerSystems.get_states_types — Method

get_states_types(value::AVRFixed) -> Array{PowerSystems.StateTypes.StateType,1}

Get AVRFixed states_types.

PowerSystems.set_V_ref! — Method

set_V_ref!(value::AVRFixed, val::Any) -> Any

Set AVRFixed V_ref.

PowerSystems.set_Vf! — Method

set_Vf!(value::AVRFixed, val::Any) -> Any

Set AVRFixed Vf.

PowerSystems.set_ext! — Method

set_ext!(value::AVRFixed, val::Any) -> Any

Set AVRFixed ext.

PowerSystems.set_internal! — Method

set_internal!(value::AVRFixed, val::Any) -> Any

Set AVRFixed internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::AVRFixed, val::Any) -> Any

Set AVRFixed n_states.

PowerSystems.set_states! — Method

set_states!(value::AVRFixed, val::Any) -> Any

Set AVRFixed states.

PowerSystems.set_states_types! — Method

set_states_types!(value::AVRFixed, val::Any) -> Any

Set AVRFixed states_types.

AVRSimple

PowerSystems.AVRSimple — Type

mutable struct AVRSimple <: AVR
    Kv::Float64
    V_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    states_types::Vector{StateTypes.StateType}
    internal::InfrastructureSystemsInternal
end

Parameters of a simple proportional AVR in the derivative of EMF i.e. an integrator controller on EMF

Arguments

Kv::Float64: Proportional Gain, validation range: (0, nothing)

V_ref::Float64: Reference Voltage Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

Vf: field voltage

n_states::Int64: Fixed AVR has 1 states

states_types::Vector{StateTypes.StateType}: Simple AVR has 1 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_Kv — Method

get_Kv(value::AVRSimple) -> Float64

Get AVRSimple Kv.

PowerSystems.get_V_ref — Method

get_V_ref(value::AVRSimple) -> Float64

Get AVRSimple V_ref.

PowerSystems.get_ext — Method

get_ext(value::AVRSimple) -> Dict{String,Any}

Get AVRSimple ext.

PowerSystems.get_internal — Method

get_internal(value::AVRSimple) -> InfrastructureSystems.InfrastructureSystemsInternal

Get AVRSimple internal.

PowerSystems.get_n_states — Method

get_n_states(value::AVRSimple) -> Int64

Get AVRSimple n_states.

PowerSystems.get_states — Method

get_states(value::AVRSimple) -> Array{Symbol,1}

Get AVRSimple states.

PowerSystems.get_states_types — Method

get_states_types(value::AVRSimple) -> Array{PowerSystems.StateTypes.StateType,1}

Get AVRSimple states_types.

PowerSystems.set_Kv! — Method

set_Kv!(value::AVRSimple, val::Any) -> Any

Set AVRSimple Kv.

PowerSystems.set_V_ref! — Method

set_V_ref!(value::AVRSimple, val::Any) -> Any

Set AVRSimple V_ref.

PowerSystems.set_ext! — Method

set_ext!(value::AVRSimple, val::Any) -> Any

Set AVRSimple ext.

PowerSystems.set_internal! — Method

set_internal!(value::AVRSimple, val::Any) -> Any

Set AVRSimple internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::AVRSimple, val::Any) -> Any

Set AVRSimple n_states.

PowerSystems.set_states! — Method

set_states!(value::AVRSimple, val::Any) -> Any

Set AVRSimple states.

PowerSystems.set_states_types! — Method

set_states_types!(value::AVRSimple, val::Any) -> Any

Set AVRSimple states_types.

AVRTypeI

PowerSystems.AVRTypeI — Type

mutable struct AVRTypeI <: AVR
    Ka::Float64
    Ke::Float64
    Kf::Float64
    Ta::Float64
    Te::Float64
    Tf::Float64
    Tr::Float64
    Vr_max::Float64
    Vr_min::Float64
    Ae::Float64
    Be::Float64
    V_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    states_types::Vector{StateTypes.StateType}
    internal::InfrastructureSystemsInternal
end

Parameters of an Automatic Voltage Regulator Type I - Resembles IEEE Type DC1

Arguments

Ka::Float64: Amplifier Gain, validation range: (0, nothing)

Ke::Float64: Field circuit integral deviation, validation range: (0, nothing)

Kf::Float64: Stabilizer Gain in s * pu/pu, validation range: (0, nothing)

Ta::Float64: Amplifier Time Constant in s, validation range: (0, nothing)

Te::Float64: Field Circuit Time Constant in s, validation range: (0, nothing)

Tf::Float64: Stabilizer Time Constant in s, validation range: (0, nothing)

Tr::Float64: Voltage Measurement Time Constant in s, validation range: (0, nothing)

Vr_max::Float64: Maximum regulator voltage in pu, validation range: (0, nothing)

Vr_min::Float64: Minimum regulator voltage in pu, validation range: (0, nothing)

Ae::Float64: 1st ceiling coefficient, validation range: (0, nothing)

Be::Float64: 2nd ceiling coefficient, validation range: (0, nothing)

V_ref::Float64: Reference Voltage Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

Vf: Voltage field,
Vr1: Amplifier State,
Vr2: Stabilizing Feedback State,
Vm: Measured voltage

n_states::Int64: The AVR Type I has 4 states

states_types::Vector{StateTypes.StateType}: AVR Type I has 4 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_Ae — Method

get_Ae(value::AVRTypeI) -> Float64

Get AVRTypeI Ae.

PowerSystems.get_Be — Method

get_Be(value::AVRTypeI) -> Float64

Get AVRTypeI Be.

PowerSystems.get_Ka — Method

get_Ka(value::AVRTypeI) -> Float64

Get AVRTypeI Ka.

PowerSystems.get_Ke — Method

get_Ke(value::AVRTypeI) -> Float64

Get AVRTypeI Ke.

PowerSystems.get_Kf — Method

get_Kf(value::AVRTypeI) -> Float64

Get AVRTypeI Kf.

PowerSystems.get_Ta — Method

get_Ta(value::AVRTypeI) -> Float64

Get AVRTypeI Ta.

PowerSystems.get_Te — Method

get_Te(value::AVRTypeI) -> Float64

Get AVRTypeI Te.

PowerSystems.get_Tf — Method

get_Tf(value::AVRTypeI) -> Float64

Get AVRTypeI Tf.

PowerSystems.get_Tr — Method

get_Tr(value::AVRTypeI) -> Float64

Get AVRTypeI Tr.

PowerSystems.get_V_ref — Method

get_V_ref(value::AVRTypeI) -> Float64

Get AVRTypeI V_ref.

PowerSystems.get_Vr_max — Method

get_Vr_max(value::AVRTypeI) -> Float64

Get AVRTypeI Vr_max.

PowerSystems.get_Vr_min — Method

get_Vr_min(value::AVRTypeI) -> Float64

Get AVRTypeI Vr_min.

PowerSystems.get_ext — Method

get_ext(value::AVRTypeI) -> Dict{String,Any}

Get AVRTypeI ext.

PowerSystems.get_internal — Method

get_internal(value::AVRTypeI) -> InfrastructureSystems.InfrastructureSystemsInternal

Get AVRTypeI internal.

PowerSystems.get_n_states — Method

get_n_states(value::AVRTypeI) -> Int64

Get AVRTypeI n_states.

PowerSystems.get_states — Method

get_states(value::AVRTypeI) -> Array{Symbol,1}

Get AVRTypeI states.

PowerSystems.get_states_types — Method

get_states_types(value::AVRTypeI) -> Array{PowerSystems.StateTypes.StateType,1}

Get AVRTypeI states_types.

PowerSystems.set_Ae! — Method

set_Ae!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Ae.

PowerSystems.set_Be! — Method

set_Be!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Be.

PowerSystems.set_Ka! — Method

set_Ka!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Ka.

PowerSystems.set_Ke! — Method

set_Ke!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Ke.

PowerSystems.set_Kf! — Method

set_Kf!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Kf.

PowerSystems.set_Ta! — Method

set_Ta!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Ta.

PowerSystems.set_Te! — Method

set_Te!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Te.

PowerSystems.set_Tf! — Method

set_Tf!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Tf.

PowerSystems.set_Tr! — Method

set_Tr!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Tr.

PowerSystems.set_V_ref! — Method

set_V_ref!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI V_ref.

PowerSystems.set_Vr_max! — Method

set_Vr_max!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Vr_max.

PowerSystems.set_Vr_min! — Method

set_Vr_min!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI Vr_min.

PowerSystems.set_ext! — Method

set_ext!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI ext.

PowerSystems.set_internal! — Method

set_internal!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI n_states.

PowerSystems.set_states! — Method

set_states!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI states.

PowerSystems.set_states_types! — Method

set_states_types!(value::AVRTypeI, val::Any) -> Any

Set AVRTypeI states_types.

AVRTypeII

PowerSystems.AVRTypeII — Type

mutable struct AVRTypeII <: AVR
    K0::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    T4::Float64
    Te::Float64
    Tr::Float64
    Vr_max::Float64
    Vr_min::Float64
    Ae::Float64
    Be::Float64
    V_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    states_types::Vector{StateTypes.StateType}
    internal::InfrastructureSystemsInternal
end

Parameters of an Automatic Voltage Regulator Type II - Typical static exciter model

Arguments

K0::Float64: Regulator Gain, validation range: (0, nothing)

T1::Float64: First Pole in s, validation range: (0, nothing)

T2::Float64: First zero in s, validation range: (0, nothing)

T3::Float64: First Pole in s, validation range: (0, nothing)

T4::Float64: First zero in s, validation range: (0, nothing)

Te::Float64: Field Circuit Time Constant in s, validation range: (0, nothing)

Tr::Float64: Voltage Measurement Time Constant in s, validation range: (0, nothing)

Vr_max::Float64: Maximum regulator voltage in pu, validation range: (0, nothing)

Vr_min::Float64: Minimum regulator voltage in pu, validation range: (0, nothing)

Ae::Float64: 1st ceiling coefficient, validation range: (0, nothing)

Be::Float64: 2nd ceiling coefficient, validation range: (0, nothing)

V_ref::Float64: Reference Voltage Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

Vf: Voltage field,
Vr1: First Lead-Lag state,
Vr2: Second lead-lag state,
Vm: Measured voltage

n_states::Int64: AVR Type II has 4 states

states_types::Vector{StateTypes.StateType}: AVR Type II has 4 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_Ae — Method

get_Ae(value::AVRTypeII) -> Float64

Get AVRTypeII Ae.

PowerSystems.get_Be — Method

get_Be(value::AVRTypeII) -> Float64

Get AVRTypeII Be.

PowerSystems.get_K0 — Method

get_K0(value::AVRTypeII) -> Float64

Get AVRTypeII K0.

PowerSystems.get_T1 — Method

get_T1(value::AVRTypeII) -> Float64

Get AVRTypeII T1.

PowerSystems.get_T2 — Method

get_T2(value::AVRTypeII) -> Float64

Get AVRTypeII T2.

PowerSystems.get_T3 — Method

get_T3(value::AVRTypeII) -> Float64

Get AVRTypeII T3.

PowerSystems.get_T4 — Method

get_T4(value::AVRTypeII) -> Float64

Get AVRTypeII T4.

PowerSystems.get_Te — Method

get_Te(value::AVRTypeII) -> Float64

Get AVRTypeII Te.

PowerSystems.get_Tr — Method

get_Tr(value::AVRTypeII) -> Float64

Get AVRTypeII Tr.

PowerSystems.get_V_ref — Method

get_V_ref(value::AVRTypeII) -> Float64

Get AVRTypeII V_ref.

PowerSystems.get_Vr_max — Method

get_Vr_max(value::AVRTypeII) -> Float64

Get AVRTypeII Vr_max.

PowerSystems.get_Vr_min — Method

get_Vr_min(value::AVRTypeII) -> Float64

Get AVRTypeII Vr_min.

PowerSystems.get_ext — Method

get_ext(value::AVRTypeII) -> Dict{String,Any}

Get AVRTypeII ext.

PowerSystems.get_internal — Method

get_internal(value::AVRTypeII) -> InfrastructureSystems.InfrastructureSystemsInternal

Get AVRTypeII internal.

PowerSystems.get_n_states — Method

get_n_states(value::AVRTypeII) -> Int64

Get AVRTypeII n_states.

PowerSystems.get_states — Method

get_states(value::AVRTypeII) -> Array{Symbol,1}

Get AVRTypeII states.

PowerSystems.get_states_types — Method

get_states_types(value::AVRTypeII) -> Array{PowerSystems.StateTypes.StateType,1}

Get AVRTypeII states_types.

PowerSystems.set_Ae! — Method

set_Ae!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII Ae.

PowerSystems.set_Be! — Method

set_Be!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII Be.

PowerSystems.set_K0! — Method

set_K0!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII K0.

PowerSystems.set_T1! — Method

set_T1!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII T1.

PowerSystems.set_T2! — Method

set_T2!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII T2.

PowerSystems.set_T3! — Method

set_T3!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII T3.

PowerSystems.set_T4! — Method

set_T4!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII T4.

PowerSystems.set_Te! — Method

set_Te!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII Te.

PowerSystems.set_Tr! — Method

set_Tr!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII Tr.

PowerSystems.set_V_ref! — Method

set_V_ref!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII V_ref.

PowerSystems.set_Vr_max! — Method

set_Vr_max!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII Vr_max.

PowerSystems.set_Vr_min! — Method

set_Vr_min!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII Vr_min.

PowerSystems.set_ext! — Method

set_ext!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII ext.

PowerSystems.set_internal! — Method

set_internal!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII n_states.

PowerSystems.set_states! — Method

set_states!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII states.

PowerSystems.set_states_types! — Method

set_states_types!(value::AVRTypeII, val::Any) -> Any

Set AVRTypeII states_types.

ModifiedAC1A

PowerSystems.ModifiedAC1A — Type

mutable struct ModifiedAC1A <: AVR
    Tr::Float64
    Tb::Float64
    Tc::Float64
    Ka::Float64
    Ta::Float64
    Vr_lim::Tuple{Float64, Float64}
    Te::Float64
    Kf::Float64
    Tf::Float64
    Kc::Float64
    Kd::Float64
    Ke::Float64
    E_sat::Tuple{Float64, Float64}
    Se::Tuple{Float64, Float64}
    V_ref::Float64
    saturation_coeffs::Tuple{Float64, Float64}
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    states_types::Vector{StateTypes.StateType}
    internal::InfrastructureSystemsInternal
end

Parameters of Modified Type AC1A Excitacion System. EXAC1 in PSSE and PSLF

Arguments

Tr::Float64: Regulator input filter time constant in s, validation range: (0, nothing)

Tb::Float64: Regulator denominator (lag) time constant in s, validation range: (0, nothing)

Tc::Float64: Regulator numerator (lead) time constant in s, validation range: (0, nothing)

Ka::Float64: Regulator output gain, validation range: (0, nothing)

Ta::Float64: Regulator output time constant in s, validation range: (0, nothing)

Vr_lim::Tuple{Float64, Float64}: Limits for regulator output (Vrmin, Vrmax)

Te::Float64: Exciter field time constant in s, validation range: ("eps()", nothing)

Kf::Float64: Rate feedback excitation system stabilizer gain, validation range: (0, nothing)

Tf::Float64: Rate feedback time constant, validation range: ("eps()", nothing)

Kc::Float64: Rectifier loading factor proportional to commutating reactance, validation range: (0, nothing)

Kd::Float64: Demagnetizing factor, function of exciter alternator reactances, validation range: (0, nothing)

Ke::Float64: Exciter field proportional constant, validation range: (0, nothing)

E_sat::Tuple{Float64, Float64}: Exciter output voltage for saturation factor: (E1, E2)

Se::Tuple{Float64, Float64}: Exciter saturation factor at exciter output voltage: (Se(E1), Se(E2))

V_ref::Float64: Reference Voltage Set-point, validation range: (0, nothing)

saturation_coeffs::Tuple{Float64, Float64}: Coefficients (A,B) of the function: Se(V) = B(V - A)^2/V

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

Vm: Sensed terminal voltage,
Vr1: Lead-lag state,
Vr2: Regulator output state,
Ve: Integrator output state,
Vr3: Feedback output state

n_states::Int64

states_types::Vector{StateTypes.StateType}: ModifiedAC1A has 5 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_E_sat — Method

get_E_sat(value::ModifiedAC1A) -> Tuple{Float64,Float64}

Get ModifiedAC1A E_sat.

PowerSystems.get_Ka — Method

get_Ka(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Ka.

PowerSystems.get_Kc — Method

get_Kc(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Kc.

PowerSystems.get_Kd — Method

get_Kd(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Kd.

PowerSystems.get_Ke — Method

get_Ke(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Ke.

PowerSystems.get_Kf — Method

get_Kf(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Kf.

PowerSystems.get_Se — Method

get_Se(value::ModifiedAC1A) -> Tuple{Float64,Float64}

Get ModifiedAC1A Se.

PowerSystems.get_Ta — Method

get_Ta(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Ta.

PowerSystems.get_Tb — Method

get_Tb(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Tb.

PowerSystems.get_Tc — Method

get_Tc(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Tc.

PowerSystems.get_Te — Method

get_Te(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Te.

PowerSystems.get_Tf — Method

get_Tf(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Tf.

PowerSystems.get_Tr — Method

get_Tr(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A Tr.

PowerSystems.get_V_ref — Method

get_V_ref(value::ModifiedAC1A) -> Float64

Get ModifiedAC1A V_ref.

PowerSystems.get_Vr_lim — Method

get_Vr_lim(value::ModifiedAC1A) -> Tuple{Float64,Float64}

Get ModifiedAC1A Vr_lim.

PowerSystems.get_ext — Method

get_ext(value::ModifiedAC1A) -> Dict{String,Any}

Get ModifiedAC1A ext.

PowerSystems.get_internal — Method

get_internal(value::ModifiedAC1A) -> InfrastructureSystems.InfrastructureSystemsInternal

Get ModifiedAC1A internal.

PowerSystems.get_n_states — Method

get_n_states(value::ModifiedAC1A) -> Int64

Get ModifiedAC1A n_states.

PowerSystems.get_saturation_coeffs — Method

get_saturation_coeffs(value::ModifiedAC1A) -> Tuple{Float64,Float64}

Get ModifiedAC1A saturation_coeffs.

PowerSystems.get_states — Method

get_states(value::ModifiedAC1A) -> Array{Symbol,1}

Get ModifiedAC1A states.

PowerSystems.get_states_types — Method

get_states_types(value::ModifiedAC1A) -> Array{PowerSystems.StateTypes.StateType,1}

Get ModifiedAC1A states_types.

PowerSystems.set_E_sat! — Method

set_E_sat!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A E_sat.

PowerSystems.set_Ka! — Method

set_Ka!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Ka.

PowerSystems.set_Kc! — Method

set_Kc!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Kc.

PowerSystems.set_Kd! — Method

set_Kd!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Kd.

PowerSystems.set_Ke! — Method

set_Ke!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Ke.

PowerSystems.set_Kf! — Method

set_Kf!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Kf.

PowerSystems.set_Se! — Method

set_Se!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Se.

PowerSystems.set_Ta! — Method

set_Ta!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Ta.

PowerSystems.set_Tb! — Method

set_Tb!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Tb.

PowerSystems.set_Tc! — Method

set_Tc!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Tc.

PowerSystems.set_Te! — Method

set_Te!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Te.

PowerSystems.set_Tf! — Method

set_Tf!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Tf.

PowerSystems.set_Tr! — Method

set_Tr!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Tr.

PowerSystems.set_V_ref! — Method

set_V_ref!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A V_ref.

PowerSystems.set_Vr_lim! — Method

set_Vr_lim!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A Vr_lim.

PowerSystems.set_ext! — Method

set_ext!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A ext.

PowerSystems.set_internal! — Method

set_internal!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A n_states.

PowerSystems.set_saturation_coeffs! — Method

set_saturation_coeffs!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A saturation_coeffs.

PowerSystems.set_states! — Method

set_states!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A states.

PowerSystems.set_states_types! — Method

set_states_types!(value::ModifiedAC1A, val::Any) -> Any

Set ModifiedAC1A states_types.

ST1A

PowerSystems.ST1A — Type

mutable struct ST1A <: AVR
    UEL_flags::Int64
    PSS_flags::Int64
    Tr::Float64
    Vi_lim::Tuple{Float64, Float64}
    Tc::Float64
    Tb::Float64
    Tc1::Float64
    Tb1::Float64
    Ka::Float64
    Ta::Float64
    Va_lim::Tuple{Float64, Float64}
    Vr_lim::Tuple{Float64, Float64}
    Kc::Float64
    Kf::Float64
    Tf::Float64
    K_lr::Float64
    I_lr::Float64
    V_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    states_types::Vector{StateTypes.StateType}
    internal::InfrastructureSystemsInternal
end

This excitation system supplies power through a transformer from the generator terminals and its regulated by a controlled rectifier (via thyristors). Parameters of IEEE Std 421.5 Type ST1A Excitacion System. ESST1A in PSSE and PSLF

Arguments

UEL_flags::Int64: Code input for Underexcitization limiter (UEL) entry. Not supported., validation range: (1, 3)

PSS_flags::Int64: Code input for Power System Stabilizer (PSS) or (VOS) entry., validation range: (1, 2)

Tr::Float64: Regulator input filter time constant in s, validation range: (0, nothing)

Vi_lim::Tuple{Float64, Float64}: Voltage error limits (regulator input) (Vimin, Vimax)

Tc::Float64: First regulator denominator (lead) time constant in s, validation range: (0, nothing)

Tb::Float64: First regulator denominator (lag) time constant in s, validation range: (0, nothing)

Tc1::Float64: Second regulator denominator (lead) time constant in s, validation range: (0, nothing)

Tb1::Float64: Second regulator denominator (lead) time constant in s, validation range: (0, nothing)

Ka::Float64: Voltage regulator gain, validation range: (0, nothing)

Ta::Float64: Voltage regulator time constant in s, validation range: (0, nothing)

Va_lim::Tuple{Float64, Float64}: Limits for regulator output (Vamin, Vamax)

Vr_lim::Tuple{Float64, Float64}: Limits for exciter output (Vrmin, Vrmax)

Kc::Float64: Rectifier loading factor proportional to commutating reactance, validation range: (0, nothing)

Kf::Float64: Rate feedback gain, validation range: (0, nothing)

Tf::Float64: Rate feedback time constant in s, validation range: ("eps()", nothing)

K_lr::Float64: Exciter output current limiter gain, validation range: (0, nothing)

I_lr::Float64: Exciter output current limit reference, validation range: (0, nothing)

V_ref::Float64: Reference Voltage Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

Vm: Sensed terminal voltage,
Vr1: First Lead-lag state,
Vr2: Second lead-lag state,
Va: Regulator output state,
Vr3: Feedback output state

n_states::Int64: ST1A has 5 states

states_types::Vector{StateTypes.StateType}: ST1A has 5 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_I_lr — Method

get_I_lr(value::ST1A) -> Float64

Get ST1A I_lr.

PowerSystems.get_K_lr — Method

get_K_lr(value::ST1A) -> Float64

Get ST1A K_lr.

PowerSystems.get_Ka — Method

get_Ka(value::ST1A) -> Float64

Get ST1A Ka.

PowerSystems.get_Kc — Method

get_Kc(value::ST1A) -> Float64

Get ST1A Kc.

PowerSystems.get_Kf — Method

get_Kf(value::ST1A) -> Float64

Get ST1A Kf.

PowerSystems.get_PSS_flags — Method

get_PSS_flags(value::ST1A) -> Int64

Get ST1A PSS_flags.

PowerSystems.get_Ta — Method

get_Ta(value::ST1A) -> Float64

Get ST1A Ta.

PowerSystems.get_Tb — Method

get_Tb(value::ST1A) -> Float64

Get ST1A Tb.

PowerSystems.get_Tb1 — Method

get_Tb1(value::ST1A) -> Float64

Get ST1A Tb1.

PowerSystems.get_Tc — Method

get_Tc(value::ST1A) -> Float64

Get ST1A Tc.

PowerSystems.get_Tc1 — Method

get_Tc1(value::ST1A) -> Float64

Get ST1A Tc1.

PowerSystems.get_Tf — Method

get_Tf(value::ST1A) -> Float64

Get ST1A Tf.

PowerSystems.get_Tr — Method

get_Tr(value::ST1A) -> Float64

Get ST1A Tr.

PowerSystems.get_UEL_flags — Method

get_UEL_flags(value::ST1A) -> Int64

Get ST1A UEL_flags.

PowerSystems.get_V_ref — Method

get_V_ref(value::ST1A) -> Float64

Get ST1A V_ref.

PowerSystems.get_Va_lim — Method

get_Va_lim(value::ST1A) -> Tuple{Float64,Float64}

Get ST1A Va_lim.

PowerSystems.get_Vi_lim — Method

get_Vi_lim(value::ST1A) -> Tuple{Float64,Float64}

Get ST1A Vi_lim.

PowerSystems.get_Vr_lim — Method

get_Vr_lim(value::ST1A) -> Tuple{Float64,Float64}

Get ST1A Vr_lim.

PowerSystems.get_ext — Method

get_ext(value::ST1A) -> Dict{String,Any}

Get ST1A ext.

PowerSystems.get_internal — Method

get_internal(value::ST1A) -> InfrastructureSystems.InfrastructureSystemsInternal

Get ST1A internal.

PowerSystems.get_n_states — Method

get_n_states(value::ST1A) -> Int64

Get ST1A n_states.

PowerSystems.get_states — Method

get_states(value::ST1A) -> Array{Symbol,1}

Get ST1A states.

PowerSystems.get_states_types — Method

get_states_types(value::ST1A) -> Array{PowerSystems.StateTypes.StateType,1}

Get ST1A states_types.

PowerSystems.set_I_lr! — Method

set_I_lr!(value::ST1A, val::Any) -> Any

Set ST1A I_lr.

PowerSystems.set_K_lr! — Method

set_K_lr!(value::ST1A, val::Any) -> Any

Set ST1A K_lr.

PowerSystems.set_Ka! — Method

set_Ka!(value::ST1A, val::Any) -> Any

Set ST1A Ka.

PowerSystems.set_Kc! — Method

set_Kc!(value::ST1A, val::Any) -> Any

Set ST1A Kc.

PowerSystems.set_Kf! — Method

set_Kf!(value::ST1A, val::Any) -> Any

Set ST1A Kf.

PowerSystems.set_PSS_flags! — Method

set_PSS_flags!(value::ST1A, val::Any) -> Any

Set ST1A PSS_flags.

PowerSystems.set_Ta! — Method

set_Ta!(value::ST1A, val::Any) -> Any

Set ST1A Ta.

PowerSystems.set_Tb! — Method

set_Tb!(value::ST1A, val::Any) -> Any

Set ST1A Tb.

PowerSystems.set_Tb1! — Method

set_Tb1!(value::ST1A, val::Any) -> Any

Set ST1A Tb1.

PowerSystems.set_Tc! — Method

set_Tc!(value::ST1A, val::Any) -> Any

Set ST1A Tc.

PowerSystems.set_Tc1! — Method

set_Tc1!(value::ST1A, val::Any) -> Any

Set ST1A Tc1.

PowerSystems.set_Tf! — Method

set_Tf!(value::ST1A, val::Any) -> Any

Set ST1A Tf.

PowerSystems.set_Tr! — Method

set_Tr!(value::ST1A, val::Any) -> Any

Set ST1A Tr.

PowerSystems.set_UEL_flags! — Method

set_UEL_flags!(value::ST1A, val::Any) -> Any

Set ST1A UEL_flags.

PowerSystems.set_V_ref! — Method

set_V_ref!(value::ST1A, val::Any) -> Any

Set ST1A V_ref.

PowerSystems.set_Va_lim! — Method

set_Va_lim!(value::ST1A, val::Any) -> Any

Set ST1A Va_lim.

PowerSystems.set_Vi_lim! — Method

set_Vi_lim!(value::ST1A, val::Any) -> Any

Set ST1A Vi_lim.

PowerSystems.set_Vr_lim! — Method

set_Vr_lim!(value::ST1A, val::Any) -> Any

Set ST1A Vr_lim.

PowerSystems.set_ext! — Method

set_ext!(value::ST1A, val::Any) -> Any

Set ST1A ext.

PowerSystems.set_internal! — Method

set_internal!(value::ST1A, val::Any) -> Any

Set ST1A internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::ST1A, val::Any) -> Any

Set ST1A n_states.

PowerSystems.set_states! — Method

set_states!(value::ST1A, val::Any) -> Any

Set ST1A states.

PowerSystems.set_states_types! — Method

set_states_types!(value::ST1A, val::Any) -> Any

Set ST1A states_types.