ActiveConstantPowerLoad

PowerSystems.ActiveConstantPowerLoad — Type

mutable struct ActiveConstantPowerLoad <: DynamicInjection
    name::String
    r_load::Float64
    c_dc::Float64
    rf::Float64
    lf::Float64
    cf::Float64
    rg::Float64
    lg::Float64
    kp_pll::Float64
    ki_pll::Float64
    kpv::Float64
    kiv::Float64
    kpc::Float64
    kic::Float64
    base_power::Float64
    ext::Dict{String, Any}
    P_ref::Float64
    Q_ref::Float64
    V_ref::Float64
    ω_ref::Float64
    is_filter_differential::Int
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 12-states active power load based on the paper, "Dynamic Stability of a Microgrid With an Active Load."

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

r_load::Float64: DC-side resistor, validation range: (0, nothing)

c_dc::Float64: DC-side capacitor, validation range: (0, nothing)

rf::Float64: Converter side filter resistance, validation range: (0, nothing)

lf::Float64: Converter side filter inductance, validation range: (0, nothing)

cf::Float64: AC Converter filter capacitance, validation range: (0, nothing)

rg::Float64: Network side filter resistance, validation range: (0, nothing)

lg::Float64: Network side filter inductance, validation range: (0, nothing)

kp_pll::Float64: Proportional constant for PI-PLL block, validation range: (0, nothing)

ki_pll::Float64: Integral constant for PI-PLL block, validation range: (0, nothing)

kpv::Float64: Proportional constant for Voltage Control block, validation range: (0, nothing)

kiv::Float64: Integral constant for Voltage Control block, validation range: (0, nothing)

kpc::Float64: Proportional constant for Current Control block, validation range: (0, nothing)

kic::Float64: Integral constant for Current Control block, validation range: (0, nothing)

base_power::Float64: Base power of the unit (MVA) for per unitization, validation range: (0.0001, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

P_ref::Float64: Reference active power (pu)

Q_ref::Float64: Reference reactive power (pu)

V_ref::Float64: Reference voltage (pu)

ω_ref::Float64: Reference frequency (pu)

is_filter_differential::Int: Boolean to decide if filter states are differential or algebraic

states::Vector{Symbol}: (Do not modify.) The states are:

θ_pll: PLL deviation angle, 
ϵ_pll: PLL integrator state, 
η: DC-voltage controller integrator state, 
v_dc: DC voltage at the capacitor, 
γd: d-axis Current controller integrator state, 
γq: q-axis Current controller integrator state, 
ir_cnv: Real current out of the converter,
ii_cnv: Imaginary current out of the converter,
vr_filter: Real voltage at the filter's capacitor,
vi_filter: Imaginary voltage at the filter's capacitor,
ir_filter: Real current out of the filter,
ii_filter: Imaginary current out of the filter

n_states::Int: (Do not modify.) ActiveConstantPowerLoad has 12 states

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::ActiveConstantPowerLoad) -> String

Get ActiveConstantPowerLoad name.

PowerSystems.get_P_ref — Method

get_P_ref(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad P_ref.

PowerSystems.get_Q_ref — Method

get_Q_ref(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad Q_ref.

PowerSystems.get_V_ref — Method

get_V_ref(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad V_ref.

PowerSystems.get_base_power — Method

get_base_power(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad base_power.

PowerSystems.get_c_dc — Method

get_c_dc(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad c_dc.

PowerSystems.get_cf — Method

get_cf(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad cf.

PowerSystems.get_ext — Method

get_ext(value::ActiveConstantPowerLoad) -> Dict{String, Any}

Get ActiveConstantPowerLoad ext.

PowerSystems.get_is_filter_differential — Method

get_is_filter_differential(
    value::ActiveConstantPowerLoad
) -> Int64

Get ActiveConstantPowerLoad is_filter_differential.

PowerSystems.get_ki_pll — Method

get_ki_pll(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad ki_pll.

PowerSystems.get_kic — Method

get_kic(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad kic.

PowerSystems.get_kiv — Method

get_kiv(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad kiv.

PowerSystems.get_kp_pll — Method

get_kp_pll(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad kp_pll.

PowerSystems.get_kpc — Method

get_kpc(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad kpc.

PowerSystems.get_kpv — Method

get_kpv(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad kpv.

PowerSystems.get_lf — Method

get_lf(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad lf.

PowerSystems.get_lg — Method

get_lg(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad lg.

PowerSystems.get_n_states — Method

get_n_states(value::ActiveConstantPowerLoad) -> Int64

Get ActiveConstantPowerLoad n_states.

PowerSystems.get_r_load — Method

get_r_load(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad r_load.

PowerSystems.get_rf — Method

get_rf(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad rf.

PowerSystems.get_rg — Method

get_rg(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad rg.

PowerSystems.get_states — Method

get_states(value::ActiveConstantPowerLoad) -> Vector{Symbol}

Get ActiveConstantPowerLoad states.

PowerSystems.get_ω_ref — Method

get_ω_ref(value::ActiveConstantPowerLoad) -> Float64

Get ActiveConstantPowerLoad ω_ref.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad P_ref.

PowerSystems.set_Q_ref! — Method

set_Q_ref!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad Q_ref.

PowerSystems.set_V_ref! — Method

set_V_ref!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad V_ref.

PowerSystems.set_base_power! — Method

set_base_power!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad base_power.

PowerSystems.set_c_dc! — Method

set_c_dc!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad c_dc.

PowerSystems.set_cf! — Method

set_cf!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad cf.

PowerSystems.set_ext! — Method

set_ext!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad ext.

PowerSystems.set_is_filter_differential! — Method

set_is_filter_differential!(
    value::ActiveConstantPowerLoad,
    val
) -> Any

Set ActiveConstantPowerLoad is_filter_differential.

PowerSystems.set_ki_pll! — Method

set_ki_pll!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad ki_pll.

PowerSystems.set_kic! — Method

set_kic!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad kic.

PowerSystems.set_kiv! — Method

set_kiv!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad kiv.

PowerSystems.set_kp_pll! — Method

set_kp_pll!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad kp_pll.

PowerSystems.set_kpc! — Method

set_kpc!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad kpc.

PowerSystems.set_kpv! — Method

set_kpv!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad kpv.

PowerSystems.set_lf! — Method

set_lf!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad lf.

PowerSystems.set_lg! — Method

set_lg!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad lg.

PowerSystems.set_r_load! — Method

set_r_load!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad r_load.

PowerSystems.set_rf! — Method

set_rf!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad rf.

PowerSystems.set_rg! — Method

set_rg!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad rg.

PowerSystems.set_ω_ref! — Method

set_ω_ref!(value::ActiveConstantPowerLoad, val) -> Any

Set ActiveConstantPowerLoad ω_ref.