AggregateDistributedGenerationA

PowerSystems.AggregateDistributedGenerationA — Type

mutable struct AggregateDistributedGenerationA <: DynamicInjection
    name::String
    Pf_Flag::Int
    Freq_Flag::Int
    PQ_Flag::Int
    Gen_Flag::Int
    Vtrip_Flag::Int
    Ftrip_Flag::Int
    T_rv::Float64
    Trf::Float64
    dbd_pnts::Tuple{Float64, Float64}
    K_qv::Float64
    Tp::Float64
    T_iq::Float64
    D_dn::Float64
    D_up::Float64
    fdbd_pnts::Tuple{Float64, Float64}
    fe_lim::MinMax
    P_lim::MinMax
    dP_lim::MinMax
    Tpord::Float64
    Kpg::Float64
    Kig::Float64
    I_max::Float64
    vl_pnts::Vector{Tuple{Float64,Float64}}
    vh_pnts::Vector{Tuple{Float64,Float64}}
    Vrfrac::Float64
    fl::Float64
    fh::Float64
    tfl::Float64
    tfh::Float64
    Tg::Float64
    rrpwr::Float64
    Tv::Float64
    Vpr::Float64
    Iq_lim::MinMax
    V_ref::Float64
    Pfa_ref::Float64
    ω_ref::Float64
    Q_ref::Float64
    P_ref::Float64
    base_power::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of the DERA1 model in PSS/E

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

Pf_Flag::Int: Flag for Power Factor Control, validation range: (0, 1)

Freq_Flag::Int: Flag to enable/disable frequency control, validation range: (0, 1)

PQ_Flag::Int: Flag used to enforce maximum current, validation range: (0, 1)

Gen_Flag::Int: Flag to specify generator or storage, validation range: (0, 1)

Vtrip_Flag::Int: Flag to enable/disable voltage trip logic, validation range: (0, 1)

Ftrip_Flag::Int: Flag to enable/disable frequency trip logic, validation range: (0, 1)

T_rv::Float64: Voltage measurement transducer time constant, validation range: (0, nothing)

Trf::Float64: Frequency measurement transducer time constant, validation range: (0, nothing)

dbd_pnts::Tuple{Float64, Float64}: Voltage deadband thresholds (dbd1, dbd2)

K_qv::Float64: Proportional voltage control gain (pu), validation range: (0, nothing)

Tp::Float64: Power measurement transducer time constant, validation range: (0, nothing)

T_iq::Float64: Time constant for low-pass filter for state q_V when QFlag = 0, validation range: (0, nothing)

D_dn::Float64: Reciprocal of droop for over-frequency conditions (>0) (pu), validation range: (0, nothing)

D_up::Float64: Reciprocal of droop for under-frequency conditions <=0) (pu), validation range: (0, nothing)

fdbd_pnts::Tuple{Float64, Float64}: Frequency control deadband thresholds (fdbd1, fdbd2)

fe_lim::MinMax: Frequency error limits (femin, femax)

P_lim::MinMax: Power limits (Pmin, Pmax)

dP_lim::MinMax: Power reference ramp rate limits (dPmin, dPmax)

Tpord::Float64: Power filter time constant, validation range: (0, nothing)

Kpg::Float64: PI controller proportional gain (pu), validation range: (0, nothing)

Kig::Float64: PI controller integral gain (pu), validation range: (0, nothing)

I_max::Float64: Maximum limit on total converter current (pu), validation range: (0, nothing)

vl_pnts::Vector{Tuple{Float64,Float64}}: Low voltage cutout points [(tv10, vl0), (tv11, vl1)]

vh_pnts::Vector{Tuple{Float64,Float64}}: High voltage cutout points [(tvh0, vh0), (tvh1, vh1)]

Vrfrac::Float64: Fraction of device that recovers after voltage comes back to within vl1 < V < vh1 (0 <= Vrfrac <= 1), validation range: (0, 1)

fl::Float64: Inverter frequency break-point for low frequency cut-out (Hz), validation range: (0, nothing)

fh::Float64: Inverter frequency break-point for high frequency cut-out (Hz), validation range: (0, nothing)

tfl::Float64: Low frequency cut-out timer corresponding to frequency fl (s), validation range: (0, nothing)

tfh::Float64: High frequency cut-out timer corresponding to frequency fh (s), validation range: (0, nothing)

Tg::Float64: Current control time constant (to represent behavior of inner control loops) (> 0) (s), validation range: (0, nothing)

rrpwr::Float64: Ramp rate for real power increase following a fault (pu/s), validation range: (0, nothing)

Tv::Float64: Time constant on the output of the multiplier (s), validation range: (0, nothing)

Vpr::Float64: Voltage below which frequency tripping is disabled (pu), validation range: (0, nothing)

Iq_lim::MinMax: Reactive current injection limits (Iqll, Iqhl)

V_ref::Float64: (default: 1.0) User defined voltage reference. If 0, PowerSimulationsDynamics.jl initializes to initial terminal voltage, validation range: (0, nothing)

Pfa_ref::Float64: (default: 0.0) Reference power factor, validation range: (0, nothing)

ω_ref::Float64: (default: 1.0) Reference Frequency (pu), validation range: (0, nothing)

Q_ref::Float64: (default: 0.0) Reference reactive power, in pu, validation range: (0, nothing)

P_ref::Float64: (default: 1.0) Reference active power, in pu, validation range: (0, nothing)

base_power::Float64: (default: 100.0) Base power (MVA) for per unitization

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

states::Vector{Symbol}: (Do not modify.) The states of AggregateDistributedGenerationA depends on the Flags

n_states::Int: (Do not modify.) The states of AggregateDistributedGenerationA depends on the Flags

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::AggregateDistributedGenerationA) -> String

Get AggregateDistributedGenerationA name.

PowerSystems.get_D_dn — Method

get_D_dn(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA D_dn.

PowerSystems.get_D_up — Method

get_D_up(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA D_up.

PowerSystems.get_Freq_Flag — Method

get_Freq_Flag(
    value::AggregateDistributedGenerationA
) -> Int64

Get AggregateDistributedGenerationA Freq_Flag.

PowerSystems.get_Ftrip_Flag — Method

get_Ftrip_Flag(
    value::AggregateDistributedGenerationA
) -> Int64

Get AggregateDistributedGenerationA Ftrip_Flag.

PowerSystems.get_Gen_Flag — Method

get_Gen_Flag(
    value::AggregateDistributedGenerationA
) -> Int64

Get AggregateDistributedGenerationA Gen_Flag.

PowerSystems.get_I_max — Method

get_I_max(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA I_max.

PowerSystems.get_Iq_lim — Method

get_Iq_lim(
    value::AggregateDistributedGenerationA
) -> @NamedTuple{min::Float64, max::Float64}

Get AggregateDistributedGenerationA Iq_lim.

PowerSystems.get_K_qv — Method

get_K_qv(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA K_qv.

PowerSystems.get_Kig — Method

get_Kig(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA Kig.

PowerSystems.get_Kpg — Method

get_Kpg(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA Kpg.

PowerSystems.get_PQ_Flag — Method

get_PQ_Flag(value::AggregateDistributedGenerationA) -> Int64

Get AggregateDistributedGenerationA PQ_Flag.

PowerSystems.get_P_lim — Method

get_P_lim(
    value::AggregateDistributedGenerationA
) -> @NamedTuple{min::Float64, max::Float64}

Get AggregateDistributedGenerationA P_lim.

PowerSystems.get_P_ref — Method

get_P_ref(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA P_ref.

PowerSystems.get_Pf_Flag — Method

get_Pf_Flag(value::AggregateDistributedGenerationA) -> Int64

Get AggregateDistributedGenerationA Pf_Flag.

PowerSystems.get_Pfa_ref — Method

get_Pfa_ref(
    value::AggregateDistributedGenerationA
) -> Float64

Get AggregateDistributedGenerationA Pfa_ref.

PowerSystems.get_Q_ref — Method

get_Q_ref(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA Q_ref.

PowerSystems.get_T_iq — Method

get_T_iq(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA T_iq.

PowerSystems.get_T_rv — Method

get_T_rv(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA T_rv.

PowerSystems.get_Tg — Method

get_Tg(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA Tg.

PowerSystems.get_Tp — Method

get_Tp(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA Tp.

PowerSystems.get_Tpord — Method

get_Tpord(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA Tpord.

PowerSystems.get_Trf — Method

get_Trf(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA Trf.

PowerSystems.get_Tv — Method

get_Tv(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA Tv.

PowerSystems.get_V_ref — Method

get_V_ref(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA V_ref.

PowerSystems.get_Vpr — Method

get_Vpr(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA Vpr.

PowerSystems.get_Vrfrac — Method

get_Vrfrac(
    value::AggregateDistributedGenerationA
) -> Float64

Get AggregateDistributedGenerationA Vrfrac.

PowerSystems.get_Vtrip_Flag — Method

get_Vtrip_Flag(
    value::AggregateDistributedGenerationA
) -> Int64

Get AggregateDistributedGenerationA Vtrip_Flag.

PowerSystems.get_base_power — Method

get_base_power(
    value::AggregateDistributedGenerationA
) -> Float64

Get AggregateDistributedGenerationA base_power.

PowerSystems.get_dP_lim — Method

get_dP_lim(
    value::AggregateDistributedGenerationA
) -> @NamedTuple{min::Float64, max::Float64}

Get AggregateDistributedGenerationA dP_lim.

PowerSystems.get_dbd_pnts — Method

get_dbd_pnts(
    value::AggregateDistributedGenerationA
) -> Tuple{Float64, Float64}

Get AggregateDistributedGenerationA dbd_pnts.

PowerSystems.get_ext — Method

get_ext(
    value::AggregateDistributedGenerationA
) -> Dict{String, Any}

Get AggregateDistributedGenerationA ext.

PowerSystems.get_fdbd_pnts — Method

get_fdbd_pnts(
    value::AggregateDistributedGenerationA
) -> Tuple{Float64, Float64}

Get AggregateDistributedGenerationA fdbd_pnts.

PowerSystems.get_fe_lim — Method

get_fe_lim(
    value::AggregateDistributedGenerationA
) -> @NamedTuple{min::Float64, max::Float64}

Get AggregateDistributedGenerationA fe_lim.

PowerSystems.get_fh — Method

get_fh(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA fh.

PowerSystems.get_fl — Method

get_fl(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA fl.

PowerSystems.get_n_states — Method

get_n_states(
    value::AggregateDistributedGenerationA
) -> Int64

Get AggregateDistributedGenerationA n_states.

PowerSystems.get_rrpwr — Method

get_rrpwr(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA rrpwr.

PowerSystems.get_states — Method

get_states(
    value::AggregateDistributedGenerationA
) -> Vector{Symbol}

Get AggregateDistributedGenerationA states.

PowerSystems.get_tfh — Method

get_tfh(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA tfh.

PowerSystems.get_tfl — Method

get_tfl(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA tfl.

PowerSystems.get_vh_pnts — Method

get_vh_pnts(
    value::AggregateDistributedGenerationA
) -> Vector{Tuple{Float64, Float64}}

Get AggregateDistributedGenerationA vh_pnts.

PowerSystems.get_vl_pnts — Method

get_vl_pnts(
    value::AggregateDistributedGenerationA
) -> Vector{Tuple{Float64, Float64}}

Get AggregateDistributedGenerationA vl_pnts.

PowerSystems.get_ω_ref — Method

get_ω_ref(value::AggregateDistributedGenerationA) -> Float64

Get AggregateDistributedGenerationA ω_ref.

PowerSystems.set_D_dn! — Method

set_D_dn!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA D_dn.

PowerSystems.set_D_up! — Method

set_D_up!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA D_up.

PowerSystems.set_Freq_Flag! — Method

set_Freq_Flag!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Freq_Flag.

PowerSystems.set_Ftrip_Flag! — Method

set_Ftrip_Flag!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Ftrip_Flag.

PowerSystems.set_Gen_Flag! — Method

set_Gen_Flag!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Gen_Flag.

PowerSystems.set_I_max! — Method

set_I_max!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA I_max.

PowerSystems.set_Iq_lim! — Method

set_Iq_lim!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Iq_lim.

PowerSystems.set_K_qv! — Method

set_K_qv!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA K_qv.

PowerSystems.set_Kig! — Method

set_Kig!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA Kig.

PowerSystems.set_Kpg! — Method

set_Kpg!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA Kpg.

PowerSystems.set_PQ_Flag! — Method

set_PQ_Flag!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA PQ_Flag.

PowerSystems.set_P_lim! — Method

set_P_lim!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA P_lim.

PowerSystems.set_P_ref! — Method

set_P_ref!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA P_ref.

PowerSystems.set_Pf_Flag! — Method

set_Pf_Flag!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Pf_Flag.

PowerSystems.set_Pfa_ref! — Method

set_Pfa_ref!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Pfa_ref.

PowerSystems.set_Q_ref! — Method

set_Q_ref!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Q_ref.

PowerSystems.set_T_iq! — Method

set_T_iq!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA T_iq.

PowerSystems.set_T_rv! — Method

set_T_rv!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA T_rv.

PowerSystems.set_Tg! — Method

set_Tg!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA Tg.

PowerSystems.set_Tp! — Method

set_Tp!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA Tp.

PowerSystems.set_Tpord! — Method

set_Tpord!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Tpord.

PowerSystems.set_Trf! — Method

set_Trf!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA Trf.

PowerSystems.set_Tv! — Method

set_Tv!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA Tv.

PowerSystems.set_V_ref! — Method

set_V_ref!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA V_ref.

PowerSystems.set_Vpr! — Method

set_Vpr!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA Vpr.

PowerSystems.set_Vrfrac! — Method

set_Vrfrac!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Vrfrac.

PowerSystems.set_Vtrip_Flag! — Method

set_Vtrip_Flag!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA Vtrip_Flag.

PowerSystems.set_base_power! — Method

set_base_power!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA base_power.

PowerSystems.set_dP_lim! — Method

set_dP_lim!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA dP_lim.

PowerSystems.set_dbd_pnts! — Method

set_dbd_pnts!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA dbd_pnts.

PowerSystems.set_ext! — Method

set_ext!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA ext.

PowerSystems.set_fdbd_pnts! — Method

set_fdbd_pnts!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA fdbd_pnts.

PowerSystems.set_fe_lim! — Method

set_fe_lim!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA fe_lim.

PowerSystems.set_fh! — Method

set_fh!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA fh.

PowerSystems.set_fl! — Method

set_fl!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA fl.

PowerSystems.set_rrpwr! — Method

set_rrpwr!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA rrpwr.

PowerSystems.set_tfh! — Method

set_tfh!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA tfh.

PowerSystems.set_tfl! — Method

set_tfl!(value::AggregateDistributedGenerationA, val) -> Any

Set AggregateDistributedGenerationA tfl.

PowerSystems.set_vh_pnts! — Method

set_vh_pnts!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA vh_pnts.

PowerSystems.set_vl_pnts! — Method

set_vl_pnts!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA vl_pnts.

PowerSystems.set_ω_ref! — Method

set_ω_ref!(
    value::AggregateDistributedGenerationA,
    val
) -> Any

Set AggregateDistributedGenerationA ω_ref.