Area

PowerSystems.Area — Type

mutable struct Area <: AggregationTopology
    name::String
    peak_active_power::Float64
    peak_reactive_power::Float64
    load_response::Float64
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A collection of buses for control purposes.

The Area can be specified when defining each ACBus or DCBus in the area

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

peak_active_power::Float64: (default: 0.0) Peak active power in the area

peak_reactive_power::Float64: (default: 0.0) Peak reactive power in the area

load_response::Float64: (default: 0.0) Load-frequency damping parameter modeling how much the load in the area changes due to changes in frequency (MW/Hz). Example here.

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::Area) -> String

Get Area name.

PowerSystems.get_ext — Method

get_ext(value::Area) -> Dict{String, Any}

Get Area ext.

PowerSystems.get_load_response — Method

get_load_response(value::Area) -> Float64

Get Area load_response.

PowerSystems.get_peak_active_power — Method

get_peak_active_power(value::Area) -> Float64

Get Area peak_active_power.

PowerSystems.get_peak_reactive_power — Method

get_peak_reactive_power(value::Area) -> Float64

Get Area peak_reactive_power.

PowerSystems.set_ext! — Method

set_ext!(value::Area, val) -> Any

Set Area ext.

PowerSystems.set_load_response! — Method

set_load_response!(value::Area, val) -> Any

Set Area load_response.

PowerSystems.set_peak_active_power! — Method

set_peak_active_power!(value::Area, val) -> Any

Set Area peak_active_power.

PowerSystems.set_peak_reactive_power! — Method

set_peak_reactive_power!(value::Area, val) -> Any

Set Area peak_reactive_power.