CSVGN1

PowerSystems.CSVGN1 — Type

mutable struct CSVGN1 <: DynamicInjection
    name::String
    K::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    T4::Float64
    T5::Float64
    Rmin::Float64
    Vmax::Float64
    Vmin::Float64
    CBase::Float64
    base_power::Float64
    ext::Dict{String, Any}
    R_th::Float64
    X_th::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of static shunt compensator: CSVGN1 in PSSE

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

K::Float64: Gain in pu (DEVICE_BASE), validation range: (0, nothing)

T1::Float64: Time constant in s, validation range: (0, nothing)

T2::Float64: Time constant in s, validation range: (0, nothing)

T3::Float64: Time constant in s, validation range: (eps(), nothing)

T4::Float64: Time constant in s, validation range: (0, nothing)

T5::Float64: Time constant in s, validation range: (0, nothing)

Rmin::Float64: Reactor minimum Mvar, validation range: (0, nothing)

Vmax::Float64: Maximum voltage in pu, validation range: (0, nothing)

Vmin::Float64: Minimum voltage in pu, validation range: (0, nothing)

CBase::Float64: Capacitor (MVAR), validation range: (0, nothing)

base_power::Float64: Base power of the unit (MVA) for per unitization, validation range: (1.0e-6, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

R_th::Float64: Source Thevenin resistance

X_th::Float64: Source Thevenin reactance

states::Vector{Symbol}: (Do not modify.) The states are:

thy: thyristor,
vr1: regulator output 1,
vr2: regulator output 2

n_states::Int: (Do not modify.) CSVGN1 has 3 states

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::CSVGN1) -> String

Get CSVGN1 name.

PowerSystems.get_CBase — Method

get_CBase(value::CSVGN1) -> Float64

Get CSVGN1 CBase.

PowerSystems.get_K — Method

get_K(value::CSVGN1) -> Float64

Get CSVGN1 K.

PowerSystems.get_R_th — Method

get_R_th(value::CSVGN1) -> Float64

Get CSVGN1 R_th.

PowerSystems.get_Rmin — Method

get_Rmin(value::CSVGN1) -> Float64

Get CSVGN1 Rmin.

PowerSystems.get_T1 — Method

get_T1(value::CSVGN1) -> Float64

Get CSVGN1 T1.

PowerSystems.get_T2 — Method

get_T2(value::CSVGN1) -> Float64

Get CSVGN1 T2.

PowerSystems.get_T3 — Method

get_T3(value::CSVGN1) -> Float64

Get CSVGN1 T3.

PowerSystems.get_T4 — Method

get_T4(value::CSVGN1) -> Float64

Get CSVGN1 T4.

PowerSystems.get_T5 — Method

get_T5(value::CSVGN1) -> Float64

Get CSVGN1 T5.

PowerSystems.get_Vmax — Method

get_Vmax(value::CSVGN1) -> Float64

Get CSVGN1 Vmax.

PowerSystems.get_Vmin — Method

get_Vmin(value::CSVGN1) -> Float64

Get CSVGN1 Vmin.

PowerSystems.get_X_th — Method

get_X_th(value::CSVGN1) -> Float64

Get CSVGN1 X_th.

PowerSystems.get_base_power — Method

get_base_power(value::CSVGN1) -> Float64

Get CSVGN1 base_power.

PowerSystems.get_ext — Method

get_ext(value::CSVGN1) -> Dict{String, Any}

Get CSVGN1 ext.

PowerSystems.get_n_states — Method

get_n_states(value::CSVGN1) -> Int64

Get CSVGN1 n_states.

PowerSystems.get_states — Method

get_states(value::CSVGN1) -> Vector{Symbol}

Get CSVGN1 states.

PowerSystems.set_CBase! — Method

set_CBase!(value::CSVGN1, val) -> Any

Set CSVGN1 CBase.

PowerSystems.set_K! — Method

set_K!(value::CSVGN1, val) -> Any

Set CSVGN1 K.

PowerSystems.set_R_th! — Method

set_R_th!(value::CSVGN1, val) -> Any

Set CSVGN1 R_th.

PowerSystems.set_Rmin! — Method

set_Rmin!(value::CSVGN1, val) -> Any

Set CSVGN1 Rmin.

PowerSystems.set_T1! — Method

set_T1!(value::CSVGN1, val) -> Any

Set CSVGN1 T1.

PowerSystems.set_T2! — Method

set_T2!(value::CSVGN1, val) -> Any

Set CSVGN1 T2.

PowerSystems.set_T3! — Method

set_T3!(value::CSVGN1, val) -> Any

Set CSVGN1 T3.

PowerSystems.set_T4! — Method

set_T4!(value::CSVGN1, val) -> Any

Set CSVGN1 T4.

PowerSystems.set_T5! — Method

set_T5!(value::CSVGN1, val) -> Any

Set CSVGN1 T5.

PowerSystems.set_Vmax! — Method

set_Vmax!(value::CSVGN1, val) -> Any

Set CSVGN1 Vmax.

PowerSystems.set_Vmin! — Method

set_Vmin!(value::CSVGN1, val) -> Any

Set CSVGN1 Vmin.

PowerSystems.set_X_th! — Method

set_X_th!(value::CSVGN1, val) -> Any

Set CSVGN1 X_th.

PowerSystems.set_base_power! — Method

set_base_power!(value::CSVGN1, val) -> Any

Set CSVGN1 base_power.

PowerSystems.set_ext! — Method

set_ext!(value::CSVGN1, val) -> Any

Set CSVGN1 ext.