Filter

LCFilter

PowerSystems.LCFilter — Type

mutable struct LCFilter <: Filter
    lf::Float64
    rf::Float64
    cf::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a LCL filter outside the converter

Arguments

lf::Float64: filter inductance, validation range: (0, nothing)

rf::Float64: filter resistance, validation range: (0, nothing)

cf::Float64: filter capacitance, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

states::Vector{Symbol}: (Do not modify.) The states of the LCFilter model are:

ir_filter: Real current out of the filter,
ii_filter: Imaginary current out of the filter

n_states::Int: (Do not modify.) LCFilter has two states

PowerSystems.get_cf — Method

get_cf(value::LCFilter) -> Float64

Get LCFilter cf.

PowerSystems.get_ext — Method

get_ext(value::LCFilter) -> Dict{String, Any}

Get LCFilter ext.

PowerSystems.get_lf — Method

get_lf(value::LCFilter) -> Float64

Get LCFilter lf.

PowerSystems.get_n_states — Method

get_n_states(value::LCFilter) -> Int64

Get LCFilter n_states.

PowerSystems.get_rf — Method

get_rf(value::LCFilter) -> Float64

Get LCFilter rf.

PowerSystems.get_states — Method

get_states(value::LCFilter) -> Vector{Symbol}

Get LCFilter states.

PowerSystems.set_cf! — Method

set_cf!(value::LCFilter, val) -> Any

Set LCFilter cf.

PowerSystems.set_ext! — Method

set_ext!(value::LCFilter, val) -> Any

Set LCFilter ext.

PowerSystems.set_lf! — Method

set_lf!(value::LCFilter, val) -> Any

Set LCFilter lf.

PowerSystems.set_rf! — Method

set_rf!(value::LCFilter, val) -> Any

Set LCFilter rf.

LCLFilter

PowerSystems.LCLFilter — Type

mutable struct LCLFilter <: Filter
    lf::Float64
    rf::Float64
    cf::Float64
    lg::Float64
    rg::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a LCL filter outside the converter, the states are in the grid's reference frame

Arguments

lf::Float64: Series inductance in p.u. of converter filter, validation range: (0, nothing)

rf::Float64: Series resistance in p.u. of converter filter, validation range: (0, nothing)

cf::Float64: Shunt capacitance in p.u. of converter filter, validation range: (0, nothing)

lg::Float64: Series inductance in p.u. of converter filter to the grid, validation range: (0, nothing)

rg::Float64: Series resistance in p.u. of converter filter to the grid, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

states::Vector{Symbol}: (Do not modify.) The states of the LCLFilter model are:

ir_cnv: Real current out of the converter,
ii_cnv: Imaginary current out of the converter,
vr_filter: Real voltage at the filter's capacitor,
vi_filter: Imaginary voltage at the filter's capacitor,
ir_filter: Real current out of the filter,
ii_filter: Imaginary current out of the filter

n_states::Int: (Do not modify.) LCLFilter has 6 states

PowerSystems.get_cf — Method

get_cf(value::LCLFilter) -> Float64

Get LCLFilter cf.

PowerSystems.get_ext — Method

get_ext(value::LCLFilter) -> Dict{String, Any}

Get LCLFilter ext.

PowerSystems.get_lf — Method

get_lf(value::LCLFilter) -> Float64

Get LCLFilter lf.

PowerSystems.get_lg — Method

get_lg(value::LCLFilter) -> Float64

Get LCLFilter lg.

PowerSystems.get_n_states — Method

get_n_states(value::LCLFilter) -> Int64

Get LCLFilter n_states.

PowerSystems.get_rf — Method

get_rf(value::LCLFilter) -> Float64

Get LCLFilter rf.

PowerSystems.get_rg — Method

get_rg(value::LCLFilter) -> Float64

Get LCLFilter rg.

PowerSystems.get_states — Method

get_states(value::LCLFilter) -> Vector{Symbol}

Get LCLFilter states.

PowerSystems.set_cf! — Method

set_cf!(value::LCLFilter, val) -> Any

Set LCLFilter cf.

PowerSystems.set_ext! — Method

set_ext!(value::LCLFilter, val) -> Any

Set LCLFilter ext.

PowerSystems.set_lf! — Method

set_lf!(value::LCLFilter, val) -> Any

Set LCLFilter lf.

PowerSystems.set_lg! — Method

set_lg!(value::LCLFilter, val) -> Any

Set LCLFilter lg.

PowerSystems.set_rf! — Method

set_rf!(value::LCLFilter, val) -> Any

Set LCLFilter rf.

PowerSystems.set_rg! — Method

set_rg!(value::LCLFilter, val) -> Any

Set LCLFilter rg.

RLFilter

PowerSystems.RLFilter — Type

mutable struct RLFilter <: Filter
    rf::Float64
    lf::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of RL series filter in algebraic representation

Arguments

rf::Float64: Series resistance in p.u. of converter filter to the grid, validation range: (0, nothing)

lf::Float64: Series inductance in p.u. of converter filter to the grid, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

states::Vector{Symbol}: (Do not modify.) RLFilter has zero states

n_states::Int: (Do not modify.) RLFilter has zero states

PowerSystems.get_ext — Method

get_ext(value::RLFilter) -> Dict{String, Any}

Get RLFilter ext.

PowerSystems.get_lf — Method

get_lf(value::RLFilter) -> Float64

Get RLFilter lf.

PowerSystems.get_n_states — Method

get_n_states(value::RLFilter) -> Int64

Get RLFilter n_states.

PowerSystems.get_rf — Method

get_rf(value::RLFilter) -> Float64

Get RLFilter rf.

PowerSystems.get_states — Method

get_states(value::RLFilter) -> Vector{Symbol}

Get RLFilter states.

PowerSystems.set_ext! — Method

set_ext!(value::RLFilter, val) -> Any

Set RLFilter ext.

PowerSystems.set_lf! — Method

set_lf!(value::RLFilter, val) -> Any

Set RLFilter lf.

PowerSystems.set_rf! — Method

set_rf!(value::RLFilter, val) -> Any

Set RLFilter rf.