FrequencyEstimator

FixedFrequency

PowerSystems.FixedFrequency — Type

mutable struct FixedFrequency <: FrequencyEstimator
    frequency::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a Fixed Frequency Estimator (i.e. no PLL)

Arguments

frequency::Float64: (default: 1.0) Reference Frequency (pu)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

states::Vector{Symbol}: (Do not modify.) FixedFrequency has no states

n_states::Int: (Do not modify.) FixedFrequency has no states

PowerSystems.get_ext — Method

get_ext(value::FixedFrequency) -> Dict{String, Any}

Get FixedFrequency ext.

PowerSystems.get_frequency — Method

get_frequency(value::FixedFrequency) -> Float64

Get FixedFrequency frequency.

PowerSystems.get_n_states — Method

get_n_states(value::FixedFrequency) -> Int64

Get FixedFrequency n_states.

PowerSystems.get_states — Method

get_states(value::FixedFrequency) -> Vector{Symbol}

Get FixedFrequency states.

PowerSystems.set_ext! — Method

set_ext!(value::FixedFrequency, val) -> Any

Set FixedFrequency ext.

PowerSystems.set_frequency! — Method

set_frequency!(value::FixedFrequency, val) -> Any

Set FixedFrequency frequency.

KauraPLL

PowerSystems.KauraPLL — Type

mutable struct KauraPLL <: FrequencyEstimator
    ω_lp::Float64
    kp_pll::Float64
    ki_pll::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a Phase-Locked Loop (PLL) based on "Operation of a phase locked loop system under distorted utility conditions" by Vikram Kaura, and Vladimir Blasko

Arguments

ω_lp::Float64: PLL low-pass filter frequency (rad/sec), validation range: (0, nothing)

kp_pll::Float64: PLL proportional gain, validation range: (0, nothing)

ki_pll::Float64: PLL integral gain, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

states::Vector{Symbol}: (Do not modify.) The states of the KauraPLL model are:

vd_pll: d-axis of the measured voltage in the PLL synchronous reference frame (SRF),
vq_pll: q-axis of the measured voltage in the PLL SRF,
ε_pll: Integrator state of the PI controller,
θ_pll: Phase angle displacement in the PLL SRF

n_states::Int: (Do not modify.) KauraPLL has 4 states

PowerSystems.get_ext — Method

get_ext(value::KauraPLL) -> Dict{String, Any}

Get KauraPLL ext.

PowerSystems.get_ki_pll — Method

get_ki_pll(value::KauraPLL) -> Float64

Get KauraPLL ki_pll.

PowerSystems.get_kp_pll — Method

get_kp_pll(value::KauraPLL) -> Float64

Get KauraPLL kp_pll.

PowerSystems.get_n_states — Method

get_n_states(value::KauraPLL) -> Int64

Get KauraPLL n_states.

PowerSystems.get_states — Method

get_states(value::KauraPLL) -> Vector{Symbol}

Get KauraPLL states.

PowerSystems.get_ω_lp — Method

get_ω_lp(value::KauraPLL) -> Float64

Get KauraPLL ω_lp.

PowerSystems.set_ext! — Method

set_ext!(value::KauraPLL, val) -> Any

Set KauraPLL ext.

PowerSystems.set_ki_pll! — Method

set_ki_pll!(value::KauraPLL, val) -> Any

Set KauraPLL ki_pll.

PowerSystems.set_kp_pll! — Method

set_kp_pll!(value::KauraPLL, val) -> Any

Set KauraPLL kp_pll.

PowerSystems.set_ω_lp! — Method

set_ω_lp!(value::KauraPLL, val) -> Any

Set KauraPLL ω_lp.

ReducedOrderPLL

PowerSystems.ReducedOrderPLL — Type

mutable struct ReducedOrderPLL <: FrequencyEstimator
    ω_lp::Float64
    kp_pll::Float64
    ki_pll::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a Phase-Locked Loop (PLL) based on "Reduced-order Structure-preserving Model for Parallel-connected Three-phase Grid-tied Inverters."

Arguments

ω_lp::Float64: PLL low-pass filter frequency (rad/sec), validation range: (0, nothing)

kp_pll::Float64: PLL proportional gain, validation range: (0, nothing)

ki_pll::Float64: PLL integral gain, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

states::Vector{Symbol}: (Do not modify.) The states of the ReducedOrderPLL model are:

vq_pll: q-axis of the measured voltage in the PLL synchronous reference frame (SRF),
ε_pll: Integrator state of the PI controller,
θ_pll: Phase angle displacement in the PLL SRF

n_states::Int: (Do not modify.) ReducedOrderPLL has 3 states

PowerSystems.get_ext — Method

get_ext(value::ReducedOrderPLL) -> Dict{String, Any}

Get ReducedOrderPLL ext.

PowerSystems.get_ki_pll — Method

get_ki_pll(value::ReducedOrderPLL) -> Float64

Get ReducedOrderPLL ki_pll.

PowerSystems.get_kp_pll — Method

get_kp_pll(value::ReducedOrderPLL) -> Float64

Get ReducedOrderPLL kp_pll.

PowerSystems.get_n_states — Method

get_n_states(value::ReducedOrderPLL) -> Int64

Get ReducedOrderPLL n_states.

PowerSystems.get_states — Method

get_states(value::ReducedOrderPLL) -> Vector{Symbol}

Get ReducedOrderPLL states.

PowerSystems.get_ω_lp — Method

get_ω_lp(value::ReducedOrderPLL) -> Float64

Get ReducedOrderPLL ω_lp.

PowerSystems.set_ext! — Method

set_ext!(value::ReducedOrderPLL, val) -> Any

Set ReducedOrderPLL ext.

PowerSystems.set_ki_pll! — Method

set_ki_pll!(value::ReducedOrderPLL, val) -> Any

Set ReducedOrderPLL ki_pll.

PowerSystems.set_kp_pll! — Method

set_kp_pll!(value::ReducedOrderPLL, val) -> Any

Set ReducedOrderPLL kp_pll.

PowerSystems.set_ω_lp! — Method

set_ω_lp!(value::ReducedOrderPLL, val) -> Any

Set ReducedOrderPLL ω_lp.