HydroDispatch

PowerSystems.HydroDispatch — Type

mutable struct HydroDispatch <: HydroGen
    name::String
    available::Bool
    bus::ACBus
    active_power::Float64
    reactive_power::Float64
    rating::Float64
    prime_mover_type::PrimeMovers
    active_power_limits::MinMax
    reactive_power_limits::Union{Nothing, MinMax}
    ramp_limits::Union{Nothing, UpDown}
    time_limits::Union{Nothing, UpDown}
    base_power::Float64
    operation_cost::Union{HydroGenerationCost, MarketBidCost}
    services::Vector{Service}
    dynamic_injector::Union{Nothing, DynamicInjection}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A hydropower generator without a reservoir, suitable for modeling run-of-river hydropower.

For hydro generators with an upper reservoir, see HydroEnergyReservoir

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

bus::ACBus: Bus that this component is connected to

active_power::Float64: Initial active power set point of the unit in MW. For power flow, this is the steady state operating point of the system. For production cost modeling, this may or may not be used as the initial starting point for the solver, depending on the solver used

reactive_power::Float64: Initial reactive power set point of the unit (MVAR), validation range: reactive_power_limits

rating::Float64: Maximum output power rating of the unit (MVA), validation range: (0, nothing)

prime_mover_type::PrimeMovers: Prime mover technology according to EIA 923. Options are listed here

active_power_limits::MinMax: Minimum and maximum stable active power levels (MW), validation range: (0, nothing)

reactive_power_limits::Union{Nothing, MinMax}: Minimum and maximum reactive power limits. Set to Nothing if not applicable

ramp_limits::Union{Nothing, UpDown}: ramp up and ramp down limits in MW/min, validation range: (0, nothing)

time_limits::Union{Nothing, UpDown}: Minimum up and Minimum down time limits in hours, validation range: (0, nothing)

base_power::Float64: Base power of the unit (MVA) for per unitization, validation range: (0, nothing)

operation_cost::Union{HydroGenerationCost, MarketBidCost}: (default: HydroGenerationCost(nothing)) OperationalCost of generation

services::Vector{Service}: (default: Device[]) Services that this device contributes to

dynamic_injector::Union{Nothing, DynamicInjection}: (default: nothing) corresponding dynamic injection device

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::HydroDispatch) -> String

Get HydroDispatch name.

PowerSystems.get_active_power — Method

get_active_power(value::HydroDispatch) -> Any

Get HydroDispatch active_power.

PowerSystems.get_active_power_limits — Method

get_active_power_limits(
    value::HydroDispatch
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get HydroDispatch active_power_limits.

PowerSystems.get_available — Method

get_available(value::HydroDispatch) -> Bool

Get HydroDispatch available.

PowerSystems.get_base_power — Method

get_base_power(value::HydroDispatch) -> Float64

Get HydroDispatch base_power.

PowerSystems.get_bus — Method

get_bus(value::HydroDispatch) -> ACBus

Get HydroDispatch bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(
    value::HydroDispatch
) -> Union{Nothing, DynamicInjection}

Get HydroDispatch dynamic_injector.

PowerSystems.get_ext — Method

get_ext(value::HydroDispatch) -> Dict{String, Any}

Get HydroDispatch ext.

PowerSystems.get_operation_cost — Method

get_operation_cost(
    value::HydroDispatch
) -> Union{HydroGenerationCost, MarketBidCost}

Get HydroDispatch operation_cost.

PowerSystems.get_prime_mover_type — Method

get_prime_mover_type(value::HydroDispatch) -> PrimeMovers

Get HydroDispatch prime_mover_type.

PowerSystems.get_ramp_limits — Method

get_ramp_limits(
    value::HydroDispatch
) -> Union{Nothing, NamedTuple{(:up, :down), <:Tuple{Any, Any}}}

Get HydroDispatch ramp_limits.

PowerSystems.get_rating — Method

get_rating(value::HydroDispatch) -> Any

Get HydroDispatch rating.

PowerSystems.get_reactive_power — Method

get_reactive_power(value::HydroDispatch) -> Any

Get HydroDispatch reactive_power.

PowerSystems.get_reactive_power_limits — Method

get_reactive_power_limits(
    value::HydroDispatch
) -> Union{Nothing, NamedTuple{(:min, :max), <:Tuple{Any, Any}}}

Get HydroDispatch reactive_power_limits.

PowerSystems.get_services — Method

get_services(value::HydroDispatch) -> Vector{Service}

Get HydroDispatch services.

PowerSystems.get_time_limits — Method

get_time_limits(
    value::HydroDispatch
) -> Union{Nothing, @NamedTuple{up::Float64, down::Float64}}

Get HydroDispatch time_limits.

PowerSystems.set_active_power! — Method

set_active_power!(value::HydroDispatch, val) -> Any

Set HydroDispatch active_power.

PowerSystems.set_active_power_limits! — Method

set_active_power_limits!(value::HydroDispatch, val) -> Any

Set HydroDispatch active_power_limits.

PowerSystems.set_available! — Method

set_available!(value::HydroDispatch, val) -> Any

Set HydroDispatch available.

PowerSystems.set_base_power! — Method

set_base_power!(value::HydroDispatch, val) -> Any

Set HydroDispatch base_power.

PowerSystems.set_bus! — Method

set_bus!(value::HydroDispatch, val) -> Any

Set HydroDispatch bus.

PowerSystems.set_ext! — Method

set_ext!(value::HydroDispatch, val) -> Any

Set HydroDispatch ext.

PowerSystems.set_operation_cost! — Method

set_operation_cost!(value::HydroDispatch, val) -> Any

Set HydroDispatch operation_cost.

PowerSystems.set_prime_mover_type! — Method

set_prime_mover_type!(value::HydroDispatch, val) -> Any

Set HydroDispatch prime_mover_type.

PowerSystems.set_ramp_limits! — Method

set_ramp_limits!(value::HydroDispatch, val) -> Any

Set HydroDispatch ramp_limits.

PowerSystems.set_rating! — Method

set_rating!(value::HydroDispatch, val) -> Any

Set HydroDispatch rating.

PowerSystems.set_reactive_power! — Method

set_reactive_power!(value::HydroDispatch, val) -> Any

Set HydroDispatch reactive_power.

PowerSystems.set_reactive_power_limits! — Method

set_reactive_power_limits!(value::HydroDispatch, val) -> Any

Set HydroDispatch reactive_power_limits.

PowerSystems.set_services! — Method

set_services!(value::HydroDispatch, val) -> Any

Set HydroDispatch services.

PowerSystems.set_time_limits! — Method

set_time_limits!(value::HydroDispatch, val) -> Any

Set HydroDispatch time_limits.