Line

PowerSystems.Line — Type

mutable struct Line <: ACBranch
    name::String
    available::Bool
    active_power_flow::Float64
    reactive_power_flow::Float64
    arc::Arc
    r::Float64
    x::Float64
    b::FromTo
    rating::Float64
    angle_limits::MinMax
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

An AC transmission line

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow::Float64: Initial condition of active power flow on the line (MW)

reactive_power_flow::Float64: Initial condition of reactive power flow on the line (MVAR)

arc::Arc: An Arc defining this line from a bus to another bus

r::Float64: Resistance in pu (SYSTEM_BASE), validation range: (0, 4)

x::Float64: Reactance in pu (SYSTEM_BASE), validation range: (0, 4)

b::FromTo: Shunt susceptance in pu (SYSTEM_BASE), specified both on the from and to ends of the line. These are commonly modeled with the same value, validation range: (0, 100)

rating::Float64: Thermal rating (MVA). Flow on the line must be between -rating and rating. When defining a line before it is attached to a System, rating must be in per-unit divided by the base power of the System it will be attached to

angle_limits::MinMax: Minimum and maximum angle limits (radians), validation range: (-1.571, 1.571)

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::Line) -> String

Get Line name.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::Line) -> Any

Get Line active_power_flow.

PowerSystems.get_angle_limits — Method

get_angle_limits(
    value::Line
) -> @NamedTuple{min::Float64, max::Float64}

Get Line angle_limits.

PowerSystems.get_arc — Method

get_arc(value::Line) -> Arc

Get Line arc.

PowerSystems.get_available — Method

get_available(value::Line) -> Bool

Get Line available.

PowerSystems.get_b — Method

get_b(
    value::Line
) -> @NamedTuple{from::Float64, to::Float64}

Get Line b.

PowerSystems.get_ext — Method

get_ext(value::Line) -> Dict{String, Any}

Get Line ext.

PowerSystems.get_r — Method

get_r(value::Line) -> Float64

Get Line r.

PowerSystems.get_rating — Method

get_rating(value::Line) -> Any

Get Line rating.

PowerSystems.get_reactive_power_flow — Method

get_reactive_power_flow(value::Line) -> Any

Get Line reactive_power_flow.

PowerSystems.get_services — Method

get_services(value::Line) -> Vector{Service}

Get Line services.

PowerSystems.get_x — Method

get_x(value::Line) -> Float64

Get Line x.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(value::Line, val) -> Any

Set Line active_power_flow.

PowerSystems.set_angle_limits! — Method

set_angle_limits!(value::Line, val) -> Any

Set Line angle_limits.

PowerSystems.set_arc! — Method

set_arc!(value::Line, val) -> Any

Set Line arc.

PowerSystems.set_available! — Method

set_available!(value::Line, val) -> Any

Set Line available.

PowerSystems.set_b! — Method

set_b!(value::Line, val) -> Any

Set Line b.

PowerSystems.set_ext! — Method

set_ext!(value::Line, val) -> Any

Set Line ext.

PowerSystems.set_r! — Method

set_r!(value::Line, val) -> Any

Set Line r.

PowerSystems.set_rating! — Method

set_rating!(value::Line, val) -> Any

Set Line rating.

PowerSystems.set_reactive_power_flow! — Method

set_reactive_power_flow!(value::Line, val) -> Any

Set Line reactive_power_flow.

PowerSystems.set_services! — Method

set_services!(value::Line, val) -> Any

Set Line services.

PowerSystems.set_x! — Method

set_x!(value::Line, val) -> Any

Set Line x.