MarketBidCost

PowerSystems.MarketBidCost — Type

mutable struct MarketBidCost <: OperationalCost
    no_load::Float64
    start_up::NamedTuple{(:hot, :warm, :cold), NTuple{3, Float64}}
    shut_down::Float64
    variable::Union{Nothing, IS.TimeSeriesKey}
    ancillary_services::Vector{Service}
end

Data Structure Operational Cost to reflect market bids of energy and ancilliary services. Compatible with most US Market bidding mechanisms

Arguments

no_load::Float64: no load cost

start_up::NamedTuple{(:hot, :warm, :cold), NTuple{3, Float64}}: start-up cost at different stages of the thermal cycle. Warm is also refered as intermediate in some markets

shut_down::Float64: shut-down cost, validation range: (0, nothing), action if invalid: warn

variable::Union{Nothing, IS.TimeSeriesKey}: Variable Cost TimeSeriesKey

ancillary_services::Vector{Service}: Bids for the ancillary services

PowerSystems.get_ancillary_services — Method

get_ancillary_services(
    value::MarketBidCost
) -> Vector{Service}

Get MarketBidCost ancillary_services.

PowerSystems.get_no_load — Method

get_no_load(value::MarketBidCost) -> Float64

Get MarketBidCost no_load.

PowerSystems.get_shut_down — Method

get_shut_down(value::MarketBidCost) -> Float64

Get MarketBidCost shut_down.

PowerSystems.get_start_up — Method

get_start_up(
    value::MarketBidCost
) -> NamedTuple{(:hot, :warm, :cold), Tuple{Float64, Float64, Float64}}

Get MarketBidCost start_up.

PowerSystems.get_variable — Method

get_variable(
    value::MarketBidCost
) -> Union{Nothing, InfrastructureSystems.TimeSeriesKey}

Get MarketBidCost variable.

PowerSystems.set_ancillary_services! — Method

set_ancillary_services!(value::MarketBidCost, val) -> Any

Set MarketBidCost ancillary_services.

PowerSystems.set_no_load! — Method

set_no_load!(value::MarketBidCost, val) -> Any

Set MarketBidCost no_load.

PowerSystems.set_shut_down! — Method

set_shut_down!(value::MarketBidCost, val) -> Any

Set MarketBidCost shut_down.

PowerSystems.set_start_up! — Method

set_start_up!(value::MarketBidCost, val) -> Any

Set MarketBidCost start_up.

PowerSystems.set_variable! — Method

set_variable!(value::MarketBidCost, val) -> Any

Set MarketBidCost variable.