MultiStartCost

PowerSystems.MultiStartCost — Type

mutable struct MultiStartCost <: OperationalCost
    variable::VariableCost
    no_load::Float64
    fixed::Float64
    start_up::NamedTuple{(:hot, :warm, :cold), NTuple{3, Float64}}
    shut_down::Float64
end

Data Structure Operational Cost Data which includes fixed, variable cost, multiple start up cost and stop costs.

Arguments

variable::VariableCost: variable cost

no_load::Float64: no load cost

fixed::Float64: fixed cost

start_up::NamedTuple{(:hot, :warm, :cold), NTuple{3, Float64}}: start-up cost

shut_down::Float64: shut-down cost, validation range: (0, nothing), action if invalid: warn

PowerSystems.get_fixed — Method

get_fixed(value::MultiStartCost) -> Float64

Get MultiStartCost fixed.

PowerSystems.get_no_load — Method

get_no_load(value::MultiStartCost) -> Float64

Get MultiStartCost no_load.

PowerSystems.get_shut_down — Method

get_shut_down(value::MultiStartCost) -> Float64

Get MultiStartCost shut_down.

PowerSystems.get_start_up — Method

get_start_up(
    value::MultiStartCost
) -> NamedTuple{(:hot, :warm, :cold), Tuple{Float64, Float64, Float64}}

Get MultiStartCost start_up.

PowerSystems.get_variable — Method

get_variable(value::MultiStartCost) -> VariableCost

Get MultiStartCost variable.

PowerSystems.set_fixed! — Method

set_fixed!(value::MultiStartCost, val) -> Any

Set MultiStartCost fixed.

PowerSystems.set_no_load! — Method

set_no_load!(value::MultiStartCost, val) -> Any

Set MultiStartCost no_load.

PowerSystems.set_shut_down! — Method

set_shut_down!(value::MultiStartCost, val) -> Any

Set MultiStartCost shut_down.

PowerSystems.set_start_up! — Method

set_start_up!(value::MultiStartCost, val) -> Any

Set MultiStartCost start_up.

PowerSystems.set_variable! — Method

set_variable!(value::MultiStartCost, val) -> Any

Set MultiStartCost variable.