PSS

IEEEST

PowerSystems.IEEEST — Type

mutable struct IEEEST <: PSS
    input_code::Int
    remote_bus_control::Int
    A1::Float64
    A2::Float64
    A3::Float64
    A4::Float64
    A5::Float64
    A6::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    T4::Float64
    T5::Float64
    T6::Float64
    Ks::Float64
    Ls_lim::Tuple{Float64, Float64}
    Vcu::Float64
    Vcl::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

IEEE Stabilizing Model PSS.

Arguments

input_code::Int: Code input for stabilizer, validation range: (1, 6), action if invalid: error

remote_bus_control::Int: Remote Bus number for control.

A1::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

A2::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

A3::Float64: Filter coefficient, validation range: (0, nothing)

A4::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

A5::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

A6::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

T1::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T2::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T3::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T4::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T5::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T6::Float64: Time constant, validation range: (eps(), 2.0), action if invalid: error

Ks::Float64: Proportional gain, validation range: (0, nothing), action if invalid: warn

Ls_lim::Tuple{Float64, Float64}: PSS output limits for regulator output (Ls_min, Ls_max)

Vcu::Float64: Cutoff limiter upper bound, validation range: (0, 1.25), action if invalid: warn

Vcl::Float64: Cutoff limiter lower bound, validation range: (0, 1.0), action if invalid: warn

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

x_p1: 1st filter integration,
x_p2: 2nd filter integration, 
x_p3: 3rd filter integration, 
x_p4: 4rd filter integration, 
x_p5: T1/T2 lead-lag integrator, 
x_p6: T3/T4 lead-lag integrator, 
:x_p7 last integer,

n_states::Int: IEEEST has 7 states

states_types::Vector{StateTypes}: IEEEST has 7 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_A1 — Method

get_A1(value::IEEEST) -> Float64

Get IEEEST A1.

PowerSystems.get_A2 — Method

get_A2(value::IEEEST) -> Float64

Get IEEEST A2.

PowerSystems.get_A3 — Method

get_A3(value::IEEEST) -> Float64

Get IEEEST A3.

PowerSystems.get_A4 — Method

get_A4(value::IEEEST) -> Float64

Get IEEEST A4.

PowerSystems.get_A5 — Method

get_A5(value::IEEEST) -> Float64

Get IEEEST A5.

PowerSystems.get_A6 — Method

get_A6(value::IEEEST) -> Float64

Get IEEEST A6.

PowerSystems.get_Ks — Method

get_Ks(value::IEEEST) -> Float64

Get IEEEST Ks.

PowerSystems.get_Ls_lim — Method

get_Ls_lim(value::IEEEST) -> Tuple{Float64, Float64}

Get IEEEST Ls_lim.

PowerSystems.get_T1 — Method

get_T1(value::IEEEST) -> Float64

Get IEEEST T1.

PowerSystems.get_T2 — Method

get_T2(value::IEEEST) -> Float64

Get IEEEST T2.

PowerSystems.get_T3 — Method

get_T3(value::IEEEST) -> Float64

Get IEEEST T3.

PowerSystems.get_T4 — Method

get_T4(value::IEEEST) -> Float64

Get IEEEST T4.

PowerSystems.get_T5 — Method

get_T5(value::IEEEST) -> Float64

Get IEEEST T5.

PowerSystems.get_T6 — Method

get_T6(value::IEEEST) -> Float64

Get IEEEST T6.

PowerSystems.get_Vcl — Method

get_Vcl(value::IEEEST) -> Float64

Get IEEEST Vcl.

PowerSystems.get_Vcu — Method

get_Vcu(value::IEEEST) -> Float64

Get IEEEST Vcu.

PowerSystems.get_ext — Method

get_ext(value::IEEEST) -> Dict{String, Any}

Get IEEEST ext.

PowerSystems.get_input_code — Method

get_input_code(value::IEEEST) -> Int64

Get IEEEST input_code.

PowerSystems.get_n_states — Method

get_n_states(value::IEEEST) -> Int64

Get IEEEST n_states.

PowerSystems.get_remote_bus_control — Method

get_remote_bus_control(value::IEEEST) -> Int64

Get IEEEST remote_bus_control.

PowerSystems.get_states — Method

get_states(value::IEEEST) -> Vector{Symbol}

Get IEEEST states.

PowerSystems.get_states_types — Method

get_states_types(value::IEEEST) -> Vector{StateTypes}

Get IEEEST states_types.

PowerSystems.set_A1! — Method

set_A1!(value::IEEEST, val) -> Any

Set IEEEST A1.

PowerSystems.set_A2! — Method

set_A2!(value::IEEEST, val) -> Any

Set IEEEST A2.

PowerSystems.set_A3! — Method

set_A3!(value::IEEEST, val) -> Any

Set IEEEST A3.

PowerSystems.set_A4! — Method

set_A4!(value::IEEEST, val) -> Any

Set IEEEST A4.

PowerSystems.set_A5! — Method

set_A5!(value::IEEEST, val) -> Any

Set IEEEST A5.

PowerSystems.set_A6! — Method

set_A6!(value::IEEEST, val) -> Any

Set IEEEST A6.

PowerSystems.set_Ks! — Method

set_Ks!(value::IEEEST, val) -> Any

Set IEEEST Ks.

PowerSystems.set_Ls_lim! — Method

set_Ls_lim!(value::IEEEST, val) -> Any

Set IEEEST Ls_lim.

PowerSystems.set_T1! — Method

set_T1!(value::IEEEST, val) -> Any

Set IEEEST T1.

PowerSystems.set_T2! — Method

set_T2!(value::IEEEST, val) -> Any

Set IEEEST T2.

PowerSystems.set_T3! — Method

set_T3!(value::IEEEST, val) -> Any

Set IEEEST T3.

PowerSystems.set_T4! — Method

set_T4!(value::IEEEST, val) -> Any

Set IEEEST T4.

PowerSystems.set_T5! — Method

set_T5!(value::IEEEST, val) -> Any

Set IEEEST T5.

PowerSystems.set_T6! — Method

set_T6!(value::IEEEST, val) -> Any

Set IEEEST T6.

PowerSystems.set_Vcl! — Method

set_Vcl!(value::IEEEST, val) -> Any

Set IEEEST Vcl.

PowerSystems.set_Vcu! — Method

set_Vcu!(value::IEEEST, val) -> Any

Set IEEEST Vcu.

PowerSystems.set_ext! — Method

set_ext!(value::IEEEST, val) -> Any

Set IEEEST ext.

PowerSystems.set_input_code! — Method

set_input_code!(value::IEEEST, val) -> Any

Set IEEEST input_code.

PowerSystems.set_remote_bus_control! — Method

set_remote_bus_control!(value::IEEEST, val) -> Any

Set IEEEST remote_bus_control.

PowerSystems.set_states_types! — Method

set_states_types!(value::IEEEST, val) -> Any

Set IEEEST states_types.

PSS2A

PowerSystems.PSS2A — Type

mutable struct PSS2A <: PSS
    input_code_1::Int
    remote_bus_control_1::Int
    input_code_2::Int
    remote_bus_control_2::Int
    M_rtf::Int
    N_rtf::Int
    Tw1::Float64
    Tw2::Float64
    T6::Float64
    Tw3::Float64
    Tw4::Float64
    T7::Float64
    Ks2::Float64
    Ks3::Float64
    T8::Float64
    T9::Float64
    Ks1::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    T4::Float64
    Vst_lim::Tuple{Float64, Float64}
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

IEEE Dual-Input Stabilizer Model

Arguments

input_code_1::Int: First Input Code for stabilizer, validation range: (1, 6), action if invalid: error

remote_bus_control_1::Int: First Input Remote Bus number for control.

input_code_2::Int: Second Input Code for stabilizer, validation range: (1, 6), action if invalid: error

remote_bus_control_2::Int: Second Input Remote Bus number for control.

M_rtf::Int: M parameter for ramp tracking filter, validation range: (0, 8), action if invalid: error

N_rtf::Int: N parameter for ramp tracking filter, validation range: (0, 8), action if invalid: error

Tw1::Float64: Time constant for first washout filter for first input, validation range: (eps(), nothing), action if invalid: warn

Tw2::Float64: Time constant for second washout filter for first input, validation range: (0, nothing), action if invalid: warn

T6::Float64: Time constant for low-pass filter for first input, validation range: (0, nothing), action if invalid: warn

Tw3::Float64: Time constant for first washout filter for second input, validation range: (eps(), nothing), action if invalid: warn

Tw4::Float64: Time constant for second washout filter for second input, validation range: (0, nothing), action if invalid: warn

T7::Float64: Time constant for low-pass filter for second input, validation range: (0, nothing), action if invalid: warn

Ks2::Float64: Gain for low-pass filter for second input, validation range: (0, nothing), action if invalid: warn

Ks3::Float64: Gain for second input, validation range: (0, nothing), action if invalid: warn

T8::Float64: Time constant for ramp tracking filter, validation range: (0, nothing), action if invalid: warn

T9::Float64: Time constant for ramp tracking filter, validation range: (eps(), nothing), action if invalid: warn

Ks1::Float64: Gain before lead-lag blocks, validation range: (0, nothing), action if invalid: warn

T1::Float64: Time constant for first lead-lag block, validation range: (0, nothing), action if invalid: warn

T2::Float64: Time constant for first lead-lag block, validation range: (0, nothing), action if invalid: warn

T3::Float64: Time constant for second lead-lag block, validation range: (0, nothing), action if invalid: warn

T4::Float64: Time constant for second lead-lag block, validation range: (0, nothing), action if invalid: warn

Vst_lim::Tuple{Float64, Float64}: PSS output limits (Vst_min, Vst_max)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

x_p1: 1st washout 1st input, 
x_p2: 2nd washout 1st input, 
x_p3: transducer 1st input, 
x_p4: 1st washout 2nd input, 
x_p5: 2nd washout 2nd input, 
x_p6: transducer 2nd input, 
x_p7: ramp tracking filter state 1, 
x_p8: ramp tracking filter state 2, 
x_p9: ramp tracking filter state 3, 
x_p10: ramp tracking filter state 4, 
x_p11: ramp tracking filter state 5, 
x_p12: ramp tracking filter state 6, 
x_p13: ramp tracking filter state 7, 
x_p14: ramp tracking filter state 8, 
x_p15: 1st lead-lag, 
x_p16: 2nd lead-lag,

n_states::Int: IEEEST has 16 states

states_types::Vector{StateTypes}: IEEEST has 16 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_Ks1 — Method

get_Ks1(value::PSS2A) -> Float64

Get PSS2A Ks1.

PowerSystems.get_Ks2 — Method

get_Ks2(value::PSS2A) -> Float64

Get PSS2A Ks2.

PowerSystems.get_Ks3 — Method

get_Ks3(value::PSS2A) -> Float64

Get PSS2A Ks3.

PowerSystems.get_M_rtf — Method

get_M_rtf(value::PSS2A) -> Int64

Get PSS2A M_rtf.

PowerSystems.get_N_rtf — Method

get_N_rtf(value::PSS2A) -> Int64

Get PSS2A N_rtf.

PowerSystems.get_T1 — Method

get_T1(value::PSS2A) -> Float64

Get PSS2A T1.

PowerSystems.get_T2 — Method

get_T2(value::PSS2A) -> Float64

Get PSS2A T2.

PowerSystems.get_T3 — Method

get_T3(value::PSS2A) -> Float64

Get PSS2A T3.

PowerSystems.get_T4 — Method

get_T4(value::PSS2A) -> Float64

Get PSS2A T4.

PowerSystems.get_T6 — Method

get_T6(value::PSS2A) -> Float64

Get PSS2A T6.

PowerSystems.get_T7 — Method

get_T7(value::PSS2A) -> Float64

Get PSS2A T7.

PowerSystems.get_T8 — Method

get_T8(value::PSS2A) -> Float64

Get PSS2A T8.

PowerSystems.get_T9 — Method

get_T9(value::PSS2A) -> Float64

Get PSS2A T9.

PowerSystems.get_Tw1 — Method

get_Tw1(value::PSS2A) -> Float64

Get PSS2A Tw1.

PowerSystems.get_Tw2 — Method

get_Tw2(value::PSS2A) -> Float64

Get PSS2A Tw2.

PowerSystems.get_Tw3 — Method

get_Tw3(value::PSS2A) -> Float64

Get PSS2A Tw3.

PowerSystems.get_Tw4 — Method

get_Tw4(value::PSS2A) -> Float64

Get PSS2A Tw4.

PowerSystems.get_Vst_lim — Method

get_Vst_lim(value::PSS2A) -> Tuple{Float64, Float64}

Get PSS2A Vst_lim.

PowerSystems.get_ext — Method

get_ext(value::PSS2A) -> Dict{String, Any}

Get PSS2A ext.

PowerSystems.get_input_code_1 — Method

get_input_code_1(value::PSS2A) -> Int64

Get PSS2A input_code_1.

PowerSystems.get_input_code_2 — Method

get_input_code_2(value::PSS2A) -> Int64

Get PSS2A input_code_2.

PowerSystems.get_n_states — Method

get_n_states(value::PSS2A) -> Int64

Get PSS2A n_states.

PowerSystems.get_remote_bus_control_1 — Method

get_remote_bus_control_1(value::PSS2A) -> Int64

Get PSS2A remote_bus_control_1.

PowerSystems.get_remote_bus_control_2 — Method

get_remote_bus_control_2(value::PSS2A) -> Int64

Get PSS2A remote_bus_control_2.

PowerSystems.get_states — Method

get_states(value::PSS2A) -> Vector{Symbol}

Get PSS2A states.

PowerSystems.get_states_types — Method

get_states_types(value::PSS2A) -> Vector{StateTypes}

Get PSS2A states_types.

PowerSystems.set_Ks1! — Method

set_Ks1!(value::PSS2A, val) -> Any

Set PSS2A Ks1.

PowerSystems.set_Ks2! — Method

set_Ks2!(value::PSS2A, val) -> Any

Set PSS2A Ks2.

PowerSystems.set_Ks3! — Method

set_Ks3!(value::PSS2A, val) -> Any

Set PSS2A Ks3.

PowerSystems.set_M_rtf! — Method

set_M_rtf!(value::PSS2A, val) -> Any

Set PSS2A M_rtf.

PowerSystems.set_N_rtf! — Method

set_N_rtf!(value::PSS2A, val) -> Any

Set PSS2A N_rtf.

PowerSystems.set_T1! — Method

set_T1!(value::PSS2A, val) -> Any

Set PSS2A T1.

PowerSystems.set_T2! — Method

set_T2!(value::PSS2A, val) -> Any

Set PSS2A T2.

PowerSystems.set_T3! — Method

set_T3!(value::PSS2A, val) -> Any

Set PSS2A T3.

PowerSystems.set_T4! — Method

set_T4!(value::PSS2A, val) -> Any

Set PSS2A T4.

PowerSystems.set_T6! — Method

set_T6!(value::PSS2A, val) -> Any

Set PSS2A T6.

PowerSystems.set_T7! — Method

set_T7!(value::PSS2A, val) -> Any

Set PSS2A T7.

PowerSystems.set_T8! — Method

set_T8!(value::PSS2A, val) -> Any

Set PSS2A T8.

PowerSystems.set_T9! — Method

set_T9!(value::PSS2A, val) -> Any

Set PSS2A T9.

PowerSystems.set_Tw1! — Method

set_Tw1!(value::PSS2A, val) -> Any

Set PSS2A Tw1.

PowerSystems.set_Tw2! — Method

set_Tw2!(value::PSS2A, val) -> Any

Set PSS2A Tw2.

PowerSystems.set_Tw3! — Method

set_Tw3!(value::PSS2A, val) -> Any

Set PSS2A Tw3.

PowerSystems.set_Tw4! — Method

set_Tw4!(value::PSS2A, val) -> Any

Set PSS2A Tw4.

PowerSystems.set_Vst_lim! — Method

set_Vst_lim!(value::PSS2A, val) -> Any

Set PSS2A Vst_lim.

PowerSystems.set_ext! — Method

set_ext!(value::PSS2A, val) -> Any

Set PSS2A ext.

PowerSystems.set_input_code_1! — Method

set_input_code_1!(value::PSS2A, val) -> Any

Set PSS2A input_code_1.

PowerSystems.set_input_code_2! — Method

set_input_code_2!(value::PSS2A, val) -> Any

Set PSS2A input_code_2.

PowerSystems.set_remote_bus_control_1! — Method

set_remote_bus_control_1!(value::PSS2A, val) -> Any

Set PSS2A remote_bus_control_1.

PowerSystems.set_remote_bus_control_2! — Method

set_remote_bus_control_2!(value::PSS2A, val) -> Any

Set PSS2A remote_bus_control_2.

PowerSystems.set_states_types! — Method

set_states_types!(value::PSS2A, val) -> Any

Set PSS2A states_types.

PSS2B

PowerSystems.PSS2B — Type

mutable struct PSS2B <: PSS
    input_code_1::Int
    remote_bus_control_1::Int
    input_code_2::Int
    remote_bus_control_2::Int
    M_rtf::Int
    N_rtf::Int
    Tw1::Float64
    Tw2::Float64
    T6::Float64
    Tw3::Float64
    Tw4::Float64
    T7::Float64
    Ks2::Float64
    Ks3::Float64
    T8::Float64
    T9::Float64
    Ks1::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    T4::Float64
    T10::Float64
    T11::Float64
    Vs1_lim::Tuple{Float64, Float64}
    Vs2_lim::Tuple{Float64, Float64}
    Vst_lim::Tuple{Float64, Float64}
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

IEEE 421.5 2005 PSS2B IEEE Dual-Input Stabilizer Model

Arguments

input_code_1::Int: First Input Code for stabilizer, validation range: (1, 6), action if invalid: error

remote_bus_control_1::Int: First Input Remote Bus number for control.

input_code_2::Int: Second Input Code for stabilizer, validation range: (1, 6), action if invalid: error

remote_bus_control_2::Int: Second Input Remote Bus number for control.

M_rtf::Int: M parameter for ramp tracking filter, validation range: (0, 8), action if invalid: error

N_rtf::Int: N parameter for ramp tracking filter, validation range: (0, 8), action if invalid: error

Tw1::Float64: Time constant for first washout filter for first input, validation range: (eps(), nothing), action if invalid: warn

Tw2::Float64: Time constant for second washout filter for first input, validation range: (0, nothing), action if invalid: warn

T6::Float64: Time constant for low-pass filter for first input, validation range: (0, nothing), action if invalid: warn

Tw3::Float64: Time constant for first washout filter for second input, validation range: (eps(), nothing), action if invalid: warn

Tw4::Float64: Time constant for second washout filter for second input, validation range: (0, nothing), action if invalid: warn

T7::Float64: Time constant for low-pass filter for second input, validation range: (0, nothing), action if invalid: warn

Ks2::Float64: Gain for low-pass filter for second input, validation range: (0, nothing), action if invalid: warn

Ks3::Float64: Gain for second input, validation range: (0, nothing), action if invalid: warn

T8::Float64: Time constant for ramp tracking filter, validation range: (0, nothing), action if invalid: warn

T9::Float64: Time constant for ramp tracking filter, validation range: (eps(), nothing), action if invalid: warn

Ks1::Float64: Gain before lead-lag blocks, validation range: (0, nothing), action if invalid: warn

T1::Float64: Time constant for first lead-lag block, validation range: (0, nothing), action if invalid: warn

T2::Float64: Time constant for first lead-lag block, validation range: (0, nothing), action if invalid: warn

T3::Float64: Time constant for second lead-lag block, validation range: (0, nothing), action if invalid: warn

T4::Float64: Time constant for second lead-lag block, validation range: (0, nothing), action if invalid: warn

T10::Float64: Time constant for third lead-lag block, validation range: (0, nothing), action if invalid: warn

T11::Float64: Time constant for third lead-lag block, validation range: (0, nothing), action if invalid: warn

Vs1_lim::Tuple{Float64, Float64}: First input limits (Vs1_min, Vs1_max)

Vs2_lim::Tuple{Float64, Float64}: Second input limits (Vs2_min, Vs2_max)

Vst_lim::Tuple{Float64, Float64}: PSS output limits (Vst_min, Vst_max)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

x_p1: 1st washout 1st input, 
x_p2: 2nd washout 1st input, 
x_p3: transducer 1st input, 
x_p4: 1st washout 2nd input, 
x_p5: 2nd washout 2nd input, 
x_p6: transducer 2nd input, 
x_p7: ramp tracking filter state 1, 
x_p8: ramp tracking filter state 2, 
x_p9: ramp tracking filter state 3, 
x_p10: ramp tracking filter state 4, 
x_p11: ramp tracking filter state 5, 
x_p12: ramp tracking filter state 6, 
x_p13: ramp tracking filter state 7, 
x_p14: ramp tracking filter state 8, 
x_p15: 1st lead-lag, 
x_p16: 2nd lead-lag, 
x_p17: 3rd lead-lag,

n_states::Int: IEEEST has 17 states

states_types::Vector{StateTypes}: IEEEST has 17 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_Ks1 — Method

get_Ks1(value::PSS2B) -> Float64

Get PSS2B Ks1.

PowerSystems.get_Ks2 — Method

get_Ks2(value::PSS2B) -> Float64

Get PSS2B Ks2.

PowerSystems.get_Ks3 — Method

get_Ks3(value::PSS2B) -> Float64

Get PSS2B Ks3.

PowerSystems.get_M_rtf — Method

get_M_rtf(value::PSS2B) -> Int64

Get PSS2B M_rtf.

PowerSystems.get_N_rtf — Method

get_N_rtf(value::PSS2B) -> Int64

Get PSS2B N_rtf.

PowerSystems.get_T1 — Method

get_T1(value::PSS2B) -> Float64

Get PSS2B T1.

PowerSystems.get_T10 — Method

get_T10(value::PSS2B) -> Float64

Get PSS2B T10.

PowerSystems.get_T11 — Method

get_T11(value::PSS2B) -> Float64

Get PSS2B T11.

PowerSystems.get_T2 — Method

get_T2(value::PSS2B) -> Float64

Get PSS2B T2.

PowerSystems.get_T3 — Method

get_T3(value::PSS2B) -> Float64

Get PSS2B T3.

PowerSystems.get_T4 — Method

get_T4(value::PSS2B) -> Float64

Get PSS2B T4.

PowerSystems.get_T6 — Method

get_T6(value::PSS2B) -> Float64

Get PSS2B T6.

PowerSystems.get_T7 — Method

get_T7(value::PSS2B) -> Float64

Get PSS2B T7.

PowerSystems.get_T8 — Method

get_T8(value::PSS2B) -> Float64

Get PSS2B T8.

PowerSystems.get_T9 — Method

get_T9(value::PSS2B) -> Float64

Get PSS2B T9.

PowerSystems.get_Tw1 — Method

get_Tw1(value::PSS2B) -> Float64

Get PSS2B Tw1.

PowerSystems.get_Tw2 — Method

get_Tw2(value::PSS2B) -> Float64

Get PSS2B Tw2.

PowerSystems.get_Tw3 — Method

get_Tw3(value::PSS2B) -> Float64

Get PSS2B Tw3.

PowerSystems.get_Tw4 — Method

get_Tw4(value::PSS2B) -> Float64

Get PSS2B Tw4.

PowerSystems.get_Vs1_lim — Method

get_Vs1_lim(value::PSS2B) -> Tuple{Float64, Float64}

Get PSS2B Vs1_lim.

PowerSystems.get_Vs2_lim — Method

get_Vs2_lim(value::PSS2B) -> Tuple{Float64, Float64}

Get PSS2B Vs2_lim.

PowerSystems.get_Vst_lim — Method

get_Vst_lim(value::PSS2B) -> Tuple{Float64, Float64}

Get PSS2B Vst_lim.

PowerSystems.get_ext — Method

get_ext(value::PSS2B) -> Dict{String, Any}

Get PSS2B ext.

PowerSystems.get_input_code_1 — Method

get_input_code_1(value::PSS2B) -> Int64

Get PSS2B input_code_1.

PowerSystems.get_input_code_2 — Method

get_input_code_2(value::PSS2B) -> Int64

Get PSS2B input_code_2.

PowerSystems.get_n_states — Method

get_n_states(value::PSS2B) -> Int64

Get PSS2B n_states.

PowerSystems.get_remote_bus_control_1 — Method

get_remote_bus_control_1(value::PSS2B) -> Int64

Get PSS2B remote_bus_control_1.

PowerSystems.get_remote_bus_control_2 — Method

get_remote_bus_control_2(value::PSS2B) -> Int64

Get PSS2B remote_bus_control_2.

PowerSystems.get_states — Method

get_states(value::PSS2B) -> Vector{Symbol}

Get PSS2B states.

PowerSystems.get_states_types — Method

get_states_types(value::PSS2B) -> Vector{StateTypes}

Get PSS2B states_types.

PowerSystems.set_Ks1! — Method

set_Ks1!(value::PSS2B, val) -> Any

Set PSS2B Ks1.

PowerSystems.set_Ks2! — Method

set_Ks2!(value::PSS2B, val) -> Any

Set PSS2B Ks2.

PowerSystems.set_Ks3! — Method

set_Ks3!(value::PSS2B, val) -> Any

Set PSS2B Ks3.

PowerSystems.set_M_rtf! — Method

set_M_rtf!(value::PSS2B, val) -> Any

Set PSS2B M_rtf.

PowerSystems.set_N_rtf! — Method

set_N_rtf!(value::PSS2B, val) -> Any

Set PSS2B N_rtf.

PowerSystems.set_T1! — Method

set_T1!(value::PSS2B, val) -> Any

Set PSS2B T1.

PowerSystems.set_T10! — Method

set_T10!(value::PSS2B, val) -> Any

Set PSS2B T10.

PowerSystems.set_T11! — Method

set_T11!(value::PSS2B, val) -> Any

Set PSS2B T11.

PowerSystems.set_T2! — Method

set_T2!(value::PSS2B, val) -> Any

Set PSS2B T2.

PowerSystems.set_T3! — Method

set_T3!(value::PSS2B, val) -> Any

Set PSS2B T3.

PowerSystems.set_T4! — Method

set_T4!(value::PSS2B, val) -> Any

Set PSS2B T4.

PowerSystems.set_T6! — Method

set_T6!(value::PSS2B, val) -> Any

Set PSS2B T6.

PowerSystems.set_T7! — Method

set_T7!(value::PSS2B, val) -> Any

Set PSS2B T7.

PowerSystems.set_T8! — Method

set_T8!(value::PSS2B, val) -> Any

Set PSS2B T8.

PowerSystems.set_T9! — Method

set_T9!(value::PSS2B, val) -> Any

Set PSS2B T9.

PowerSystems.set_Tw1! — Method

set_Tw1!(value::PSS2B, val) -> Any

Set PSS2B Tw1.

PowerSystems.set_Tw2! — Method

set_Tw2!(value::PSS2B, val) -> Any

Set PSS2B Tw2.

PowerSystems.set_Tw3! — Method

set_Tw3!(value::PSS2B, val) -> Any

Set PSS2B Tw3.

PowerSystems.set_Tw4! — Method

set_Tw4!(value::PSS2B, val) -> Any

Set PSS2B Tw4.

PowerSystems.set_Vs1_lim! — Method

set_Vs1_lim!(value::PSS2B, val) -> Any

Set PSS2B Vs1_lim.

PowerSystems.set_Vs2_lim! — Method

set_Vs2_lim!(value::PSS2B, val) -> Any

Set PSS2B Vs2_lim.

PowerSystems.set_Vst_lim! — Method

set_Vst_lim!(value::PSS2B, val) -> Any

Set PSS2B Vst_lim.

PowerSystems.set_ext! — Method

set_ext!(value::PSS2B, val) -> Any

Set PSS2B ext.

PowerSystems.set_input_code_1! — Method

set_input_code_1!(value::PSS2B, val) -> Any

Set PSS2B input_code_1.

PowerSystems.set_input_code_2! — Method

set_input_code_2!(value::PSS2B, val) -> Any

Set PSS2B input_code_2.

PowerSystems.set_remote_bus_control_1! — Method

set_remote_bus_control_1!(value::PSS2B, val) -> Any

Set PSS2B remote_bus_control_1.

PowerSystems.set_remote_bus_control_2! — Method

set_remote_bus_control_2!(value::PSS2B, val) -> Any

Set PSS2B remote_bus_control_2.

PowerSystems.set_states_types! — Method

set_states_types!(value::PSS2B, val) -> Any

Set PSS2B states_types.

PSS2C

PowerSystems.PSS2C — Type

mutable struct PSS2C <: PSS
    input_code_1::Int
    remote_bus_control_1::Int
    input_code_2::Int
    remote_bus_control_2::Int
    M_rtf::Int
    N_rtf::Int
    Tw1::Float64
    Tw2::Float64
    T6::Float64
    Tw3::Float64
    Tw4::Float64
    T7::Float64
    Ks2::Float64
    Ks3::Float64
    T8::Float64
    T9::Float64
    Ks1::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    T4::Float64
    T10::Float64
    T11::Float64
    Vs1_lim::Tuple{Float64, Float64}
    Vs2_lim::Tuple{Float64, Float64}
    Vst_lim::Tuple{Float64, Float64}
    T12::Float64
    T13::Float64
    PSS_Hysteresis_param::Tuple{Float64, Float64}
    Xcomp::Float64
    Tcomp::Float64
    hysteresis_binary_logic::Int
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

IEEE 421.5 2016 PSS2C IEEE Dual-Input Stabilizer Model

Arguments

input_code_1::Int: First Input Code for stabilizer, validation range: (1, 7), action if invalid: error

remote_bus_control_1::Int: First Input Remote Bus number for control.

input_code_2::Int: Second Input Code for stabilizer, validation range: (1, 6), action if invalid: error

remote_bus_control_2::Int: Second Input Remote Bus number for control.

M_rtf::Int: M parameter for ramp tracking filter, validation range: (0, 8), action if invalid: error

N_rtf::Int: N parameter for ramp tracking filter, validation range: (0, 8), action if invalid: error

Tw1::Float64: Time constant for first washout filter for first input, validation range: (eps(), nothing), action if invalid: warn

Tw2::Float64: Time constant for second washout filter for first input, validation range: (0, nothing), action if invalid: warn

T6::Float64: Time constant for low-pass filter for first input, validation range: (0, nothing), action if invalid: warn

Tw3::Float64: Time constant for first washout filter for second input, validation range: (eps(), nothing), action if invalid: warn

Tw4::Float64: Time constant for second washout filter for second input, validation range: (0, nothing), action if invalid: warn

T7::Float64: Time constant for low-pass filter for second input, validation range: (0, nothing), action if invalid: warn

Ks2::Float64: Gain for low-pass filter for second input, validation range: (0, nothing), action if invalid: warn

Ks3::Float64: Gain for second input, validation range: (0, nothing), action if invalid: warn

T8::Float64: Time constant for ramp tracking filter, validation range: (0, nothing), action if invalid: warn

T9::Float64: Time constant for ramp tracking filter, validation range: (eps(), nothing), action if invalid: warn

Ks1::Float64: Gain before lead-lag blocks, validation range: (0, nothing), action if invalid: warn

T1::Float64: Time constant for first lead-lag block, validation range: (0, nothing), action if invalid: warn

T2::Float64: Time constant for first lead-lag block, validation range: (0, nothing), action if invalid: warn

T3::Float64: Time constant for second lead-lag block, validation range: (0, nothing), action if invalid: warn

T4::Float64: Time constant for second lead-lag block, validation range: (0, nothing), action if invalid: warn

T10::Float64: Time constant for third lead-lag block, validation range: (0, nothing), action if invalid: warn

T11::Float64: Time constant for third lead-lag block, validation range: (0, nothing), action if invalid: warn

Vs1_lim::Tuple{Float64, Float64}: First input limits (Vs1_min, Vs1_max)

Vs2_lim::Tuple{Float64, Float64}: Second input limits (Vs2_min, Vs2_max)

Vst_lim::Tuple{Float64, Float64}: PSS output limits (Vst_min, Vst_max)

T12::Float64: Time constant for fourth lead-lag block, validation range: (0, nothing), action if invalid: warn

T13::Float64: Time constant for fourth lead-lag block, validation range: (0, nothing), action if invalid: warn

PSS_Hysteresis_param::Tuple{Float64, Float64}: PSS output hysteresis parameters (PSSOFF, PSSON)

Xcomp::Float64: Stator Leakage Reactance, validation range: (0, nothing)

Tcomp::Float64: Time measured with compensated frequency, validation range: (eps(), nothing), action if invalid: error

hysteresis_binary_logic::Int: Hysteresis memory variable

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

x_p1: 1st washout 1st input, 
x_p2: 2nd washout 1st input, 
x_p3: transducer 1st input, 
x_p4: 1st washout 2nd input, 
x_p5: 2nd washout 2nd input, 
x_p6: transducer 2nd input, 
x_p7: ramp tracking filter state 1, 
x_p8: ramp tracking filter state 2, 
x_p9: ramp tracking filter state 3, 
x_p10: ramp tracking filter state 4, 
x_p11: ramp tracking filter state 5, 
x_p12: ramp tracking filter state 6, 
x_p13: ramp tracking filter state 7, 
x_p14: ramp tracking filter state 8, 
x_p15: 1st lead-lag, 
x_p16: 2nd lead-lag, 
x_p17: 3rd lead-lag, 
x_p18: 4th lead-lag, 
x_p19: washout block for compensated frequency,

n_states::Int: IEEEST has 19 states

states_types::Vector{StateTypes}: IEEEST has 19 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_Ks1 — Method

get_Ks1(value::PSS2C) -> Float64

Get PSS2C Ks1.

PowerSystems.get_Ks2 — Method

get_Ks2(value::PSS2C) -> Float64

Get PSS2C Ks2.

PowerSystems.get_Ks3 — Method

get_Ks3(value::PSS2C) -> Float64

Get PSS2C Ks3.

PowerSystems.get_M_rtf — Method

get_M_rtf(value::PSS2C) -> Int64

Get PSS2C M_rtf.

PowerSystems.get_N_rtf — Method

get_N_rtf(value::PSS2C) -> Int64

Get PSS2C N_rtf.

PowerSystems.get_PSS_Hysteresis_param — Method

get_PSS_Hysteresis_param(
    value::PSS2C
) -> Tuple{Float64, Float64}

Get PSS2C PSS_Hysteresis_param.

PowerSystems.get_T1 — Method

get_T1(value::PSS2C) -> Float64

Get PSS2C T1.

PowerSystems.get_T10 — Method

get_T10(value::PSS2C) -> Float64

Get PSS2C T10.

PowerSystems.get_T11 — Method

get_T11(value::PSS2C) -> Float64

Get PSS2C T11.

PowerSystems.get_T12 — Method

get_T12(value::PSS2C) -> Float64

Get PSS2C T12.

PowerSystems.get_T13 — Method

get_T13(value::PSS2C) -> Float64

Get PSS2C T13.

PowerSystems.get_T2 — Method

get_T2(value::PSS2C) -> Float64

Get PSS2C T2.

PowerSystems.get_T3 — Method

get_T3(value::PSS2C) -> Float64

Get PSS2C T3.

PowerSystems.get_T4 — Method

get_T4(value::PSS2C) -> Float64

Get PSS2C T4.

PowerSystems.get_T6 — Method

get_T6(value::PSS2C) -> Float64

Get PSS2C T6.

PowerSystems.get_T7 — Method

get_T7(value::PSS2C) -> Float64

Get PSS2C T7.

PowerSystems.get_T8 — Method

get_T8(value::PSS2C) -> Float64

Get PSS2C T8.

PowerSystems.get_T9 — Method

get_T9(value::PSS2C) -> Float64

Get PSS2C T9.

PowerSystems.get_Tcomp — Method

get_Tcomp(value::PSS2C) -> Float64

Get PSS2C Tcomp.

PowerSystems.get_Tw1 — Method

get_Tw1(value::PSS2C) -> Float64

Get PSS2C Tw1.

PowerSystems.get_Tw2 — Method

get_Tw2(value::PSS2C) -> Float64

Get PSS2C Tw2.

PowerSystems.get_Tw3 — Method

get_Tw3(value::PSS2C) -> Float64

Get PSS2C Tw3.

PowerSystems.get_Tw4 — Method

get_Tw4(value::PSS2C) -> Float64

Get PSS2C Tw4.

PowerSystems.get_Vs1_lim — Method

get_Vs1_lim(value::PSS2C) -> Tuple{Float64, Float64}

Get PSS2C Vs1_lim.

PowerSystems.get_Vs2_lim — Method

get_Vs2_lim(value::PSS2C) -> Tuple{Float64, Float64}

Get PSS2C Vs2_lim.

PowerSystems.get_Vst_lim — Method

get_Vst_lim(value::PSS2C) -> Tuple{Float64, Float64}

Get PSS2C Vst_lim.

PowerSystems.get_Xcomp — Method

get_Xcomp(value::PSS2C) -> Float64

Get PSS2C Xcomp.

PowerSystems.get_ext — Method

get_ext(value::PSS2C) -> Dict{String, Any}

Get PSS2C ext.

PowerSystems.get_hysteresis_binary_logic — Method

get_hysteresis_binary_logic(value::PSS2C) -> Int64

Get PSS2C hysteresis_binary_logic.

PowerSystems.get_input_code_1 — Method

get_input_code_1(value::PSS2C) -> Int64

Get PSS2C input_code_1.

PowerSystems.get_input_code_2 — Method

get_input_code_2(value::PSS2C) -> Int64

Get PSS2C input_code_2.

PowerSystems.get_n_states — Method

get_n_states(value::PSS2C) -> Int64

Get PSS2C n_states.

PowerSystems.get_remote_bus_control_1 — Method

get_remote_bus_control_1(value::PSS2C) -> Int64

Get PSS2C remote_bus_control_1.

PowerSystems.get_remote_bus_control_2 — Method

get_remote_bus_control_2(value::PSS2C) -> Int64

Get PSS2C remote_bus_control_2.

PowerSystems.get_states — Method

get_states(value::PSS2C) -> Vector{Symbol}

Get PSS2C states.

PowerSystems.get_states_types — Method

get_states_types(value::PSS2C) -> Vector{StateTypes}

Get PSS2C states_types.

PowerSystems.set_Ks1! — Method

set_Ks1!(value::PSS2C, val) -> Any

Set PSS2C Ks1.

PowerSystems.set_Ks2! — Method

set_Ks2!(value::PSS2C, val) -> Any

Set PSS2C Ks2.

PowerSystems.set_Ks3! — Method

set_Ks3!(value::PSS2C, val) -> Any

Set PSS2C Ks3.

PowerSystems.set_M_rtf! — Method

set_M_rtf!(value::PSS2C, val) -> Any

Set PSS2C M_rtf.

PowerSystems.set_N_rtf! — Method

set_N_rtf!(value::PSS2C, val) -> Any

Set PSS2C N_rtf.

PowerSystems.set_PSS_Hysteresis_param! — Method

set_PSS_Hysteresis_param!(value::PSS2C, val) -> Any

Set PSS2C PSS_Hysteresis_param.

PowerSystems.set_T1! — Method

set_T1!(value::PSS2C, val) -> Any

Set PSS2C T1.

PowerSystems.set_T10! — Method

set_T10!(value::PSS2C, val) -> Any

Set PSS2C T10.

PowerSystems.set_T11! — Method

set_T11!(value::PSS2C, val) -> Any

Set PSS2C T11.

PowerSystems.set_T12! — Method

set_T12!(value::PSS2C, val) -> Any

Set PSS2C T12.

PowerSystems.set_T13! — Method

set_T13!(value::PSS2C, val) -> Any

Set PSS2C T13.

PowerSystems.set_T2! — Method

set_T2!(value::PSS2C, val) -> Any

Set PSS2C T2.

PowerSystems.set_T3! — Method

set_T3!(value::PSS2C, val) -> Any

Set PSS2C T3.

PowerSystems.set_T4! — Method

set_T4!(value::PSS2C, val) -> Any

Set PSS2C T4.

PowerSystems.set_T6! — Method

set_T6!(value::PSS2C, val) -> Any

Set PSS2C T6.

PowerSystems.set_T7! — Method

set_T7!(value::PSS2C, val) -> Any

Set PSS2C T7.

PowerSystems.set_T8! — Method

set_T8!(value::PSS2C, val) -> Any

Set PSS2C T8.

PowerSystems.set_T9! — Method

set_T9!(value::PSS2C, val) -> Any

Set PSS2C T9.

PowerSystems.set_Tcomp! — Method

set_Tcomp!(value::PSS2C, val) -> Any

Set PSS2C Tcomp.

PowerSystems.set_Tw1! — Method

set_Tw1!(value::PSS2C, val) -> Any

Set PSS2C Tw1.

PowerSystems.set_Tw2! — Method

set_Tw2!(value::PSS2C, val) -> Any

Set PSS2C Tw2.

PowerSystems.set_Tw3! — Method

set_Tw3!(value::PSS2C, val) -> Any

Set PSS2C Tw3.

PowerSystems.set_Tw4! — Method

set_Tw4!(value::PSS2C, val) -> Any

Set PSS2C Tw4.

PowerSystems.set_Vs1_lim! — Method

set_Vs1_lim!(value::PSS2C, val) -> Any

Set PSS2C Vs1_lim.

PowerSystems.set_Vs2_lim! — Method

set_Vs2_lim!(value::PSS2C, val) -> Any

Set PSS2C Vs2_lim.

PowerSystems.set_Vst_lim! — Method

set_Vst_lim!(value::PSS2C, val) -> Any

Set PSS2C Vst_lim.

PowerSystems.set_Xcomp! — Method

set_Xcomp!(value::PSS2C, val) -> Any

Set PSS2C Xcomp.

PowerSystems.set_ext! — Method

set_ext!(value::PSS2C, val) -> Any

Set PSS2C ext.

PowerSystems.set_hysteresis_binary_logic! — Method

set_hysteresis_binary_logic!(value::PSS2C, val) -> Any

Set PSS2C hysteresis_binary_logic.

PowerSystems.set_input_code_1! — Method

set_input_code_1!(value::PSS2C, val) -> Any

Set PSS2C input_code_1.

PowerSystems.set_input_code_2! — Method

set_input_code_2!(value::PSS2C, val) -> Any

Set PSS2C input_code_2.

PowerSystems.set_remote_bus_control_1! — Method

set_remote_bus_control_1!(value::PSS2C, val) -> Any

Set PSS2C remote_bus_control_1.

PowerSystems.set_remote_bus_control_2! — Method

set_remote_bus_control_2!(value::PSS2C, val) -> Any

Set PSS2C remote_bus_control_2.

PowerSystems.set_states_types! — Method

set_states_types!(value::PSS2C, val) -> Any

Set PSS2C states_types.

PSSFixed

PowerSystems.PSSFixed — Type

mutable struct PSSFixed <: PSS
    V_pss::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of a PSS that returns a fixed voltage to add to the reference for the AVR

Arguments

V_pss::Float64: Fixed voltage stabilization signal, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}

n_states::Int: PSSFixed has no states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_V_pss — Method

get_V_pss(value::PSSFixed) -> Float64

Get PSSFixed V_pss.

PowerSystems.get_ext — Method

get_ext(value::PSSFixed) -> Dict{String, Any}

Get PSSFixed ext.

PowerSystems.get_n_states — Method

get_n_states(value::PSSFixed) -> Int64

Get PSSFixed n_states.

PowerSystems.get_states — Method

get_states(value::PSSFixed) -> Vector{Symbol}

Get PSSFixed states.

PowerSystems.set_V_pss! — Method

set_V_pss!(value::PSSFixed, val) -> Any

Set PSSFixed V_pss.

PowerSystems.set_ext! — Method

set_ext!(value::PSSFixed, val) -> Any

Set PSSFixed ext.

PSSSimple

PowerSystems.PSSSimple — Type

mutable struct PSSSimple <: PSS
    K_ω::Float64
    K_p::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of a PSS that returns a proportional droop voltage to add to the reference for the AVR

Arguments

K_ω::Float64: Proportional gain for frequency, validation range: (0, nothing)

K_p::Float64: Proportional gain for active power, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}

n_states::Int: PSSSimple has no states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_K_p — Method

get_K_p(value::PSSSimple) -> Float64

Get PSSSimple K_p.

PowerSystems.get_K_ω — Method

get_K_ω(value::PSSSimple) -> Float64

Get PSSSimple K_ω.

PowerSystems.get_ext — Method

get_ext(value::PSSSimple) -> Dict{String, Any}

Get PSSSimple ext.

PowerSystems.get_n_states — Method

get_n_states(value::PSSSimple) -> Int64

Get PSSSimple n_states.

PowerSystems.get_states — Method

get_states(value::PSSSimple) -> Vector{Symbol}

Get PSSSimple states.

PowerSystems.set_K_p! — Method

set_K_p!(value::PSSSimple, val) -> Any

Set PSSSimple K_p.

PowerSystems.set_K_ω! — Method

set_K_ω!(value::PSSSimple, val) -> Any

Set PSSSimple K_ω.

PowerSystems.set_ext! — Method

set_ext!(value::PSSSimple, val) -> Any

Set PSSSimple ext.

STAB1

PowerSystems.STAB1 — Type

mutable struct STAB1 <: PSS
    KT::Float64
    T::Float64
    T1T3::Float64
    T3::Float64
    T2T4::Float64
    T4::Float64
    H_lim::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

Speed-Sensitive Stabilizing Model

Arguments

KT::Float64: K/T for washout filter, validation range: (0, nothing), action if invalid: warn

T::Float64: Time constant for washout filter, validation range: (0.01, nothing), action if invalid: warn

T1T3::Float64: Time constant division T1/T3, validation range: (0, nothing)

T3::Float64: Time constant, validation range: (0.01, nothing), action if invalid: warn

T2T4::Float64: Time constant division T2/T4, validation range: (0, nothing), action if invalid: warn

T4::Float64: Time constant, validation range: (0.01, nothing), action if invalid: warn

H_lim::Float64: PSS output limit, validation range: (0, 0.5), action if invalid: warn

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

x_p1: washout filter,
x_p2: T1/T3 lead-lag block, 
x_p3: T2/T4 lead-lag block,

n_states::Int: STAB1 has 3 states

states_types::Vector{StateTypes}: STAB1 has 3 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_H_lim — Method

get_H_lim(value::STAB1) -> Float64

Get STAB1 H_lim.

PowerSystems.get_KT — Method

get_KT(value::STAB1) -> Float64

Get STAB1 KT.

PowerSystems.get_T — Method

get_T(value::STAB1) -> Float64

Get STAB1 T.

PowerSystems.get_T1T3 — Method

get_T1T3(value::STAB1) -> Float64

Get STAB1 T1T3.

PowerSystems.get_T2T4 — Method

get_T2T4(value::STAB1) -> Float64

Get STAB1 T2T4.

PowerSystems.get_T3 — Method

get_T3(value::STAB1) -> Float64

Get STAB1 T3.

PowerSystems.get_T4 — Method

get_T4(value::STAB1) -> Float64

Get STAB1 T4.

PowerSystems.get_ext — Method

get_ext(value::STAB1) -> Dict{String, Any}

Get STAB1 ext.

PowerSystems.get_n_states — Method

get_n_states(value::STAB1) -> Int64

Get STAB1 n_states.

PowerSystems.get_states — Method

get_states(value::STAB1) -> Vector{Symbol}

Get STAB1 states.

PowerSystems.get_states_types — Method

get_states_types(value::STAB1) -> Vector{StateTypes}

Get STAB1 states_types.

PowerSystems.set_H_lim! — Method

set_H_lim!(value::STAB1, val) -> Any

Set STAB1 H_lim.

PowerSystems.set_KT! — Method

set_KT!(value::STAB1, val) -> Any

Set STAB1 KT.

PowerSystems.set_T! — Method

set_T!(value::STAB1, val) -> Any

Set STAB1 T.

PowerSystems.set_T1T3! — Method

set_T1T3!(value::STAB1, val) -> Any

Set STAB1 T1T3.

PowerSystems.set_T2T4! — Method

set_T2T4!(value::STAB1, val) -> Any

Set STAB1 T2T4.

PowerSystems.set_T3! — Method

set_T3!(value::STAB1, val) -> Any

Set STAB1 T3.

PowerSystems.set_T4! — Method

set_T4!(value::STAB1, val) -> Any

Set STAB1 T4.

PowerSystems.set_ext! — Method

set_ext!(value::STAB1, val) -> Any

Set STAB1 ext.

PowerSystems.set_states_types! — Method

set_states_types!(value::STAB1, val) -> Any

Set STAB1 states_types.