PeriodicVariableSource

PowerSystems.PeriodicVariableSource — Type

mutable struct PeriodicVariableSource <: StaticInjection
    name::String
    available::Bool
    bus::Bus
    initial_active_power::Float64
    initial_reactive_power::Float64
    R_th::Float64
    X_th::Float64
    internal_voltage_bias::Float64
    internal_voltage_frequencies::Vector{Float64}
    internal_voltage_coefficients::Vector{Tuple{Float64,Float64}}
    internal_angle_bias::Float64
    internal_angle_frequencies::Vector{Float64}
    internal_angle_coefficients::Vector{Tuple{Float64,Float64}}
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

This struct acts as an infinity bus with time varying phasor values magnitude and angle V(t) heta(t). Time varying functions are represented using fourier series

Arguments

name::String

available::Bool

bus::Bus

initial_active_power::Float64

initial_reactive_power::Float64

R_th::Float64: Source Thevenin resistance, validation range: (0, nothing)

X_th::Float64: Source Thevenin reactance, validation range: (0, nothing)

internal_voltage_bias::Float64: a0 term of the Fourier Series for the voltage

internal_voltage_frequencies::Vector{Float64}: Frequencies in radians/s

internal_voltage_coefficients::Vector{Tuple{Float64,Float64}}: Coefficients for terms n > 1. First component corresponds to sin and second component to cos

internal_angle_bias::Float64: a0 term of the Fourier Series for the angle

internal_angle_frequencies::Vector{Float64}: Frequencies in radians/s

internal_angle_coefficients::Vector{Tuple{Float64,Float64}}: Coefficients for terms n > 1. First component corresponds to sin and second component to cos

services::Vector{Service}: Services that this device contributes to

ext::Dict{String, Any}

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

InfrastructureSystems.get_name — Method

get_name(value::PeriodicVariableSource) -> String

Get PeriodicVariableSource name.

InfrastructureSystems.set_name! — Method

set_name!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource name.

PowerSystems.get_R_th — Method

get_R_th(value::PeriodicVariableSource) -> Float64

Get PeriodicVariableSource R_th.

PowerSystems.get_X_th — Method

get_X_th(value::PeriodicVariableSource) -> Float64

Get PeriodicVariableSource X_th.

PowerSystems.get_available — Method

get_available(value::PeriodicVariableSource) -> Bool

Get PeriodicVariableSource available.

PowerSystems.get_bus — Method

get_bus(value::PeriodicVariableSource) -> Bus

Get PeriodicVariableSource bus.

PowerSystems.get_ext — Method

get_ext(value::PeriodicVariableSource) -> Dict{String, Any}

Get PeriodicVariableSource ext.

PowerSystems.get_initial_active_power — Method

get_initial_active_power(value::PeriodicVariableSource) -> Float64

Get PeriodicVariableSource initial_active_power.

PowerSystems.get_initial_reactive_power — Method

get_initial_reactive_power(value::PeriodicVariableSource) -> Float64

Get PeriodicVariableSource initial_reactive_power.

PowerSystems.get_internal_angle_bias — Method

get_internal_angle_bias(value::PeriodicVariableSource) -> Float64

Get PeriodicVariableSource internal_angle_bias.

PowerSystems.get_internal_angle_coefficients — Method

get_internal_angle_coefficients(value::PeriodicVariableSource) -> Vector{Tuple{Float64, Float64}}

Get PeriodicVariableSource internal_angle_coefficients.

PowerSystems.get_internal_angle_frequencies — Method

get_internal_angle_frequencies(value::PeriodicVariableSource) -> Vector{Float64}

Get PeriodicVariableSource internal_angle_frequencies.

PowerSystems.get_internal_voltage_bias — Method

get_internal_voltage_bias(value::PeriodicVariableSource) -> Float64

Get PeriodicVariableSource internal_voltage_bias.

PowerSystems.get_internal_voltage_coefficients — Method

get_internal_voltage_coefficients(value::PeriodicVariableSource) -> Vector{Tuple{Float64, Float64}}

Get PeriodicVariableSource internal_voltage_coefficients.

PowerSystems.get_internal_voltage_frequencies — Method

get_internal_voltage_frequencies(value::PeriodicVariableSource) -> Vector{Float64}

Get PeriodicVariableSource internal_voltage_frequencies.

PowerSystems.get_services — Method

get_services(value::PeriodicVariableSource) -> Vector{Service}

Get PeriodicVariableSource services.

PowerSystems.set_R_th! — Method

set_R_th!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource R_th.

PowerSystems.set_X_th! — Method

set_X_th!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource X_th.

PowerSystems.set_available! — Method

set_available!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource available.

PowerSystems.set_bus! — Method

set_bus!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource bus.

PowerSystems.set_ext! — Method

set_ext!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource ext.

PowerSystems.set_initial_active_power! — Method

set_initial_active_power!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource initial_active_power.

PowerSystems.set_initial_reactive_power! — Method

set_initial_reactive_power!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource initial_reactive_power.

PowerSystems.set_internal_angle_bias! — Method

set_internal_angle_bias!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource internal_angle_bias.

PowerSystems.set_internal_angle_coefficients! — Method

set_internal_angle_coefficients!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource internal_angle_coefficients.

PowerSystems.set_internal_angle_frequencies! — Method

set_internal_angle_frequencies!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource internal_angle_frequencies.

PowerSystems.set_internal_voltage_bias! — Method

set_internal_voltage_bias!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource internal_voltage_bias.

PowerSystems.set_internal_voltage_coefficients! — Method

set_internal_voltage_coefficients!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource internal_voltage_coefficients.

PowerSystems.set_internal_voltage_frequencies! — Method

set_internal_voltage_frequencies!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource internal_voltage_frequencies.

PowerSystems.set_services! — Method

set_services!(value::PeriodicVariableSource, val) -> Any

Set PeriodicVariableSource services.