PhaseShiftingTransformer

PowerSystems.PhaseShiftingTransformer — Type

mutable struct PhaseShiftingTransformer <: TwoWindingTransformer
    name::String
    available::Bool
    active_power_flow::Float64
    reactive_power_flow::Float64
    arc::Arc
    r::Float64
    x::Float64
    primary_shunt::Complex{Float64}
    tap::Float64
    α::Float64
    rating::Union{Nothing, Float64}
    base_power::Float64
    base_voltage_primary::Union{Nothing, Float64}
    base_voltage_secondary::Union{Nothing, Float64}
    rating_b::Union{Nothing, Float64}
    rating_c::Union{Nothing, Float64}
    phase_angle_limits::MinMax
    control_objective::TransformerControlObjective
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A phase-shifting transformer regulating the phase angle between two buses to control active power flow in the system.

The model uses an equivalent circuit assuming the impedance is on the High Voltage Side of the transformer. The model allocates the iron losses and magnetizing susceptance to the primary side

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow::Float64: Initial condition of active power flow through the transformer (MW)

reactive_power_flow::Float64: Initial condition of reactive power flow through the transformer (MVAR)

arc::Arc: An Arc defining this transformer from a bus to another bus

r::Float64: Resistance in pu (SYSTEM_BASE), validation range: (0, 4)

x::Float64: Reactance in pu (SYSTEM_BASE), validation range: (-2, 4)

primary_shunt::Complex{Float64}: Primary shunt admittance in pu (SYSTEM_BASE)

tap::Float64: Normalized tap changer position for voltage control, varying between 0 and 2, with 1 centered at the nominal voltage, validation range: (0, 2)

α::Float64: Initial condition of phase shift (radians) between the from and to buses , validation range: (-1.571, 1.571)

rating::Union{Nothing, Float64}: Thermal rating (MVA). Flow through the transformer must be between -rating and rating. When defining a transformer before it is attached to a System, rating must be in pu (SYSTEM_BASE) using the base power of the System it will be attached to, validation range: (0, nothing)

base_power::Float64: Base power (MVA) for per unitization, validation range: (0.0001, nothing)

base_voltage_primary::Union{Nothing, Float64}: (default: get_base_voltage(get_from(arc))) Primary base voltage in kV, validation range: (0, nothing)

base_voltage_secondary::Union{Nothing, Float64}: (default: get_base_voltage(get_to(arc))) Secondary base voltage in kV, validation range: (0, nothing)

rating_b::Union{Nothing, Float64}: (default: nothing) Second current rating; entered in MVA.

rating_c::Union{Nothing, Float64}: (default: nothing) Third current rating; entered in MVA.

phase_angle_limits::MinMax: (default: (min=-3.1416, max=3.1416)) Minimum and maximum phase angle limits (radians)

control_objective::TransformerControlObjective: (default: TransformerControlObjective.UNDEFINED) Control objective for the tap changer for power flow calculations. See TransformerControlObjective

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::PhaseShiftingTransformer) -> Bool

Get PhaseShiftingTransformer available.

InfrastructureSystems.get_name — Method

get_name(value::PhaseShiftingTransformer) -> String

Get PhaseShiftingTransformer name.

InfrastructureSystems.set_available! — Method

set_available!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer available.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(
    value::PhaseShiftingTransformer
) -> Float64

Get PhaseShiftingTransformer active_power_flow.

PowerSystems.get_arc — Method

get_arc(value::PhaseShiftingTransformer) -> Arc

Get PhaseShiftingTransformer arc.

PowerSystems.get_base_power — Method

get_base_power(value::PhaseShiftingTransformer) -> Float64

Get PhaseShiftingTransformer base_power.

PowerSystems.get_base_voltage_primary — Method

get_base_voltage_primary(
    value::PhaseShiftingTransformer
) -> Union{Nothing, Float64}

Get PhaseShiftingTransformer base_voltage_primary.

PowerSystems.get_base_voltage_secondary — Method

get_base_voltage_secondary(
    value::PhaseShiftingTransformer
) -> Union{Nothing, Float64}

Get PhaseShiftingTransformer base_voltage_secondary.

PowerSystems.get_control_objective — Method

get_control_objective(
    value::PhaseShiftingTransformer
) -> PowerSystems.TransformerControlObjectiveModule.TransformerControlObjective

Get PhaseShiftingTransformer control_objective.

PowerSystems.get_ext — Method

get_ext(
    value::PhaseShiftingTransformer
) -> Dict{String, Any}

Get PhaseShiftingTransformer ext.

PowerSystems.get_phase_angle_limits — Method

get_phase_angle_limits(
    value::PhaseShiftingTransformer
) -> @NamedTuple{min::Float64, max::Float64}

Get PhaseShiftingTransformer phase_angle_limits.

PowerSystems.get_primary_shunt — Method

get_primary_shunt(
    value::PhaseShiftingTransformer
) -> ComplexF64

Get PhaseShiftingTransformer primary_shunt.

PowerSystems.get_r — Method

get_r(value::PhaseShiftingTransformer) -> Float64

Get PhaseShiftingTransformer r.

PowerSystems.get_rating — Method

get_rating(
    value::PhaseShiftingTransformer
) -> Union{Nothing, Float64}

Get PhaseShiftingTransformer rating.

PowerSystems.get_rating_b — Method

get_rating_b(
    value::PhaseShiftingTransformer
) -> Union{Nothing, Float64}

Get PhaseShiftingTransformer rating_b.

PowerSystems.get_rating_c — Method

get_rating_c(
    value::PhaseShiftingTransformer
) -> Union{Nothing, Float64}

Get PhaseShiftingTransformer rating_c.

PowerSystems.get_reactive_power_flow — Method

get_reactive_power_flow(
    value::PhaseShiftingTransformer
) -> Float64

Get PhaseShiftingTransformer reactive_power_flow.

PowerSystems.get_services — Method

get_services(
    value::PhaseShiftingTransformer
) -> Vector{Service}

Get PhaseShiftingTransformer services.

PowerSystems.get_tap — Method

get_tap(value::PhaseShiftingTransformer) -> Float64

Get PhaseShiftingTransformer tap.

PowerSystems.get_x — Method

get_x(value::PhaseShiftingTransformer) -> Float64

Get PhaseShiftingTransformer x.

PowerSystems.get_α — Method

get_α(value::PhaseShiftingTransformer) -> Float64

Get PhaseShiftingTransformer α.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(
    value::PhaseShiftingTransformer,
    val
) -> Any

Set PhaseShiftingTransformer active_power_flow.

PowerSystems.set_arc! — Method

set_arc!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer arc.

PowerSystems.set_base_power! — Method

set_base_power!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer base_power.

PowerSystems.set_base_voltage_primary! — Method

set_base_voltage_primary!(
    value::PhaseShiftingTransformer,
    val
) -> Any

Set PhaseShiftingTransformer base_voltage_primary.

PowerSystems.set_base_voltage_secondary! — Method

set_base_voltage_secondary!(
    value::PhaseShiftingTransformer,
    val
) -> Any

Set PhaseShiftingTransformer base_voltage_secondary.

PowerSystems.set_control_objective! — Method

set_control_objective!(
    value::PhaseShiftingTransformer,
    val
) -> Any

Set PhaseShiftingTransformer control_objective.

PowerSystems.set_ext! — Method

set_ext!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer ext.

PowerSystems.set_phase_angle_limits! — Method

set_phase_angle_limits!(
    value::PhaseShiftingTransformer,
    val
) -> Any

Set PhaseShiftingTransformer phase_angle_limits.

PowerSystems.set_primary_shunt! — Method

set_primary_shunt!(
    value::PhaseShiftingTransformer,
    val
) -> Any

Set PhaseShiftingTransformer primary_shunt.

PowerSystems.set_r! — Method

set_r!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer r.

PowerSystems.set_rating! — Method

set_rating!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer rating.

PowerSystems.set_rating_b! — Method

set_rating_b!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer rating_b.

PowerSystems.set_rating_c! — Method

set_rating_c!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer rating_c.

PowerSystems.set_reactive_power_flow! — Method

set_reactive_power_flow!(
    value::PhaseShiftingTransformer,
    val
) -> Any

Set PhaseShiftingTransformer reactive_power_flow.

PowerSystems.set_services! — Method

set_services!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer services.

PowerSystems.set_tap! — Method

set_tap!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer tap.

PowerSystems.set_x! — Method

set_x!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer x.

PowerSystems.set_α! — Method

set_α!(value::PhaseShiftingTransformer, val) -> Any

Set PhaseShiftingTransformer α.