PowerLoad

PowerSystems.PowerLoad — Type

mutable struct PowerLoad <: StaticLoad
    name::String
    available::Bool
    bus::ACBus
    active_power::Float64
    reactive_power::Float64
    base_power::Float64
    max_active_power::Float64
    max_reactive_power::Float64
    services::Vector{Service}
    dynamic_injector::Union{Nothing, DynamicInjection}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A static power load, most commonly used for operational models such as power flow and operational optimizations.

This load consumes a set amount of power (set by active_power for a power flow simulation or a max_active_power time series for an operational simulation). For loads that can be compensated for load interruptions through demand response programs, see InterruptiblePowerLoad. For voltage-dependent loads used in dynamics modeling, see StandardLoad

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

bus::ACBus: Bus that this component is connected to

active_power::Float64: Initial steady-state active power demand (MW)

reactive_power::Float64: Initial steady-state reactive power demand (MVAR)

base_power::Float64: Base power (MVA) for per unitization, validation range: (0, nothing)

max_active_power::Float64: Maximum active power (MW) that this load can demand

max_reactive_power::Float64: Maximum reactive power (MVAR) that this load can demand

services::Vector{Service}: (default: Device[]) Services that this device contributes to

dynamic_injector::Union{Nothing, DynamicInjection}: (default: nothing) corresponding dynamic injection device

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::PowerLoad) -> String

Get PowerLoad name.

PowerSystems.get_active_power — Method

get_active_power(value::PowerLoad) -> Any

Get PowerLoad active_power.

PowerSystems.get_available — Method

get_available(value::PowerLoad) -> Bool

Get PowerLoad available.

PowerSystems.get_base_power — Method

get_base_power(value::PowerLoad) -> Float64

Get PowerLoad base_power.

PowerSystems.get_bus — Method

get_bus(value::PowerLoad) -> ACBus

Get PowerLoad bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(
    value::PowerLoad
) -> Union{Nothing, DynamicInjection}

Get PowerLoad dynamic_injector.

PowerSystems.get_ext — Method

get_ext(value::PowerLoad) -> Dict{String, Any}

Get PowerLoad ext.

PowerSystems.get_max_active_power — Method

get_max_active_power(value::PowerLoad) -> Any

Get PowerLoad max_active_power.

PowerSystems.get_max_reactive_power — Method

get_max_reactive_power(value::PowerLoad) -> Any

Get PowerLoad max_reactive_power.

PowerSystems.get_reactive_power — Method

get_reactive_power(value::PowerLoad) -> Any

Get PowerLoad reactive_power.

PowerSystems.get_services — Method

get_services(value::PowerLoad) -> Vector{Service}

Get PowerLoad services.

PowerSystems.set_active_power! — Method

set_active_power!(value::PowerLoad, val) -> Any

Set PowerLoad active_power.

PowerSystems.set_available! — Method

set_available!(value::PowerLoad, val) -> Any

Set PowerLoad available.

PowerSystems.set_base_power! — Method

set_base_power!(value::PowerLoad, val) -> Any

Set PowerLoad base_power.

PowerSystems.set_bus! — Method

set_bus!(value::PowerLoad, val) -> Any

Set PowerLoad bus.

PowerSystems.set_ext! — Method

set_ext!(value::PowerLoad, val) -> Any

Set PowerLoad ext.

PowerSystems.set_max_active_power! — Method

set_max_active_power!(value::PowerLoad, val) -> Any

Set PowerLoad max_active_power.

PowerSystems.set_max_reactive_power! — Method

set_max_reactive_power!(value::PowerLoad, val) -> Any

Set PowerLoad max_reactive_power.

PowerSystems.set_reactive_power! — Method

set_reactive_power!(value::PowerLoad, val) -> Any

Set PowerLoad reactive_power.

PowerSystems.set_services! — Method

set_services!(value::PowerLoad, val) -> Any

Set PowerLoad services.