RenewableDispatch

PowerSystems.RenewableDispatch — Type

mutable struct RenewableDispatch <: RenewableGen
    name::String
    available::Bool
    bus::ACBus
    active_power::Float64
    reactive_power::Float64
    rating::Float64
    prime_mover_type::PrimeMovers
    reactive_power_limits::Union{Nothing, MinMax}
    power_factor::Float64
    operation_cost::Union{RenewableGenerationCost, MarketBidCost}
    base_power::Float64
    services::Vector{Service}
    dynamic_injector::Union{Nothing, DynamicInjection}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A renewable (e.g., wind or solar) generator whose output can be curtailed to satisfy power system constraints.

These generators can also participate in reserves markets, including upwards reserves by proactively curtailing some available power (based on its max_active_power time series). Example uses include: a utility-scale wind or solar generator whose PPA allows curtailment. For non-curtailable or must-take renewables, see RenewableNonDispatch.

Renewable generators do not have a max_active_power parameter, which is instead calculated when calling get_max_active_power()

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

bus::ACBus: Bus that this component is connected to

active_power::Float64: Initial active power set point of the unit in MW. For power flow, this is the steady state operating point of the system. For production cost modeling, this may or may not be used as the initial starting point for the solver, depending on the solver used

reactive_power::Float64: Initial reactive power set point of the unit (MVAR), used in some production cost modeling simulations. To set the reactive power in a load flow, use power_factor

rating::Float64: Maximum output power rating of the unit (MVA), validation range: (0, nothing)

prime_mover_type::PrimeMovers: Prime mover technology according to EIA 923. Options are listed here

reactive_power_limits::Union{Nothing, MinMax}: Minimum and maximum reactive power limits, used in some production cost model simulations and in power flow if the unit is connected to a PV bus. Set to nothing if not applicable

power_factor::Float64: Power factor [0, 1] set-point, used in some production cost modeling and in load flow if the unit is connected to a PQ bus, validation range: (0, 1)

operation_cost::Union{RenewableGenerationCost, MarketBidCost}: OperationalCost of generation

base_power::Float64: Base power of the unit (MVA) for per unitization, validation range: (0, nothing)

services::Vector{Service}: (default: Device[]) Services that this device contributes to

dynamic_injector::Union{Nothing, DynamicInjection}: (default: nothing) corresponding dynamic injection device

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::RenewableDispatch) -> String

Get RenewableDispatch name.

PowerSystems.get_active_power — Method

get_active_power(value::RenewableDispatch) -> Any

Get RenewableDispatch active_power.

PowerSystems.get_available — Method

get_available(value::RenewableDispatch) -> Bool

Get RenewableDispatch available.

PowerSystems.get_base_power — Method

get_base_power(value::RenewableDispatch) -> Float64

Get RenewableDispatch base_power.

PowerSystems.get_bus — Method

get_bus(value::RenewableDispatch) -> ACBus

Get RenewableDispatch bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(
    value::RenewableDispatch
) -> Union{Nothing, DynamicInjection}

Get RenewableDispatch dynamic_injector.

PowerSystems.get_ext — Method

get_ext(value::RenewableDispatch) -> Dict{String, Any}

Get RenewableDispatch ext.

PowerSystems.get_operation_cost — Method

get_operation_cost(
    value::RenewableDispatch
) -> Union{MarketBidCost, RenewableGenerationCost}

Get RenewableDispatch operation_cost.

PowerSystems.get_power_factor — Method

get_power_factor(value::RenewableDispatch) -> Float64

Get RenewableDispatch power_factor.

PowerSystems.get_prime_mover_type — Method

get_prime_mover_type(
    value::RenewableDispatch
) -> PrimeMovers

Get RenewableDispatch prime_mover_type.

PowerSystems.get_rating — Method

get_rating(value::RenewableDispatch) -> Any

Get RenewableDispatch rating.

PowerSystems.get_reactive_power — Method

get_reactive_power(value::RenewableDispatch) -> Any

Get RenewableDispatch reactive_power.

PowerSystems.get_reactive_power_limits — Method

get_reactive_power_limits(
    value::RenewableDispatch
) -> Union{Nothing, NamedTuple{(:min, :max), <:Tuple{Any, Any}}}

Get RenewableDispatch reactive_power_limits.

PowerSystems.get_services — Method

get_services(value::RenewableDispatch) -> Vector{Service}

Get RenewableDispatch services.

PowerSystems.set_active_power! — Method

set_active_power!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch active_power.

PowerSystems.set_available! — Method

set_available!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch available.

PowerSystems.set_base_power! — Method

set_base_power!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch base_power.

PowerSystems.set_bus! — Method

set_bus!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch bus.

PowerSystems.set_ext! — Method

set_ext!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch ext.

PowerSystems.set_operation_cost! — Method

set_operation_cost!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch operation_cost.

PowerSystems.set_power_factor! — Method

set_power_factor!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch power_factor.

PowerSystems.set_prime_mover_type! — Method

set_prime_mover_type!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch prime_mover_type.

PowerSystems.set_rating! — Method

set_rating!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch rating.

PowerSystems.set_reactive_power! — Method

set_reactive_power!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch reactive_power.

PowerSystems.set_reactive_power_limits! — Method

set_reactive_power_limits!(
    value::RenewableDispatch,
    val
) -> Any

Set RenewableDispatch reactive_power_limits.

PowerSystems.set_services! — Method

set_services!(value::RenewableDispatch, val) -> Any

Set RenewableDispatch services.