RenewableNonDispatch

PowerSystems.RenewableNonDispatch — Type

mutable struct RenewableNonDispatch <: RenewableGen
    name::String
    available::Bool
    bus::ACBus
    active_power::Float64
    reactive_power::Float64
    rating::Float64
    prime_mover_type::PrimeMovers
    power_factor::Float64
    base_power::Float64
    services::Vector{Service}
    dynamic_injector::Union{Nothing, DynamicInjection}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A non-dispatchable (i.e., non-curtailable or must-take) renewable generator.

Its output is equal to its max_active_power time series by default. Example use: an aggregation of behind-the-meter distributed energy resources like rooftop solar. For curtailable or downward dispatachable generation, see RenewableDispatch.

Renewable generators do not have a max_active_power parameter, which is instead calculated when calling get_max_active_power()

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

bus::ACBus: Bus that this component is connected to

active_power::Float64: Initial active power set point of the unit in MW. For power flow, this is the steady state operating point of the system. For production cost modeling, this may or may not be used as the initial starting point for the solver, depending on the solver used

reactive_power::Float64: Initial reactive power set point of the unit (MVAR), used in some production cost modeling simulations. To set the reactive power in a load flow, use power_factor

rating::Float64: Maximum AC side output power rating of the unit. Stored in per unit of the device and not to be confused with base_power, validation range: (0, nothing)

prime_mover_type::PrimeMovers: Prime mover technology according to EIA 923. Options are listed here

power_factor::Float64: Power factor [0, 1] set-point, used in some production cost modeling and in load flow if the unit is connected to a PQ bus, validation range: (0, 1)

base_power::Float64: Base power of the unit (MVA) for per unitization, validation range: (0.0001, nothing)

services::Vector{Service}: (default: Device[]) Services that this device contributes to

dynamic_injector::Union{Nothing, DynamicInjection}: (default: nothing) corresponding dynamic injection device

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::RenewableNonDispatch) -> Bool

Get RenewableNonDispatch available.

InfrastructureSystems.get_name — Method

get_name(value::RenewableNonDispatch) -> String

Get RenewableNonDispatch name.

InfrastructureSystems.set_available! — Method

set_available!(value::RenewableNonDispatch, val) -> Any

Set RenewableNonDispatch available.

PowerSystems.get_active_power — Method

get_active_power(value::RenewableNonDispatch) -> Float64

Get RenewableNonDispatch active_power.

PowerSystems.get_base_power — Method

get_base_power(value::RenewableNonDispatch) -> Float64

Get RenewableNonDispatch base_power.

PowerSystems.get_bus — Method

get_bus(value::RenewableNonDispatch) -> ACBus

Get RenewableNonDispatch bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(
    value::RenewableNonDispatch
) -> Union{Nothing, DynamicInjection}

Get RenewableNonDispatch dynamic_injector.

PowerSystems.get_ext — Method

get_ext(value::RenewableNonDispatch) -> Dict{String, Any}

Get RenewableNonDispatch ext.

PowerSystems.get_power_factor — Method

get_power_factor(value::RenewableNonDispatch) -> Float64

Get RenewableNonDispatch power_factor.

PowerSystems.get_prime_mover_type — Method

get_prime_mover_type(
    value::RenewableNonDispatch
) -> PrimeMovers

Get RenewableNonDispatch prime_mover_type.

PowerSystems.get_rating — Method

get_rating(value::RenewableNonDispatch) -> Float64

Get RenewableNonDispatch rating.

PowerSystems.get_reactive_power — Method

get_reactive_power(value::RenewableNonDispatch) -> Float64

Get RenewableNonDispatch reactive_power.

PowerSystems.get_services — Method

get_services(value::RenewableNonDispatch) -> Vector{Service}

Get RenewableNonDispatch services.

PowerSystems.set_active_power! — Method

set_active_power!(value::RenewableNonDispatch, val) -> Any

Set RenewableNonDispatch active_power.

PowerSystems.set_base_power! — Method

set_base_power!(value::RenewableNonDispatch, val) -> Any

Set RenewableNonDispatch base_power.

PowerSystems.set_bus! — Method

set_bus!(value::RenewableNonDispatch, val) -> Any

Set RenewableNonDispatch bus.

PowerSystems.set_ext! — Method

set_ext!(value::RenewableNonDispatch, val) -> Any

Set RenewableNonDispatch ext.

PowerSystems.set_power_factor! — Method

set_power_factor!(value::RenewableNonDispatch, val) -> Any

Set RenewableNonDispatch power_factor.

PowerSystems.set_prime_mover_type! — Method

set_prime_mover_type!(
    value::RenewableNonDispatch,
    val
) -> Any

Set RenewableNonDispatch prime_mover_type.

PowerSystems.set_rating! — Method

set_rating!(value::RenewableNonDispatch, val) -> Any

Set RenewableNonDispatch rating.

PowerSystems.set_reactive_power! — Method

set_reactive_power!(value::RenewableNonDispatch, val) -> Any

Set RenewableNonDispatch reactive_power.

PowerSystems.set_services! — Method

set_services!(value::RenewableNonDispatch, val) -> Any

Set RenewableNonDispatch services.