SimplifiedSingleCageInductionMachine

PowerSystems.SimplifiedSingleCageInductionMachine — Type

mutable struct SimplifiedSingleCageInductionMachine <: DynamicInjection
    name::String
    R_s::Float64
    R_r::Float64
    X_ls::Float64
    X_lr::Float64
    X_m::Float64
    H::Float64
    A::Float64
    B::Float64
    base_power::Float64
    ext::Dict{String, Any}
    C::Float64
    τ_ref::Float64
    B_shunt::Float64
    X_ss::Float64
    X_rr::Float64
    X_p::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 3-states three-phase single cage induction machine with quadratic torque-speed relationship

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

R_s::Float64: Armature stator resistance, validation range: (0, nothing)

R_r::Float64: Rotor resistance, validation range: (0, nothing)

X_ls::Float64: Stator Leakage Reactance, validation range: (0, nothing)

X_lr::Float64: Rotor Leakage Reactance, validation range: (0, nothing)

X_m::Float64: Stator-Rotor Mutual Reactance, validation range: (0, nothing)

H::Float64: Motor Inertia Constant [s], validation range: (0, nothing)

A::Float64: Torque-Speed Quadratic Term, validation range: (0, 1)

B::Float64: Torque-Speed Linear Term, validation range: (0, 1)

base_power::Float64: Base power of the unit (MVA) for per unitization, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

C::Float64: (Do not modify.) Torque-Speed Constant Term

τ_ref::Float64: Reference torque parameter

B_shunt::Float64: Susceptance Initialization Corrector Term

X_ss::Float64: (Do not modify.) Stator self reactance

X_rr::Float64: (Do not modify.) Rotor self reactance

X_p::Float64: (Do not modify.) Transient reactance

states::Vector{Symbol}: (Do not modify.) The states are:

ψ_qr: rotor flux in the q-axis,
ψ_dr: rotor flux in the d-axis, 
ωr: Rotor speed [pu],

n_states::Int: (Do not modify.) SimplifiedSingleCageInductionMachine has 3 states

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(
    value::SimplifiedSingleCageInductionMachine
) -> String

Get SimplifiedSingleCageInductionMachine name.

PowerSystems.get_A — Method

get_A(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine A.

PowerSystems.get_B — Method

get_B(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine B.

PowerSystems.get_B_shunt — Method

get_B_shunt(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine B_shunt.

PowerSystems.get_C — Method

get_C(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine C.

PowerSystems.get_H — Method

get_H(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine H.

PowerSystems.get_R_r — Method

get_R_r(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine R_r.

PowerSystems.get_R_s — Method

get_R_s(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine R_s.

PowerSystems.get_X_lr — Method

get_X_lr(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine X_lr.

PowerSystems.get_X_ls — Method

get_X_ls(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine X_ls.

PowerSystems.get_X_m — Method

get_X_m(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine X_m.

PowerSystems.get_X_p — Method

get_X_p(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine X_p.

PowerSystems.get_X_rr — Method

get_X_rr(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine X_rr.

PowerSystems.get_X_ss — Method

get_X_ss(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine X_ss.

PowerSystems.get_base_power — Method

get_base_power(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine base_power.

PowerSystems.get_ext — Method

get_ext(
    value::SimplifiedSingleCageInductionMachine
) -> Dict{String, Any}

Get SimplifiedSingleCageInductionMachine ext.

PowerSystems.get_n_states — Method

get_n_states(
    value::SimplifiedSingleCageInductionMachine
) -> Int64

Get SimplifiedSingleCageInductionMachine n_states.

PowerSystems.get_states — Method

get_states(
    value::SimplifiedSingleCageInductionMachine
) -> Vector{Symbol}

Get SimplifiedSingleCageInductionMachine states.

PowerSystems.get_τ_ref — Method

get_τ_ref(
    value::SimplifiedSingleCageInductionMachine
) -> Float64

Get SimplifiedSingleCageInductionMachine τ_ref.

PowerSystems.set_A! — Method

set_A!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine A.

PowerSystems.set_B! — Method

set_B!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine B.

PowerSystems.set_B_shunt! — Method

set_B_shunt!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine B_shunt.

PowerSystems.set_C! — Method

set_C!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine C.

PowerSystems.set_H! — Method

set_H!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine H.

PowerSystems.set_R_r! — Method

set_R_r!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine R_r.

PowerSystems.set_R_s! — Method

set_R_s!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine R_s.

PowerSystems.set_X_lr! — Method

set_X_lr!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine X_lr.

PowerSystems.set_X_ls! — Method

set_X_ls!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine X_ls.

PowerSystems.set_X_m! — Method

set_X_m!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine X_m.

PowerSystems.set_X_p! — Method

set_X_p!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine X_p.

PowerSystems.set_X_rr! — Method

set_X_rr!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine X_rr.

PowerSystems.set_X_ss! — Method

set_X_ss!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine X_ss.

PowerSystems.set_base_power! — Method

set_base_power!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine base_power.

PowerSystems.set_ext! — Method

set_ext!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine ext.

PowerSystems.set_τ_ref! — Method

set_τ_ref!(
    value::SimplifiedSingleCageInductionMachine,
    val
) -> Any

Set SimplifiedSingleCageInductionMachine τ_ref.