SwitchedAdmittance

PowerSystems.SwitchedAdmittance — Type

mutable struct SwitchedAdmittance <: ElectricLoad
    name::String
    available::Bool
    bus::ACBus
    Y::Complex{Float64}
    number_of_steps::Int
    Y_increase::Complex{Float64}
    dynamic_injector::Union{Nothing, DynamicInjection}
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A switched admittance, with discrete steps to adjust the admittance.

Most often used in power flow studies, iterating over the steps to see impacts of admittance on the results. Total admittance is calculated as: Y + number_of_steps * Y_increase

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

bus::ACBus: Bus that this component is connected to

Y::Complex{Float64}: Initial admittance at N = 0

number_of_steps::Int: (default: 0) Number of steps for adjustable shunt

Y_increase::Complex{Float64}: (default: 0) Admittance increment for each of step increase

dynamic_injector::Union{Nothing, DynamicInjection}: (default: nothing) corresponding dynamic injection model for admittance

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude. See Adding additional fields

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::SwitchedAdmittance) -> String

Get SwitchedAdmittance name.

PowerSystems.get_Y — Method

get_Y(value::SwitchedAdmittance) -> ComplexF64

Get SwitchedAdmittance Y.

PowerSystems.get_Y_increase — Method

get_Y_increase(value::SwitchedAdmittance) -> ComplexF64

Get SwitchedAdmittance Y_increase.

PowerSystems.get_available — Method

get_available(value::SwitchedAdmittance) -> Bool

Get SwitchedAdmittance available.

PowerSystems.get_bus — Method

get_bus(value::SwitchedAdmittance) -> ACBus

Get SwitchedAdmittance bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(
    value::SwitchedAdmittance
) -> Union{Nothing, DynamicInjection}

Get SwitchedAdmittance dynamic_injector.

PowerSystems.get_ext — Method

get_ext(value::SwitchedAdmittance) -> Dict{String, Any}

Get SwitchedAdmittance ext.

PowerSystems.get_number_of_steps — Method

get_number_of_steps(value::SwitchedAdmittance) -> Int64

Get SwitchedAdmittance number_of_steps.

PowerSystems.get_services — Method

get_services(value::SwitchedAdmittance) -> Vector{Service}

Get SwitchedAdmittance services.

PowerSystems.set_Y! — Method

set_Y!(value::SwitchedAdmittance, val) -> Any

Set SwitchedAdmittance Y.

PowerSystems.set_Y_increase! — Method

set_Y_increase!(value::SwitchedAdmittance, val) -> Any

Set SwitchedAdmittance Y_increase.

PowerSystems.set_available! — Method

set_available!(value::SwitchedAdmittance, val) -> Any

Set SwitchedAdmittance available.

PowerSystems.set_bus! — Method

set_bus!(value::SwitchedAdmittance, val) -> Any

Set SwitchedAdmittance bus.

PowerSystems.set_ext! — Method

set_ext!(value::SwitchedAdmittance, val) -> Any

Set SwitchedAdmittance ext.

PowerSystems.set_number_of_steps! — Method

set_number_of_steps!(value::SwitchedAdmittance, val) -> Any

Set SwitchedAdmittance number_of_steps.

PowerSystems.set_services! — Method

set_services!(value::SwitchedAdmittance, val) -> Any

Set SwitchedAdmittance services.