SynchronousCondenser

PowerSystems.SynchronousCondenser — Type

mutable struct SynchronousCondenser <: StaticInjection
    name::String
    available::Bool
    bus::ACBus
    reactive_power::Float64
    rating::Float64
    reactive_power_limits::Union{Nothing, MinMax}
    base_power::Float64
    active_power_losses::Float64
    services::Vector{Service}
    dynamic_injector::Union{Nothing, DynamicInjection}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A Synchronous Machine connected to the system to provide inertia or reactive power support

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

bus::ACBus: Bus that this component is connected to

reactive_power::Float64: Initial reactive power set point of the unit (MVAR), validation range: reactive_power_limits

rating::Float64: Maximum AC side output power rating of the unit. Stored in per unit of the device and not to be confused with base_power, validation range: (0, nothing)

reactive_power_limits::Union{Nothing, MinMax}: Minimum and maximum reactive power limits. Set to Nothing if not applicable

base_power::Float64: Base power of the unit (MVA) for per unitization, validation range: (0.0001, nothing)

active_power_losses::Float64: (default: 0.0) Active Power Loss incurred by having the unit online., validation range: (0, nothing)

services::Vector{Service}: (default: Device[]) Services that this device contributes to

dynamic_injector::Union{Nothing, DynamicInjection}: (default: nothing) corresponding dynamic injection device

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::SynchronousCondenser) -> Bool

Get SynchronousCondenser available.

InfrastructureSystems.get_name — Method

get_name(value::SynchronousCondenser) -> String

Get SynchronousCondenser name.

InfrastructureSystems.set_available! — Method

set_available!(value::SynchronousCondenser, val) -> Any

Set SynchronousCondenser available.

PowerSystems.get_active_power_losses — Method

get_active_power_losses(
    value::SynchronousCondenser
) -> Float64

Get SynchronousCondenser active_power_losses.

PowerSystems.get_base_power — Method

get_base_power(value::SynchronousCondenser) -> Float64

Get SynchronousCondenser base_power.

PowerSystems.get_bus — Method

get_bus(value::SynchronousCondenser) -> ACBus

Get SynchronousCondenser bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(
    value::SynchronousCondenser
) -> Union{Nothing, DynamicInjection}

Get SynchronousCondenser dynamic_injector.

PowerSystems.get_ext — Method

get_ext(value::SynchronousCondenser) -> Dict{String, Any}

Get SynchronousCondenser ext.

PowerSystems.get_rating — Method

get_rating(value::SynchronousCondenser) -> Float64

Get SynchronousCondenser rating.

PowerSystems.get_reactive_power — Method

get_reactive_power(value::SynchronousCondenser) -> Float64

Get SynchronousCondenser reactive_power.

PowerSystems.get_reactive_power_limits — Method

get_reactive_power_limits(
    value::SynchronousCondenser
) -> Union{Nothing, @NamedTuple{min::Float64, max::Float64}}

Get SynchronousCondenser reactive_power_limits.

PowerSystems.get_services — Method

get_services(value::SynchronousCondenser) -> Vector{Service}

Get SynchronousCondenser services.

PowerSystems.set_active_power_losses! — Method

set_active_power_losses!(
    value::SynchronousCondenser,
    val
) -> Any

Set SynchronousCondenser active_power_losses.

PowerSystems.set_base_power! — Method

set_base_power!(value::SynchronousCondenser, val) -> Any

Set SynchronousCondenser base_power.

PowerSystems.set_bus! — Method

set_bus!(value::SynchronousCondenser, val) -> Any

Set SynchronousCondenser bus.

PowerSystems.set_ext! — Method

set_ext!(value::SynchronousCondenser, val) -> Any

Set SynchronousCondenser ext.

PowerSystems.set_rating! — Method

set_rating!(value::SynchronousCondenser, val) -> Any

Set SynchronousCondenser rating.

PowerSystems.set_reactive_power! — Method

set_reactive_power!(value::SynchronousCondenser, val) -> Any

Set SynchronousCondenser reactive_power.

PowerSystems.set_reactive_power_limits! — Method

set_reactive_power_limits!(
    value::SynchronousCondenser,
    val
) -> Any

Set SynchronousCondenser reactive_power_limits.

PowerSystems.set_services! — Method

set_services!(value::SynchronousCondenser, val) -> Any

Set SynchronousCondenser services.