TapTransformer

PowerSystems.TapTransformer — Type

mutable struct TapTransformer <: TwoWindingTransformer
    name::String
    available::Bool
    active_power_flow::Float64
    reactive_power_flow::Float64
    arc::Arc
    r::Float64
    x::Float64
    primary_shunt::Complex{Float64}
    tap::Float64
    rating::Union{Nothing, Float64}
    base_power::Float64
    base_voltage_primary::Union{Nothing, Float64}
    base_voltage_secondary::Union{Nothing, Float64}
    rating_b::Union{Nothing, Float64}
    rating_c::Union{Nothing, Float64}
    winding_group_number::WindingGroupNumber
    control_objective::TransformerControlObjective
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A 2-winding transformer, with a tap changer for variable turns ratio.

The model uses an equivalent circuit assuming the impedance is on the High Voltage Side of the transformer. The model allocates the iron losses and magnetizing susceptance to the primary side

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow::Float64: Initial condition of active power flow through the transformer (MW)

reactive_power_flow::Float64: Initial condition of reactive power flow through the transformer (MVAR)

arc::Arc: An Arc defining this transformer from a bus to another bus

r::Float64: Resistance in p.u. (SYSTEM_BASE), validation range: (-2, 2)

x::Float64: Reactance in p.u. (SYSTEM_BASE), validation range: (-2, 4)

primary_shunt::Complex{Float64}: Primary shunt admittance in pu (SYSTEM_BASE)

tap::Float64: Normalized tap changer position for voltage control, varying between 0 and 2, with 1 centered at the nominal voltage, validation range: (0, 2)

rating::Union{Nothing, Float64}: Thermal rating (MVA). Flow through the transformer must be between -rating. When defining a transformer before it is attached to a System, rating must be in pu (SYSTEM_BASE) using the base power of the System it will be attached to, validation range: (0, nothing)

base_power::Float64: Base power (MVA) for per unitization, validation range: (0, nothing)

base_voltage_primary::Union{Nothing, Float64}: (default: get_base_voltage(get_from(arc))) Primary base voltage in kV, validation range: (0, nothing)

base_voltage_secondary::Union{Nothing, Float64}: (default: get_base_voltage(get_to(arc))) Secondary base voltage in kV, validation range: (0, nothing)

rating_b::Union{Nothing, Float64}: (default: nothing) Second current rating; entered in MVA.

rating_c::Union{Nothing, Float64}: (default: nothing) Third current rating; entered in MVA.

winding_group_number::WindingGroupNumber: (default: WindingGroupNumber.UNDEFINED) Vector group number ('clock number') indicating fixed phase shift (radians) between the from and to buses due to the connection group configuration

control_objective::TransformerControlObjective: (default: TransformerControlObjective.UNDEFINED) Control objective for the tap changer for power flow calculations. See TransformerControlObjective

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::TapTransformer) -> Bool

Get TapTransformer available.

InfrastructureSystems.get_name — Method

get_name(value::TapTransformer) -> String

Get TapTransformer name.

InfrastructureSystems.set_available! — Method

set_available!(value::TapTransformer, val) -> Any

Set TapTransformer available.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::TapTransformer) -> Float64

Get TapTransformer active_power_flow.

PowerSystems.get_arc — Method

get_arc(value::TapTransformer) -> Arc

Get TapTransformer arc.

PowerSystems.get_base_power — Method

get_base_power(value::TapTransformer) -> Float64

Get TapTransformer base_power.

PowerSystems.get_base_voltage_primary — Method

get_base_voltage_primary(
    value::TapTransformer
) -> Union{Nothing, Float64}

Get TapTransformer base_voltage_primary.

PowerSystems.get_base_voltage_secondary — Method

get_base_voltage_secondary(
    value::TapTransformer
) -> Union{Nothing, Float64}

Get TapTransformer base_voltage_secondary.

PowerSystems.get_control_objective — Method

get_control_objective(
    value::TapTransformer
) -> PowerSystems.TransformerControlObjectiveModule.TransformerControlObjective

Get TapTransformer control_objective.

PowerSystems.get_ext — Method

get_ext(value::TapTransformer) -> Dict{String, Any}

Get TapTransformer ext.

PowerSystems.get_primary_shunt — Method

get_primary_shunt(value::TapTransformer) -> ComplexF64

Get TapTransformer primary_shunt.

PowerSystems.get_r — Method

get_r(value::TapTransformer) -> Float64

Get TapTransformer r.

PowerSystems.get_rating — Method

get_rating(value::TapTransformer) -> Union{Nothing, Float64}

Get TapTransformer rating.

PowerSystems.get_rating_b — Method

get_rating_b(
    value::TapTransformer
) -> Union{Nothing, Float64}

Get TapTransformer rating_b.

PowerSystems.get_rating_c — Method

get_rating_c(
    value::TapTransformer
) -> Union{Nothing, Float64}

Get TapTransformer rating_c.

PowerSystems.get_reactive_power_flow — Method

get_reactive_power_flow(value::TapTransformer) -> Float64

Get TapTransformer reactive_power_flow.

PowerSystems.get_services — Method

get_services(value::TapTransformer) -> Vector{Service}

Get TapTransformer services.

PowerSystems.get_tap — Method

get_tap(value::TapTransformer) -> Float64

Get TapTransformer tap.

PowerSystems.get_winding_group_number — Method

get_winding_group_number(
    value::TapTransformer
) -> WindingGroupNumber

Get TapTransformer winding_group_number.

PowerSystems.get_x — Method

get_x(value::TapTransformer) -> Float64

Get TapTransformer x.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(value::TapTransformer, val) -> Any

Set TapTransformer active_power_flow.

PowerSystems.set_arc! — Method

set_arc!(value::TapTransformer, val) -> Any

Set TapTransformer arc.

PowerSystems.set_base_power! — Method

set_base_power!(value::TapTransformer, val) -> Any

Set TapTransformer base_power.

PowerSystems.set_base_voltage_primary! — Method

set_base_voltage_primary!(value::TapTransformer, val) -> Any

Set TapTransformer base_voltage_primary.

PowerSystems.set_base_voltage_secondary! — Method

set_base_voltage_secondary!(
    value::TapTransformer,
    val
) -> Any

Set TapTransformer base_voltage_secondary.

PowerSystems.set_control_objective! — Method

set_control_objective!(value::TapTransformer, val) -> Any

Set TapTransformer control_objective.

PowerSystems.set_ext! — Method

set_ext!(value::TapTransformer, val) -> Any

Set TapTransformer ext.

PowerSystems.set_primary_shunt! — Method

set_primary_shunt!(value::TapTransformer, val) -> Any

Set TapTransformer primary_shunt.

PowerSystems.set_r! — Method

set_r!(value::TapTransformer, val) -> Any

Set TapTransformer r.

PowerSystems.set_rating! — Method

set_rating!(value::TapTransformer, val) -> Any

Set TapTransformer rating.

PowerSystems.set_rating_b! — Method

set_rating_b!(value::TapTransformer, val) -> Any

Set TapTransformer rating_b.

PowerSystems.set_rating_c! — Method

set_rating_c!(value::TapTransformer, val) -> Any

Set TapTransformer rating_c.

PowerSystems.set_reactive_power_flow! — Method

set_reactive_power_flow!(value::TapTransformer, val) -> Any

Set TapTransformer reactive_power_flow.

PowerSystems.set_services! — Method

set_services!(value::TapTransformer, val) -> Any

Set TapTransformer services.

PowerSystems.set_tap! — Method

set_tap!(value::TapTransformer, val) -> Any

Set TapTransformer tap.

PowerSystems.set_winding_group_number! — Method

set_winding_group_number!(value::TapTransformer, val) -> Any

Set TapTransformer winding_group_number.

PowerSystems.set_x! — Method

set_x!(value::TapTransformer, val) -> Any

Set TapTransformer x.