ThermalMultiStart

PowerSystems.ThermalMultiStart — Type

mutable struct ThermalMultiStart <: ThermalGen
    name::String
    available::Bool
    status::Bool
    bus::Bus
    active_power::Float64
    reactive_power::Float64
    rating::Float64
    prime_mover::PrimeMovers.PrimeMover
    fuel::ThermalFuels.ThermalFuel
    active_power_limits::NamedTuple{(:min, :max), Tuple{Float64, Float64}}
    reactive_power_limits::Union{Nothing, Min_Max}
    ramp_limits::Union{Nothing, NamedTuple{(:up, :down), Tuple{Float64, Float64}}}
    power_trajectory::Union{Nothing, NamedTuple{(:startup, :shutdown), Tuple{Float64, Float64}}}
    time_limits::Union{Nothing, NamedTuple{(:up, :down), Tuple{Float64, Float64}}}
    start_time_limits::Union{Nothing, NamedTuple{(:hot, :warm, :cold), Tuple{Float64, Float64, Float64}}}
    start_types::Int
    operation_cost::OperationalCost
    base_power::Float64
    services::Vector{Service}
    time_at_status::Float64
    must_run::Bool
    dynamic_injector::Union{Nothing, DynamicInjection}
    ext::Dict{String, Any}
    time_series_container::InfrastructureSystems.TimeSeriesContainer
    internal::InfrastructureSystemsInternal
end

Data Structure for thermal generation technologies.

Arguments

name::String

available::Bool

status::Bool

bus::Bus

active_power::Float64, validation range: active_power_limits, action if invalid: warn

reactive_power::Float64, validation range: reactive_power_limits, action if invalid: warn

rating::Float64: Thermal limited MVA Power Output of the unit. <= Capacity, validation range: (0, nothing), action if invalid: error

prime_mover::PrimeMovers.PrimeMover: Prime mover technology according to EIA 923

fuel::ThermalFuels.ThermalFuel: Prime mover fuel according to EIA 923

active_power_limits::NamedTuple{(:min, :max), Tuple{Float64, Float64}}

reactive_power_limits::Union{Nothing, Min_Max}

ramp_limits::Union{Nothing, NamedTuple{(:up, :down), Tuple{Float64, Float64}}}, validation range: (0, nothing), action if invalid: error

power_trajectory::Union{Nothing, NamedTuple{(:startup, :shutdown), Tuple{Float64, Float64}}}: Power trajectory the unit will take during the start-up and shut-down ramp process, validation range: (0, nothing), action if invalid: error

time_limits::Union{Nothing, NamedTuple{(:up, :down), Tuple{Float64, Float64}}}: Minimum up and Minimum down time limits in hours, validation range: (0, nothing), action if invalid: error

start_time_limits::Union{Nothing, NamedTuple{(:hot, :warm, :cold), Tuple{Float64, Float64, Float64}}}: Time limits for start-up based on turbine temperature in hours

start_types::Int: Number of start-up based on turbine temperature, validation range: (1, 3), action if invalid: error

operation_cost::OperationalCost

base_power::Float64: Base power of the unit in MVA, validation range: (0, nothing), action if invalid: warn

services::Vector{Service}: Services that this device contributes to

time_at_status::Float64

must_run::Bool

dynamic_injector::Union{Nothing, DynamicInjection}: corresponding dynamic injection device

ext::Dict{String, Any}

time_series_container::InfrastructureSystems.TimeSeriesContainer: internal time_series storage

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_active_power — Method

get_active_power(value::ThermalMultiStart) -> Any

Get ThermalMultiStart active_power.

PowerSystems.get_active_power_limits — Method

get_active_power_limits(value::ThermalMultiStart) -> NamedTuple{(:min, :max),_A} where _A<:Tuple

Get ThermalMultiStart active_power_limits.

PowerSystems.get_available — Method

get_available(value::ThermalMultiStart) -> Bool

Get ThermalMultiStart available.

PowerSystems.get_base_power — Method

get_base_power(value::ThermalMultiStart) -> Float64

Get ThermalMultiStart base_power.

PowerSystems.get_bus — Method

get_bus(value::ThermalMultiStart) -> Bus

Get ThermalMultiStart bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(value::ThermalMultiStart) -> Union{Nothing, DynamicInjection}

Get ThermalMultiStart dynamic_injector.

PowerSystems.get_ext — Method

get_ext(value::ThermalMultiStart) -> Dict{String,Any}

Get ThermalMultiStart ext.

PowerSystems.get_fuel — Method

get_fuel(value::ThermalMultiStart) -> PowerSystems.ThermalFuels.ThermalFuel

Get ThermalMultiStart fuel.

PowerSystems.get_must_run — Method

get_must_run(value::ThermalMultiStart) -> Bool

Get ThermalMultiStart must_run.

PowerSystems.get_operation_cost — Method

get_operation_cost(value::ThermalMultiStart) -> PowerSystems.OperationalCost

Get ThermalMultiStart operation_cost.

PowerSystems.get_power_trajectory — Method

get_power_trajectory(value::ThermalMultiStart) -> Union{Nothing, NamedTuple{(:startup, :shutdown),Tuple{Float64,Float64}}}

Get ThermalMultiStart power_trajectory.

PowerSystems.get_prime_mover — Method

get_prime_mover(value::ThermalMultiStart) -> PowerSystems.PrimeMovers.PrimeMover

Get ThermalMultiStart prime_mover.

PowerSystems.get_ramp_limits — Method

get_ramp_limits(value::ThermalMultiStart) -> Union{Nothing, NamedTuple{(:up, :down),_A} where _A<:Tuple}

Get ThermalMultiStart ramp_limits.

PowerSystems.get_rating — Method

get_rating(value::ThermalMultiStart) -> Any

Get ThermalMultiStart rating.

PowerSystems.get_reactive_power — Method

get_reactive_power(value::ThermalMultiStart) -> Any

Get ThermalMultiStart reactive_power.

PowerSystems.get_reactive_power_limits — Method

get_reactive_power_limits(value::ThermalMultiStart) -> Union{Nothing, NamedTuple{(:min, :max),_A} where _A<:Tuple}

Get ThermalMultiStart reactive_power_limits.

PowerSystems.get_services — Method

get_services(value::ThermalMultiStart) -> Array{Service,1}

Get ThermalMultiStart services.

PowerSystems.get_start_time_limits — Method

get_start_time_limits(value::ThermalMultiStart) -> Union{Nothing, NamedTuple{(:hot, :warm, :cold),Tuple{Float64,Float64,Float64}}}

Get ThermalMultiStart start_time_limits.

PowerSystems.get_start_types — Method

get_start_types(value::ThermalMultiStart) -> Int64

Get ThermalMultiStart start_types.

PowerSystems.get_status — Method

get_status(value::ThermalMultiStart) -> Bool

Get ThermalMultiStart status.

PowerSystems.get_time_at_status — Method

get_time_at_status(value::ThermalMultiStart) -> Float64

Get ThermalMultiStart time_at_status.

PowerSystems.get_time_limits — Method

get_time_limits(value::ThermalMultiStart) -> Union{Nothing, NamedTuple{(:up, :down),Tuple{Float64,Float64}}}

Get ThermalMultiStart time_limits.

PowerSystems.set_active_power! — Method

set_active_power!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart active_power.

PowerSystems.set_active_power_limits! — Method

set_active_power_limits!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart active_power_limits.

PowerSystems.set_available! — Method

set_available!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart available.

PowerSystems.set_base_power! — Method

set_base_power!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart base_power.

PowerSystems.set_bus! — Method

set_bus!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart bus.

PowerSystems.set_ext! — Method

set_ext!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart ext.

PowerSystems.set_fuel! — Method

set_fuel!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart fuel.

PowerSystems.set_must_run! — Method

set_must_run!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart must_run.

PowerSystems.set_operation_cost! — Method

set_operation_cost!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart operation_cost.

PowerSystems.set_power_trajectory! — Method

set_power_trajectory!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart power_trajectory.

PowerSystems.set_prime_mover! — Method

set_prime_mover!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart prime_mover.

PowerSystems.set_ramp_limits! — Method

set_ramp_limits!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart ramp_limits.

PowerSystems.set_rating! — Method

set_rating!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart rating.

PowerSystems.set_reactive_power! — Method

set_reactive_power!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart reactive_power.

PowerSystems.set_reactive_power_limits! — Method

set_reactive_power_limits!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart reactive_power_limits.

PowerSystems.set_services! — Method

set_services!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart services.

PowerSystems.set_start_time_limits! — Method

set_start_time_limits!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart start_time_limits.

PowerSystems.set_start_types! — Method

set_start_types!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart start_types.

PowerSystems.set_status! — Method

set_status!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart status.

PowerSystems.set_time_at_status! — Method

set_time_at_status!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart time_at_status.

PowerSystems.set_time_limits! — Method

set_time_limits!(value::ThermalMultiStart, val::Any) -> Any

Set ThermalMultiStart time_limits.