Transformer3W

PowerSystems.Transformer3W — Type

mutable struct Transformer3W <: ThreeWindingTransformer
    name::String
    available::Bool
    primary_star_arc::Arc
    secondary_star_arc::Arc
    tertiary_star_arc::Arc
    star_bus::ACBus
    active_power_flow_primary::Float64
    reactive_power_flow_primary::Float64
    active_power_flow_secondary::Float64
    reactive_power_flow_secondary::Float64
    active_power_flow_tertiary::Float64
    reactive_power_flow_tertiary::Float64
    r_primary::Float64
    x_primary::Float64
    r_secondary::Float64
    x_secondary::Float64
    r_tertiary::Float64
    x_tertiary::Float64
    rating::Union{Nothing, Float64}
    r_12::Float64
    x_12::Float64
    r_23::Float64
    x_23::Float64
    r_13::Float64
    x_13::Float64
    base_power_12::Float64
    base_power_23::Float64
    base_power_13::Float64
    base_voltage_primary::Union{Nothing, Float64}
    base_voltage_secondary::Union{Nothing, Float64}
    base_voltage_tertiary::Union{Nothing, Float64}
    g::Float64
    b::Float64
    primary_turns_ratio::Float64
    secondary_turns_ratio::Float64
    tertiary_turns_ratio::Float64
    available_primary::Bool
    available_secondary::Bool
    available_tertiary::Bool
    rating_primary::Float64
    rating_secondary::Float64
    rating_tertiary::Float64
    primary_group_number::WindingGroupNumber
    secondary_group_number::WindingGroupNumber
    tertiary_group_number::WindingGroupNumber
    control_objective_primary::TransformerControlObjective
    control_objective_secondary::TransformerControlObjective
    control_objective_tertiary::TransformerControlObjective
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A 3-winding transformer.

The model uses an equivalent star model with a star (hidden) bus. The user must transform the data to use CW = CZ = CM = 1 and COD1 = COD2 = COD3 = 0 (no voltage control) if taken from a PSS/E 3W transformer model. Three equivalent impedances (connecting each side to the star bus) are required to define the model. Shunt conductance (iron losses) and magnetizing susceptance can be considered from the star bus to ground. The model is described in Chapter 3.6 in J.D. Glover, M.S. Sarma and T. Overbye: Power Systems Analysis and Design.

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

primary_star_arc::Arc: An Arc defining this transformer from a primary bus to the star bus

secondary_star_arc::Arc: An Arc defining this transformer from a secondary bus to the star bus

tertiary_star_arc::Arc: An Arc defining this transformer from a tertiary bus to the star bus

star_bus::ACBus: Star (hidden) Bus that this component (equivalent model) is connected to

active_power_flow_primary::Float64: Initial condition of active power flow through the transformer primary side to star (hidden) bus (MW)

reactive_power_flow_primary::Float64: Initial condition of reactive power flow through the transformer primary side to star (hidden) bus (MW)

active_power_flow_secondary::Float64: Initial condition of active power flow through the transformer secondary side to star (hidden) bus (MW)

reactive_power_flow_secondary::Float64: Initial condition of reactive power flow through the transformer secondary side to star (hidden) bus (MW)

active_power_flow_tertiary::Float64: Initial condition of active power flow through the transformer tertiary side to star (hidden) bus (MW)

reactive_power_flow_tertiary::Float64: Initial condition of reactive power flow through the transformer tertiary side to star (hidden) bus (MW)

r_primary::Float64: Equivalent resistance in pu (SYSTEM_BASE) from primary to star (hidden) bus., validation range: (-2, 4)

x_primary::Float64: Equivalent reactance in pu (SYSTEM_BASE) from primary to star (hidden) bus., validation range: (-2, 4)

r_secondary::Float64: Equivalent resistance in pu (SYSTEM_BASE) from secondary to star (hidden) bus., validation range: (-2, 4)

x_secondary::Float64: Equivalent reactance in pu (SYSTEM_BASE) from secondary to star (hidden) bus., validation range: (-2, 4)

r_tertiary::Float64: Equivalent resistance in pu (SYSTEM_BASE) from tertiary to star (hidden) bus., validation range: (-2, 4)

x_tertiary::Float64: Equivalent reactance in pu (SYSTEM_BASE) from tertiary to star (hidden) bus., validation range: (-2, 4)

rating::Union{Nothing, Float64}: Thermal rating (MVA). Flow through the transformer must be between -rating and rating. When defining a transformer before it is attached to a System, rating must be in pu (SYSTEM_BASE) using the base power of the System it will be attached to, validation range: (0, nothing)

r_12::Float64: Measured resistance in pu (SYSTEM_BASE) from primary to secondary windings (R1-2 with CZ = 1 in PSS/E)., validation range: (0, 4)

x_12::Float64: Measured reactance in pu (SYSTEM_BASE) from primary to secondary windings (X1-2 with CZ = 1 in PSS/E)., validation range: (0, 4)

r_23::Float64: Measured resistance in pu (SYSTEM_BASE) from secondary to tertiary windings (R2-3 with CZ = 1 in PSS/E)., validation range: (0, 4)

x_23::Float64: Measured reactance in pu (SYSTEM_BASE) from secondary to tertiary windings (X2-3 with CZ = 1 in PSS/E)., validation range: (0, 4)

r_13::Float64: Measured resistance in pu (SYSTEM_BASE) from primary to tertiary windings (R1-3 with CZ = 1 in PSS/E)., validation range: (0, 4)

x_13::Float64: Measured reactance in pu (SYSTEM_BASE) from primary to tertiary windings (X1-3 with CZ = 1 in PSS/E)., validation range: (0, 4)

base_power_12::Float64: Base power (MVA) for per unitization for primary-secondary windings., validation range: (0, nothing)

base_power_23::Float64: Base power (MVA) for per unitization for secondary-tertiary windings., validation range: (0, nothing)

base_power_13::Float64: Base power (MVA) for per unitization for primary-tertiary windings., validation range: (0, nothing)

base_voltage_primary::Union{Nothing, Float64}: (default: get_base_voltage(get_from(primary_star_arc))) Primary base voltage in kV, validation range: (0, nothing)

base_voltage_secondary::Union{Nothing, Float64}: (default: get_base_voltage(get_from(secondary_star_arc))) Secondary base voltage in kV, validation range: (0, nothing)

base_voltage_tertiary::Union{Nothing, Float64}: (default: get_base_voltage(get_from(tertiary_star_arc))) Tertiary base voltage in kV, validation range: (0, nothing)

g::Float64: (default: 0.0) Shunt conductance in pu (SYSTEM_BASE) from star (hidden) bus to ground (MAG1 in PSS/E).

b::Float64: (default: 0.0) Shunt susceptance in pu (SYSTEM_BASE) from star (hidden) bus to ground (MAG2 in PSS/E).

primary_turns_ratio::Float64: (default: 1.0) Primary side off-nominal turns ratio in p.u. with respect to connected primary bus (WINDV1 with CW = 1 in PSS/E).

secondary_turns_ratio::Float64: (default: 1.0) Secondary side off-nominal turns ratio in p.u. with respect to connected secondary bus (WINDV2 with CW = 1 in PSS/E).

tertiary_turns_ratio::Float64: (default: 1.0) Tertiary side off-nominal turns ratio in p.u. with respect to connected tertiary bus (WINDV3 with CW = 1 in PSS/E).

available_primary::Bool: (default: true) Status if primary winding is available or not.

available_secondary::Bool: (default: true) Status if primary winding is available or not.

available_tertiary::Bool: (default: true) Status if primary winding is available or not.

rating_primary::Float64: (default: 0.0) Rating (in MVA) for primary winding.

rating_secondary::Float64: (default: 0.0) Rating (in MVA) for secondary winding.

rating_tertiary::Float64: (default: 0.0) Rating (in MVA) for tertiary winding.

primary_group_number::WindingGroupNumber: (default: WindingGroupNumber.UNDEFINED) Vector group number ('clock number') indicating fixed phase shift (radians) between the from and to buses due to the connection group configuration

secondary_group_number::WindingGroupNumber: (default: WindingGroupNumber.UNDEFINED) Vector group number ('clock number') indicating fixed phase shift (radians) between the from and to buses due to the connection group configuration

tertiary_group_number::WindingGroupNumber: (default: WindingGroupNumber.UNDEFINED) Vector group number ('clock number') indicating fixed phase shift (radians) between the from and to buses due to the connection group configuration

control_objective_primary::TransformerControlObjective: (default: TransformerControlObjective.UNDEFINED) Control objective for the tap changer for winding 1. See TransformerControlObjective

control_objective_secondary::TransformerControlObjective: (default: TransformerControlObjective.UNDEFINED) Control objective for the tap changer for winding 2. See TransformerControlObjective

control_objective_tertiary::TransformerControlObjective: (default: TransformerControlObjective.UNDEFINED) Control objective for the tap changer for winding 3. See TransformerControlObjective

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::Transformer3W) -> Bool

Get Transformer3W available.

InfrastructureSystems.get_name — Method

get_name(value::Transformer3W) -> String

Get Transformer3W name.

InfrastructureSystems.set_available! — Method

set_available!(value::Transformer3W, val) -> Any

Set Transformer3W available.

PowerSystems.get_active_power_flow_primary — Method

get_active_power_flow_primary(
    value::Transformer3W
) -> Float64

Get Transformer3W active_power_flow_primary.

PowerSystems.get_active_power_flow_secondary — Method

get_active_power_flow_secondary(
    value::Transformer3W
) -> Float64

Get Transformer3W active_power_flow_secondary.

PowerSystems.get_active_power_flow_tertiary — Method

get_active_power_flow_tertiary(
    value::Transformer3W
) -> Float64

Get Transformer3W active_power_flow_tertiary.

PowerSystems.get_available_primary — Method

get_available_primary(value::Transformer3W) -> Bool

Get Transformer3W available_primary.

PowerSystems.get_available_secondary — Method

get_available_secondary(value::Transformer3W) -> Bool

Get Transformer3W available_secondary.

PowerSystems.get_available_tertiary — Method

get_available_tertiary(value::Transformer3W) -> Bool

Get Transformer3W available_tertiary.

PowerSystems.get_b — Method

get_b(value::Transformer3W) -> Float64

Get Transformer3W b.

PowerSystems.get_base_power_12 — Method

get_base_power_12(value::Transformer3W) -> Float64

Get Transformer3W base_power_12.

PowerSystems.get_base_power_13 — Method

get_base_power_13(value::Transformer3W) -> Float64

Get Transformer3W base_power_13.

PowerSystems.get_base_power_23 — Method

get_base_power_23(value::Transformer3W) -> Float64

Get Transformer3W base_power_23.

PowerSystems.get_base_voltage_primary — Method

get_base_voltage_primary(
    value::Transformer3W
) -> Union{Nothing, Float64}

Get Transformer3W base_voltage_primary.

PowerSystems.get_base_voltage_secondary — Method

get_base_voltage_secondary(
    value::Transformer3W
) -> Union{Nothing, Float64}

Get Transformer3W base_voltage_secondary.

PowerSystems.get_base_voltage_tertiary — Method

get_base_voltage_tertiary(
    value::Transformer3W
) -> Union{Nothing, Float64}

Get Transformer3W base_voltage_tertiary.

PowerSystems.get_control_objective_primary — Method

get_control_objective_primary(
    value::Transformer3W
) -> PowerSystems.TransformerControlObjectiveModule.TransformerControlObjective

Get Transformer3W control_objective_primary.

PowerSystems.get_control_objective_secondary — Method

get_control_objective_secondary(
    value::Transformer3W
) -> PowerSystems.TransformerControlObjectiveModule.TransformerControlObjective

Get Transformer3W control_objective_secondary.

PowerSystems.get_control_objective_tertiary — Method

get_control_objective_tertiary(
    value::Transformer3W
) -> PowerSystems.TransformerControlObjectiveModule.TransformerControlObjective

Get Transformer3W control_objective_tertiary.

PowerSystems.get_ext — Method

get_ext(value::Transformer3W) -> Dict{String, Any}

Get Transformer3W ext.

PowerSystems.get_g — Method

get_g(value::Transformer3W) -> Float64

Get Transformer3W g.

PowerSystems.get_primary_group_number — Method

get_primary_group_number(
    value::Transformer3W
) -> WindingGroupNumber

Get Transformer3W primary_group_number.

PowerSystems.get_primary_star_arc — Method

get_primary_star_arc(value::Transformer3W) -> Arc

Get Transformer3W primary_star_arc.

PowerSystems.get_primary_turns_ratio — Method

get_primary_turns_ratio(value::Transformer3W) -> Float64

Get Transformer3W primary_turns_ratio.

PowerSystems.get_r_12 — Method

get_r_12(value::Transformer3W) -> Float64

Get Transformer3W r_12.

PowerSystems.get_r_13 — Method

get_r_13(value::Transformer3W) -> Float64

Get Transformer3W r_13.

PowerSystems.get_r_23 — Method

get_r_23(value::Transformer3W) -> Float64

Get Transformer3W r_23.

PowerSystems.get_r_primary — Method

get_r_primary(value::Transformer3W) -> Float64

Get Transformer3W r_primary.

PowerSystems.get_r_secondary — Method

get_r_secondary(value::Transformer3W) -> Float64

Get Transformer3W r_secondary.

PowerSystems.get_r_tertiary — Method

get_r_tertiary(value::Transformer3W) -> Float64

Get Transformer3W r_tertiary.

PowerSystems.get_rating — Method

get_rating(value::Transformer3W) -> Union{Nothing, Float64}

Get Transformer3W rating.

PowerSystems.get_rating_primary — Method

get_rating_primary(value::Transformer3W) -> Float64

Get Transformer3W rating_primary.

PowerSystems.get_rating_secondary — Method

get_rating_secondary(value::Transformer3W) -> Float64

Get Transformer3W rating_secondary.

PowerSystems.get_rating_tertiary — Method

get_rating_tertiary(value::Transformer3W) -> Float64

Get Transformer3W rating_tertiary.

PowerSystems.get_reactive_power_flow_primary — Method

get_reactive_power_flow_primary(
    value::Transformer3W
) -> Float64

Get Transformer3W reactive_power_flow_primary.

PowerSystems.get_reactive_power_flow_secondary — Method

get_reactive_power_flow_secondary(
    value::Transformer3W
) -> Float64

Get Transformer3W reactive_power_flow_secondary.

PowerSystems.get_reactive_power_flow_tertiary — Method

get_reactive_power_flow_tertiary(
    value::Transformer3W
) -> Float64

Get Transformer3W reactive_power_flow_tertiary.

PowerSystems.get_secondary_group_number — Method

get_secondary_group_number(
    value::Transformer3W
) -> WindingGroupNumber

Get Transformer3W secondary_group_number.

PowerSystems.get_secondary_star_arc — Method

get_secondary_star_arc(value::Transformer3W) -> Arc

Get Transformer3W secondary_star_arc.

PowerSystems.get_secondary_turns_ratio — Method

get_secondary_turns_ratio(value::Transformer3W) -> Float64

Get Transformer3W secondary_turns_ratio.

PowerSystems.get_services — Method

get_services(value::Transformer3W) -> Vector{Service}

Get Transformer3W services.

PowerSystems.get_star_bus — Method

get_star_bus(value::Transformer3W) -> ACBus

Get Transformer3W star_bus.

PowerSystems.get_tertiary_group_number — Method

get_tertiary_group_number(
    value::Transformer3W
) -> WindingGroupNumber

Get Transformer3W tertiary_group_number.

PowerSystems.get_tertiary_star_arc — Method

get_tertiary_star_arc(value::Transformer3W) -> Arc

Get Transformer3W tertiary_star_arc.

PowerSystems.get_tertiary_turns_ratio — Method

get_tertiary_turns_ratio(value::Transformer3W) -> Float64

Get Transformer3W tertiary_turns_ratio.

PowerSystems.get_x_12 — Method

get_x_12(value::Transformer3W) -> Float64

Get Transformer3W x_12.

PowerSystems.get_x_13 — Method

get_x_13(value::Transformer3W) -> Float64

Get Transformer3W x_13.

PowerSystems.get_x_23 — Method

get_x_23(value::Transformer3W) -> Float64

Get Transformer3W x_23.

PowerSystems.get_x_primary — Method

get_x_primary(value::Transformer3W) -> Float64

Get Transformer3W x_primary.

PowerSystems.get_x_secondary — Method

get_x_secondary(value::Transformer3W) -> Float64

Get Transformer3W x_secondary.

PowerSystems.get_x_tertiary — Method

get_x_tertiary(value::Transformer3W) -> Float64

Get Transformer3W x_tertiary.

PowerSystems.set_active_power_flow_primary! — Method

set_active_power_flow_primary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W active_power_flow_primary.

PowerSystems.set_active_power_flow_secondary! — Method

set_active_power_flow_secondary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W active_power_flow_secondary.

PowerSystems.set_active_power_flow_tertiary! — Method

set_active_power_flow_tertiary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W active_power_flow_tertiary.

PowerSystems.set_available_primary! — Method

set_available_primary!(value::Transformer3W, val) -> Any

Set Transformer3W available_primary.

PowerSystems.set_available_secondary! — Method

set_available_secondary!(value::Transformer3W, val) -> Any

Set Transformer3W available_secondary.

PowerSystems.set_available_tertiary! — Method

set_available_tertiary!(value::Transformer3W, val) -> Any

Set Transformer3W available_tertiary.

PowerSystems.set_b! — Method

set_b!(value::Transformer3W, val) -> Any

Set Transformer3W b.

PowerSystems.set_base_power_12! — Method

set_base_power_12!(value::Transformer3W, val) -> Any

Set Transformer3W base_power_12.

PowerSystems.set_base_power_13! — Method

set_base_power_13!(value::Transformer3W, val) -> Any

Set Transformer3W base_power_13.

PowerSystems.set_base_power_23! — Method

set_base_power_23!(value::Transformer3W, val) -> Any

Set Transformer3W base_power_23.

PowerSystems.set_base_voltage_primary! — Method

set_base_voltage_primary!(value::Transformer3W, val) -> Any

Set Transformer3W base_voltage_primary.

PowerSystems.set_base_voltage_secondary! — Method

set_base_voltage_secondary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W base_voltage_secondary.

PowerSystems.set_base_voltage_tertiary! — Method

set_base_voltage_tertiary!(value::Transformer3W, val) -> Any

Set Transformer3W base_voltage_tertiary.

PowerSystems.set_control_objective_primary! — Method

set_control_objective_primary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W control_objective_primary.

PowerSystems.set_control_objective_secondary! — Method

set_control_objective_secondary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W control_objective_secondary.

PowerSystems.set_control_objective_tertiary! — Method

set_control_objective_tertiary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W control_objective_tertiary.

PowerSystems.set_ext! — Method

set_ext!(value::Transformer3W, val) -> Any

Set Transformer3W ext.

PowerSystems.set_g! — Method

set_g!(value::Transformer3W, val) -> Any

Set Transformer3W g.

PowerSystems.set_primary_group_number! — Method

set_primary_group_number!(value::Transformer3W, val) -> Any

Set Transformer3W primary_group_number.

PowerSystems.set_primary_star_arc! — Method

set_primary_star_arc!(value::Transformer3W, val) -> Any

Set Transformer3W primary_star_arc.

PowerSystems.set_primary_turns_ratio! — Method

set_primary_turns_ratio!(value::Transformer3W, val) -> Any

Set Transformer3W primary_turns_ratio.

PowerSystems.set_r_12! — Method

set_r_12!(value::Transformer3W, val) -> Any

Set Transformer3W r_12.

PowerSystems.set_r_13! — Method

set_r_13!(value::Transformer3W, val) -> Any

Set Transformer3W r_13.

PowerSystems.set_r_23! — Method

set_r_23!(value::Transformer3W, val) -> Any

Set Transformer3W r_23.

PowerSystems.set_r_primary! — Method

set_r_primary!(value::Transformer3W, val) -> Any

Set Transformer3W r_primary.

PowerSystems.set_r_secondary! — Method

set_r_secondary!(value::Transformer3W, val) -> Any

Set Transformer3W r_secondary.

PowerSystems.set_r_tertiary! — Method

set_r_tertiary!(value::Transformer3W, val) -> Any

Set Transformer3W r_tertiary.

PowerSystems.set_rating! — Method

set_rating!(value::Transformer3W, val) -> Any

Set Transformer3W rating.

PowerSystems.set_rating_primary! — Method

set_rating_primary!(value::Transformer3W, val) -> Any

Set Transformer3W rating_primary.

PowerSystems.set_rating_secondary! — Method

set_rating_secondary!(value::Transformer3W, val) -> Any

Set Transformer3W rating_secondary.

PowerSystems.set_rating_tertiary! — Method

set_rating_tertiary!(value::Transformer3W, val) -> Any

Set Transformer3W rating_tertiary.

PowerSystems.set_reactive_power_flow_primary! — Method

set_reactive_power_flow_primary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W reactive_power_flow_primary.

PowerSystems.set_reactive_power_flow_secondary! — Method

set_reactive_power_flow_secondary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W reactive_power_flow_secondary.

PowerSystems.set_reactive_power_flow_tertiary! — Method

set_reactive_power_flow_tertiary!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W reactive_power_flow_tertiary.

PowerSystems.set_secondary_group_number! — Method

set_secondary_group_number!(
    value::Transformer3W,
    val
) -> Any

Set Transformer3W secondary_group_number.

PowerSystems.set_secondary_star_arc! — Method

set_secondary_star_arc!(value::Transformer3W, val) -> Any

Set Transformer3W secondary_star_arc.

PowerSystems.set_secondary_turns_ratio! — Method

set_secondary_turns_ratio!(value::Transformer3W, val) -> Any

Set Transformer3W secondary_turns_ratio.

PowerSystems.set_services! — Method

set_services!(value::Transformer3W, val) -> Any

Set Transformer3W services.

PowerSystems.set_star_bus! — Method

set_star_bus!(value::Transformer3W, val) -> Any

Set Transformer3W star_bus.

PowerSystems.set_tertiary_group_number! — Method

set_tertiary_group_number!(value::Transformer3W, val) -> Any

Set Transformer3W tertiary_group_number.

PowerSystems.set_tertiary_star_arc! — Method

set_tertiary_star_arc!(value::Transformer3W, val) -> Any

Set Transformer3W tertiary_star_arc.

PowerSystems.set_tertiary_turns_ratio! — Method

set_tertiary_turns_ratio!(value::Transformer3W, val) -> Any

Set Transformer3W tertiary_turns_ratio.

PowerSystems.set_x_12! — Method

set_x_12!(value::Transformer3W, val) -> Any

Set Transformer3W x_12.

PowerSystems.set_x_13! — Method

set_x_13!(value::Transformer3W, val) -> Any

Set Transformer3W x_13.

PowerSystems.set_x_23! — Method

set_x_23!(value::Transformer3W, val) -> Any

Set Transformer3W x_23.

PowerSystems.set_x_primary! — Method

set_x_primary!(value::Transformer3W, val) -> Any

Set Transformer3W x_primary.

PowerSystems.set_x_secondary! — Method

set_x_secondary!(value::Transformer3W, val) -> Any

Set Transformer3W x_secondary.

PowerSystems.set_x_tertiary! — Method

set_x_tertiary!(value::Transformer3W, val) -> Any

Set Transformer3W x_tertiary.