TurbineGov

DEGOV

PowerSystems.DEGOV — Type

mutable struct DEGOV <: TurbineGov
    T1::Float64
    T2::Float64
    T3::Float64
    K::Float64
    T4::Float64
    T5::Float64
    T6::Float64
    Td::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

Parameters Woodward Diesel Governor Model. DEGOV in PowerWorld.

Arguments

T1::Float64: Governor mechanism time constant, validation range: (eps(), 100)

T2::Float64: Turbine power time constant, validation range: (eps(), 100)

T3::Float64: Turbine exhaust temperature time constant, validation range: (eps(), 100)

K::Float64: Governor gain (reciprocal of droop), validation range: (eps(), 100)

T4::Float64: Governor lead time constant, validation range: (eps(), 100)

T5::Float64: Governor lag time constant, validation range: (eps(), 100)

T6::Float64: Actuator time constant, validation range: (eps(), 100)

Td::Float64: Engine time delay, validation range: (eps(), 100)

P_ref::Float64: Reference Load Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the DEGOV model are:

x_ecb1: Electric control box 1,
x_ecb2: Electric control box 2,
x_a1: Actuator 1,
x_a2: Actuator 2,
x_a3: Actuator 3,

n_states::Int: DEGOV has 5 states

states_types::Vector{StateTypes}: DEGOV has 5 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_K — Method

get_K(value::DEGOV) -> Float64

Get DEGOV K.

PowerSystems.get_P_ref — Method

get_P_ref(value::DEGOV) -> Float64

Get DEGOV P_ref.

PowerSystems.get_T1 — Method

get_T1(value::DEGOV) -> Float64

Get DEGOV T1.

PowerSystems.get_T2 — Method

get_T2(value::DEGOV) -> Float64

Get DEGOV T2.

PowerSystems.get_T3 — Method

get_T3(value::DEGOV) -> Float64

Get DEGOV T3.

PowerSystems.get_T4 — Method

get_T4(value::DEGOV) -> Float64

Get DEGOV T4.

PowerSystems.get_T5 — Method

get_T5(value::DEGOV) -> Float64

Get DEGOV T5.

PowerSystems.get_T6 — Method

get_T6(value::DEGOV) -> Float64

Get DEGOV T6.

PowerSystems.get_Td — Method

get_Td(value::DEGOV) -> Float64

Get DEGOV Td.

PowerSystems.get_ext — Method

get_ext(value::DEGOV) -> Dict{String, Any}

Get DEGOV ext.

PowerSystems.get_n_states — Method

get_n_states(value::DEGOV) -> Int64

Get DEGOV n_states.

PowerSystems.get_states — Method

get_states(value::DEGOV) -> Vector{Symbol}

Get DEGOV states.

PowerSystems.get_states_types — Method

get_states_types(value::DEGOV) -> Vector{StateTypes}

Get DEGOV states_types.

PowerSystems.set_K! — Method

set_K!(value::DEGOV, val) -> Any

Set DEGOV K.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::DEGOV, val) -> Any

Set DEGOV P_ref.

PowerSystems.set_T1! — Method

set_T1!(value::DEGOV, val) -> Any

Set DEGOV T1.

PowerSystems.set_T2! — Method

set_T2!(value::DEGOV, val) -> Any

Set DEGOV T2.

PowerSystems.set_T3! — Method

set_T3!(value::DEGOV, val) -> Any

Set DEGOV T3.

PowerSystems.set_T4! — Method

set_T4!(value::DEGOV, val) -> Any

Set DEGOV T4.

PowerSystems.set_T5! — Method

set_T5!(value::DEGOV, val) -> Any

Set DEGOV T5.

PowerSystems.set_T6! — Method

set_T6!(value::DEGOV, val) -> Any

Set DEGOV T6.

PowerSystems.set_Td! — Method

set_Td!(value::DEGOV, val) -> Any

Set DEGOV Td.

PowerSystems.set_ext! — Method

set_ext!(value::DEGOV, val) -> Any

Set DEGOV ext.

PowerSystems.set_states_types! — Method

set_states_types!(value::DEGOV, val) -> Any

Set DEGOV states_types.

GasTG

PowerSystems.GasTG — Type

mutable struct GasTG <: TurbineGov
    R::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    AT::Float64
    Kt::Float64
    V_lim::Tuple{Float64, Float64}
    D_turb::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

Parameters of Gas Turbine-Governor. GAST in PSSE and GAST_PTI in PowerWorld.

Arguments

R::Float64: Speed droop parameter, validation range: (eps(), 0.1)

T1::Float64: Governor time constant in s, validation range: (eps(), 0.5)

T2::Float64: Combustion chamber time constant, validation range: (eps(), 0.5)

T3::Float64: Load limit time constant (exhaust gas measurement time), validation range: (eps(), 5)

AT::Float64: Ambient temperature load limit, validation range: (0, 1)

Kt::Float64: Load limit feedback gain, validation range: (0, 5)

V_lim::Tuple{Float64, Float64}: Operational control limits on fuel valve opening (Vmin, Vmax)

D_turb::Float64: Speed damping coefficient of gas turbine rotor, validation range: (0, 0.5)

P_ref::Float64: Reference Load Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the GAST model are:

x_g1: Fuel valve opening,
x_g2: Fuel flow,
x_g3: Exhaust temperature load

n_states::Int: GasTG has 3 states

states_types::Vector{StateTypes}: GAST has 3 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_AT — Method

get_AT(value::GasTG) -> Float64

Get GasTG AT.

PowerSystems.get_D_turb — Method

get_D_turb(value::GasTG) -> Float64

Get GasTG D_turb.

PowerSystems.get_Kt — Method

get_Kt(value::GasTG) -> Float64

Get GasTG Kt.

PowerSystems.get_P_ref — Method

get_P_ref(value::GasTG) -> Float64

Get GasTG P_ref.

PowerSystems.get_R — Method

get_R(value::GasTG) -> Float64

Get GasTG R.

PowerSystems.get_T1 — Method

get_T1(value::GasTG) -> Float64

Get GasTG T1.

PowerSystems.get_T2 — Method

get_T2(value::GasTG) -> Float64

Get GasTG T2.

PowerSystems.get_T3 — Method

get_T3(value::GasTG) -> Float64

Get GasTG T3.

PowerSystems.get_V_lim — Method

get_V_lim(value::GasTG) -> Tuple{Float64, Float64}

Get GasTG V_lim.

PowerSystems.get_ext — Method

get_ext(value::GasTG) -> Dict{String, Any}

Get GasTG ext.

PowerSystems.get_n_states — Method

get_n_states(value::GasTG) -> Int64

Get GasTG n_states.

PowerSystems.get_states — Method

get_states(value::GasTG) -> Vector{Symbol}

Get GasTG states.

PowerSystems.get_states_types — Method

get_states_types(value::GasTG) -> Vector{StateTypes}

Get GasTG states_types.

PowerSystems.set_AT! — Method

set_AT!(value::GasTG, val) -> Any

Set GasTG AT.

PowerSystems.set_D_turb! — Method

set_D_turb!(value::GasTG, val) -> Any

Set GasTG D_turb.

PowerSystems.set_Kt! — Method

set_Kt!(value::GasTG, val) -> Any

Set GasTG Kt.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::GasTG, val) -> Any

Set GasTG P_ref.

PowerSystems.set_R! — Method

set_R!(value::GasTG, val) -> Any

Set GasTG R.

PowerSystems.set_T1! — Method

set_T1!(value::GasTG, val) -> Any

Set GasTG T1.

PowerSystems.set_T2! — Method

set_T2!(value::GasTG, val) -> Any

Set GasTG T2.

PowerSystems.set_T3! — Method

set_T3!(value::GasTG, val) -> Any

Set GasTG T3.

PowerSystems.set_V_lim! — Method

set_V_lim!(value::GasTG, val) -> Any

Set GasTG V_lim.

PowerSystems.set_ext! — Method

set_ext!(value::GasTG, val) -> Any

Set GasTG ext.

PowerSystems.set_states_types! — Method

set_states_types!(value::GasTG, val) -> Any

Set GasTG states_types.

GeneralGovModel

PowerSystems.GeneralGovModel — Type

mutable struct GeneralGovModel <: TurbineGov
    Rselect::Int
    fuel_flag::Int
    R::Float64
    Tpelec::Float64
    speed_error_signal::MinMax
    Kp_gov::Float64
    Ki_gov::Float64
    Kd_gov::Float64
    Td_gov::Float64
    valve_position_limits::MinMax
    T_act::Float64
    K_turb::Float64
    Wf_nl::Float64
    Tb::Float64
    Tc::Float64
    T_eng::Float64
    Tf_load::Float64
    Kp_load::Float64
    Ki_load::Float64
    Ld_ref::Float64
    Dm::Float64
    R_open::Float64
    R_close::Float64
    Ki_mw::Float64
    A_set::Float64
    Ka::Float64
    Ta::Float64
    T_rate::Float64
    db::Float64
    Tsa::Float64
    Tsb::Float64
    R_lim::UpDown
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

GE General Governor/Turbine Model. The GeneralGovModel (GGOV1) model is a general purpose governor model used for a variety of prime movers controlled by proportional-integral-derivative (PID) governors including gas turbines.

Arguments

Rselect::Int: Feedback signal for governor droop, validation range: (-2, 1), action if invalid: error

fuel_flag::Int: Flag Switch for fuel source characteristic, validation range: (0, 1), action if invalid: error

R::Float64: Speed droop parameter, validation range: (eps(), nothing), action if invalid: warn

Tpelec::Float64: Electrical power transducer time constant, seconds, validation range: (eps(), nothing), action if invalid: warn

speed_error_signal::MinMax: Speed error signal limits

Kp_gov::Float64: Governor proportional gain, validation range: (0, nothing), action if invalid: warn

Ki_gov::Float64: Governor integral gain, validation range: (0, nothing), action if invalid: warn

Kd_gov::Float64: Governor derivative gain, validation range: (0, nothing), action if invalid: warn

Td_gov::Float64: Governor derivative time constant, validation range: (0, nothing), action if invalid: warn

valve_position_limits::MinMax: Valve position limits

T_act::Float64: Actuator time constant, validation range: (0, nothing), action if invalid: warn

K_turb::Float64: Turbine gain, validation range: (0, nothing), action if invalid: warn

Wf_nl::Float64: No load fuel flow, pu, validation range: (0, nothing), action if invalid: warn

Tb::Float64: Turbine lag time constant, sec, validation range: (0, nothing), action if invalid: warn

Tc::Float64: Turbine lead time constant, sec, validation range: (0, nothing), action if invalid: warn

T_eng::Float64: Transport lag time constant for diesel engine, sec, validation range: (0, nothing), action if invalid: warn

Tf_load::Float64: Load limiter time constant, validation range: (0, nothing), action if invalid: warn

Kp_load::Float64: Load limiter proportional gain for PI controller, validation range: (0, nothing), action if invalid: warn

Ki_load::Float64: Load integral gain for PI controller, validation range: (0, nothing), action if invalid: warn

Ld_ref::Float64: Load limiter integral gain for PI controller, validation range: (0, nothing), action if invalid: warn

Dm::Float64: Mechanical damping coefficient, pu, validation range: (0, nothing), action if invalid: warn

R_open::Float64: Maximum valve opening rate, pu/sec, validation range: (0, nothing), action if invalid: warn

R_close::Float64: Maximum valve closing rate, pu/sec, validation range: (0, nothing), action if invalid: warn

Ki_mw::Float64: Power controller (reset) gain, validation range: (0, nothing), action if invalid: warn

A_set::Float64: Acceleration limiter setpoint, pu/sec, validation range: (0, nothing), action if invalid: warn

Ka::Float64: Acceleration limiter gain, validation range: (0, nothing), action if invalid: warn

Ta::Float64: Acceleration limiter time constant , validation range: (eps(), nothing), action if invalid: error

T_rate::Float64: Turbine rating, validation range: (0, nothing), action if invalid: warn

db::Float64: Speed governor deadband, validation range: (0, nothing), action if invalid: warn

Tsa::Float64: Temperature detection lead time constant, validation range: (0, nothing), action if invalid: warn

Tsb::Float64: Temperature detection lag time constant, validation range: (0, nothing), action if invalid: warn

R_lim::UpDown: Maximum rate of load increa

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the GGOV1 model are:

Pe: Machine Electrical Power Measurement,
x_g1: Governor differential control,
x_g2: Governor integral control, 
x_g3: Turbine actuator, 
x_g4: Turbine Lead-Lag, 
x_g5: Turbine load limiter measurement, 
x_g6: Turbine Load Limiter Integral Control, 
x_g7: Supervisory Load Control, 
x_g8: Acceleration Control, 
x_g9 Temperature Detection Lead - Lag:

n_states::Int: GeneralGovModel has 10 states

states_types::Vector{StateTypes}: GGOV1 has 10 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_A_set — Method

get_A_set(value::GeneralGovModel) -> Float64

Get GeneralGovModel A_set.

PowerSystems.get_Dm — Method

get_Dm(value::GeneralGovModel) -> Float64

Get GeneralGovModel Dm.

PowerSystems.get_K_turb — Method

get_K_turb(value::GeneralGovModel) -> Float64

Get GeneralGovModel K_turb.

PowerSystems.get_Ka — Method

get_Ka(value::GeneralGovModel) -> Float64

Get GeneralGovModel Ka.

PowerSystems.get_Kd_gov — Method

get_Kd_gov(value::GeneralGovModel) -> Float64

Get GeneralGovModel Kd_gov.

PowerSystems.get_Ki_gov — Method

get_Ki_gov(value::GeneralGovModel) -> Float64

Get GeneralGovModel Ki_gov.

PowerSystems.get_Ki_load — Method

get_Ki_load(value::GeneralGovModel) -> Float64

Get GeneralGovModel Ki_load.

PowerSystems.get_Ki_mw — Method

get_Ki_mw(value::GeneralGovModel) -> Float64

Get GeneralGovModel Ki_mw.

PowerSystems.get_Kp_gov — Method

get_Kp_gov(value::GeneralGovModel) -> Float64

Get GeneralGovModel Kp_gov.

PowerSystems.get_Kp_load — Method

get_Kp_load(value::GeneralGovModel) -> Float64

Get GeneralGovModel Kp_load.

PowerSystems.get_Ld_ref — Method

get_Ld_ref(value::GeneralGovModel) -> Float64

Get GeneralGovModel Ld_ref.

PowerSystems.get_P_ref — Method

get_P_ref(value::GeneralGovModel) -> Float64

Get GeneralGovModel P_ref.

PowerSystems.get_R — Method

get_R(value::GeneralGovModel) -> Float64

Get GeneralGovModel R.

PowerSystems.get_R_close — Method

get_R_close(value::GeneralGovModel) -> Float64

Get GeneralGovModel R_close.

PowerSystems.get_R_lim — Method

get_R_lim(
    value::GeneralGovModel
) -> @NamedTuple{up::Float64, down::Float64}

Get GeneralGovModel R_lim.

PowerSystems.get_R_open — Method

get_R_open(value::GeneralGovModel) -> Float64

Get GeneralGovModel R_open.

PowerSystems.get_Rselect — Method

get_Rselect(value::GeneralGovModel) -> Int64

Get GeneralGovModel Rselect.

PowerSystems.get_T_act — Method

get_T_act(value::GeneralGovModel) -> Float64

Get GeneralGovModel T_act.

PowerSystems.get_T_eng — Method

get_T_eng(value::GeneralGovModel) -> Float64

Get GeneralGovModel T_eng.

PowerSystems.get_T_rate — Method

get_T_rate(value::GeneralGovModel) -> Float64

Get GeneralGovModel T_rate.

PowerSystems.get_Ta — Method

get_Ta(value::GeneralGovModel) -> Float64

Get GeneralGovModel Ta.

PowerSystems.get_Tb — Method

get_Tb(value::GeneralGovModel) -> Float64

Get GeneralGovModel Tb.

PowerSystems.get_Tc — Method

get_Tc(value::GeneralGovModel) -> Float64

Get GeneralGovModel Tc.

PowerSystems.get_Td_gov — Method

get_Td_gov(value::GeneralGovModel) -> Float64

Get GeneralGovModel Td_gov.

PowerSystems.get_Tf_load — Method

get_Tf_load(value::GeneralGovModel) -> Float64

Get GeneralGovModel Tf_load.

PowerSystems.get_Tpelec — Method

get_Tpelec(value::GeneralGovModel) -> Float64

Get GeneralGovModel Tpelec.

PowerSystems.get_Tsa — Method

get_Tsa(value::GeneralGovModel) -> Float64

Get GeneralGovModel Tsa.

PowerSystems.get_Tsb — Method

get_Tsb(value::GeneralGovModel) -> Float64

Get GeneralGovModel Tsb.

PowerSystems.get_Wf_nl — Method

get_Wf_nl(value::GeneralGovModel) -> Float64

Get GeneralGovModel Wf_nl.

PowerSystems.get_db — Method

get_db(value::GeneralGovModel) -> Float64

Get GeneralGovModel db.

PowerSystems.get_ext — Method

get_ext(value::GeneralGovModel) -> Dict{String, Any}

Get GeneralGovModel ext.

PowerSystems.get_fuel_flag — Method

get_fuel_flag(value::GeneralGovModel) -> Int64

Get GeneralGovModel fuel_flag.

PowerSystems.get_n_states — Method

get_n_states(value::GeneralGovModel) -> Int64

Get GeneralGovModel n_states.

PowerSystems.get_speed_error_signal — Method

get_speed_error_signal(
    value::GeneralGovModel
) -> @NamedTuple{min::Float64, max::Float64}

Get GeneralGovModel speed_error_signal.

PowerSystems.get_states — Method

get_states(value::GeneralGovModel) -> Vector{Symbol}

Get GeneralGovModel states.

PowerSystems.get_states_types — Method

get_states_types(
    value::GeneralGovModel
) -> Vector{StateTypes}

Get GeneralGovModel states_types.

PowerSystems.get_valve_position_limits — Method

get_valve_position_limits(
    value::GeneralGovModel
) -> @NamedTuple{min::Float64, max::Float64}

Get GeneralGovModel valve_position_limits.

PowerSystems.set_A_set! — Method

set_A_set!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel A_set.

PowerSystems.set_Dm! — Method

set_Dm!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Dm.

PowerSystems.set_K_turb! — Method

set_K_turb!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel K_turb.

PowerSystems.set_Ka! — Method

set_Ka!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Ka.

PowerSystems.set_Kd_gov! — Method

set_Kd_gov!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Kd_gov.

PowerSystems.set_Ki_gov! — Method

set_Ki_gov!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Ki_gov.

PowerSystems.set_Ki_load! — Method

set_Ki_load!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Ki_load.

PowerSystems.set_Ki_mw! — Method

set_Ki_mw!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Ki_mw.

PowerSystems.set_Kp_gov! — Method

set_Kp_gov!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Kp_gov.

PowerSystems.set_Kp_load! — Method

set_Kp_load!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Kp_load.

PowerSystems.set_Ld_ref! — Method

set_Ld_ref!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Ld_ref.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel P_ref.

PowerSystems.set_R! — Method

set_R!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel R.

PowerSystems.set_R_close! — Method

set_R_close!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel R_close.

PowerSystems.set_R_lim! — Method

set_R_lim!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel R_lim.

PowerSystems.set_R_open! — Method

set_R_open!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel R_open.

PowerSystems.set_Rselect! — Method

set_Rselect!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Rselect.

PowerSystems.set_T_act! — Method

set_T_act!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel T_act.

PowerSystems.set_T_eng! — Method

set_T_eng!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel T_eng.

PowerSystems.set_T_rate! — Method

set_T_rate!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel T_rate.

PowerSystems.set_Ta! — Method

set_Ta!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Ta.

PowerSystems.set_Tb! — Method

set_Tb!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Tb.

PowerSystems.set_Tc! — Method

set_Tc!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Tc.

PowerSystems.set_Td_gov! — Method

set_Td_gov!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Td_gov.

PowerSystems.set_Tf_load! — Method

set_Tf_load!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Tf_load.

PowerSystems.set_Tpelec! — Method

set_Tpelec!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Tpelec.

PowerSystems.set_Tsa! — Method

set_Tsa!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Tsa.

PowerSystems.set_Tsb! — Method

set_Tsb!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Tsb.

PowerSystems.set_Wf_nl! — Method

set_Wf_nl!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel Wf_nl.

PowerSystems.set_db! — Method

set_db!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel db.

PowerSystems.set_ext! — Method

set_ext!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel ext.

PowerSystems.set_fuel_flag! — Method

set_fuel_flag!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel fuel_flag.

PowerSystems.set_speed_error_signal! — Method

set_speed_error_signal!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel speed_error_signal.

PowerSystems.set_states_types! — Method

set_states_types!(value::GeneralGovModel, val) -> Any

Set GeneralGovModel states_types.

PowerSystems.set_valve_position_limits! — Method

set_valve_position_limits!(
    value::GeneralGovModel,
    val
) -> Any

Set GeneralGovModel valve_position_limits.

HydroTurbineGov

PowerSystems.HydroTurbineGov — Type

mutable struct HydroTurbineGov <: TurbineGov
    R::Float64
    r::Float64
    Tr::Float64
    Tf::Float64
    Tg::Float64
    VELM::Float64
    gate_position_limits::MinMax
    Tw::Float64
    At::Float64
    D_T::Float64
    q_nl::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

Hydro Turbine-Governor.

Arguments

R::Float64: Permanent droop parameter, validation range: (0, 0.1), action if invalid: warn

r::Float64: Temporary Droop, validation range: (0, 2), action if invalid: warn

Tr::Float64: Governor time constant, validation range: (eps(), 30), action if invalid: error

Tf::Float64: Filter Time constant, validation range: (eps(), 0.1)

Tg::Float64: Servo time constant, validation range: (eps(), 1), action if invalid: error

VELM::Float64: gate velocity limit, validation range: (eps(), 0.3), action if invalid: error

gate_position_limits::MinMax: Gate position limits

Tw::Float64: water time constant, validation range: (eps(), 3), action if invalid: error

At::Float64: Turbine gain, validation range: (0.8, 1.5), action if invalid: warn

D_T::Float64: Turbine Damping, validation range: (0, 0.5), action if invalid: warn

q_nl::Float64: No-power flow, validation range: (0, nothing), action if invalid: warn

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the HydroTurbineGov model are:

x_g1: filter_output,
x_g2: desired gate, 
x_g3: gate opening, 
x_g4: turbine flow

n_states::Int: HYGOV has 4 states

states_types::Vector{StateTypes}: HYGOV has 4 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_At — Method

get_At(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov At.

PowerSystems.get_D_T — Method

get_D_T(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov D_T.

PowerSystems.get_P_ref — Method

get_P_ref(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov P_ref.

PowerSystems.get_R — Method

get_R(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov R.

PowerSystems.get_Tf — Method

get_Tf(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov Tf.

PowerSystems.get_Tg — Method

get_Tg(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov Tg.

PowerSystems.get_Tr — Method

get_Tr(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov Tr.

PowerSystems.get_Tw — Method

get_Tw(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov Tw.

PowerSystems.get_VELM — Method

get_VELM(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov VELM.

PowerSystems.get_ext — Method

get_ext(value::HydroTurbineGov) -> Dict{String, Any}

Get HydroTurbineGov ext.

PowerSystems.get_gate_position_limits — Method

get_gate_position_limits(
    value::HydroTurbineGov
) -> @NamedTuple{min::Float64, max::Float64}

Get HydroTurbineGov gate_position_limits.

PowerSystems.get_n_states — Method

get_n_states(value::HydroTurbineGov) -> Int64

Get HydroTurbineGov n_states.

PowerSystems.get_q_nl — Method

get_q_nl(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov q_nl.

PowerSystems.get_r — Method

get_r(value::HydroTurbineGov) -> Float64

Get HydroTurbineGov r.

PowerSystems.get_states — Method

get_states(value::HydroTurbineGov) -> Vector{Symbol}

Get HydroTurbineGov states.

PowerSystems.get_states_types — Method

get_states_types(
    value::HydroTurbineGov
) -> Vector{StateTypes}

Get HydroTurbineGov states_types.

PowerSystems.set_At! — Method

set_At!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov At.

PowerSystems.set_D_T! — Method

set_D_T!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov D_T.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov P_ref.

PowerSystems.set_R! — Method

set_R!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov R.

PowerSystems.set_Tf! — Method

set_Tf!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov Tf.

PowerSystems.set_Tg! — Method

set_Tg!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov Tg.

PowerSystems.set_Tr! — Method

set_Tr!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov Tr.

PowerSystems.set_Tw! — Method

set_Tw!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov Tw.

PowerSystems.set_VELM! — Method

set_VELM!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov VELM.

PowerSystems.set_ext! — Method

set_ext!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov ext.

PowerSystems.set_gate_position_limits! — Method

set_gate_position_limits!(
    value::HydroTurbineGov,
    val
) -> Any

Set HydroTurbineGov gate_position_limits.

PowerSystems.set_q_nl! — Method

set_q_nl!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov q_nl.

PowerSystems.set_r! — Method

set_r!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov r.

PowerSystems.set_states_types! — Method

set_states_types!(value::HydroTurbineGov, val) -> Any

Set HydroTurbineGov states_types.

IEEETurbineGov1

PowerSystems.IEEETurbineGov1 — Type

mutable struct IEEETurbineGov1 <: TurbineGov
    K::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    U0::Float64
    U_c::Float64
    valve_position_limits::MinMax
    T4::Float64
    K1::Float64
    K2::Float64
    T5::Float64
    K3::Float64
    K4::Float64
    T6::Float64
    K5::Float64
    K6::Float64
    T7::Float64
    K7::Float64
    K8::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

IEEE Type 1 Speed-Governing Model

Arguments

K::Float64: Governor Gain, validation range: (5, 30), action if invalid: warn

T1::Float64: Input Filter Lag, validation range: (0, 5), action if invalid: warn

T2::Float64: Input Filter Lead, validation range: (0, 10), action if invalid: warn

T3::Float64: Valve position Time Constant, validation range: (eps(), 1), action if invalid: error

U0::Float64: Maximum Valve Opening Rate, validation range: (0.01, 0.03), action if invalid: warn

U_c::Float64: Maximum Valve closing rate, validation range: (-0.3, 0), action if invalid: warn

valve_position_limits::MinMax: Valve position limits in MW

T4::Float64: Time Constant inlet steam, validation range: (0, 1), action if invalid: warn

K1::Float64: Fraction of high presure shaft power, validation range: (-2, 1), action if invalid: warn

K2::Float64: Fraction of low presure shaft power, validation range: (0, nothing), action if invalid: warn

T5::Float64: Time constant for second boiler pass, validation range: (0, 10), action if invalid: warn

K3::Float64: Fraction of high presure shaft power second boiler pass, validation range: (0, 0.5), action if invalid: warn

K4::Float64: Fraction of low presure shaft power second boiler pass, validation range: (0, 0.5), action if invalid: warn

T6::Float64: Time constant for third boiler pass, validation range: (0, 10), action if invalid: warn

K5::Float64: Fraction of high presure shaft power third boiler pass, validation range: (0, 0.35), action if invalid: warn

K6::Float64: Fraction of low presure shaft power third boiler pass, validation range: (0, 0.55), action if invalid: warn

T7::Float64: Time constant for fourth boiler pass, validation range: (0, 10), action if invalid: warn

K7::Float64: Fraction of high presure shaft power fourth boiler pass, validation range: (0, 0.3), action if invalid: warn

K8::Float64: Fraction of low presure shaft power fourth boiler pass, validation range: (0, 0.3), action if invalid: warn

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the IEEETurbineGov model are:

x_g1: First Governor integrator,
x_g2: Governor output,
x_g3: First Turbine integrator, 
x_g4: Second Turbine Integrator, 
x_g5: Third Turbine Integrator, 
x_g6: Fourth Turbine Integrator,

n_states::Int: IEEEG1 has 6 states

states_types::Vector{StateTypes}: IEEEG1 has 6 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_K — Method

get_K(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 K.

PowerSystems.get_K1 — Method

get_K1(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 K1.

PowerSystems.get_K2 — Method

get_K2(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 K2.

PowerSystems.get_K3 — Method

get_K3(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 K3.

PowerSystems.get_K4 — Method

get_K4(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 K4.

PowerSystems.get_K5 — Method

get_K5(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 K5.

PowerSystems.get_K6 — Method

get_K6(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 K6.

PowerSystems.get_K7 — Method

get_K7(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 K7.

PowerSystems.get_K8 — Method

get_K8(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 K8.

PowerSystems.get_P_ref — Method

get_P_ref(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 P_ref.

PowerSystems.get_T1 — Method

get_T1(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 T1.

PowerSystems.get_T2 — Method

get_T2(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 T2.

PowerSystems.get_T3 — Method

get_T3(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 T3.

PowerSystems.get_T4 — Method

get_T4(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 T4.

PowerSystems.get_T5 — Method

get_T5(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 T5.

PowerSystems.get_T6 — Method

get_T6(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 T6.

PowerSystems.get_T7 — Method

get_T7(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 T7.

PowerSystems.get_U0 — Method

get_U0(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 U0.

PowerSystems.get_U_c — Method

get_U_c(value::IEEETurbineGov1) -> Float64

Get IEEETurbineGov1 U_c.

PowerSystems.get_ext — Method

get_ext(value::IEEETurbineGov1) -> Dict{String, Any}

Get IEEETurbineGov1 ext.

PowerSystems.get_n_states — Method

get_n_states(value::IEEETurbineGov1) -> Int64

Get IEEETurbineGov1 n_states.

PowerSystems.get_states — Method

get_states(value::IEEETurbineGov1) -> Vector{Symbol}

Get IEEETurbineGov1 states.

PowerSystems.get_states_types — Method

get_states_types(
    value::IEEETurbineGov1
) -> Vector{StateTypes}

Get IEEETurbineGov1 states_types.

PowerSystems.get_valve_position_limits — Method

get_valve_position_limits(
    value::IEEETurbineGov1
) -> @NamedTuple{min::Float64, max::Float64}

Get IEEETurbineGov1 valve_position_limits.

PowerSystems.set_K! — Method

set_K!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 K.

PowerSystems.set_K1! — Method

set_K1!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 K1.

PowerSystems.set_K2! — Method

set_K2!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 K2.

PowerSystems.set_K3! — Method

set_K3!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 K3.

PowerSystems.set_K4! — Method

set_K4!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 K4.

PowerSystems.set_K5! — Method

set_K5!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 K5.

PowerSystems.set_K6! — Method

set_K6!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 K6.

PowerSystems.set_K7! — Method

set_K7!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 K7.

PowerSystems.set_K8! — Method

set_K8!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 K8.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 P_ref.

PowerSystems.set_T1! — Method

set_T1!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 T1.

PowerSystems.set_T2! — Method

set_T2!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 T2.

PowerSystems.set_T3! — Method

set_T3!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 T3.

PowerSystems.set_T4! — Method

set_T4!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 T4.

PowerSystems.set_T5! — Method

set_T5!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 T5.

PowerSystems.set_T6! — Method

set_T6!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 T6.

PowerSystems.set_T7! — Method

set_T7!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 T7.

PowerSystems.set_U0! — Method

set_U0!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 U0.

PowerSystems.set_U_c! — Method

set_U_c!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 U_c.

PowerSystems.set_ext! — Method

set_ext!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 ext.

PowerSystems.set_states_types! — Method

set_states_types!(value::IEEETurbineGov1, val) -> Any

Set IEEETurbineGov1 states_types.

PowerSystems.set_valve_position_limits! — Method

set_valve_position_limits!(
    value::IEEETurbineGov1,
    val
) -> Any

Set IEEETurbineGov1 valve_position_limits.

SteamTurbineGov1

PowerSystems.SteamTurbineGov1 — Type

mutable struct SteamTurbineGov1 <: TurbineGov
    R::Float64
    T1::Float64
    valve_position_limits::MinMax
    T2::Float64
    T3::Float64
    D_T::Float64
    DB_h::Float64
    DB_l::Float64
    T_rate::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

Steam Turbine-Governor. This model considers both TGOV1 or TGOV1DU in PSS/E.

Arguments

R::Float64: Droop parameter, validation range: (0, 0.1), action if invalid: warn

T1::Float64: Governor time constant, validation range: (eps(), 0.5), action if invalid: error

valve_position_limits::MinMax: Valve position limits

T2::Float64: Lead Lag Lead Time constant , validation range: (0, nothing), action if invalid: warn

T3::Float64: Lead Lag Lag Time constant , validation range: (eps(), 10), action if invalid: error

D_T::Float64: Turbine Damping, validation range: (0, 0.5), action if invalid: warn

DB_h::Float64: Deadband for overspeed, validation range: (0, nothing), action if invalid: warn

DB_l::Float64: Deadband for underspeed, validation range: (nothing, 0), action if invalid: warn

T_rate::Float64: Turbine Rate (MW). If zero, generator base is used., validation range: (0, nothing), action if invalid: warn

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the SteamTurbineGov1 model are:

x_g1: Valve Opening,
x_g2: Lead-lag state

n_states::Int: TGOV1 has 2 states

states_types::Vector{StateTypes}: TGOV1 has 2 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_DB_h — Method

get_DB_h(value::SteamTurbineGov1) -> Float64

Get SteamTurbineGov1 DB_h.

PowerSystems.get_DB_l — Method

get_DB_l(value::SteamTurbineGov1) -> Float64

Get SteamTurbineGov1 DB_l.

PowerSystems.get_D_T — Method

get_D_T(value::SteamTurbineGov1) -> Float64

Get SteamTurbineGov1 D_T.

PowerSystems.get_P_ref — Method

get_P_ref(value::SteamTurbineGov1) -> Float64

Get SteamTurbineGov1 P_ref.

PowerSystems.get_R — Method

get_R(value::SteamTurbineGov1) -> Float64

Get SteamTurbineGov1 R.

PowerSystems.get_T1 — Method

get_T1(value::SteamTurbineGov1) -> Float64

Get SteamTurbineGov1 T1.

PowerSystems.get_T2 — Method

get_T2(value::SteamTurbineGov1) -> Float64

Get SteamTurbineGov1 T2.

PowerSystems.get_T3 — Method

get_T3(value::SteamTurbineGov1) -> Float64

Get SteamTurbineGov1 T3.

PowerSystems.get_T_rate — Method

get_T_rate(value::SteamTurbineGov1) -> Float64

Get SteamTurbineGov1 T_rate.

PowerSystems.get_ext — Method

get_ext(value::SteamTurbineGov1) -> Dict{String, Any}

Get SteamTurbineGov1 ext.

PowerSystems.get_n_states — Method

get_n_states(value::SteamTurbineGov1) -> Int64

Get SteamTurbineGov1 n_states.

PowerSystems.get_states — Method

get_states(value::SteamTurbineGov1) -> Vector{Symbol}

Get SteamTurbineGov1 states.

PowerSystems.get_states_types — Method

get_states_types(
    value::SteamTurbineGov1
) -> Vector{StateTypes}

Get SteamTurbineGov1 states_types.

PowerSystems.get_valve_position_limits — Method

get_valve_position_limits(
    value::SteamTurbineGov1
) -> @NamedTuple{min::Float64, max::Float64}

Get SteamTurbineGov1 valve_position_limits.

PowerSystems.set_DB_h! — Method

set_DB_h!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 DB_h.

PowerSystems.set_DB_l! — Method

set_DB_l!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 DB_l.

PowerSystems.set_D_T! — Method

set_D_T!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 D_T.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 P_ref.

PowerSystems.set_R! — Method

set_R!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 R.

PowerSystems.set_T1! — Method

set_T1!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 T1.

PowerSystems.set_T2! — Method

set_T2!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 T2.

PowerSystems.set_T3! — Method

set_T3!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 T3.

PowerSystems.set_T_rate! — Method

set_T_rate!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 T_rate.

PowerSystems.set_ext! — Method

set_ext!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 ext.

PowerSystems.set_states_types! — Method

set_states_types!(value::SteamTurbineGov1, val) -> Any

Set SteamTurbineGov1 states_types.

PowerSystems.set_valve_position_limits! — Method

set_valve_position_limits!(
    value::SteamTurbineGov1,
    val
) -> Any

Set SteamTurbineGov1 valve_position_limits.

TGFixed

PowerSystems.TGFixed — Type

mutable struct TGFixed <: TurbineGov
    efficiency::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of a fixed Turbine Governor that returns a fixed mechanical torque given by the product of P_ref*efficiency

Arguments

efficiency::Float64: Efficiency factor that multiplies P_ref, validation range: (0, nothing)

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}

n_states::Int: TGFixed has no states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_P_ref — Method

get_P_ref(value::TGFixed) -> Float64

Get TGFixed P_ref.

PowerSystems.get_efficiency — Method

get_efficiency(value::TGFixed) -> Float64

Get TGFixed efficiency.

PowerSystems.get_ext — Method

get_ext(value::TGFixed) -> Dict{String, Any}

Get TGFixed ext.

PowerSystems.get_n_states — Method

get_n_states(value::TGFixed) -> Int64

Get TGFixed n_states.

PowerSystems.get_states — Method

get_states(value::TGFixed) -> Vector{Symbol}

Get TGFixed states.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::TGFixed, val) -> Any

Set TGFixed P_ref.

PowerSystems.set_efficiency! — Method

set_efficiency!(value::TGFixed, val) -> Any

Set TGFixed efficiency.

PowerSystems.set_ext! — Method

set_ext!(value::TGFixed, val) -> Any

Set TGFixed ext.

TGTypeI

PowerSystems.TGTypeI — Type

mutable struct TGTypeI <: TurbineGov
    R::Float64
    Ts::Float64
    Tc::Float64
    T3::Float64
    T4::Float64
    T5::Float64
    valve_position_limits::MinMax
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of a Turbine Governor Type I.

Arguments

R::Float64: Droop parameter, validation range: (0, nothing)

Ts::Float64: Governor time constant, validation range: (0, nothing)

Tc::Float64: Servo time constant, validation range: (0, nothing)

T3::Float64: Transient gain time constant, validation range: (0, nothing)

T4::Float64: Power fraction time constant, validation range: (0, nothing)

T5::Float64: Reheat time constant, validation range: (0, nothing)

valve_position_limits::MinMax: Valve position limits in MW

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the TGTypeI model are:

x_g1: Governor state,
x_g2: Servo state,
x_g3: Reheat state

n_states::Int: TGTypeI has 3 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_P_ref — Method

get_P_ref(value::TGTypeI) -> Float64

Get TGTypeI P_ref.

PowerSystems.get_R — Method

get_R(value::TGTypeI) -> Float64

Get TGTypeI R.

PowerSystems.get_T3 — Method

get_T3(value::TGTypeI) -> Float64

Get TGTypeI T3.

PowerSystems.get_T4 — Method

get_T4(value::TGTypeI) -> Float64

Get TGTypeI T4.

PowerSystems.get_T5 — Method

get_T5(value::TGTypeI) -> Float64

Get TGTypeI T5.

PowerSystems.get_Tc — Method

get_Tc(value::TGTypeI) -> Float64

Get TGTypeI Tc.

PowerSystems.get_Ts — Method

get_Ts(value::TGTypeI) -> Float64

Get TGTypeI Ts.

PowerSystems.get_ext — Method

get_ext(value::TGTypeI) -> Dict{String, Any}

Get TGTypeI ext.

PowerSystems.get_n_states — Method

get_n_states(value::TGTypeI) -> Int64

Get TGTypeI n_states.

PowerSystems.get_states — Method

get_states(value::TGTypeI) -> Vector{Symbol}

Get TGTypeI states.

PowerSystems.get_valve_position_limits — Method

get_valve_position_limits(
    value::TGTypeI
) -> @NamedTuple{min::Float64, max::Float64}

Get TGTypeI valve_position_limits.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::TGTypeI, val) -> Any

Set TGTypeI P_ref.

PowerSystems.set_R! — Method

set_R!(value::TGTypeI, val) -> Any

Set TGTypeI R.

PowerSystems.set_T3! — Method

set_T3!(value::TGTypeI, val) -> Any

Set TGTypeI T3.

PowerSystems.set_T4! — Method

set_T4!(value::TGTypeI, val) -> Any

Set TGTypeI T4.

PowerSystems.set_T5! — Method

set_T5!(value::TGTypeI, val) -> Any

Set TGTypeI T5.

PowerSystems.set_Tc! — Method

set_Tc!(value::TGTypeI, val) -> Any

Set TGTypeI Tc.

PowerSystems.set_Ts! — Method

set_Ts!(value::TGTypeI, val) -> Any

Set TGTypeI Ts.

PowerSystems.set_ext! — Method

set_ext!(value::TGTypeI, val) -> Any

Set TGTypeI ext.

PowerSystems.set_valve_position_limits! — Method

set_valve_position_limits!(value::TGTypeI, val) -> Any

Set TGTypeI valve_position_limits.

TGTypeII

PowerSystems.TGTypeII — Type

mutable struct TGTypeII <: TurbineGov
    R::Float64
    T1::Float64
    T2::Float64
    τ_limits::MinMax
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of a Turbine Governor Type II.

Arguments

R::Float64: Droop parameter, validation range: (0, nothing)

T1::Float64: Transient gain time constant, validation range: (0, nothing)

T2::Float64: Power fraction time constant, validation range: (0, nothing)

τ_limits::MinMax: Power into the governor limits

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the TGTypeI model are:

x_g1: lead-lag state

n_states::Int: TGTypeII has 1 state

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_P_ref — Method

get_P_ref(value::TGTypeII) -> Float64

Get TGTypeII P_ref.

PowerSystems.get_R — Method

get_R(value::TGTypeII) -> Float64

Get TGTypeII R.

PowerSystems.get_T1 — Method

get_T1(value::TGTypeII) -> Float64

Get TGTypeII T1.

PowerSystems.get_T2 — Method

get_T2(value::TGTypeII) -> Float64

Get TGTypeII T2.

PowerSystems.get_ext — Method

get_ext(value::TGTypeII) -> Dict{String, Any}

Get TGTypeII ext.

PowerSystems.get_n_states — Method

get_n_states(value::TGTypeII) -> Int64

Get TGTypeII n_states.

PowerSystems.get_states — Method

get_states(value::TGTypeII) -> Vector{Symbol}

Get TGTypeII states.

PowerSystems.get_τ_limits — Method

get_τ_limits(
    value::TGTypeII
) -> @NamedTuple{min::Float64, max::Float64}

Get TGTypeII τ_limits.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::TGTypeII, val) -> Any

Set TGTypeII P_ref.

PowerSystems.set_R! — Method

set_R!(value::TGTypeII, val) -> Any

Set TGTypeII R.

PowerSystems.set_T1! — Method

set_T1!(value::TGTypeII, val) -> Any

Set TGTypeII T1.

PowerSystems.set_T2! — Method

set_T2!(value::TGTypeII, val) -> Any

Set TGTypeII T2.

PowerSystems.set_ext! — Method

set_ext!(value::TGTypeII, val) -> Any

Set TGTypeII ext.

PowerSystems.set_τ_limits! — Method

set_τ_limits!(value::TGTypeII, val) -> Any

Set TGTypeII τ_limits.