TwoTerminalLCCLine

PowerSystems.TwoTerminalLCCLine — Type

mutable struct TwoTerminalLCCLine <: TwoTerminalHVDC
    name::String
    available::Bool
    arc::Arc
    active_power_flow::Float64
    r::Float64
    transfer_setpoint::Float64
    scheduled_dc_voltage::Float64
    rectifier_bridges::Int
    rectifier_delay_angle_limits::MinMax
    rectifier_rc::Float64
    rectifier_xc::Float64
    rectifier_base_voltage::Float64
    inverter_bridges::Int
    inverter_extinction_angle_limits::MinMax
    inverter_rc::Float64
    inverter_xc::Float64
    inverter_base_voltage::Float64
    power_mode::Bool
    switch_mode_voltage::Float64
    compounding_resistance::Float64
    min_compounding_voltage::Float64
    rectifier_transformer_ratio::Float64
    rectifier_tap_setting::Float64
    rectifier_tap_limits::MinMax
    rectifier_tap_step::Float64
    rectifier_delay_angle::Float64
    rectifier_capacitor_reactance::Float64
    inverter_transformer_ratio::Float64
    inverter_tap_setting::Float64
    inverter_tap_limits::MinMax
    inverter_tap_step::Float64
    inverter_extinction_angle::Float64
    inverter_capacitor_reactance::Float64
    active_power_limits_from::MinMax
    active_power_limits_to::MinMax
    reactive_power_limits_from::MinMax
    reactive_power_limits_to::MinMax
    loss::Union{LinearCurve, PiecewiseIncrementalCurve}
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A Non-Capacitor Line Commutated Converter (LCC)-HVDC transmission line.

As implemented in PSS/E.

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

arc::Arc: An Arc defining this line from a rectifier bus to an inverter bus. The rectifier bus must be specified in the from bus and inverter bus in the to bus.

active_power_flow::Float64: Initial condition of active power flow on the line (MW)

r::Float64: Series resistance of the DC line in pu (SYSTEM_BASE)

transfer_setpoint::Float64: Desired set-point of power. If power_mode = true this value is in MW units, and if power_mode = false is in Amperes units. This parameter must not be specified in per-unit. A positive value represents the desired consumed power at the rectifier bus, while a negative value represents the desired power at the inverter bus (i.e. the absolute value of transfer_setpoint is the generated power at the inverter bus).

scheduled_dc_voltage::Float64: Scheduled compounded DC voltage in kV. By default this parameter is the scheduled DC voltage in the inverter bus This parameter must not be specified in per-unit.

rectifier_bridges::Int: Number of bridges in series in the rectifier side.

rectifier_delay_angle_limits::MinMax: Minimum and maximum rectifier firing delay angle (α) (radians)

rectifier_rc::Float64: Rectifier commutating transformer resistance per bridge in system p.u. (SYSTEM_BASE)

rectifier_xc::Float64: Rectifier commutating transformer reactance per bridge in system p.u. (SYSTEM_BASE)

rectifier_base_voltage::Float64: Rectifier primary base AC voltage in kV, entered in kV.

inverter_bridges::Int: Number of bridges in series in the inverter side.

inverter_extinction_angle_limits::MinMax: Minimum and maximum inverter extinction angle (γ) (radians)

inverter_rc::Float64: Inverter commutating transformer resistance per bridge in system p.u. (SYSTEM_BASE)

inverter_xc::Float64: Inverter commutating transformer reactance per bridge in system p.u. (SYSTEM_BASE)

inverter_base_voltage::Float64: Inverter primary base AC voltage in kV, entered in kV.

power_mode::Bool: (default: true) Boolean flag to identify if the LCC line is in power mode or current mode. If power_mode = true, setpoint values must be specified in MW, and if power_mode = false setpoint values must be specified in Amperes.

switch_mode_voltage::Float64: (default: 0.0) Mode switch DC voltage, in kV. This parameter must not be added in per-unit. If LCC line is in power mode control, and DC voltage falls below this value, the line switch to current mode control.

compounding_resistance::Float64: (default: 0.0) Compounding Resistance, in ohms. This parameter is for control of the DC voltage in the rectifier or inverter end. For inverter DC voltage control, the paremeter is set to zero; for rectifier DC voltage control, the paremeter is set to the DC line resistance; otherwise, set to a fraction of the DC line resistance.

min_compounding_voltage::Float64: (default: 0.0) Minimum compounded voltage, in kV. This parameter must not be added in per-unit. Only used in constant gamma operation (γmin = γmax), and the AC transformer is used to control the DC voltage.

rectifier_transformer_ratio::Float64: (default: 1.0) Rectifier transformer ratio between the primary and secondary side AC voltages.

rectifier_tap_setting::Float64: (default: 1.0) Rectifier transformer tap setting.

rectifier_tap_limits::MinMax: (default: (min=0.51, max=1.5)) Minimum and maximum rectifier tap limits as a ratio between the primary and secondary side AC voltages.

rectifier_tap_step::Float64: (default: 0.00625) Rectifier transformer tap step value

rectifier_delay_angle::Float64: (default: 0.0) Rectifier firing delay angle (α).

rectifier_capacitor_reactance::Float64: (default: 0.0) Commutating rectifier capacitor reactance magnitude per bridge, in system p.u. (SYSTEM_BASE).

inverter_transformer_ratio::Float64: (default: 1.0) Inverter transformer ratio between the primary and secondary side AC voltages.

inverter_tap_setting::Float64: (default: 1.0) Inverter transformer tap setting.

inverter_tap_limits::MinMax: (default: (min=0.51, max=1.5)) Minimum and maximum inverter tap limits as a ratio between the primary and secondary side AC voltages.

inverter_tap_step::Float64: (default: 0.00625) Inverter transformer tap step value.

inverter_extinction_angle::Float64: (default: 0.0) Inverter extinction angle (γ).

inverter_capacitor_reactance::Float64: (default: 0.0) Commutating inverter capacitor reactance magnitude per bridge, in system p.u. (SYSTEM_BASE).

active_power_limits_from::MinMax: (default: (min=0.0, max=0.0)) Minimum and maximum active power flows to the FROM node (MW)

active_power_limits_to::MinMax: (default: (min=0.0, max=0.0)) Minimum and maximum active power flows to the TO node (MW)

reactive_power_limits_from::MinMax: (default: (min=0.0, max=0.0)) Minimum and maximum reactive power limits to the FROM node (MVAR)

reactive_power_limits_to::MinMax: (default: (min=0.0, max=0.0)) Minimum and maximum reactive power limits to the TO node (MVAR)

loss::Union{LinearCurve, PiecewiseIncrementalCurve}: (default: LinearCurve(0.0)) A generic loss model coefficients. It accepts a linear model with a constant loss (MW) and a proportional loss rate (MW of loss per MW of flow). It also accepts a Piecewise loss, with N segments to specify different proportional losses for different segments.

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::TwoTerminalLCCLine) -> Bool

Get TwoTerminalLCCLine available.

InfrastructureSystems.get_name — Method

get_name(value::TwoTerminalLCCLine) -> String

Get TwoTerminalLCCLine name.

InfrastructureSystems.set_available! — Method

set_available!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine available.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::TwoTerminalLCCLine) -> Float64

Get TwoTerminalLCCLine active_power_flow.

PowerSystems.get_active_power_limits_from — Method

get_active_power_limits_from(
    value::TwoTerminalLCCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalLCCLine active_power_limits_from.

PowerSystems.get_active_power_limits_to — Method

get_active_power_limits_to(
    value::TwoTerminalLCCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalLCCLine active_power_limits_to.

PowerSystems.get_arc — Method

get_arc(value::TwoTerminalLCCLine) -> Arc

Get TwoTerminalLCCLine arc.

PowerSystems.get_compounding_resistance — Method

get_compounding_resistance(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine compounding_resistance.

PowerSystems.get_ext — Method

get_ext(value::TwoTerminalLCCLine) -> Dict{String, Any}

Get TwoTerminalLCCLine ext.

PowerSystems.get_inverter_base_voltage — Method

get_inverter_base_voltage(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine inverter_base_voltage.

PowerSystems.get_inverter_bridges — Method

get_inverter_bridges(value::TwoTerminalLCCLine) -> Int64

Get TwoTerminalLCCLine inverter_bridges.

PowerSystems.get_inverter_capacitor_reactance — Method

get_inverter_capacitor_reactance(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine inverter_capacitor_reactance.

PowerSystems.get_inverter_extinction_angle — Method

get_inverter_extinction_angle(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine inverter_extinction_angle.

PowerSystems.get_inverter_extinction_angle_limits — Method

get_inverter_extinction_angle_limits(
    value::TwoTerminalLCCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalLCCLine inverter_extinction_angle_limits.

PowerSystems.get_inverter_rc — Method

get_inverter_rc(value::TwoTerminalLCCLine) -> Float64

Get TwoTerminalLCCLine inverter_rc.

PowerSystems.get_inverter_tap_limits — Method

get_inverter_tap_limits(
    value::TwoTerminalLCCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalLCCLine inverter_tap_limits.

PowerSystems.get_inverter_tap_setting — Method

get_inverter_tap_setting(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine inverter_tap_setting.

PowerSystems.get_inverter_tap_step — Method

get_inverter_tap_step(value::TwoTerminalLCCLine) -> Float64

Get TwoTerminalLCCLine inverter_tap_step.

PowerSystems.get_inverter_transformer_ratio — Method

get_inverter_transformer_ratio(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine inverter_transformer_ratio.

PowerSystems.get_inverter_xc — Method

get_inverter_xc(value::TwoTerminalLCCLine) -> Float64

Get TwoTerminalLCCLine inverter_xc.

PowerSystems.get_loss — Method

get_loss(
    value::TwoTerminalLCCLine
) -> Union{LinearCurve, PiecewiseIncrementalCurve}

Get TwoTerminalLCCLine loss.

PowerSystems.get_min_compounding_voltage — Method

get_min_compounding_voltage(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine min_compounding_voltage.

PowerSystems.get_power_mode — Method

get_power_mode(value::TwoTerminalLCCLine) -> Bool

Get TwoTerminalLCCLine power_mode.

PowerSystems.get_r — Method

get_r(value::TwoTerminalLCCLine) -> Float64

Get TwoTerminalLCCLine r.

PowerSystems.get_reactive_power_limits_from — Method

get_reactive_power_limits_from(
    value::TwoTerminalLCCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalLCCLine reactive_power_limits_from.

PowerSystems.get_reactive_power_limits_to — Method

get_reactive_power_limits_to(
    value::TwoTerminalLCCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalLCCLine reactive_power_limits_to.

PowerSystems.get_rectifier_base_voltage — Method

get_rectifier_base_voltage(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine rectifier_base_voltage.

PowerSystems.get_rectifier_bridges — Method

get_rectifier_bridges(value::TwoTerminalLCCLine) -> Int64

Get TwoTerminalLCCLine rectifier_bridges.

PowerSystems.get_rectifier_capacitor_reactance — Method

get_rectifier_capacitor_reactance(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine rectifier_capacitor_reactance.

PowerSystems.get_rectifier_delay_angle — Method

get_rectifier_delay_angle(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine rectifier_delay_angle.

PowerSystems.get_rectifier_delay_angle_limits — Method

get_rectifier_delay_angle_limits(
    value::TwoTerminalLCCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalLCCLine rectifier_delay_angle_limits.

PowerSystems.get_rectifier_rc — Method

get_rectifier_rc(value::TwoTerminalLCCLine) -> Float64

Get TwoTerminalLCCLine rectifier_rc.

PowerSystems.get_rectifier_tap_limits — Method

get_rectifier_tap_limits(
    value::TwoTerminalLCCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalLCCLine rectifier_tap_limits.

PowerSystems.get_rectifier_tap_setting — Method

get_rectifier_tap_setting(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine rectifier_tap_setting.

PowerSystems.get_rectifier_tap_step — Method

get_rectifier_tap_step(value::TwoTerminalLCCLine) -> Float64

Get TwoTerminalLCCLine rectifier_tap_step.

PowerSystems.get_rectifier_transformer_ratio — Method

get_rectifier_transformer_ratio(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine rectifier_transformer_ratio.

PowerSystems.get_rectifier_xc — Method

get_rectifier_xc(value::TwoTerminalLCCLine) -> Float64

Get TwoTerminalLCCLine rectifier_xc.

PowerSystems.get_scheduled_dc_voltage — Method

get_scheduled_dc_voltage(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine scheduled_dc_voltage.

PowerSystems.get_services — Method

get_services(value::TwoTerminalLCCLine) -> Vector{Service}

Get TwoTerminalLCCLine services.

PowerSystems.get_switch_mode_voltage — Method

get_switch_mode_voltage(
    value::TwoTerminalLCCLine
) -> Float64

Get TwoTerminalLCCLine switch_mode_voltage.

PowerSystems.get_transfer_setpoint — Method

get_transfer_setpoint(value::TwoTerminalLCCLine) -> Float64

Get TwoTerminalLCCLine transfer_setpoint.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine active_power_flow.

PowerSystems.set_active_power_limits_from! — Method

set_active_power_limits_from!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine active_power_limits_from.

PowerSystems.set_active_power_limits_to! — Method

set_active_power_limits_to!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine active_power_limits_to.

PowerSystems.set_arc! — Method

set_arc!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine arc.

PowerSystems.set_compounding_resistance! — Method

set_compounding_resistance!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine compounding_resistance.

PowerSystems.set_ext! — Method

set_ext!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine ext.

PowerSystems.set_inverter_base_voltage! — Method

set_inverter_base_voltage!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine inverter_base_voltage.

PowerSystems.set_inverter_bridges! — Method

set_inverter_bridges!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine inverter_bridges.

PowerSystems.set_inverter_capacitor_reactance! — Method

set_inverter_capacitor_reactance!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine inverter_capacitor_reactance.

PowerSystems.set_inverter_extinction_angle! — Method

set_inverter_extinction_angle!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine inverter_extinction_angle.

PowerSystems.set_inverter_extinction_angle_limits! — Method

set_inverter_extinction_angle_limits!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine inverter_extinction_angle_limits.

PowerSystems.set_inverter_rc! — Method

set_inverter_rc!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine inverter_rc.

PowerSystems.set_inverter_tap_limits! — Method

set_inverter_tap_limits!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine inverter_tap_limits.

PowerSystems.set_inverter_tap_setting! — Method

set_inverter_tap_setting!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine inverter_tap_setting.

PowerSystems.set_inverter_tap_step! — Method

set_inverter_tap_step!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine inverter_tap_step.

PowerSystems.set_inverter_transformer_ratio! — Method

set_inverter_transformer_ratio!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine inverter_transformer_ratio.

PowerSystems.set_inverter_xc! — Method

set_inverter_xc!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine inverter_xc.

PowerSystems.set_loss! — Method

set_loss!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine loss.

PowerSystems.set_min_compounding_voltage! — Method

set_min_compounding_voltage!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine min_compounding_voltage.

PowerSystems.set_power_mode! — Method

set_power_mode!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine power_mode.

PowerSystems.set_r! — Method

set_r!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine r.

PowerSystems.set_reactive_power_limits_from! — Method

set_reactive_power_limits_from!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine reactive_power_limits_from.

PowerSystems.set_reactive_power_limits_to! — Method

set_reactive_power_limits_to!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine reactive_power_limits_to.

PowerSystems.set_rectifier_base_voltage! — Method

set_rectifier_base_voltage!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine rectifier_base_voltage.

PowerSystems.set_rectifier_bridges! — Method

set_rectifier_bridges!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine rectifier_bridges.

PowerSystems.set_rectifier_capacitor_reactance! — Method

set_rectifier_capacitor_reactance!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine rectifier_capacitor_reactance.

PowerSystems.set_rectifier_delay_angle! — Method

set_rectifier_delay_angle!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine rectifier_delay_angle.

PowerSystems.set_rectifier_delay_angle_limits! — Method

set_rectifier_delay_angle_limits!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine rectifier_delay_angle_limits.

PowerSystems.set_rectifier_rc! — Method

set_rectifier_rc!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine rectifier_rc.

PowerSystems.set_rectifier_tap_limits! — Method

set_rectifier_tap_limits!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine rectifier_tap_limits.

PowerSystems.set_rectifier_tap_setting! — Method

set_rectifier_tap_setting!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine rectifier_tap_setting.

PowerSystems.set_rectifier_tap_step! — Method

set_rectifier_tap_step!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine rectifier_tap_step.

PowerSystems.set_rectifier_transformer_ratio! — Method

set_rectifier_transformer_ratio!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine rectifier_transformer_ratio.

PowerSystems.set_rectifier_xc! — Method

set_rectifier_xc!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine rectifier_xc.

PowerSystems.set_scheduled_dc_voltage! — Method

set_scheduled_dc_voltage!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine scheduled_dc_voltage.

PowerSystems.set_services! — Method

set_services!(value::TwoTerminalLCCLine, val) -> Any

Set TwoTerminalLCCLine services.

PowerSystems.set_switch_mode_voltage! — Method

set_switch_mode_voltage!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine switch_mode_voltage.

PowerSystems.set_transfer_setpoint! — Method

set_transfer_setpoint!(
    value::TwoTerminalLCCLine,
    val
) -> Any

Set TwoTerminalLCCLine transfer_setpoint.