TwoTerminalVSCDCLine

PowerSystems.TwoTerminalVSCDCLine — Type

mutable struct TwoTerminalVSCDCLine <: ACBranch
    name::String
    available::Bool
    active_power_flow::Float64
    arc::Arc
    rectifier_tap_limits::MinMax
    rectifier_xrc::Float64
    rectifier_firing_angle::MinMax
    inverter_tap_limits::MinMax
    inverter_xrc::Float64
    inverter_extinction_angle::MinMax
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A Voltage Sourced Converter (VSC)-HVDC transmission line.

As implemented in "Power System Modelling and Scripting" by Federico Milano, Chapter 18, Page 397. This model is suitable for dynamic simulations

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow::Float64: Initial condition of active power flow on the line (MW)

arc::Arc: An Arc defining this line from a bus to another bus

rectifier_tap_limits::MinMax: Minimum and maximum rectifier tap limits as a ratio between the primary and secondary side voltages

rectifier_xrc::Float64: Rectifier commutation reactance in p.u. (DEVICE_BASE)

rectifier_firing_angle::MinMax: Minimum and maximum rectifier firing angle (α) (radians)

inverter_tap_limits::MinMax: Minimum and maximum inverter tap limits as a ratio between the primary and secondary side voltages

inverter_xrc::Float64: Inverter commutation reactance in p.u. (DEVICE_BASE)

inverter_extinction_angle::MinMax: Minimum and maximum inverter extinction angle (γ) (radians)

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::TwoTerminalVSCDCLine) -> String

Get TwoTerminalVSCDCLine name.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::TwoTerminalVSCDCLine) -> Any

Get TwoTerminalVSCDCLine active_power_flow.

PowerSystems.get_arc — Method

get_arc(value::TwoTerminalVSCDCLine) -> Arc

Get TwoTerminalVSCDCLine arc.

PowerSystems.get_available — Method

get_available(value::TwoTerminalVSCDCLine) -> Bool

Get TwoTerminalVSCDCLine available.

PowerSystems.get_ext — Method

get_ext(value::TwoTerminalVSCDCLine) -> Dict{String, Any}

Get TwoTerminalVSCDCLine ext.

PowerSystems.get_inverter_extinction_angle — Method

get_inverter_extinction_angle(
    value::TwoTerminalVSCDCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCDCLine inverter_extinction_angle.

PowerSystems.get_inverter_tap_limits — Method

get_inverter_tap_limits(
    value::TwoTerminalVSCDCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCDCLine inverter_tap_limits.

PowerSystems.get_inverter_xrc — Method

get_inverter_xrc(value::TwoTerminalVSCDCLine) -> Float64

Get TwoTerminalVSCDCLine inverter_xrc.

PowerSystems.get_rectifier_firing_angle — Method

get_rectifier_firing_angle(
    value::TwoTerminalVSCDCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCDCLine rectifier_firing_angle.

PowerSystems.get_rectifier_tap_limits — Method

get_rectifier_tap_limits(
    value::TwoTerminalVSCDCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCDCLine rectifier_tap_limits.

PowerSystems.get_rectifier_xrc — Method

get_rectifier_xrc(value::TwoTerminalVSCDCLine) -> Float64

Get TwoTerminalVSCDCLine rectifier_xrc.

PowerSystems.get_services — Method

get_services(value::TwoTerminalVSCDCLine) -> Vector{Service}

Get TwoTerminalVSCDCLine services.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(
    value::TwoTerminalVSCDCLine,
    val
) -> Any

Set TwoTerminalVSCDCLine active_power_flow.

PowerSystems.set_arc! — Method

set_arc!(value::TwoTerminalVSCDCLine, val) -> Any

Set TwoTerminalVSCDCLine arc.

PowerSystems.set_available! — Method

set_available!(value::TwoTerminalVSCDCLine, val) -> Any

Set TwoTerminalVSCDCLine available.

PowerSystems.set_ext! — Method

set_ext!(value::TwoTerminalVSCDCLine, val) -> Any

Set TwoTerminalVSCDCLine ext.

PowerSystems.set_inverter_extinction_angle! — Method

set_inverter_extinction_angle!(
    value::TwoTerminalVSCDCLine,
    val
) -> Any

Set TwoTerminalVSCDCLine inverter_extinction_angle.

PowerSystems.set_inverter_tap_limits! — Method

set_inverter_tap_limits!(
    value::TwoTerminalVSCDCLine,
    val
) -> Any

Set TwoTerminalVSCDCLine inverter_tap_limits.

PowerSystems.set_inverter_xrc! — Method

set_inverter_xrc!(value::TwoTerminalVSCDCLine, val) -> Any

Set TwoTerminalVSCDCLine inverter_xrc.

PowerSystems.set_rectifier_firing_angle! — Method

set_rectifier_firing_angle!(
    value::TwoTerminalVSCDCLine,
    val
) -> Any

Set TwoTerminalVSCDCLine rectifier_firing_angle.

PowerSystems.set_rectifier_tap_limits! — Method

set_rectifier_tap_limits!(
    value::TwoTerminalVSCDCLine,
    val
) -> Any

Set TwoTerminalVSCDCLine rectifier_tap_limits.

PowerSystems.set_rectifier_xrc! — Method

set_rectifier_xrc!(value::TwoTerminalVSCDCLine, val) -> Any

Set TwoTerminalVSCDCLine rectifier_xrc.

PowerSystems.set_services! — Method

set_services!(value::TwoTerminalVSCDCLine, val) -> Any

Set TwoTerminalVSCDCLine services.