TwoTerminalVSCLine

PowerSystems.TwoTerminalVSCLine — Type

mutable struct TwoTerminalVSCLine <: TwoTerminalHVDC
    name::String
    available::Bool
    arc::Arc
    active_power_flow::Float64
    rating::Float64
    active_power_limits_from::MinMax
    active_power_limits_to::MinMax
    g::Float64
    dc_current::Float64
    reactive_power_from::Float64
    dc_voltage_control_from::Bool
    ac_voltage_control_from::Bool
    dc_setpoint_from::Float64
    ac_setpoint_from::Float64
    converter_loss_from::Union{LinearCurve, QuadraticCurve}
    max_dc_current_from::Float64
    rating_from::Float64
    reactive_power_limits_from::MinMax
    power_factor_weighting_fraction_from::Float64
    voltage_limits_from::MinMax
    reactive_power_to::Float64
    dc_voltage_control_to::Bool
    ac_voltage_control_to::Bool
    dc_setpoint_to::Float64
    ac_setpoint_to::Float64
    converter_loss_to::Union{LinearCurve, QuadraticCurve}
    max_dc_current_to::Float64
    rating_to::Float64
    reactive_power_limits_to::MinMax
    power_factor_weighting_fraction_to::Float64
    voltage_limits_to::MinMax
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A High Voltage Voltage-Source Converter DC line, which must be connected to an ACBus on each end.

This model is appropriate for operational simulations with a linearized DC power flow approximation with losses using a voltage-current model. For modeling a DC network, see TModelHVDCLine

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

arc::Arc: An Arc defining this line from a bus to another bus

active_power_flow::Float64: Initial condition of active power flowing from the from-bus to the to-bus in DC.

rating::Float64: Maximum output power rating of the converter (MVA), validation range: (0, nothing)

active_power_limits_from::MinMax: Minimum and maximum active power flows to the FROM node (MW)

active_power_limits_to::MinMax: Minimum and maximum active power flows to the TO node (MW)

g::Float64: (default: 0.0) Series conductance of the DC line in pu (SYSTEM_BASE)

dc_current::Float64: (default: 0.0) DC current (A) on the converter flowing in the DC line, from from bus to to bus.

reactive_power_from::Float64: (default: 0.0) Initial condition of reactive power flowing into the from-bus.

dc_voltage_control_from::Bool: (default: true) Converter control type in the from bus converter. Set true for DC Voltage Control (set DC voltage on the DC side of the converter), and false for power demand in the converter.

ac_voltage_control_from::Bool: (default: true) Converter control type in the from bus converter. Set true for AC Voltage Control (set AC voltage on the AC side of the converter), and false for fixed power AC factor.

dc_setpoint_from::Float64: (default: 0.0) Converter DC setpoint in the from bus converter. If voltage_control_from = true this number is the DC voltage on the DC side of the converter, entered in kV. If voltage_control_from = false, this value is the power demand in MW, if positive the converter is supplying power to the AC network at the from bus; if negative, the converter is withdrawing power from the AC network at the from bus.

ac_setpoint_from::Float64: (default: 1.0) Converter AC setpoint in the from bus converter. If voltage_control_from = true this number is the AC voltage on the AC side of the converter, entered in per unit. If voltage_control_from = false, this value is the power factor setpoint.

converter_loss_from::Union{LinearCurve, QuadraticCurve}: (default: LinearCurve(0.0)) Loss model coefficients in the from bus converter. It accepts a linear model or quadratic. Same converter data is used in both ends.

max_dc_current_from::Float64: (default: 1e8) Maximum stable dc current limits (A).

rating_from::Float64: (default: 1e8) Converter rating in MVA in the from bus.

reactive_power_limits_from::MinMax: (default: (min=0.0, max=0.0)) Limits on the Reactive Power at the from side.

power_factor_weighting_fraction_from::Float64: (default: 1.0) Power weighting factor fraction used in reducing the active power order and either the reactive power order when the converter rating is violated. When is 0.0, only the active power is reduced; when is 1.0, only the reactive power is reduced; otherwise, a weighted reduction of both active and reactive power is applied., validation range: (0, 1)

voltage_limits_from::MinMax: (default: (min=0.0, max=999.9)) Limits on the Voltage at the DC from Bus in [per unit](@ref per_unit.

reactive_power_to::Float64: (default: 0.0) Initial condition of reactive power flowing into the to-bus.

dc_voltage_control_to::Bool: (default: true) Converter control type in the to bus converter. Set true for DC Voltage Control (set DC voltage on the DC side of the converter), and false for power demand in the converter.

ac_voltage_control_to::Bool: (default: true) Converter control type in the to bus converter. Set true for AC Voltage Control (set AC voltage on the AC side of the converter), and false for fixed power AC factor.

dc_setpoint_to::Float64: (default: 0.0) Converter DC setpoint in the to bus converter. If voltage_control_to = true this number is the DC voltage on the DC side of the converter, entered in kV. If voltage_control_to = false, this value is the power demand in MW, if positive the converter is supplying power to the AC network at the to bus; if negative, the converter is withdrawing power from the AC network at the to bus.

ac_setpoint_to::Float64: (default: 1.0) Converter AC setpoint in the to bus converter. If voltage_control_to = true this number is the AC voltage on the AC side of the converter, entered in per unit. If voltage_control_to = false, this value is the power factor setpoint.

converter_loss_to::Union{LinearCurve, QuadraticCurve}: (default: LinearCurve(0.0)) Loss model coefficients in the to bus converter. It accepts a linear model or quadratic. Same converter data is used in both ends.

max_dc_current_to::Float64: (default: 1e8) Maximum stable dc current limits (A).

rating_to::Float64: (default: 1e8) Converter rating in MVA in the to bus.

reactive_power_limits_to::MinMax: (default: (min=0.0, max=0.0)) Limits on the Reactive Power at the to side.

power_factor_weighting_fraction_to::Float64: (default: 1.0) Power weighting factor fraction used in reducing the active power order and either the reactive power order when the converter rating is violated. When is 0.0, only the active power is reduced; when is 1.0, only the reactive power is reduced; otherwise, a weighted reduction of both active and reactive power is applied., validation range: (0, 1)

voltage_limits_to::MinMax: (default: (min=0.0, max=999.9)) Limits on the Voltage at the DC to Bus.

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::TwoTerminalVSCLine) -> Bool

Get TwoTerminalVSCLine available.

InfrastructureSystems.get_name — Method

get_name(value::TwoTerminalVSCLine) -> String

Get TwoTerminalVSCLine name.

InfrastructureSystems.set_available! — Method

set_available!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine available.

PowerSystems.get_ac_setpoint_from — Method

get_ac_setpoint_from(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine ac_setpoint_from.

PowerSystems.get_ac_setpoint_to — Method

get_ac_setpoint_to(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine ac_setpoint_to.

PowerSystems.get_ac_voltage_control_from — Method

get_ac_voltage_control_from(
    value::TwoTerminalVSCLine
) -> Bool

Get TwoTerminalVSCLine ac_voltage_control_from.

PowerSystems.get_ac_voltage_control_to — Method

get_ac_voltage_control_to(value::TwoTerminalVSCLine) -> Bool

Get TwoTerminalVSCLine ac_voltage_control_to.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine active_power_flow.

PowerSystems.get_active_power_limits_from — Method

get_active_power_limits_from(
    value::TwoTerminalVSCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCLine active_power_limits_from.

PowerSystems.get_active_power_limits_to — Method

get_active_power_limits_to(
    value::TwoTerminalVSCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCLine active_power_limits_to.

PowerSystems.get_arc — Method

get_arc(value::TwoTerminalVSCLine) -> Arc

Get TwoTerminalVSCLine arc.

PowerSystems.get_converter_loss_from — Method

get_converter_loss_from(
    value::TwoTerminalVSCLine
) -> Union{LinearCurve, QuadraticCurve}

Get TwoTerminalVSCLine converter_loss_from.

PowerSystems.get_converter_loss_to — Method

get_converter_loss_to(
    value::TwoTerminalVSCLine
) -> Union{LinearCurve, QuadraticCurve}

Get TwoTerminalVSCLine converter_loss_to.

PowerSystems.get_dc_current — Method

get_dc_current(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine dc_current.

PowerSystems.get_dc_setpoint_from — Method

get_dc_setpoint_from(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine dc_setpoint_from.

PowerSystems.get_dc_setpoint_to — Method

get_dc_setpoint_to(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine dc_setpoint_to.

PowerSystems.get_dc_voltage_control_from — Method

get_dc_voltage_control_from(
    value::TwoTerminalVSCLine
) -> Bool

Get TwoTerminalVSCLine dc_voltage_control_from.

PowerSystems.get_dc_voltage_control_to — Method

get_dc_voltage_control_to(value::TwoTerminalVSCLine) -> Bool

Get TwoTerminalVSCLine dc_voltage_control_to.

PowerSystems.get_ext — Method

get_ext(value::TwoTerminalVSCLine) -> Dict{String, Any}

Get TwoTerminalVSCLine ext.

PowerSystems.get_g — Method

get_g(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine g.

PowerSystems.get_max_dc_current_from — Method

get_max_dc_current_from(
    value::TwoTerminalVSCLine
) -> Float64

Get TwoTerminalVSCLine max_dc_current_from.

PowerSystems.get_max_dc_current_to — Method

get_max_dc_current_to(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine max_dc_current_to.

PowerSystems.get_power_factor_weighting_fraction_from — Method

get_power_factor_weighting_fraction_from(
    value::TwoTerminalVSCLine
) -> Float64

Get TwoTerminalVSCLine power_factor_weighting_fraction_from.

PowerSystems.get_power_factor_weighting_fraction_to — Method

get_power_factor_weighting_fraction_to(
    value::TwoTerminalVSCLine
) -> Float64

Get TwoTerminalVSCLine power_factor_weighting_fraction_to.

PowerSystems.get_rating — Method

get_rating(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine rating.

PowerSystems.get_rating_from — Method

get_rating_from(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine rating_from.

PowerSystems.get_rating_to — Method

get_rating_to(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine rating_to.

PowerSystems.get_reactive_power_from — Method

get_reactive_power_from(
    value::TwoTerminalVSCLine
) -> Float64

Get TwoTerminalVSCLine reactive_power_from.

PowerSystems.get_reactive_power_limits_from — Method

get_reactive_power_limits_from(
    value::TwoTerminalVSCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCLine reactive_power_limits_from.

PowerSystems.get_reactive_power_limits_to — Method

get_reactive_power_limits_to(
    value::TwoTerminalVSCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCLine reactive_power_limits_to.

PowerSystems.get_reactive_power_to — Method

get_reactive_power_to(value::TwoTerminalVSCLine) -> Float64

Get TwoTerminalVSCLine reactive_power_to.

PowerSystems.get_services — Method

get_services(value::TwoTerminalVSCLine) -> Vector{Service}

Get TwoTerminalVSCLine services.

PowerSystems.get_voltage_limits_from — Method

get_voltage_limits_from(
    value::TwoTerminalVSCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCLine voltage_limits_from.

PowerSystems.get_voltage_limits_to — Method

get_voltage_limits_to(
    value::TwoTerminalVSCLine
) -> @NamedTuple{min::Float64, max::Float64}

Get TwoTerminalVSCLine voltage_limits_to.

PowerSystems.set_ac_setpoint_from! — Method

set_ac_setpoint_from!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine ac_setpoint_from.

PowerSystems.set_ac_setpoint_to! — Method

set_ac_setpoint_to!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine ac_setpoint_to.

PowerSystems.set_ac_voltage_control_from! — Method

set_ac_voltage_control_from!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine ac_voltage_control_from.

PowerSystems.set_ac_voltage_control_to! — Method

set_ac_voltage_control_to!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine ac_voltage_control_to.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine active_power_flow.

PowerSystems.set_active_power_limits_from! — Method

set_active_power_limits_from!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine active_power_limits_from.

PowerSystems.set_active_power_limits_to! — Method

set_active_power_limits_to!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine active_power_limits_to.

PowerSystems.set_arc! — Method

set_arc!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine arc.

PowerSystems.set_converter_loss_from! — Method

set_converter_loss_from!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine converter_loss_from.

PowerSystems.set_converter_loss_to! — Method

set_converter_loss_to!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine converter_loss_to.

PowerSystems.set_dc_current! — Method

set_dc_current!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine dc_current.

PowerSystems.set_dc_setpoint_from! — Method

set_dc_setpoint_from!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine dc_setpoint_from.

PowerSystems.set_dc_setpoint_to! — Method

set_dc_setpoint_to!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine dc_setpoint_to.

PowerSystems.set_dc_voltage_control_from! — Method

set_dc_voltage_control_from!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine dc_voltage_control_from.

PowerSystems.set_dc_voltage_control_to! — Method

set_dc_voltage_control_to!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine dc_voltage_control_to.

PowerSystems.set_ext! — Method

set_ext!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine ext.

PowerSystems.set_g! — Method

set_g!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine g.

PowerSystems.set_max_dc_current_from! — Method

set_max_dc_current_from!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine max_dc_current_from.

PowerSystems.set_max_dc_current_to! — Method

set_max_dc_current_to!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine max_dc_current_to.

PowerSystems.set_power_factor_weighting_fraction_from! — Method

set_power_factor_weighting_fraction_from!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine power_factor_weighting_fraction_from.

PowerSystems.set_power_factor_weighting_fraction_to! — Method

set_power_factor_weighting_fraction_to!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine power_factor_weighting_fraction_to.

PowerSystems.set_rating! — Method

set_rating!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine rating.

PowerSystems.set_rating_from! — Method

set_rating_from!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine rating_from.

PowerSystems.set_rating_to! — Method

set_rating_to!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine rating_to.

PowerSystems.set_reactive_power_from! — Method

set_reactive_power_from!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine reactive_power_from.

PowerSystems.set_reactive_power_limits_from! — Method

set_reactive_power_limits_from!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine reactive_power_limits_from.

PowerSystems.set_reactive_power_limits_to! — Method

set_reactive_power_limits_to!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine reactive_power_limits_to.

PowerSystems.set_reactive_power_to! — Method

set_reactive_power_to!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine reactive_power_to.

PowerSystems.set_services! — Method

set_services!(value::TwoTerminalVSCLine, val) -> Any

Set TwoTerminalVSCLine services.

PowerSystems.set_voltage_limits_from! — Method

set_voltage_limits_from!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine voltage_limits_from.

PowerSystems.set_voltage_limits_to! — Method

set_voltage_limits_to!(
    value::TwoTerminalVSCLine,
    val
) -> Any

Set TwoTerminalVSCLine voltage_limits_to.