HybridSystem

PowerSystems.HybridSystem — Type

Representation of a Hybrid System that collects renewable generation, load, thermal generation and storage.

Each generator is a data structure that is defined by the following components:

ThermalGen

Load

Storage

RenewableGen

PowerSystems.get_active_power — Method

get_active_power(value::HybridSystem) -> Any

Get HybridSystem active_power.

PowerSystems.get_available — Method

get_available(value::HybridSystem) -> Bool

Get HybridSystem available.

PowerSystems.get_base_power — Method

get_base_power(value::HybridSystem) -> Float64

Get HybridSystem base_power.

PowerSystems.get_bus — Method

get_bus(value::HybridSystem) -> Bus

Get HybridSystem bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(value::HybridSystem) -> Union{Nothing, DynamicInjection}

Get HybridSystem dynamic_injector.

PowerSystems.get_ext — Method

get_ext(value::HybridSystem) -> Dict{String, Any}

Get HybridSystem ext.

PowerSystems.get_input_active_power_limits — Method

get_input_active_power_limits(value::HybridSystem) -> Union{Nothing, NamedTuple{(:min, :max), _A} where _A<:Tuple{Any, Any}}

Get HybridSystem input_active_power_limits.

PowerSystems.get_operation_cost — Method

get_operation_cost(value::HybridSystem) -> PowerSystems.OperationalCost

Get HybridSystem operation_cost.

PowerSystems.get_output_active_power_limits — Method

get_output_active_power_limits(value::HybridSystem) -> Union{Nothing, NamedTuple{(:min, :max), _A} where _A<:Tuple{Any, Any}}

Get HybridSystem output_active_power_limits.

PowerSystems.get_reactive_power — Method

get_reactive_power(value::HybridSystem) -> Any

Get HybridSystem reactive_power.

PowerSystems.get_reactive_power_limits — Method

get_reactive_power_limits(value::HybridSystem) -> Union{Nothing, NamedTuple{(:min, :max), _A} where _A<:Tuple{Any, Any}}

Get HybridSystem reactive_power_limits.

PowerSystems.get_services — Method

get_services(value::HybridSystem) -> Vector{Service}

Get HybridSystem services.

PowerSystems.get_status — Method

get_status(value::HybridSystem) -> Bool

Get HybridSystem status.

PowerSystems.get_subcomponents — Method

get_subcomponents(hybrid::HybridSystem) -> Channel{Any}

Return an iterator over the subcomponents in the HybridSystem.

Examples

for subcomponent in get_subcomponents(hybrid_sys)
    @show subcomponent
end
subcomponents = collect(get_subcomponents(hybrid_sys))

PowerSystems.set_active_power! — Method

set_active_power!(value::HybridSystem, val) -> Any

Set HybridSystem active_power.

PowerSystems.set_available! — Method

set_available!(value::HybridSystem, val) -> Any

Set HybridSystem available.

PowerSystems.set_base_power! — Method

set_base_power!(value::HybridSystem, val) -> Any

Set HybridSystem base_power.

PowerSystems.set_bus! — Method

set_bus!(value::HybridSystem, val) -> Any

Set HybridSystem bus.

PowerSystems.set_ext! — Method

set_ext!(value::HybridSystem, val) -> Any

Set HybridSystem ext.

PowerSystems.set_input_active_power_limits! — Method

set_input_active_power_limits!(value::HybridSystem, val) -> Any

Set HybridSystem input_active_power_limits.

PowerSystems.set_operation_cost! — Method

set_operation_cost!(value::HybridSystem, val) -> Any

Set HybridSystem operation_cost.

PowerSystems.set_output_active_power_limits! — Method

set_output_active_power_limits!(value::HybridSystem, val) -> Any

Set HybridSystem output_active_power_limits.

PowerSystems.set_reactive_power! — Method

set_reactive_power!(value::HybridSystem, val) -> Any

Set HybridSystem reactive_power.

PowerSystems.set_reactive_power_limits! — Method

set_reactive_power_limits!(value::HybridSystem, val) -> Any

Set HybridSystem reactive_power_limits.

PowerSystems.set_services! — Method

set_services!(value::HybridSystem, val) -> Any

Set HybridSystem services.

PowerSystems.set_status! — Method

set_status!(value::HybridSystem, val) -> Any

Get HybridSystem status.