MarketBidCost

PowerSystems.MarketBidCost — Type

MarketBidCost(
    no_load_cost,
    start_up::Real,
    shut_down
) -> MarketBidCost
MarketBidCost(
    no_load_cost,
    start_up::Real,
    shut_down,
    incremental_offer_curves
) -> MarketBidCost
MarketBidCost(
    no_load_cost,
    start_up::Real,
    shut_down,
    incremental_offer_curves,
    decremental_offer_curves
) -> MarketBidCost
MarketBidCost(
    no_load_cost,
    start_up::Real,
    shut_down,
    incremental_offer_curves,
    decremental_offer_curves,
    ancillary_service_offers
) -> MarketBidCost

Accepts a single start_up value to use as the hot value, with warm and cold set to 0.0.

PowerSystems.MarketBidCost — Type

mutable struct MarketBidCost <: OperationalCost

no_load_cost::Union{Float64, TimeSeriesKey}: No load cost

start_up::Union{@NamedTuple{hot::Float64, warm::Float64, cold::Float64}, TimeSeriesKey}: Start-up cost at different stages of the thermal cycle as the unit cools after a shutdown (e.g., hot, warm, or cold starts). Warm is also referred to as intermediate in some markets. Can also accept a single value if there is only one start-up cost

shut_down::Float64: Shut-down cost

incremental_offer_curves::Union{Nothing, TimeSeriesKey, CostCurve{PiecewiseIncrementalCurve}}: Sell Offer Curves data, which can be a time series or a CostCurve using PiecewiseIncrementalCurve

decremental_offer_curves::Union{Nothing, TimeSeriesKey, CostCurve{PiecewiseIncrementalCurve}}: Buy Offer Curves data, can be a time series or a CostCurve using PiecewiseIncrementalCurve

ancillary_service_offers::Vector{Service}: Bids for the ancillary services

MarketBidCost(no_load_cost, start_up, shut_down, incremental_offer_curves, decremental_offer_curves, ancillary_service_offers)
MarketBidCost(; no_load_cost, start_up, shut_down, incremental_offer_curves, decremental_offer_curves, ancillary_service_offers)
MarketBidCost(no_load_cost, start_up::Real, shut_down, incremental_offer_curves, decremental_offer_curves, ancillary_service_offers)

An operating cost for market bids of energy and ancilliary services for any asset. Compatible with most US Market bidding mechanisms that support demand and generation side.

PowerSystems.get_ancillary_service_offers — Method

get_ancillary_service_offers(
    value::MarketBidCost
) -> Vector{Service}

Get MarketBidCost ancillary_service_offers.

PowerSystems.get_decremental_offer_curves — Method

get_decremental_offer_curves(
    value::MarketBidCost
) -> Union{Nothing, TimeSeriesKey, CostCurve{PiecewiseIncrementalCurve}}

Get MarketBidCost incremental_offer_curves.

PowerSystems.get_incremental_offer_curves — Method

get_incremental_offer_curves(
    value::MarketBidCost
) -> Union{Nothing, TimeSeriesKey, CostCurve{PiecewiseIncrementalCurve}}

Get MarketBidCost incremental_offer_curves.

PowerSystems.get_no_load_cost — Method

get_no_load_cost(
    value::MarketBidCost
) -> Union{Float64, TimeSeriesKey}

Get MarketBidCost no_load_cost.

PowerSystems.get_shut_down — Method

get_shut_down(value::MarketBidCost) -> Float64

Get MarketBidCost shut_down.

PowerSystems.get_start_up — Method

get_start_up(
    value::MarketBidCost
) -> Union{@NamedTuple{hot::Float64, warm::Float64, cold::Float64}, TimeSeriesKey}

Get MarketBidCost start_up.

PowerSystems.make_market_bid_curve — Method

make_market_bid_curve(
    data::PiecewiseStepData;
    power_units
) -> CostCurve{PiecewiseIncrementalCurve}

Make a CostCurve{PiecewiseIncrementalCurve} suitable for inclusion in a MarketBidCost from the FunctionData that might be used to store such a cost curve in a time series.

PowerSystems.make_market_bid_curve — Method

make_market_bid_curve(
    powers::Vector{Float64},
    marginal_costs::Vector{Float64};
    power_units
) -> CostCurve{PiecewiseIncrementalCurve}

Make a CostCurve{PiecewiseIncrementalCurve} suitable for inclusion in a MarketBidCost from a vector of power values, a vector of marginal costs, and an optional units system. The minimum power, and cost at minimum power, are not represented.

PowerSystems.set_ancillary_service_offers! — Method

set_ancillary_service_offers!(
    value::MarketBidCost,
    val
) -> Any

Set MarketBidCost ancillary_service_offers.

PowerSystems.set_decremental_offer_curves! — Method

set_decremental_offer_curves!(
    value::MarketBidCost,
    val
) -> Any

Set MarketBidCost incremental_offer_curves.

PowerSystems.set_incremental_offer_curves! — Method

set_incremental_offer_curves!(
    value::MarketBidCost,
    val
) -> Any

Set MarketBidCost incremental_offer_curves.

PowerSystems.set_no_load_cost! — Method

set_no_load_cost!(value::MarketBidCost, val) -> Any

Set MarketBidCost no_load_cost.

PowerSystems.set_shut_down! — Method

set_shut_down!(value::MarketBidCost, val) -> Any

Set MarketBidCost shut_down.

PowerSystems.set_start_up! — Method

set_start_up!(value::MarketBidCost, val) -> Any

Set MarketBidCost start_up.