RenewableGenerationCost

PowerSystems.RenewableGenerationCost — Type

mutable struct RenewableGenerationCost <: OperationalCost

variable::CostCurve: Variable cost represented as a CostCurve

curtailment_cost::CostCurve: (default of 0) Cost of curtailing power represented as a CostCurve

RenewableGenerationCost(variable, curtailment_cost)
RenewableGenerationCost(variable; curtailment_cost)
RenewableGenerationCost(; variable, curtailment_cost)

An operational cost of renewable generators which includes the variable cost of energy (like a PPA) and the cost of curtailing power. For example, curtailment costs can be used to represent the loss of tax incentives.

The variable cost is a required parameter, but zero(CostCurve) can be used to set it to 0.

PowerSystems.get_curtailment_cost — Method

get_curtailment_cost(
    value::RenewableGenerationCost
) -> CostCurve

Get RenewableGenerationCost curtailment_cost.

PowerSystems.get_variable — Method

get_variable(value::RenewableGenerationCost) -> CostCurve

Get RenewableGenerationCost variable.

PowerSystems.set_curtailment_cost! — Method

set_curtailment_cost!(
    value::RenewableGenerationCost,
    val
) -> Any

Set RenewableGenerationCost curtailment_cost.

PowerSystems.set_variable! — Method

set_variable!(value::RenewableGenerationCost, val) -> Any

Set RenewableGenerationCost variable.