Network Matrices

Use these matrices wisely

PowerSystems.Ybus — Type

Nodal admittance matrix (Ybus) is an N x N matrix describing a power system with N buses. It represents the nodal admittance of the buses in a power system.

The Ybus Struct is indexed using the Bus Numbers, no need for them to be sequential

PowerSystems.Ybus — Method

Ybus(branches::Any, nodes::Any; check_connectivity) -> Ybus

Builds a Ybus from a collection of buses and branches. The return is a Ybus Array indexed with the bus numbers and the branch names.

Keyword arguments

check_connectivity::Bool: Checks connectivity of the network using Goderya's algorithm

PowerSystems.PTDF — Type

Power Transfer Distribution Factors (PTDF) indicate the incremental change in real power that occurs on transmission lines due to real power injections changes at the buses.

The PTDF struct is indexed using the Bus numbers and branch names

PowerSystems.PTDF — Type

PTDF(sys::System) -> PTDF
PTDF(sys::System, dist_slack::Array{Float64,1}) -> PTDF

Builds the PTDF matrix from a system. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerSystems.PTDF — Type

PTDF(branches::Any, nodes::Any) -> PTDF
PTDF(branches::Any, nodes::Any, dist_slack::Array{Float64,1}) -> PTDF

Builds the PTDF matrix from a group of branches and nodes. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerSystems.LODF — Type

Line Outage Distribution Factors (LODFs) are a sensitivity measure of how a change in a line’s flow affects the flows on other lines in the system.

PowerSystems.LODF — Type

LODF(sys::System) -> LODF
LODF(sys::System, dist_slack::Array{Float64,1}) -> LODF

Builds the LODF matrix from a system. The return is a LOLDF array indexed with the branch name.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerSystems.LODF — Type

LODF(branches::Any, nodes::Any) -> LODF
LODF(branches::Any, nodes::Any, dist_slack::Array{Float64,1}) -> LODF

Builds the LODF matrix from a group of branches and nodes. The return is a LOLDF array indexed with the branch name.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses