System

The container of all system components is System.

This Section is under construction, refer to the # Constructing a System from raw dat section.

PowerSystems.System — Type

System

A power system defined by fields for base_power, components, and forecasts.

System(base_power)
System(components, base_power)
System(buses, generators, loads, branches, storage, base_power, services, annex; kwargs...)
System(buses, generators, loads, base_power; kwargs...)
System(file; kwargs...)
System(; buses, generators, loads, branches, storage, base_power, services, annex, kwargs...)
System(; kwargs...)

Arguments

buses::Vector{Bus}: an array of buses

generators::Vector{Generator}: an array of generators of (possibly) different types

loads::Vector{ElectricLoad}: an array of load specifications that includes timing of the loads

branches::Union{Nothing, Vector{Branch}}: an array of branches; may be nothing

storage::Union{Nothing, Vector{Storage}}: an array of storage devices; may be nothing

base_power::Float64: the base power value for the system

services::Union{Nothing, Vector{<:Service}}: an array of services; may be nothing

Keyword arguments

runchecks::Bool: Run available checks on input fields and when add_component! is called. Throws InvalidRange if an error is found.

time_series_in_memory::Bool=false: Store time series data in memory instead of HDF5.

config_path::String: specify path to validation config file