AreaInterchange

PowerSystems.AreaInterchange — Type

mutable struct AreaInterchange <: Branch
    name::String
    available::Bool
    active_power_flow::Float64
    from_area::Area
    to_area::Area
    flow_limits::FromTo_ToFrom
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

Flow exchanged between Areas. This Interchange is agnostic to the lines connecting the areas. It does not substitute Interface which is the total flow across a group of lines

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow::Float64: Initial condition of active power flow on the line (MW)

from_area::Area: Area from which the power is extracted

to_area::Area: Area to which the power is injected

flow_limits::FromTo_ToFrom: Max flow between the areas. It ignores lines and other branches totals

services::Vector{Service}: (default: Service[]) Service interfaces that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::AreaInterchange) -> Bool

Get AreaInterchange available.

InfrastructureSystems.get_name — Method

get_name(value::AreaInterchange) -> String

Get AreaInterchange name.

InfrastructureSystems.set_available! — Method

set_available!(value::AreaInterchange, val) -> Any

Set AreaInterchange available.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::AreaInterchange) -> Float64

Get AreaInterchange active_power_flow.

PowerSystems.get_ext — Method

get_ext(value::AreaInterchange) -> Dict{String, Any}

Get AreaInterchange ext.

PowerSystems.get_flow_limits — Method

get_flow_limits(
    value::AreaInterchange
) -> @NamedTuple{from_to::Float64, to_from::Float64}

Get AreaInterchange flow_limits.

PowerSystems.get_from_area — Method

get_from_area(value::AreaInterchange) -> Area

Get AreaInterchange from_area.

PowerSystems.get_services — Method

get_services(value::AreaInterchange) -> Vector{Service}

Get AreaInterchange services.

PowerSystems.get_to_area — Method

get_to_area(value::AreaInterchange) -> Area

Get AreaInterchange to_area.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(value::AreaInterchange, val) -> Any

Set AreaInterchange active_power_flow.

PowerSystems.set_ext! — Method

set_ext!(value::AreaInterchange, val) -> Any

Set AreaInterchange ext.

PowerSystems.set_flow_limits! — Method

set_flow_limits!(value::AreaInterchange, val) -> Any

Set AreaInterchange flow_limits.

PowerSystems.set_from_area! — Method

set_from_area!(value::AreaInterchange, val) -> Any

Set AreaInterchange from_area.

PowerSystems.set_services! — Method

set_services!(value::AreaInterchange, val) -> Any

Set AreaInterchange services.

PowerSystems.set_to_area! — Method

set_to_area!(value::AreaInterchange, val) -> Any

Set AreaInterchange to_area.