Converter

AverageConverter

PowerSystems.AverageConverter — Type

mutable struct AverageConverter <: Converter
    rated_voltage::Float64
    rated_current::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of an average converter model

Arguments

rated_voltage::Float64: Rated voltage (V), validation range: (0, nothing)

rated_current::Float64: Rated current (A), validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

states::Vector{Symbol}: (Do not modify.) AverageConverter has no states

n_states::Int: (Do not modify.) AverageConverter has no states

PowerSystems.get_ext — Method

get_ext(value::AverageConverter) -> Dict{String, Any}

Get AverageConverter ext.

PowerSystems.get_n_states — Method

get_n_states(value::AverageConverter) -> Int64

Get AverageConverter n_states.

PowerSystems.get_rated_current — Method

get_rated_current(value::AverageConverter) -> Float64

Get AverageConverter rated_current.

PowerSystems.get_rated_voltage — Method

get_rated_voltage(value::AverageConverter) -> Float64

Get AverageConverter rated_voltage.

PowerSystems.get_states — Method

get_states(value::AverageConverter) -> Vector{Symbol}

Get AverageConverter states.

PowerSystems.set_ext! — Method

set_ext!(value::AverageConverter, val) -> Any

Set AverageConverter ext.

PowerSystems.set_rated_current! — Method

set_rated_current!(value::AverageConverter, val) -> Any

Set AverageConverter rated_current.

PowerSystems.set_rated_voltage! — Method

set_rated_voltage!(value::AverageConverter, val) -> Any

Set AverageConverter rated_voltage.

RenewableEnergyConverterTypeA

PowerSystems.RenewableEnergyConverterTypeA — Type

mutable struct RenewableEnergyConverterTypeA <: Converter
    T_g::Float64
    Rrpwr::Float64
    Brkpt::Float64
    Zerox::Float64
    Lvpl1::Float64
    Vo_lim::Float64
    Lv_pnts::MinMax
    Io_lim::Float64
    T_fltr::Float64
    K_hv::Float64
    Iqr_lims::MinMax
    Accel::Float64
    Lvpl_sw::Int
    Q_ref::Float64
    R_source::Float64
    X_source::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a renewable energy generator/converter model, this model corresponds to REGCA1 in PSSE

Arguments

T_g::Float64: Converter time constant (s), validation range: (0, nothing)

Rrpwr::Float64: Low Voltage Power Logic (LVPL) ramp rate limit (pu/s), validation range: (0, nothing)

Brkpt::Float64: LVPL characteristic voltage 2 (pu), validation range: (0, nothing)

Zerox::Float64: LVPL characteristic voltage 1 (pu), validation range: (0, nothing)

Lvpl1::Float64: LVPL gain (pu), validation range: (0, nothing)

Vo_lim::Float64: Voltage limit for high voltage reactive current management (pu), validation range: (0, nothing)

Lv_pnts::MinMax: Voltage points for low voltage active current management (pu) (Lvpnt0, Lvpnt1)

Io_lim::Float64: Current limit (pu) for high voltage reactive current management (specified as a negative value), validation range: (nothing, 0)

T_fltr::Float64: Voltage filter time constant for low voltage active current management (s), validation range: (0, nothing)

K_hv::Float64: Overvoltage compensation gain used in the high voltage reactive current management, validation range: (0, nothing)

Iqr_lims::MinMax: Limit on rate of change for reactive current (pu/s) (Iqrmin, Iqrmax)

Accel::Float64: Acceleration factor, validation range: (0, 1)

Lvpl_sw::Int: Low voltage power logic (LVPL) switch. (0: LVPL not present, 1: LVPL present), validation range: (0, 1)

Q_ref::Float64: (default: 1.0) Initial condition of reactive power from power flow, validation range: (0, nothing)

R_source::Float64: (default: 0.0) Output resistor used for the Thevenin Equivalent, validation range: (0, nothing)

X_source::Float64: (default: 1.0e5) Output reactance used for the Thevenin Equivalent, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

states::Vector{Symbol}: (Do not modify.) The states are: Ip: Converter lag for Ipcmd, Iq: Converter lag for Iqcmd, Vmeas: Voltage filter for low voltage active current management

n_states::Int: (Do not modify.) RenewableEnergyConverterTypeA has 3 states

PowerSystems.get_Accel — Method

get_Accel(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA Accel.

PowerSystems.get_Brkpt — Method

get_Brkpt(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA Brkpt.

PowerSystems.get_Io_lim — Method

get_Io_lim(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA Io_lim.

PowerSystems.get_Iqr_lims — Method

get_Iqr_lims(
    value::RenewableEnergyConverterTypeA
) -> @NamedTuple{min::Float64, max::Float64}

Get RenewableEnergyConverterTypeA Iqr_lims.

PowerSystems.get_K_hv — Method

get_K_hv(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA K_hv.

PowerSystems.get_Lv_pnts — Method

get_Lv_pnts(
    value::RenewableEnergyConverterTypeA
) -> @NamedTuple{min::Float64, max::Float64}

Get RenewableEnergyConverterTypeA Lv_pnts.

PowerSystems.get_Lvpl1 — Method

get_Lvpl1(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA Lvpl1.

PowerSystems.get_Lvpl_sw — Method

get_Lvpl_sw(value::RenewableEnergyConverterTypeA) -> Int64

Get RenewableEnergyConverterTypeA Lvpl_sw.

PowerSystems.get_Q_ref — Method

get_Q_ref(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA Q_ref.

PowerSystems.get_R_source — Method

get_R_source(
    value::RenewableEnergyConverterTypeA
) -> Float64

Get RenewableEnergyConverterTypeA R_source.

PowerSystems.get_Rrpwr — Method

get_Rrpwr(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA Rrpwr.

PowerSystems.get_T_fltr — Method

get_T_fltr(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA T_fltr.

PowerSystems.get_T_g — Method

get_T_g(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA T_g.

PowerSystems.get_Vo_lim — Method

get_Vo_lim(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA Vo_lim.

PowerSystems.get_X_source — Method

get_X_source(
    value::RenewableEnergyConverterTypeA
) -> Float64

Get RenewableEnergyConverterTypeA X_source.

PowerSystems.get_Zerox — Method

get_Zerox(value::RenewableEnergyConverterTypeA) -> Float64

Get RenewableEnergyConverterTypeA Zerox.

PowerSystems.get_ext — Method

get_ext(
    value::RenewableEnergyConverterTypeA
) -> Dict{String, Any}

Get RenewableEnergyConverterTypeA ext.

PowerSystems.get_n_states — Method

get_n_states(value::RenewableEnergyConverterTypeA) -> Int64

Get RenewableEnergyConverterTypeA n_states.

PowerSystems.get_states — Method

get_states(
    value::RenewableEnergyConverterTypeA
) -> Vector{Symbol}

Get RenewableEnergyConverterTypeA states.

PowerSystems.set_Accel! — Method

set_Accel!(value::RenewableEnergyConverterTypeA, val) -> Any

Set RenewableEnergyConverterTypeA Accel.

PowerSystems.set_Brkpt! — Method

set_Brkpt!(value::RenewableEnergyConverterTypeA, val) -> Any

Set RenewableEnergyConverterTypeA Brkpt.

PowerSystems.set_Io_lim! — Method

set_Io_lim!(
    value::RenewableEnergyConverterTypeA,
    val
) -> Any

Set RenewableEnergyConverterTypeA Io_lim.

PowerSystems.set_Iqr_lims! — Method

set_Iqr_lims!(
    value::RenewableEnergyConverterTypeA,
    val
) -> Any

Set RenewableEnergyConverterTypeA Iqr_lims.

PowerSystems.set_K_hv! — Method

set_K_hv!(value::RenewableEnergyConverterTypeA, val) -> Any

Set RenewableEnergyConverterTypeA K_hv.

PowerSystems.set_Lv_pnts! — Method

set_Lv_pnts!(
    value::RenewableEnergyConverterTypeA,
    val
) -> Any

Set RenewableEnergyConverterTypeA Lv_pnts.

PowerSystems.set_Lvpl1! — Method

set_Lvpl1!(value::RenewableEnergyConverterTypeA, val) -> Any

Set RenewableEnergyConverterTypeA Lvpl1.

PowerSystems.set_Lvpl_sw! — Method

set_Lvpl_sw!(
    value::RenewableEnergyConverterTypeA,
    val
) -> Any

Set RenewableEnergyConverterTypeA Lvpl_sw.

PowerSystems.set_Q_ref! — Method

set_Q_ref!(value::RenewableEnergyConverterTypeA, val) -> Any

Set RenewableEnergyConverterTypeA Q_ref.

PowerSystems.set_R_source! — Method

set_R_source!(
    value::RenewableEnergyConverterTypeA,
    val
) -> Any

Set RenewableEnergyConverterTypeA R_source.

PowerSystems.set_Rrpwr! — Method

set_Rrpwr!(value::RenewableEnergyConverterTypeA, val) -> Any

Set RenewableEnergyConverterTypeA Rrpwr.

PowerSystems.set_T_fltr! — Method

set_T_fltr!(
    value::RenewableEnergyConverterTypeA,
    val
) -> Any

Set RenewableEnergyConverterTypeA T_fltr.

PowerSystems.set_T_g! — Method

set_T_g!(value::RenewableEnergyConverterTypeA, val) -> Any

Set RenewableEnergyConverterTypeA T_g.

PowerSystems.set_Vo_lim! — Method

set_Vo_lim!(
    value::RenewableEnergyConverterTypeA,
    val
) -> Any

Set RenewableEnergyConverterTypeA Vo_lim.

PowerSystems.set_X_source! — Method

set_X_source!(
    value::RenewableEnergyConverterTypeA,
    val
) -> Any

Set RenewableEnergyConverterTypeA X_source.

PowerSystems.set_Zerox! — Method

set_Zerox!(value::RenewableEnergyConverterTypeA, val) -> Any

Set RenewableEnergyConverterTypeA Zerox.

PowerSystems.set_ext! — Method

set_ext!(value::RenewableEnergyConverterTypeA, val) -> Any

Set RenewableEnergyConverterTypeA ext.

RenewableEnergyVoltageConverterTypeA

PowerSystems.RenewableEnergyVoltageConverterTypeA — Type

mutable struct RenewableEnergyVoltageConverterTypeA <: Converter
    T_g::Float64
    Rrpwr::Float64
    Brkpt::Float64
    Zerox::Float64
    Lvpl1::Float64
    Vo_lim::Float64
    Lv_pnts::MinMax
    Io_lim::Float64
    T_fltr::Float64
    K_hv::Float64
    Iqr_lims::MinMax
    Accel::Float64
    Lvpl_sw::Int
    Q_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a renewable energy generator/converter model, this model corresponds to REGCA1 in PSSE, but to be interfaced using a Voltage Source instead of a Current Source

Arguments

T_g::Float64: Converter time constant (s), validation range: (0, nothing)

Rrpwr::Float64: Low Voltage Power Logic (LVPL) ramp rate limit (pu/s), validation range: (0, nothing)

Brkpt::Float64: LVPL characteristic voltage 2 (pu), validation range: (0, nothing)

Zerox::Float64: LVPL characteristic voltage 1 (pu), validation range: (0, nothing)

Lvpl1::Float64: LVPL gain (pu), validation range: (0, nothing)

Vo_lim::Float64: Voltage limit for high voltage reactive current management (pu), validation range: (0, nothing)

Lv_pnts::MinMax: Voltage points for low voltage active current management (pu) (Lvpnt0, Lvpnt1)

Io_lim::Float64: Current limit (pu) for high voltage reactive current management (specified as a negative value), validation range: (nothing, 0)

T_fltr::Float64: Voltage filter time constant for low voltage active current management (s), validation range: (0, nothing)

K_hv::Float64: Overvoltage compensation gain used in the high voltage reactive current management, validation range: (0, nothing)

Iqr_lims::MinMax: Limit on rate of change for reactive current (pu/s) (Iqrmin, Iqrmax)

Accel::Float64: Acceleration factor, validation range: (0, 1)

Lvpl_sw::Int: Low voltage power logic (LVPL) switch. (0: LVPL not present, 1: LVPL present), validation range: (0, 1)

Q_ref::Float64: (default: 1.0) Initial condition of reactive power from power flow, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

states::Vector{Symbol}: (Do not modify.) The states are: Ip: Converter lag for Ipcmd, Iq: Converter lag for Iqcmd, Vmeas: Voltage filter for low voltage active current management

n_states::Int: (Do not modify.) RenewableEnergyVoltageConverterTypeA has 3 states

PowerSystems.get_Accel — Method

get_Accel(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA Accel.

PowerSystems.get_Brkpt — Method

get_Brkpt(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA Brkpt.

PowerSystems.get_Io_lim — Method

get_Io_lim(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA Io_lim.

PowerSystems.get_Iqr_lims — Method

get_Iqr_lims(
    value::RenewableEnergyVoltageConverterTypeA
) -> @NamedTuple{min::Float64, max::Float64}

Get RenewableEnergyVoltageConverterTypeA Iqr_lims.

PowerSystems.get_K_hv — Method

get_K_hv(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA K_hv.

PowerSystems.get_Lv_pnts — Method

get_Lv_pnts(
    value::RenewableEnergyVoltageConverterTypeA
) -> @NamedTuple{min::Float64, max::Float64}

Get RenewableEnergyVoltageConverterTypeA Lv_pnts.

PowerSystems.get_Lvpl1 — Method

get_Lvpl1(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA Lvpl1.

PowerSystems.get_Lvpl_sw — Method

get_Lvpl_sw(
    value::RenewableEnergyVoltageConverterTypeA
) -> Int64

Get RenewableEnergyVoltageConverterTypeA Lvpl_sw.

PowerSystems.get_Q_ref — Method

get_Q_ref(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA Q_ref.

PowerSystems.get_Rrpwr — Method

get_Rrpwr(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA Rrpwr.

PowerSystems.get_T_fltr — Method

get_T_fltr(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA T_fltr.

PowerSystems.get_T_g — Method

get_T_g(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA T_g.

PowerSystems.get_Vo_lim — Method

get_Vo_lim(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA Vo_lim.

PowerSystems.get_Zerox — Method

get_Zerox(
    value::RenewableEnergyVoltageConverterTypeA
) -> Float64

Get RenewableEnergyVoltageConverterTypeA Zerox.

PowerSystems.get_ext — Method

get_ext(
    value::RenewableEnergyVoltageConverterTypeA
) -> Dict{String, Any}

Get RenewableEnergyVoltageConverterTypeA ext.

PowerSystems.get_n_states — Method

get_n_states(
    value::RenewableEnergyVoltageConverterTypeA
) -> Int64

Get RenewableEnergyVoltageConverterTypeA n_states.

PowerSystems.get_states — Method

get_states(
    value::RenewableEnergyVoltageConverterTypeA
) -> Vector{Symbol}

Get RenewableEnergyVoltageConverterTypeA states.

PowerSystems.set_Accel! — Method

set_Accel!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Accel.

PowerSystems.set_Brkpt! — Method

set_Brkpt!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Brkpt.

PowerSystems.set_Io_lim! — Method

set_Io_lim!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Io_lim.

PowerSystems.set_Iqr_lims! — Method

set_Iqr_lims!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Iqr_lims.

PowerSystems.set_K_hv! — Method

set_K_hv!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA K_hv.

PowerSystems.set_Lv_pnts! — Method

set_Lv_pnts!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Lv_pnts.

PowerSystems.set_Lvpl1! — Method

set_Lvpl1!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Lvpl1.

PowerSystems.set_Lvpl_sw! — Method

set_Lvpl_sw!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Lvpl_sw.

PowerSystems.set_Q_ref! — Method

set_Q_ref!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Q_ref.

PowerSystems.set_Rrpwr! — Method

set_Rrpwr!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Rrpwr.

PowerSystems.set_T_fltr! — Method

set_T_fltr!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA T_fltr.

PowerSystems.set_T_g! — Method

set_T_g!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA T_g.

PowerSystems.set_Vo_lim! — Method

set_Vo_lim!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Vo_lim.

PowerSystems.set_Zerox! — Method

set_Zerox!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA Zerox.

PowerSystems.set_ext! — Method

set_ext!(
    value::RenewableEnergyVoltageConverterTypeA,
    val
) -> Any

Set RenewableEnergyVoltageConverterTypeA ext.