LoadZone

PowerSystems.LoadZone — Type

mutable struct LoadZone <: AggregationTopology
    name::String
    peak_active_power::Float64
    peak_reactive_power::Float64
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

A load zone for electricity price analysis.

The load zone can be specified when defining each ACBus or DCBus in the zone

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

peak_active_power::Float64: Peak active power in the zone (MW)

peak_reactive_power::Float64: Peak reactive power in the zone (MVAR)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::LoadZone) -> String

Get LoadZone name.

PowerSystems.get_ext — Method

get_ext(value::LoadZone) -> Dict{String, Any}

Get LoadZone ext.

PowerSystems.get_peak_active_power — Method

get_peak_active_power(value::LoadZone) -> Any

Get LoadZone peak_active_power.

PowerSystems.get_peak_reactive_power — Method

get_peak_reactive_power(value::LoadZone) -> Any

Get LoadZone peak_reactive_power.

PowerSystems.set_ext! — Method

set_ext!(value::LoadZone, val) -> Any

Set LoadZone ext.

PowerSystems.set_peak_active_power! — Method

set_peak_active_power!(value::LoadZone, val) -> Any

Set LoadZone peak_active_power.

PowerSystems.set_peak_reactive_power! — Method

set_peak_reactive_power!(value::LoadZone, val) -> Any

Set LoadZone peak_reactive_power.