Machine

AndersonFouadMachine

PowerSystems.AndersonFouadMachine — Type

mutable struct AndersonFouadMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xq_pp::Float64
    Td0_p::Float64
    Tq0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 6-states synchronous machine: Anderson-Fouad model

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_pp::Float64: Sub-Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

ψq: q-axis stator flux,
ψd: d-axis stator flux,
eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage,
eq_pp: q-axis subtransient voltage,
ed_pp: d-axis subtransient voltage

n_states::Int: AndersonFouadMachine has 6 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine R.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Td0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Td0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Tq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Tq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Xd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Xd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Xd_pp.

PowerSystems.get_Xq — Method

get_Xq(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Xq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Xq_p.

PowerSystems.get_Xq_pp — Method

get_Xq_pp(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachine Xq_pp.

PowerSystems.get_ext — Method

get_ext(value::AndersonFouadMachine) -> Dict{String, Any}

Get AndersonFouadMachine ext.

PowerSystems.get_n_states — Method

get_n_states(value::AndersonFouadMachine) -> Int64

Get AndersonFouadMachine n_states.

PowerSystems.get_states — Method

get_states(value::AndersonFouadMachine) -> Vector{Symbol}

Get AndersonFouadMachine states.

PowerSystems.set_R! — Method

set_R!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine R.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Td0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Td0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Tq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Tq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Xd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Xd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Xd_pp.

PowerSystems.set_Xq! — Method

set_Xq!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Xq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Xq_p.

PowerSystems.set_Xq_pp! — Method

set_Xq_pp!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine Xq_pp.

PowerSystems.set_ext! — Method

set_ext!(value::AndersonFouadMachine, val::Any) -> Any

Set AndersonFouadMachine ext.

BaseMachine

PowerSystems.BaseMachine — Type

mutable struct BaseMachine <: Machine
    R::Float64
    Xd_p::Float64
    eq_p::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of a Classic Machine: GENCLS in PSSE and PSLF

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd_p::Float64: Reactance after EMF in machine per unit, validation range: (0, nothing)

eq_p::Float64: Fixed EMF behind the impedance, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: BaseMachine has no states

n_states::Int: BaseMachine has no states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::BaseMachine) -> Float64

Get BaseMachine R.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::BaseMachine) -> Float64

Get BaseMachine Xd_p.

PowerSystems.get_eq_p — Method

get_eq_p(value::BaseMachine) -> Float64

Get BaseMachine eq_p.

PowerSystems.get_ext — Method

get_ext(value::BaseMachine) -> Dict{String, Any}

Get BaseMachine ext.

PowerSystems.get_n_states — Method

get_n_states(value::BaseMachine) -> Int64

Get BaseMachine n_states.

PowerSystems.get_states — Method

get_states(value::BaseMachine) -> Vector{Symbol}

Get BaseMachine states.

PowerSystems.set_R! — Method

set_R!(value::BaseMachine, val::Any) -> Any

Set BaseMachine R.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::BaseMachine, val::Any) -> Any

Set BaseMachine Xd_p.

PowerSystems.set_eq_p! — Method

set_eq_p!(value::BaseMachine, val::Any) -> Any

Set BaseMachine eq_p.

PowerSystems.set_ext! — Method

set_ext!(value::BaseMachine, val::Any) -> Any

Set BaseMachine ext.

FullMachine

PowerSystems.FullMachine — Type

mutable struct FullMachine <: Machine
    R::Float64
    R_f::Float64
    R_1d::Float64
    R_1q::Float64
    L_d::Float64
    L_q::Float64
    L_ad::Float64
    L_aq::Float64
    L_f1d::Float64
    L_ff::Float64
    L_1d::Float64
    L_1q::Float64
    ext::Dict{String, Any}
    inv_d_fluxlink::Array{Float64,2}
    inv_q_fluxlink::Array{Float64,2}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameter of a full order flux stator-rotor model without zero sequence flux in the stator. The derivative of stator fluxes (ψd and ψq) is NOT neglected. Only one q-axis damping circuit is considered. All parameters are in machine per unit. Refer to Chapter 3 of Power System Stability and Control by P. Kundur or Chapter 11 of Power System Dynamics: Stability and Control, by J. Machowski, J. Bialek and J. Bumby, for more details. Note that the models are somewhat different (but equivalent) due to the different Park Transformation used in both books.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

R_f::Float64: Field rotor winding resistance in per unit, validation range: (0, nothing)

R_1d::Float64: Damping rotor winding resistance on d-axis in per unit. This value is denoted as RD in Machowski., validation range: (0, nothing)

R_1q::Float64: Damping rotor winding resistance on q-axis in per unit. This value is denoted as RQ in Machowski., validation range: (0, nothing)

L_d::Float64: Inductance of fictitious damping that represent the effect of the three-phase stator winding in the d-axis of the rotor, in per unit. This value is denoted as Lad + Ll in Kundur (and Ld in Machowski)., validation range: (0, nothing)

L_q::Float64: Inductance of fictitious damping that represent the effect of the three-phase stator winding in the q-axis of the rotor, in per unit. This value is denoted as Laq + Ll in Kundur., validation range: (0, nothing)

L_ad::Float64: Mutual inductance between stator winding and rotor field (and damping) winding inductance on d-axis, in per unit, validation range: (0, nothing)

L_aq::Float64: Mutual inductance between stator winding and rotor damping winding inductance on q-axis, in per unit, validation range: (0, nothing)

L_f1d::Float64: Mutual inductance between rotor field winding and rotor damping winding inductance on d-axis, in per unit, validation range: (0, nothing)

L_ff::Float64: Field rotor winding inductance, in per unit, validation range: (0, nothing)

L_1d::Float64: Inductance of the d-axis rotor damping circuit, in per unit, validation range: (0, nothing)

L_1q::Float64: Inductance of the q-axis rotor damping circuit, in per unit, validation range: (0, nothing)

ext::Dict{String, Any}

inv_d_fluxlink::Array{Float64,2}: Equations 3.127, 3.130, 3.131 From Kundur

inv_q_fluxlink::Array{Float64,2}: Equations 3.128, 3.132 From Kundur

states::Vector{Symbol}: The states are:

ψd: d-axis stator flux,
ψq: q-axis stator flux,
ψf: field rotor flux,
ψ1d: d-axis rotor damping flux,
ψ1q: q-axis rotor damping flux

n_states::Int: FullMachine has 5 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_L_1d — Method

get_L_1d(value::FullMachine) -> Float64

Get FullMachine L_1d.

PowerSystems.get_L_1q — Method

get_L_1q(value::FullMachine) -> Float64

Get FullMachine L_1q.

PowerSystems.get_L_ad — Method

get_L_ad(value::FullMachine) -> Float64

Get FullMachine L_ad.

PowerSystems.get_L_aq — Method

get_L_aq(value::FullMachine) -> Float64

Get FullMachine L_aq.

PowerSystems.get_L_d — Method

get_L_d(value::FullMachine) -> Float64

Get FullMachine L_d.

PowerSystems.get_L_f1d — Method

get_L_f1d(value::FullMachine) -> Float64

Get FullMachine L_f1d.

PowerSystems.get_L_ff — Method

get_L_ff(value::FullMachine) -> Float64

Get FullMachine L_ff.

PowerSystems.get_L_q — Method

get_L_q(value::FullMachine) -> Float64

Get FullMachine L_q.

PowerSystems.get_R — Method

get_R(value::FullMachine) -> Float64

Get FullMachine R.

PowerSystems.get_R_1d — Method

get_R_1d(value::FullMachine) -> Float64

Get FullMachine R_1d.

PowerSystems.get_R_1q — Method

get_R_1q(value::FullMachine) -> Float64

Get FullMachine R_1q.

PowerSystems.get_R_f — Method

get_R_f(value::FullMachine) -> Float64

Get FullMachine R_f.

PowerSystems.get_ext — Method

get_ext(value::FullMachine) -> Dict{String, Any}

Get FullMachine ext.

PowerSystems.get_inv_d_fluxlink — Method

get_inv_d_fluxlink(value::FullMachine) -> Matrix{Float64}

Get FullMachine inv_d_fluxlink.

PowerSystems.get_inv_q_fluxlink — Method

get_inv_q_fluxlink(value::FullMachine) -> Matrix{Float64}

Get FullMachine inv_q_fluxlink.

PowerSystems.get_n_states — Method

get_n_states(value::FullMachine) -> Int64

Get FullMachine n_states.

PowerSystems.get_states — Method

get_states(value::FullMachine) -> Vector{Symbol}

Get FullMachine states.

PowerSystems.set_L_1d! — Method

set_L_1d!(value::FullMachine, val::Any) -> Any

Set FullMachine L_1d.

PowerSystems.set_L_1q! — Method

set_L_1q!(value::FullMachine, val::Any) -> Any

Set FullMachine L_1q.

PowerSystems.set_L_ad! — Method

set_L_ad!(value::FullMachine, val::Any) -> Any

Set FullMachine L_ad.

PowerSystems.set_L_aq! — Method

set_L_aq!(value::FullMachine, val::Any) -> Any

Set FullMachine L_aq.

PowerSystems.set_L_d! — Method

set_L_d!(value::FullMachine, val::Any) -> Any

Set FullMachine L_d.

PowerSystems.set_L_f1d! — Method

set_L_f1d!(value::FullMachine, val::Any) -> Any

Set FullMachine L_f1d.

PowerSystems.set_L_ff! — Method

set_L_ff!(value::FullMachine, val::Any) -> Any

Set FullMachine L_ff.

PowerSystems.set_L_q! — Method

set_L_q!(value::FullMachine, val::Any) -> Any

Set FullMachine L_q.

PowerSystems.set_R! — Method

set_R!(value::FullMachine, val::Any) -> Any

Set FullMachine R.

PowerSystems.set_R_1d! — Method

set_R_1d!(value::FullMachine, val::Any) -> Any

Set FullMachine R_1d.

PowerSystems.set_R_1q! — Method

set_R_1q!(value::FullMachine, val::Any) -> Any

Set FullMachine R_1q.

PowerSystems.set_R_f! — Method

set_R_f!(value::FullMachine, val::Any) -> Any

Set FullMachine R_f.

PowerSystems.set_ext! — Method

set_ext!(value::FullMachine, val::Any) -> Any

Set FullMachine ext.

PowerSystems.set_inv_d_fluxlink! — Method

set_inv_d_fluxlink!(value::FullMachine, val::Any) -> Any

Set FullMachine inv_d_fluxlink.

PowerSystems.set_inv_q_fluxlink! — Method

set_inv_q_fluxlink!(value::FullMachine, val::Any) -> Any

Set FullMachine inv_q_fluxlink.

MarconatoMachine

PowerSystems.MarconatoMachine — Type

mutable struct MarconatoMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xq_pp::Float64
    Td0_p::Float64
    Tq0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    T_AA::Float64
    ext::Dict{String, Any}
    γd::Float64
    γq::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 6-states synchronous machine: Marconato model

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_pp::Float64: Sub-Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

T_AA::Float64: Time constant of d-axis additional leakage, validation range: (0, nothing)

ext::Dict{String, Any}

γd::Float64

γq::Float64

states::Vector{Symbol}: The states are:

ψq: q-axis stator flux,
ψd: d-axis stator flux,
eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage,
eq_pp: q-axis subtransient voltage,
ed_pp: d-axis subtransient voltage

n_states::Int: MarconatoMachine has 6 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::MarconatoMachine) -> Float64

Get MarconatoMachine R.

PowerSystems.get_T_AA — Method

get_T_AA(value::MarconatoMachine) -> Float64

Get MarconatoMachine T_AA.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::MarconatoMachine) -> Float64

Get MarconatoMachine Td0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::MarconatoMachine) -> Float64

Get MarconatoMachine Td0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::MarconatoMachine) -> Float64

Get MarconatoMachine Tq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::MarconatoMachine) -> Float64

Get MarconatoMachine Tq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::MarconatoMachine) -> Float64

Get MarconatoMachine Xd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::MarconatoMachine) -> Float64

Get MarconatoMachine Xd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::MarconatoMachine) -> Float64

Get MarconatoMachine Xd_pp.

PowerSystems.get_Xq — Method

get_Xq(value::MarconatoMachine) -> Float64

Get MarconatoMachine Xq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::MarconatoMachine) -> Float64

Get MarconatoMachine Xq_p.

PowerSystems.get_Xq_pp — Method

get_Xq_pp(value::MarconatoMachine) -> Float64

Get MarconatoMachine Xq_pp.

PowerSystems.get_ext — Method

get_ext(value::MarconatoMachine) -> Dict{String, Any}

Get MarconatoMachine ext.

PowerSystems.get_n_states — Method

get_n_states(value::MarconatoMachine) -> Int64

Get MarconatoMachine n_states.

PowerSystems.get_states — Method

get_states(value::MarconatoMachine) -> Vector{Symbol}

Get MarconatoMachine states.

PowerSystems.get_γd — Method

get_γd(value::MarconatoMachine) -> Float64

Get MarconatoMachine γd.

PowerSystems.get_γq — Method

get_γq(value::MarconatoMachine) -> Float64

Get MarconatoMachine γq.

PowerSystems.set_R! — Method

set_R!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine R.

PowerSystems.set_T_AA! — Method

set_T_AA!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine T_AA.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Td0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Td0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Tq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Tq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Xd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Xd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Xd_pp.

PowerSystems.set_Xq! — Method

set_Xq!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Xq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Xq_p.

PowerSystems.set_Xq_pp! — Method

set_Xq_pp!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine Xq_pp.

PowerSystems.set_ext! — Method

set_ext!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine ext.

PowerSystems.set_γd! — Method

set_γd!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine γd.

PowerSystems.set_γq! — Method

set_γq!(value::MarconatoMachine, val::Any) -> Any

Set MarconatoMachine γq.

OneDOneQMachine

PowerSystems.OneDOneQMachine — Type

mutable struct OneDOneQMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Td0_p::Float64
    Tq0_p::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 4-states synchronous machine: Simplified Marconato model The derivative of stator fluxes (ψd and ψq) is neglected and ωψd = ψd and ωψq = ψq is assumed (i.e. ω=1.0). This is standard when transmission network dynamics is neglected.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage

n_states::Int: OneDOneQMachine has 2 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::OneDOneQMachine) -> Float64

Get OneDOneQMachine R.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::OneDOneQMachine) -> Float64

Get OneDOneQMachine Td0_p.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::OneDOneQMachine) -> Float64

Get OneDOneQMachine Tq0_p.

PowerSystems.get_Xd — Method

get_Xd(value::OneDOneQMachine) -> Float64

Get OneDOneQMachine Xd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::OneDOneQMachine) -> Float64

Get OneDOneQMachine Xd_p.

PowerSystems.get_Xq — Method

get_Xq(value::OneDOneQMachine) -> Float64

Get OneDOneQMachine Xq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::OneDOneQMachine) -> Float64

Get OneDOneQMachine Xq_p.

PowerSystems.get_ext — Method

get_ext(value::OneDOneQMachine) -> Dict{String, Any}

Get OneDOneQMachine ext.

PowerSystems.get_n_states — Method

get_n_states(value::OneDOneQMachine) -> Int64

Get OneDOneQMachine n_states.

PowerSystems.get_states — Method

get_states(value::OneDOneQMachine) -> Vector{Symbol}

Get OneDOneQMachine states.

PowerSystems.set_R! — Method

set_R!(value::OneDOneQMachine, val::Any) -> Any

Set OneDOneQMachine R.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::OneDOneQMachine, val::Any) -> Any

Set OneDOneQMachine Td0_p.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::OneDOneQMachine, val::Any) -> Any

Set OneDOneQMachine Tq0_p.

PowerSystems.set_Xd! — Method

set_Xd!(value::OneDOneQMachine, val::Any) -> Any

Set OneDOneQMachine Xd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::OneDOneQMachine, val::Any) -> Any

Set OneDOneQMachine Xd_p.

PowerSystems.set_Xq! — Method

set_Xq!(value::OneDOneQMachine, val::Any) -> Any

Set OneDOneQMachine Xq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::OneDOneQMachine, val::Any) -> Any

Set OneDOneQMachine Xq_p.

PowerSystems.set_ext! — Method

set_ext!(value::OneDOneQMachine, val::Any) -> Any

Set OneDOneQMachine ext.

RoundRotorExponential

PowerSystems.RoundRotorExponential — Type

mutable struct RoundRotorExponential <: Machine
    base_machine::RoundRotorMachine
    saturation_coeffs::Tuple{Float64, Float64}

4-states round-rotor synchronous machine with exponential saturation: IEEE Std 1110 5.3.2 (Model 2.2). GENROE model in PSSE and PSLF.

Arguments

base_machine::RoundRotorMachine: Round Rotor Machine model.

saturation_coeffs::Tuple{Float64, Float64}`: Saturation coefficients for exponential model.

RoundRotorMachine

PowerSystems.RoundRotorMachine — Type

mutable struct RoundRotorMachine <: Machine
    R::Float64
    Td0_p::Float64
    Td0_pp::Float64
    Tq0_p::Float64
    Tq0_pp::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xl::Float64
    Se::Tuple{Float64, Float64}
    ext::Dict{String, Any}
    γ_d1::Float64
    γ_q1::Float64
    γ_d2::Float64
    γ_q2::Float64
    γ_qd::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 4-states round-rotor synchronous machine with quadratic/exponential saturation: IEEE Std 1110 5.3.2 (Model 2.2). GENROU or GENROE model in PSSE and PSLF.

Arguments

R::Float64: Armature resistance, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit. Note: Xdpp = Xqpp, validation range: (0, nothing)

Xl::Float64: Stator leakage reactance, validation range: (0, nothing)

Se::Tuple{Float64, Float64}: Saturation factor at 1 and 1.2 pu flux: S(1.0) = B(|ψ_pp|-A)^2

ext::Dict{String, Any}

γ_d1::Float64: γ_d1 parameter

γ_q1::Float64: γ_q1 parameter

γ_d2::Float64: γ_d2 parameter

γ_q2::Float64: γ_q2 parameter

γ_qd::Float64: γ_qd parameter

states::Vector{Symbol}: The states are:

eq_p: q-axis generator voltage behind the transient reactance,
ed_p: d-axis generator voltage behind the transient reactance,
ψ_kd: flux linkage in the first equivalent damping circuit in the d-axis,
ψ_kq: flux linkage in the first equivalent damping circuit in the d-axis

n_states::Int: RoundRotorMachine has 4 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine R.

PowerSystems.get_Se — Method

get_Se(value::RoundRotorMachine) -> Tuple{Float64, Float64}

Get RoundRotorMachine Se.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Td0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Td0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Tq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Tq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Xd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Xd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Xd_pp.

PowerSystems.get_Xl — Method

get_Xl(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Xl.

PowerSystems.get_Xq — Method

get_Xq(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Xq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine Xq_p.

PowerSystems.get_ext — Method

get_ext(value::RoundRotorMachine) -> Dict{String, Any}

Get RoundRotorMachine ext.

PowerSystems.get_n_states — Method

get_n_states(value::RoundRotorMachine) -> Int64

Get RoundRotorMachine n_states.

PowerSystems.get_states — Method

get_states(value::RoundRotorMachine) -> Vector{Symbol}

Get RoundRotorMachine states.

PowerSystems.get_γ_d1 — Method

get_γ_d1(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine γ_d1.

PowerSystems.get_γ_d2 — Method

get_γ_d2(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine γ_d2.

PowerSystems.get_γ_q1 — Method

get_γ_q1(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine γ_q1.

PowerSystems.get_γ_q2 — Method

get_γ_q2(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine γ_q2.

PowerSystems.get_γ_qd — Method

get_γ_qd(value::RoundRotorMachine) -> Float64

Get RoundRotorMachine γ_qd.

PowerSystems.set_R! — Method

set_R!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine R.

PowerSystems.set_Se! — Method

set_Se!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Se.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Td0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Td0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Tq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Tq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Xd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Xd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Xd_pp.

PowerSystems.set_Xl! — Method

set_Xl!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Xl.

PowerSystems.set_Xq! — Method

set_Xq!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Xq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine Xq_p.

PowerSystems.set_ext! — Method

set_ext!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine ext.

PowerSystems.set_γ_d1! — Method

set_γ_d1!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine γ_d1.

PowerSystems.set_γ_d2! — Method

set_γ_d2!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine γ_d2.

PowerSystems.set_γ_q1! — Method

set_γ_q1!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine γ_q1.

PowerSystems.set_γ_q2! — Method

set_γ_q2!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine γ_q2.

PowerSystems.set_γ_qd! — Method

set_γ_qd!(value::RoundRotorMachine, val::Any) -> Any

Set RoundRotorMachine γ_qd.

RoundRotorQuadratic

PowerSystems.RoundRotorQuadratic — Type

mutable struct RoundRotorQuadratic <: Machine
    base_machine::RoundRotorMachine
    saturation_coeffs::Tuple{Float64, Float64}

4-states round-rotor synchronous machine with quadratic saturation: IEEE Std 1110 5.3.2 (Model 2.2). GENROU model in PSSE and PSLF.

Arguments

base_machine::RoundRotorMachine: Round Rotor Machine model.

saturation_coeffs::Tuple{Float64, Float64}`: Saturation coefficients for quadratic model.

SalientPoleExponential

PowerSystems.SalientPoleExponential — Type

mutable struct SalientPoleExponential <: Machine
    base_machine::SalientPoleMachine
    saturation_coeffs::Tuple{Float64, Float64}

3-states salient-pole synchronous machine with exponential saturation: IEEE Std 1110 5.3.2 (Model 2.1). GENSAE in PSSE and PSLF.

Arguments:

base_machine::SalientPoleMachine: Salient Pole Machine model.

saturation_coeffs::Tuple{Float64, Float64}`: Saturation coefficients for exponential model.

SalientPoleMachine

PowerSystems.SalientPoleMachine — Type

mutable struct SalientPoleMachine <: Machine
    R::Float64
    Td0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xd_pp::Float64
    Xl::Float64
    Se::Tuple{Float64, Float64}
    ext::Dict{String, Any}
    γ_d1::Float64
    γ_q1::Float64
    γ_d2::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 3-states salient-pole synchronous machine with quadratic/exponential saturation: IEEE Std 1110 5.3.1 (Model 2.1). GENSAL or GENSAE model in PSSE and PSLF.

Arguments

R::Float64: Armature resistance, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit. Note: Xdpp = Xqpp, validation range: (0, nothing)

Xl::Float64: Stator leakage reactance, validation range: (0, nothing)

Se::Tuple{Float64, Float64}: Saturation factor at 1 and 1.2 pu flux: Se(eqp) = B(eqp-A)^2

ext::Dict{String, Any}

γ_d1::Float64: γ_d1 parameter

γ_q1::Float64: γ_q1 parameter

γ_d2::Float64: γ_d2 parameter

states::Vector{Symbol}: The states are:

eq_p: q-axis generator voltage behind the transient reactance,
ψ_kd: flux linkage in the first equivalent damping circuit in the d-axis,
ψq_pp: phasonf of the subtransient flux linkage in the q-axis

n_states::Int: SalientPoleMachine has 3 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine R.

PowerSystems.get_Se — Method

get_Se(value::SalientPoleMachine) -> Tuple{Float64, Float64}

Get SalientPoleMachine Se.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine Td0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine Td0_pp.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine Tq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine Xd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine Xd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine Xd_pp.

PowerSystems.get_Xl — Method

get_Xl(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine Xl.

PowerSystems.get_Xq — Method

get_Xq(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine Xq.

PowerSystems.get_ext — Method

get_ext(value::SalientPoleMachine) -> Dict{String, Any}

Get SalientPoleMachine ext.

PowerSystems.get_n_states — Method

get_n_states(value::SalientPoleMachine) -> Int64

Get SalientPoleMachine n_states.

PowerSystems.get_states — Method

get_states(value::SalientPoleMachine) -> Vector{Symbol}

Get SalientPoleMachine states.

PowerSystems.get_γ_d1 — Method

get_γ_d1(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine γ_d1.

PowerSystems.get_γ_d2 — Method

get_γ_d2(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine γ_d2.

PowerSystems.get_γ_q1 — Method

get_γ_q1(value::SalientPoleMachine) -> Float64

Get SalientPoleMachine γ_q1.

PowerSystems.set_R! — Method

set_R!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine R.

PowerSystems.set_Se! — Method

set_Se!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine Se.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine Td0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine Td0_pp.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine Tq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine Xd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine Xd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine Xd_pp.

PowerSystems.set_Xl! — Method

set_Xl!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine Xl.

PowerSystems.set_Xq! — Method

set_Xq!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine Xq.

PowerSystems.set_ext! — Method

set_ext!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine ext.

PowerSystems.set_γ_d1! — Method

set_γ_d1!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine γ_d1.

PowerSystems.set_γ_d2! — Method

set_γ_d2!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine γ_d2.

PowerSystems.set_γ_q1! — Method

set_γ_q1!(value::SalientPoleMachine, val::Any) -> Any

Set SalientPoleMachine γ_q1.

SalientPoleQuadratic

PowerSystems.SalientPoleQuadratic — Type

mutable struct SalientPoleQuadratic <: Machine
    base_machine::SalientPoleMachine
    saturation_coeffs::Tuple{Float64, Float64}

3-states salient-pole synchronous machine with exponential saturation: IEEE Std 1110 5.3.2 (Model 2.1). GENSAL in PSSE and PSLF.

Arguments:

base_machine::SalientPoleMachine: Salient Pole Machine model.

saturation_coeffs::Tuple{Float64, Float64}`: Saturation coefficients for quadratic model.

SimpleAFMachine

PowerSystems.SimpleAFMachine — Type

mutable struct SimpleAFMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xq_pp::Float64
    Td0_p::Float64
    Tq0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 4-states simplified Anderson-Fouad (SimpleAFMachine) model. The derivative of stator fluxes (ψd and ψq) is neglected and ωψd = ψd and ωψq = ψq is assumed (i.e. ω=1.0). This is standard when transmission network dynamics is neglected. If transmission dynamics is considered use the full order Anderson Fouad model.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_pp::Float64: Sub-Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage,
eq_pp: q-axis subtransient voltage,
ed_pp: d-axis subtransient voltage

n_states::Int: SimpleAFMachine has 4 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine R.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Td0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Td0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Tq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Tq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Xd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Xd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Xd_pp.

PowerSystems.get_Xq — Method

get_Xq(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Xq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Xq_p.

PowerSystems.get_Xq_pp — Method

get_Xq_pp(value::SimpleAFMachine) -> Float64

Get SimpleAFMachine Xq_pp.

PowerSystems.get_ext — Method

get_ext(value::SimpleAFMachine) -> Dict{String, Any}

Get SimpleAFMachine ext.

PowerSystems.get_n_states — Method

get_n_states(value::SimpleAFMachine) -> Int64

Get SimpleAFMachine n_states.

PowerSystems.get_states — Method

get_states(value::SimpleAFMachine) -> Vector{Symbol}

Get SimpleAFMachine states.

PowerSystems.set_R! — Method

set_R!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine R.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Td0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Td0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Tq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Tq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Xd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Xd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Xd_pp.

PowerSystems.set_Xq! — Method

set_Xq!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Xq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Xq_p.

PowerSystems.set_Xq_pp! — Method

set_Xq_pp!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine Xq_pp.

PowerSystems.set_ext! — Method

set_ext!(value::SimpleAFMachine, val::Any) -> Any

Set SimpleAFMachine ext.

SimpleFullMachine

PowerSystems.SimpleFullMachine — Type

mutable struct SimpleFullMachine <: Machine
    R::Float64
    R_f::Float64
    R_1d::Float64
    R_1q::Float64
    L_d::Float64
    L_q::Float64
    L_ad::Float64
    L_aq::Float64
    L_f1d::Float64
    L_ff::Float64
    L_1d::Float64
    L_1q::Float64
    ext::Dict{String, Any}
    inv_d_fluxlink::Array{Float64,2}
    inv_q_fluxlink::Array{Float64,2}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameter of a full order flux stator-rotor model without zero sequence flux in the stator. The derivative of stator fluxes (ψd and ψq) is neglected. This is standard when transmission network dynamics is neglected. Only one q-axis damping circuit is considered. All per unit are in machine per unit. Refer to Chapter 3 of Power System Stability and Control by P. Kundur or Chapter 11 of Power System Dynamics: Stability and Control, by J. Machowski, J. Bialek and J. Bumby, for more details. Note that the models are somewhat different (but equivalent) due to the different Park Transformation used in both books.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

R_f::Float64: Field rotor winding resistance in per unit, validation range: (0, nothing)

R_1d::Float64: Damping rotor winding resistance on d-axis in per unit. This value is denoted as RD in Machowski., validation range: (0, nothing)

R_1q::Float64: Damping rotor winding resistance on q-axis in per unit. This value is denoted as RQ in Machowski., validation range: (0, nothing)

L_d::Float64: Inductance of fictitious damping that represent the effect of the three-phase stator winding in the d-axis of the rotor, in per unit. This value is denoted as Lad + Ll in Kundur (and Ld in Machowski)., validation range: (0, nothing)

L_q::Float64: Inductance of fictitious damping that represent the effect of the three-phase stator winding in the q-axis of the rotor, in per unit. This value is denoted as Laq + Ll in Kundur., validation range: (0, nothing)

L_ad::Float64: Mutual inductance between stator winding and rotor field (and damping) winding inductance on d-axis, in per unit, validation range: (0, nothing)

L_aq::Float64: Mutual inductance between stator winding and rotor damping winding inductance on q-axis, in per unit, validation range: (0, nothing)

L_f1d::Float64: Mutual inductance between rotor field winding and rotor damping winding inductance on d-axis, in per unit, validation range: (0, nothing)

L_ff::Float64: Field rotor winding inductance, in per unit, validation range: (0, nothing)

L_1d::Float64: Inductance of the d-axis rotor damping circuit, in per unit, validation range: (0, nothing)

L_1q::Float64: Inductance of the q-axis rotor damping circuit, in per unit, validation range: (0, nothing)

ext::Dict{String, Any}

inv_d_fluxlink::Array{Float64,2}: Equations 3.127, 3.130, 3.131 From Kundur

inv_q_fluxlink::Array{Float64,2}: Equations 3.128, 3.132 From Kundur

states::Vector{Symbol}: The states are:

ψf: field rotor flux,
ψ1d: d-axis rotor damping flux,
ψ1q: q-axis rotor damping flux

n_states::Int: SimpleFullMachine has 3 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_L_1d — Method

get_L_1d(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine L_1d.

PowerSystems.get_L_1q — Method

get_L_1q(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine L_1q.

PowerSystems.get_L_ad — Method

get_L_ad(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine L_ad.

PowerSystems.get_L_aq — Method

get_L_aq(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine L_aq.

PowerSystems.get_L_d — Method

get_L_d(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine L_d.

PowerSystems.get_L_f1d — Method

get_L_f1d(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine L_f1d.

PowerSystems.get_L_ff — Method

get_L_ff(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine L_ff.

PowerSystems.get_L_q — Method

get_L_q(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine L_q.

PowerSystems.get_R — Method

get_R(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine R.

PowerSystems.get_R_1d — Method

get_R_1d(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine R_1d.

PowerSystems.get_R_1q — Method

get_R_1q(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine R_1q.

PowerSystems.get_R_f — Method

get_R_f(value::SimpleFullMachine) -> Float64

Get SimpleFullMachine R_f.

PowerSystems.get_ext — Method

get_ext(value::SimpleFullMachine) -> Dict{String, Any}

Get SimpleFullMachine ext.

PowerSystems.get_inv_d_fluxlink — Method

get_inv_d_fluxlink(value::SimpleFullMachine) -> Matrix{Float64}

Get SimpleFullMachine inv_d_fluxlink.

PowerSystems.get_inv_q_fluxlink — Method

get_inv_q_fluxlink(value::SimpleFullMachine) -> Matrix{Float64}

Get SimpleFullMachine inv_q_fluxlink.

PowerSystems.get_n_states — Method

get_n_states(value::SimpleFullMachine) -> Int64

Get SimpleFullMachine n_states.

PowerSystems.get_states — Method

get_states(value::SimpleFullMachine) -> Vector{Symbol}

Get SimpleFullMachine states.

PowerSystems.set_L_1d! — Method

set_L_1d!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine L_1d.

PowerSystems.set_L_1q! — Method

set_L_1q!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine L_1q.

PowerSystems.set_L_ad! — Method

set_L_ad!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine L_ad.

PowerSystems.set_L_aq! — Method

set_L_aq!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine L_aq.

PowerSystems.set_L_d! — Method

set_L_d!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine L_d.

PowerSystems.set_L_f1d! — Method

set_L_f1d!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine L_f1d.

PowerSystems.set_L_ff! — Method

set_L_ff!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine L_ff.

PowerSystems.set_L_q! — Method

set_L_q!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine L_q.

PowerSystems.set_R! — Method

set_R!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine R.

PowerSystems.set_R_1d! — Method

set_R_1d!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine R_1d.

PowerSystems.set_R_1q! — Method

set_R_1q!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine R_1q.

PowerSystems.set_R_f! — Method

set_R_f!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine R_f.

PowerSystems.set_ext! — Method

set_ext!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine ext.

PowerSystems.set_inv_d_fluxlink! — Method

set_inv_d_fluxlink!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine inv_d_fluxlink.

PowerSystems.set_inv_q_fluxlink! — Method

set_inv_q_fluxlink!(value::SimpleFullMachine, val::Any) -> Any

Set SimpleFullMachine inv_q_fluxlink.

SimpleMarconatoMachine

PowerSystems.SimpleMarconatoMachine — Type

mutable struct SimpleMarconatoMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xq_pp::Float64
    Td0_p::Float64
    Tq0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    T_AA::Float64
    ext::Dict{String, Any}
    γd::Float64
    γq::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 4-states synchronous machine: Simplified Marconato model The derivative of stator fluxes (ψd and ψq) is neglected and ωψd = ψd and ωψq = ψq is assumed (i.e. ω=1.0). This is standard when transmission network dynamics is neglected.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_pp::Float64: Sub-Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

T_AA::Float64: Time constant of d-axis additional leakage, validation range: (0, nothing)

ext::Dict{String, Any}

γd::Float64

γq::Float64

states::Vector{Symbol}: The states are:

eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage,
eq_pp: q-axis subtransient voltage,
ed_pp: d-axis subtransient voltage

n_states::Int: SimpleMarconatoMachine has 4 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine R.

PowerSystems.get_T_AA — Method

get_T_AA(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine T_AA.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Td0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Td0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Tq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Tq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Xd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Xd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Xd_pp.

PowerSystems.get_Xq — Method

get_Xq(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Xq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Xq_p.

PowerSystems.get_Xq_pp — Method

get_Xq_pp(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine Xq_pp.

PowerSystems.get_ext — Method

get_ext(value::SimpleMarconatoMachine) -> Dict{String, Any}

Get SimpleMarconatoMachine ext.

PowerSystems.get_n_states — Method

get_n_states(value::SimpleMarconatoMachine) -> Int64

Get SimpleMarconatoMachine n_states.

PowerSystems.get_states — Method

get_states(value::SimpleMarconatoMachine) -> Vector{Symbol}

Get SimpleMarconatoMachine states.

PowerSystems.get_γd — Method

get_γd(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine γd.

PowerSystems.get_γq — Method

get_γq(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachine γq.

PowerSystems.set_R! — Method

set_R!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine R.

PowerSystems.set_T_AA! — Method

set_T_AA!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine T_AA.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Td0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Td0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Tq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Tq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Xd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Xd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Xd_pp.

PowerSystems.set_Xq! — Method

set_Xq!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Xq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Xq_p.

PowerSystems.set_Xq_pp! — Method

set_Xq_pp!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine Xq_pp.

PowerSystems.set_ext! — Method

set_ext!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine ext.

PowerSystems.set_γd! — Method

set_γd!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine γd.

PowerSystems.set_γq! — Method

set_γq!(value::SimpleMarconatoMachine, val::Any) -> Any

Set SimpleMarconatoMachine γq.