MonitoredLine

PowerSystems.MonitoredLine — Type

mutable struct MonitoredLine <: ACTransmission
    name::String
    available::Bool
    active_power_flow::Float64
    reactive_power_flow::Float64
    arc::Arc
    r::Float64
    x::Float64
    b::FromTo
    flow_limits::FromTo_ToFrom
    rating::Float64
    angle_limits::MinMax
    rating_b::Union{Nothing, Float64}
    rating_c::Union{Nothing, Float64}
    g::FromTo
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

An AC transmission line with additional power flow constraints specified by the system operator, more restrictive than the line's thermal limits.

For example, monitored lines can be used to restrict line flow following a contingency elsewhere in the network. See the flow_limits parameter. If monitoring is not needed, see Line

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow::Float64: Initial condition of active power flow on the line (MW)

reactive_power_flow::Float64: Initial condition of reactive power flow on the line (MVAR)

arc::Arc: An Arc defining this line from a bus to another bus

r::Float64: Resistance in pu (SYSTEM_BASE), validation range: (0, 4)

x::Float64: Reactance in pu (SYSTEM_BASE), validation range: (0, 4)

b::FromTo: Shunt susceptance in pu (SYSTEM_BASE), specified both on the from and to ends of the line. These are commonly modeled with the same value, validation range: (0, 2)

flow_limits::FromTo_ToFrom: Minimum and maximum permissable flow on the line (MVA), if different from the thermal rating defined in rating

rating::Float64: Thermal rating (MVA). Flow through the transformer must be between -rating and rating. When defining a line before it is attached to a System, rating must be in pu (SYSTEM_BASE) using the base power of the System it will be attached to

angle_limits::MinMax: Minimum and maximum angle limits (radians)

rating_b::Union{Nothing, Float64}: (default: nothing) Second current rating; entered in MVA.

rating_c::Union{Nothing, Float64}: (default: nothing) Third current rating; entered in MVA.

g::FromTo: (default: (from=0.0, to=0.0)) Shunt conductance in pu (SYSTEM_BASE), specified both on the from and to ends of the line. These are commonly modeled with the same value, validation range: (0, 100)

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_available — Method

get_available(value::MonitoredLine) -> Bool

Get MonitoredLine available.

InfrastructureSystems.get_name — Method

get_name(value::MonitoredLine) -> String

Get MonitoredLine name.

InfrastructureSystems.set_available! — Method

set_available!(value::MonitoredLine, val) -> Any

Set MonitoredLine available.

PowerSystems.get_active_power_flow — Method

get_active_power_flow(value::MonitoredLine) -> Float64

Get MonitoredLine active_power_flow.

PowerSystems.get_angle_limits — Method

get_angle_limits(
    value::MonitoredLine
) -> @NamedTuple{min::Float64, max::Float64}

Get MonitoredLine angle_limits.

PowerSystems.get_arc — Method

get_arc(value::MonitoredLine) -> Arc

Get MonitoredLine arc.

PowerSystems.get_b — Method

get_b(
    value::MonitoredLine
) -> @NamedTuple{from::Float64, to::Float64}

Get MonitoredLine b.

PowerSystems.get_ext — Method

get_ext(value::MonitoredLine) -> Dict{String, Any}

Get MonitoredLine ext.

PowerSystems.get_flow_limits — Method

get_flow_limits(
    value::MonitoredLine
) -> @NamedTuple{from_to::Float64, to_from::Float64}

Get MonitoredLine flow_limits.

PowerSystems.get_g — Method

get_g(
    value::MonitoredLine
) -> @NamedTuple{from::Float64, to::Float64}

Get MonitoredLine g.

PowerSystems.get_r — Method

get_r(value::MonitoredLine) -> Float64

Get MonitoredLine r.

PowerSystems.get_rating — Method

get_rating(value::MonitoredLine) -> Float64

Get MonitoredLine rating.

PowerSystems.get_rating_b — Method

get_rating_b(
    value::MonitoredLine
) -> Union{Nothing, Float64}

Get MonitoredLine rating_b.

PowerSystems.get_rating_c — Method

get_rating_c(
    value::MonitoredLine
) -> Union{Nothing, Float64}

Get MonitoredLine rating_c.

PowerSystems.get_reactive_power_flow — Method

get_reactive_power_flow(value::MonitoredLine) -> Float64

Get MonitoredLine reactive_power_flow.

PowerSystems.get_services — Method

get_services(value::MonitoredLine) -> Vector{Service}

Get MonitoredLine services.

PowerSystems.get_x — Method

get_x(value::MonitoredLine) -> Float64

Get MonitoredLine x.

PowerSystems.set_active_power_flow! — Method

set_active_power_flow!(value::MonitoredLine, val) -> Any

Set MonitoredLine active_power_flow.

PowerSystems.set_angle_limits! — Method

set_angle_limits!(value::MonitoredLine, val) -> Any

Set MonitoredLine angle_limits.

PowerSystems.set_arc! — Method

set_arc!(value::MonitoredLine, val) -> Any

Set MonitoredLine arc.

PowerSystems.set_b! — Method

set_b!(value::MonitoredLine, val) -> Any

Set MonitoredLine b.

PowerSystems.set_ext! — Method

set_ext!(value::MonitoredLine, val) -> Any

Set MonitoredLine ext.

PowerSystems.set_flow_limits! — Method

set_flow_limits!(value::MonitoredLine, val) -> Any

Set MonitoredLine flow_limits.

PowerSystems.set_g! — Method

set_g!(value::MonitoredLine, val) -> Any

Set MonitoredLine g.

PowerSystems.set_r! — Method

set_r!(value::MonitoredLine, val) -> Any

Set MonitoredLine r.

PowerSystems.set_rating! — Method

set_rating!(value::MonitoredLine, val) -> Any

Set MonitoredLine rating.

PowerSystems.set_rating_b! — Method

set_rating_b!(value::MonitoredLine, val) -> Any

Set MonitoredLine rating_b.

PowerSystems.set_rating_c! — Method

set_rating_c!(value::MonitoredLine, val) -> Any

Set MonitoredLine rating_c.

PowerSystems.set_reactive_power_flow! — Method

set_reactive_power_flow!(value::MonitoredLine, val) -> Any

Set MonitoredLine reactive_power_flow.

PowerSystems.set_services! — Method

set_services!(value::MonitoredLine, val) -> Any

Set MonitoredLine services.

PowerSystems.set_x! — Method

set_x!(value::MonitoredLine, val) -> Any

Set MonitoredLine x.