PSS

IEEEST

PowerSystems.IEEEST — Type

mutable struct IEEEST <: PSS
    input_code::Int
    remote_bus_control::Int
    A1::Float64
    A2::Float64
    A3::Float64
    A4::Float64
    A5::Float64
    A6::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    T4::Float64
    T5::Float64
    T6::Float64
    Ks::Float64
    Ls_lim::Tuple{Float64, Float64}
    Vcu::Float64
    Vcl::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

IEEE Stabilizing Model PSS.

Arguments

input_code::Int: Code input for stabilizar, validation range: (1, 6), action if invalid: error

remote_bus_control::Int: Remote Bus number for control.

A1::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

A2::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

A3::Float64: Filter coefficient, validation range: (0, nothing)

A4::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

A5::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

A6::Float64: Filter coefficient, validation range: (0, nothing), action if invalid: warn

T1::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T2::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T3::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T4::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T5::Float64: Time constant, validation range: (0, 10), action if invalid: warn

T6::Float64: Time constant, validation range: (eps(), 2.0), action if invalid: error

Ks::Float64: Proportional gain, validation range: (0, nothing), action if invalid: warn

Ls_lim::Tuple{Float64, Float64}: PSS output limits for regulator output (Ls_min, Ls_max)

Vcu::Float64: Cutoff limiter upper bound, validation range: (0, 1.25), action if invalid: warn

Vcl::Float64: Cutoff limiter lower bound, validation range: (0, 1.0), action if invalid: warn

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

x_p1: 1st filter integration,
x_p2: 2nd filter integration, 
x_p3: 3rd filter integration, 
x_p4: 4rd filter integration, 
x_p5: T1/T2 lead-lag integrator, 
x_p6: T3/T4 lead-lag integrator, 
:x_p7 last integer,

n_states::Int: IEEEST has 7 states

states_types::Vector{StateTypes}: IEEEST has 7 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_A1 — Method

get_A1(value::IEEEST) -> Float64

Get IEEEST A1.

PowerSystems.get_A2 — Method

get_A2(value::IEEEST) -> Float64

Get IEEEST A2.

PowerSystems.get_A3 — Method

get_A3(value::IEEEST) -> Float64

Get IEEEST A3.

PowerSystems.get_A4 — Method

get_A4(value::IEEEST) -> Float64

Get IEEEST A4.

PowerSystems.get_A5 — Method

get_A5(value::IEEEST) -> Float64

Get IEEEST A5.

PowerSystems.get_A6 — Method

get_A6(value::IEEEST) -> Float64

Get IEEEST A6.

PowerSystems.get_Ks — Method

get_Ks(value::IEEEST) -> Float64

Get IEEEST Ks.

PowerSystems.get_Ls_lim — Method

get_Ls_lim(value::IEEEST) -> Tuple{Float64, Float64}

Get IEEEST Ls_lim.

PowerSystems.get_T1 — Method

get_T1(value::IEEEST) -> Float64

Get IEEEST T1.

PowerSystems.get_T2 — Method

get_T2(value::IEEEST) -> Float64

Get IEEEST T2.

PowerSystems.get_T3 — Method

get_T3(value::IEEEST) -> Float64

Get IEEEST T3.

PowerSystems.get_T4 — Method

get_T4(value::IEEEST) -> Float64

Get IEEEST T4.

PowerSystems.get_T5 — Method

get_T5(value::IEEEST) -> Float64

Get IEEEST T5.

PowerSystems.get_T6 — Method

get_T6(value::IEEEST) -> Float64

Get IEEEST T6.

PowerSystems.get_Vcl — Method

get_Vcl(value::IEEEST) -> Float64

Get IEEEST Vcl.

PowerSystems.get_Vcu — Method

get_Vcu(value::IEEEST) -> Float64

Get IEEEST Vcu.

PowerSystems.get_ext — Method

get_ext(value::IEEEST) -> Dict{String, Any}

Get IEEEST ext.

PowerSystems.get_input_code — Method

get_input_code(value::IEEEST) -> Int64

Get IEEEST input_code.

PowerSystems.get_n_states — Method

get_n_states(value::IEEEST) -> Int64

Get IEEEST n_states.

PowerSystems.get_remote_bus_control — Method

get_remote_bus_control(value::IEEEST) -> Int64

Get IEEEST remote_bus_control.

PowerSystems.get_states — Method

get_states(value::IEEEST) -> Vector{Symbol}

Get IEEEST states.

PowerSystems.get_states_types — Method

get_states_types(value::IEEEST) -> Vector{StateTypes}

Get IEEEST states_types.

PowerSystems.set_A1! — Method

set_A1!(value::IEEEST, val) -> Any

Set IEEEST A1.

PowerSystems.set_A2! — Method

set_A2!(value::IEEEST, val) -> Any

Set IEEEST A2.

PowerSystems.set_A3! — Method

set_A3!(value::IEEEST, val) -> Any

Set IEEEST A3.

PowerSystems.set_A4! — Method

set_A4!(value::IEEEST, val) -> Any

Set IEEEST A4.

PowerSystems.set_A5! — Method

set_A5!(value::IEEEST, val) -> Any

Set IEEEST A5.

PowerSystems.set_A6! — Method

set_A6!(value::IEEEST, val) -> Any

Set IEEEST A6.

PowerSystems.set_Ks! — Method

set_Ks!(value::IEEEST, val) -> Any

Set IEEEST Ks.

PowerSystems.set_Ls_lim! — Method

set_Ls_lim!(value::IEEEST, val) -> Any

Set IEEEST Ls_lim.

PowerSystems.set_T1! — Method

set_T1!(value::IEEEST, val) -> Any

Set IEEEST T1.

PowerSystems.set_T2! — Method

set_T2!(value::IEEEST, val) -> Any

Set IEEEST T2.

PowerSystems.set_T3! — Method

set_T3!(value::IEEEST, val) -> Any

Set IEEEST T3.

PowerSystems.set_T4! — Method

set_T4!(value::IEEEST, val) -> Any

Set IEEEST T4.

PowerSystems.set_T5! — Method

set_T5!(value::IEEEST, val) -> Any

Set IEEEST T5.

PowerSystems.set_T6! — Method

set_T6!(value::IEEEST, val) -> Any

Set IEEEST T6.

PowerSystems.set_Vcl! — Method

set_Vcl!(value::IEEEST, val) -> Any

Set IEEEST Vcl.

PowerSystems.set_Vcu! — Method

set_Vcu!(value::IEEEST, val) -> Any

Set IEEEST Vcu.

PowerSystems.set_ext! — Method

set_ext!(value::IEEEST, val) -> Any

Set IEEEST ext.

PowerSystems.set_input_code! — Method

set_input_code!(value::IEEEST, val) -> Any

Set IEEEST input_code.

PowerSystems.set_remote_bus_control! — Method

set_remote_bus_control!(value::IEEEST, val) -> Any

Set IEEEST remote_bus_control.

PowerSystems.set_states_types! — Method

set_states_types!(value::IEEEST, val) -> Any

Set IEEEST states_types.

PSSFixed

PowerSystems.PSSFixed — Type

mutable struct PSSFixed <: PSS
    V_pss::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of a PSS that returns a fixed voltage to add to the reference for the AVR

Arguments

V_pss::Float64: Fixed voltage stabilization signal, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}

n_states::Int: PSSFixed has no states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_V_pss — Method

get_V_pss(value::PSSFixed) -> Float64

Get PSSFixed V_pss.

PowerSystems.get_ext — Method

get_ext(value::PSSFixed) -> Dict{String, Any}

Get PSSFixed ext.

PowerSystems.get_n_states — Method

get_n_states(value::PSSFixed) -> Int64

Get PSSFixed n_states.

PowerSystems.get_states — Method

get_states(value::PSSFixed) -> Vector{Symbol}

Get PSSFixed states.

PowerSystems.set_V_pss! — Method

set_V_pss!(value::PSSFixed, val) -> Any

Set PSSFixed V_pss.

PowerSystems.set_ext! — Method

set_ext!(value::PSSFixed, val) -> Any

Set PSSFixed ext.

PSSSimple

PowerSystems.PSSSimple — Type

mutable struct PSSSimple <: PSS
    K_ω::Float64
    K_p::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of a PSS that returns a proportional droop voltage to add to the reference for the AVR

Arguments

K_ω::Float64: Proportional gain for frequency, validation range: (0, nothing)

K_p::Float64: Proportional gain for active power, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}

n_states::Int: PSSSimple has no states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_K_p — Method

get_K_p(value::PSSSimple) -> Float64

Get PSSSimple K_p.

PowerSystems.get_K_ω — Method

get_K_ω(value::PSSSimple) -> Float64

Get PSSSimple K_ω.

PowerSystems.get_ext — Method

get_ext(value::PSSSimple) -> Dict{String, Any}

Get PSSSimple ext.

PowerSystems.get_n_states — Method

get_n_states(value::PSSSimple) -> Int64

Get PSSSimple n_states.

PowerSystems.get_states — Method

get_states(value::PSSSimple) -> Vector{Symbol}

Get PSSSimple states.

PowerSystems.set_K_p! — Method

set_K_p!(value::PSSSimple, val) -> Any

Set PSSSimple K_p.

PowerSystems.set_K_ω! — Method

set_K_ω!(value::PSSSimple, val) -> Any

Set PSSSimple K_ω.

PowerSystems.set_ext! — Method

set_ext!(value::PSSSimple, val) -> Any

Set PSSSimple ext.

STAB1

PowerSystems.STAB1 — Type

mutable struct STAB1 <: PSS
    KT::Float64
    T::Float64
    T1T3::Float64
    T3::Float64
    T2T4::Float64
    T4::Float64
    H_lim::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    states_types::Vector{StateTypes}
    internal::InfrastructureSystemsInternal
end

Speed-Sensitive Stabilizing Model

Arguments

KT::Float64: K/T for washout filter, validation range: (0, nothing), action if invalid: warn

T::Float64: Time constant for washout filter, validation range: (0.01, nothing), action if invalid: warn

T1T3::Float64: Time constant division T1/T3, validation range: (0, nothing)

T3::Float64: Time constant, validation range: (0.01, nothing), action if invalid: warn

T2T4::Float64: Time constant division T2/T4, validation range: (0, nothing), action if invalid: warn

T4::Float64: Time constant, validation range: (0.01, nothing), action if invalid: warn

H_lim::Float64: PSS output limit, validation range: (0, 0.5), action if invalid: warn

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

x_p1: washout filter,
x_p2: T1/T3 lead-lag block, 
x_p3: T2/T4 lead-lag block,

n_states::Int: STAB1 has 3 states

states_types::Vector{StateTypes}: STAB1 has 3 differential states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_H_lim — Method

get_H_lim(value::STAB1) -> Float64

Get STAB1 H_lim.

PowerSystems.get_KT — Method

get_KT(value::STAB1) -> Float64

Get STAB1 KT.

PowerSystems.get_T — Method

get_T(value::STAB1) -> Float64

Get STAB1 T.

PowerSystems.get_T1T3 — Method

get_T1T3(value::STAB1) -> Float64

Get STAB1 T1T3.

PowerSystems.get_T2T4 — Method

get_T2T4(value::STAB1) -> Float64

Get STAB1 T2T4.

PowerSystems.get_T3 — Method

get_T3(value::STAB1) -> Float64

Get STAB1 T3.

PowerSystems.get_T4 — Method

get_T4(value::STAB1) -> Float64

Get STAB1 T4.

PowerSystems.get_ext — Method

get_ext(value::STAB1) -> Dict{String, Any}

Get STAB1 ext.

PowerSystems.get_n_states — Method

get_n_states(value::STAB1) -> Int64

Get STAB1 n_states.

PowerSystems.get_states — Method

get_states(value::STAB1) -> Vector{Symbol}

Get STAB1 states.

PowerSystems.get_states_types — Method

get_states_types(value::STAB1) -> Vector{StateTypes}

Get STAB1 states_types.

PowerSystems.set_H_lim! — Method

set_H_lim!(value::STAB1, val) -> Any

Set STAB1 H_lim.

PowerSystems.set_KT! — Method

set_KT!(value::STAB1, val) -> Any

Set STAB1 KT.

PowerSystems.set_T! — Method

set_T!(value::STAB1, val) -> Any

Set STAB1 T.

PowerSystems.set_T1T3! — Method

set_T1T3!(value::STAB1, val) -> Any

Set STAB1 T1T3.

PowerSystems.set_T2T4! — Method

set_T2T4!(value::STAB1, val) -> Any

Set STAB1 T2T4.

PowerSystems.set_T3! — Method

set_T3!(value::STAB1, val) -> Any

Set STAB1 T3.

PowerSystems.set_T4! — Method

set_T4!(value::STAB1, val) -> Any

Set STAB1 T4.

PowerSystems.set_ext! — Method

set_ext!(value::STAB1, val) -> Any

Set STAB1 ext.

PowerSystems.set_states_types! — Method

set_states_types!(value::STAB1, val) -> Any

Set STAB1 states_types.