Source

PowerSystems.Source — Type

mutable struct Source <: StaticInjection
    name::String
    available::Bool
    bus::ACBus
    active_power::Float64
    reactive_power::Float64
    R_th::Float64
    X_th::Float64
    internal_voltage::Float64
    internal_angle::Float64
    dynamic_injector::Union{Nothing, DynamicInjection}
    services::Vector{Service}
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

An infinite bus with a constant voltage output.

Commonly used in dynamics simulations to represent a very large machine on a single bus

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

bus::ACBus: Bus that this component is connected to

active_power::Float64: Initial active power set point of the unit in MW. For power flow, this is the steady state operating point of the system. For production cost modeling, this may or may not be used as the initial starting point for the solver, depending on the solver used

reactive_power::Float64: Initial reactive power set point of the unit (MVAR)

R_th::Float64: Source Thevenin resistance, validation range: (0, nothing)

X_th::Float64: Source Thevenin reactance, validation range: (0, nothing)

internal_voltage::Float64: (default: 1.0) Internal Voltage (pu), validation range: (0, nothing)

internal_angle::Float64: (default: 0.0) Internal Angle

dynamic_injector::Union{Nothing, DynamicInjection}: (default: nothing) corresponding dynamic injection device

services::Vector{Service}: (default: Device[]) Services that this device contributes to

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude. See Adding additional fields

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::Source) -> String

Get Source name.

PowerSystems.get_R_th — Method

get_R_th(value::Source) -> Float64

Get Source R_th.

PowerSystems.get_X_th — Method

get_X_th(value::Source) -> Float64

Get Source X_th.

PowerSystems.get_active_power — Method

get_active_power(value::Source) -> Float64

Get Source active_power.

PowerSystems.get_available — Method

get_available(value::Source) -> Bool

Get Source available.

PowerSystems.get_bus — Method

get_bus(value::Source) -> ACBus

Get Source bus.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(
    value::Source
) -> Union{Nothing, DynamicInjection}

Get Source dynamic_injector.

PowerSystems.get_ext — Method

get_ext(value::Source) -> Dict{String, Any}

Get Source ext.

PowerSystems.get_internal_angle — Method

get_internal_angle(value::Source) -> Float64

Get Source internal_angle.

PowerSystems.get_internal_voltage — Method

get_internal_voltage(value::Source) -> Float64

Get Source internal_voltage.

PowerSystems.get_reactive_power — Method

get_reactive_power(value::Source) -> Float64

Get Source reactive_power.

PowerSystems.get_services — Method

get_services(value::Source) -> Vector{Service}

Get Source services.

PowerSystems.set_R_th! — Method

set_R_th!(value::Source, val) -> Any

Set Source R_th.

PowerSystems.set_X_th! — Method

set_X_th!(value::Source, val) -> Any

Set Source X_th.

PowerSystems.set_active_power! — Method

set_active_power!(value::Source, val) -> Any

Set Source active_power.

PowerSystems.set_available! — Method

set_available!(value::Source, val) -> Any

Set Source available.

PowerSystems.set_bus! — Method

set_bus!(value::Source, val) -> Any

Set Source bus.

PowerSystems.set_ext! — Method

set_ext!(value::Source, val) -> Any

Set Source ext.

PowerSystems.set_internal_angle! — Method

set_internal_angle!(value::Source, val) -> Any

Set Source internal_angle.

PowerSystems.set_internal_voltage! — Method

set_internal_voltage!(value::Source, val) -> Any

Set Source internal_voltage.

PowerSystems.set_reactive_power! — Method

set_reactive_power!(value::Source, val) -> Any

Set Source reactive_power.

PowerSystems.set_services! — Method

set_services!(value::Source, val) -> Any

Set Source services.