TransmissionInterface

PowerSystems.TransmissionInterface — Type

mutable struct TransmissionInterface <: Service
    name::String
    available::Bool
    active_power_flow_limits::MinMax
    violation_penalty::Float64
    direction_mapping::Dict{String, Int}
    internal::InfrastructureSystemsInternal
end

A collection of branches that make up an interface or corridor for the transfer of power, such as between different Areas or LoadZones.

The interface can be used to constrain the power flow across it

Arguments

name::String: Name of the component. Components of the same type (e.g., PowerLoad) must have unique names, but components of different types (e.g., PowerLoad and ACBus) can have the same name

available::Bool: Indicator of whether the component is connected and online (true) or disconnected, offline, or down (false). Unavailable components are excluded during simulations

active_power_flow_limits::MinMax: Minimum and maximum active power flow limits on the interface (MW)

violation_penalty::Float64: (default: INFINITE_COST) Penalty cost for violating the flow limits in the interface

direction_mapping::Dict{String, Int}: (default: Dict{String, Int}()) Dictionary of the line names in the interface and their direction of flow (1 or -1) relative to the flow of the interface

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

InfrastructureSystems.get_name — Method

get_name(value::TransmissionInterface) -> String

Get TransmissionInterface name.

PowerSystems.get_active_power_flow_limits — Method

get_active_power_flow_limits(
    value::TransmissionInterface
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get TransmissionInterface active_power_flow_limits.

PowerSystems.get_available — Method

get_available(value::TransmissionInterface) -> Bool

Get TransmissionInterface available.

PowerSystems.get_direction_mapping — Method

get_direction_mapping(
    value::TransmissionInterface
) -> Dict{String, Int64}

Get TransmissionInterface direction_mapping.

PowerSystems.get_violation_penalty — Method

get_violation_penalty(
    value::TransmissionInterface
) -> Float64

Get TransmissionInterface violation_penalty.

PowerSystems.set_active_power_flow_limits! — Method

set_active_power_flow_limits!(
    value::TransmissionInterface,
    val
) -> Any

Set TransmissionInterface active_power_flow_limits.

PowerSystems.set_available! — Method

set_available!(value::TransmissionInterface, val) -> Any

Set TransmissionInterface available.

PowerSystems.set_direction_mapping! — Method

set_direction_mapping!(
    value::TransmissionInterface,
    val
) -> Any

Set TransmissionInterface direction_mapping.

PowerSystems.set_violation_penalty! — Method

set_violation_penalty!(
    value::TransmissionInterface,
    val
) -> Any

Set TransmissionInterface violation_penalty.