TurbineGov

GasTG

PowerSystems.GasTG — Type

mutable struct GasTG <: TurbineGov
    R::Float64
    T1::Float64
    T2::Float64
    T3::Float64
    AT::Float64
    Kt::Float64
    V_lim::Tuple{Float64, Float64}
    D_turb::Float64
    Load_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    internal::InfrastructureSystemsInternal
end

Parameters of Gas Turbine-Governor. GAST in PSSE and GAST_PTI in PowerWorld.

Arguments

R::Float64: Speed droop parameter, validation range: ("eps()", nothing)

T1::Float64: Governor time constant in s, validation range: ("eps()", nothing)

T2::Float64: Combustion chamber time constant, validation range: ("eps()", nothing)

T3::Float64: Load limit time constant (exhaust gas measurement time), validation range: ("eps()", nothing)

AT::Float64: Ambient temperature load limit, validation range: (0, nothing)

Kt::Float64: Load limit feedback gain, validation range: (0, nothing)

V_lim::Tuple{Float64, Float64}: Operational control limits on fuel valve opening (Vmin, Vmax)

D_turb::Float64: Speed damping coefficient of gas turbine rotor, validation range: (0, nothing)

Load_ref::Float64: Reference Load Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the GAST model are:

x_g1: Fuel valve opening,
x_g2: Fuel flow,
x_g3: Exhaust temperature load

n_states::Int64: GasTG has 3 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_AT — Method

get_AT(value::GasTG) -> Float64

Get GasTG AT.

PowerSystems.get_D_turb — Method

get_D_turb(value::GasTG) -> Float64

Get GasTG D_turb.

PowerSystems.get_Kt — Method

get_Kt(value::GasTG) -> Float64

Get GasTG Kt.

PowerSystems.get_Load_ref — Method

get_Load_ref(value::GasTG) -> Float64

Get GasTG Load_ref.

PowerSystems.get_R — Method

get_R(value::GasTG) -> Float64

Get GasTG R.

PowerSystems.get_T1 — Method

get_T1(value::GasTG) -> Float64

Get GasTG T1.

PowerSystems.get_T2 — Method

get_T2(value::GasTG) -> Float64

Get GasTG T2.

PowerSystems.get_T3 — Method

get_T3(value::GasTG) -> Float64

Get GasTG T3.

PowerSystems.get_V_lim — Method

get_V_lim(value::GasTG) -> Tuple{Float64,Float64}

Get GasTG V_lim.

PowerSystems.get_ext — Method

get_ext(value::GasTG) -> Dict{String,Any}

Get GasTG ext.

PowerSystems.get_internal — Method

get_internal(value::GasTG) -> InfrastructureSystems.InfrastructureSystemsInternal

Get GasTG internal.

PowerSystems.get_n_states — Method

get_n_states(value::GasTG) -> Int64

Get GasTG n_states.

PowerSystems.get_states — Method

get_states(value::GasTG) -> Array{Symbol,1}

Get GasTG states.

PowerSystems.set_AT! — Method

set_AT!(value::GasTG, val::Any) -> Any

Set GasTG AT.

PowerSystems.set_D_turb! — Method

set_D_turb!(value::GasTG, val::Any) -> Any

Set GasTG D_turb.

PowerSystems.set_Kt! — Method

set_Kt!(value::GasTG, val::Any) -> Any

Set GasTG Kt.

PowerSystems.set_Load_ref! — Method

set_Load_ref!(value::GasTG, val::Any) -> Any

Set GasTG Load_ref.

PowerSystems.set_R! — Method

set_R!(value::GasTG, val::Any) -> Any

Set GasTG R.

PowerSystems.set_T1! — Method

set_T1!(value::GasTG, val::Any) -> Any

Set GasTG T1.

PowerSystems.set_T2! — Method

set_T2!(value::GasTG, val::Any) -> Any

Set GasTG T2.

PowerSystems.set_T3! — Method

set_T3!(value::GasTG, val::Any) -> Any

Set GasTG T3.

PowerSystems.set_V_lim! — Method

set_V_lim!(value::GasTG, val::Any) -> Any

Set GasTG V_lim.

PowerSystems.set_ext! — Method

set_ext!(value::GasTG, val::Any) -> Any

Set GasTG ext.

PowerSystems.set_internal! — Method

set_internal!(value::GasTG, val::Any) -> Any

Set GasTG internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::GasTG, val::Any) -> Any

Set GasTG n_states.

PowerSystems.set_states! — Method

set_states!(value::GasTG, val::Any) -> Any

Set GasTG states.

TGFixed

PowerSystems.TGFixed — Type

mutable struct TGFixed <: TurbineGov
    efficiency::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    internal::InfrastructureSystemsInternal
end

Parameters of a fixed Turbine Governor that returns a fixed mechanical torque given by the product of P_ref*efficiency

Arguments

efficiency::Float64: Efficiency factor that multiplies P_ref, validation range: (0, nothing)

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}

n_states::Int64: TGFixed has no states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_P_ref — Method

get_P_ref(value::TGFixed) -> Float64

Get TGFixed P_ref.

PowerSystems.get_efficiency — Method

get_efficiency(value::TGFixed) -> Float64

Get TGFixed efficiency.

PowerSystems.get_ext — Method

get_ext(value::TGFixed) -> Dict{String,Any}

Get TGFixed ext.

PowerSystems.get_internal — Method

get_internal(value::TGFixed) -> InfrastructureSystems.InfrastructureSystemsInternal

Get TGFixed internal.

PowerSystems.get_n_states — Method

get_n_states(value::TGFixed) -> Int64

Get TGFixed n_states.

PowerSystems.get_states — Method

get_states(value::TGFixed) -> Array{Symbol,1}

Get TGFixed states.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::TGFixed, val::Any) -> Any

Set TGFixed P_ref.

PowerSystems.set_efficiency! — Method

set_efficiency!(value::TGFixed, val::Any) -> Any

Set TGFixed efficiency.

PowerSystems.set_ext! — Method

set_ext!(value::TGFixed, val::Any) -> Any

Set TGFixed ext.

PowerSystems.set_internal! — Method

set_internal!(value::TGFixed, val::Any) -> Any

Set TGFixed internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::TGFixed, val::Any) -> Any

Set TGFixed n_states.

PowerSystems.set_states! — Method

set_states!(value::TGFixed, val::Any) -> Any

Set TGFixed states.

TGTypeI

PowerSystems.TGTypeI — Type

mutable struct TGTypeI <: TurbineGov
    R::Float64
    Ts::Float64
    Tc::Float64
    T3::Float64
    T4::Float64
    T5::Float64
    P_min::Float64
    P_max::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    internal::InfrastructureSystemsInternal
end

Parameters of a Turbine Governor Type I.

Arguments

R::Float64: Droop parameter, validation range: (0, nothing)

Ts::Float64: Governor time constant, validation range: (0, nothing)

Tc::Float64: Servo time constant, validation range: (0, nothing)

T3::Float64: Transient gain time constant, validation range: (0, nothing)

T4::Float64: Power fraction time constant, validation range: (0, nothing)

T5::Float64: Reheat time constant, validation range: (0, nothing)

P_min::Float64: Min Power into the Governor, validation range: (0, nothing)

P_max::Float64: Max Power into the Governor, validation range: (0, nothing)

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the TGTypeI model are:

x_g1: Governor state,
x_g2: Servo state,
x_g3: Reheat state

n_states::Int64: TGTypeI has 3 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_P_max — Method

get_P_max(value::TGTypeI) -> Float64

Get TGTypeI P_max.

PowerSystems.get_P_min — Method

get_P_min(value::TGTypeI) -> Float64

Get TGTypeI P_min.

PowerSystems.get_P_ref — Method

get_P_ref(value::TGTypeI) -> Float64

Get TGTypeI P_ref.

PowerSystems.get_R — Method

get_R(value::TGTypeI) -> Float64

Get TGTypeI R.

PowerSystems.get_T3 — Method

get_T3(value::TGTypeI) -> Float64

Get TGTypeI T3.

PowerSystems.get_T4 — Method

get_T4(value::TGTypeI) -> Float64

Get TGTypeI T4.

PowerSystems.get_T5 — Method

get_T5(value::TGTypeI) -> Float64

Get TGTypeI T5.

PowerSystems.get_Tc — Method

get_Tc(value::TGTypeI) -> Float64

Get TGTypeI Tc.

PowerSystems.get_Ts — Method

get_Ts(value::TGTypeI) -> Float64

Get TGTypeI Ts.

PowerSystems.get_ext — Method

get_ext(value::TGTypeI) -> Dict{String,Any}

Get TGTypeI ext.

PowerSystems.get_internal — Method

get_internal(value::TGTypeI) -> InfrastructureSystems.InfrastructureSystemsInternal

Get TGTypeI internal.

PowerSystems.get_n_states — Method

get_n_states(value::TGTypeI) -> Int64

Get TGTypeI n_states.

PowerSystems.get_states — Method

get_states(value::TGTypeI) -> Array{Symbol,1}

Get TGTypeI states.

PowerSystems.set_P_max! — Method

set_P_max!(value::TGTypeI, val::Any) -> Any

Set TGTypeI P_max.

PowerSystems.set_P_min! — Method

set_P_min!(value::TGTypeI, val::Any) -> Any

Set TGTypeI P_min.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::TGTypeI, val::Any) -> Any

Set TGTypeI P_ref.

PowerSystems.set_R! — Method

set_R!(value::TGTypeI, val::Any) -> Any

Set TGTypeI R.

PowerSystems.set_T3! — Method

set_T3!(value::TGTypeI, val::Any) -> Any

Set TGTypeI T3.

PowerSystems.set_T4! — Method

set_T4!(value::TGTypeI, val::Any) -> Any

Set TGTypeI T4.

PowerSystems.set_T5! — Method

set_T5!(value::TGTypeI, val::Any) -> Any

Set TGTypeI T5.

PowerSystems.set_Tc! — Method

set_Tc!(value::TGTypeI, val::Any) -> Any

Set TGTypeI Tc.

PowerSystems.set_Ts! — Method

set_Ts!(value::TGTypeI, val::Any) -> Any

Set TGTypeI Ts.

PowerSystems.set_ext! — Method

set_ext!(value::TGTypeI, val::Any) -> Any

Set TGTypeI ext.

PowerSystems.set_internal! — Method

set_internal!(value::TGTypeI, val::Any) -> Any

Set TGTypeI internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::TGTypeI, val::Any) -> Any

Set TGTypeI n_states.

PowerSystems.set_states! — Method

set_states!(value::TGTypeI, val::Any) -> Any

Set TGTypeI states.

TGTypeII

PowerSystems.TGTypeII — Type

mutable struct TGTypeII <: TurbineGov
    R::Float64
    T1::Float64
    T2::Float64
    τ_min::Float64
    τ_max::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int64
    internal::InfrastructureSystemsInternal
end

Parameters of a Turbine Governor Type II.

Arguments

R::Float64: Droop parameter, validation range: (0, nothing)

T1::Float64: Transient gain time constant, validation range: (0, nothing)

T2::Float64: Power fraction time constant, validation range: (0, nothing)

τ_min::Float64: Min Power into the Governor, validation range: (0, nothing)

τ_max::Float64: Max Power into the Governor, validation range: (0, nothing)

P_ref::Float64: Reference Power Set-point, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states of the TGTypeI model are:

x_g1: lead-lag state

n_states::Int64: TGTypeII has 1 state

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_P_ref — Method

get_P_ref(value::TGTypeII) -> Float64

Get TGTypeII P_ref.

PowerSystems.get_R — Method

get_R(value::TGTypeII) -> Float64

Get TGTypeII R.

PowerSystems.get_T1 — Method

get_T1(value::TGTypeII) -> Float64

Get TGTypeII T1.

PowerSystems.get_T2 — Method

get_T2(value::TGTypeII) -> Float64

Get TGTypeII T2.

PowerSystems.get_ext — Method

get_ext(value::TGTypeII) -> Dict{String,Any}

Get TGTypeII ext.

PowerSystems.get_internal — Method

get_internal(value::TGTypeII) -> InfrastructureSystems.InfrastructureSystemsInternal

Get TGTypeII internal.

PowerSystems.get_n_states — Method

get_n_states(value::TGTypeII) -> Int64

Get TGTypeII n_states.

PowerSystems.get_states — Method

get_states(value::TGTypeII) -> Array{Symbol,1}

Get TGTypeII states.

PowerSystems.get_τ_max — Method

get_τ_max(value::TGTypeII) -> Float64

Get TGTypeII τ_max.

PowerSystems.get_τ_min — Method

get_τ_min(value::TGTypeII) -> Float64

Get TGTypeII τ_min.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::TGTypeII, val::Any) -> Any

Set TGTypeII P_ref.

PowerSystems.set_R! — Method

set_R!(value::TGTypeII, val::Any) -> Any

Set TGTypeII R.

PowerSystems.set_T1! — Method

set_T1!(value::TGTypeII, val::Any) -> Any

Set TGTypeII T1.

PowerSystems.set_T2! — Method

set_T2!(value::TGTypeII, val::Any) -> Any

Set TGTypeII T2.

PowerSystems.set_ext! — Method

set_ext!(value::TGTypeII, val::Any) -> Any

Set TGTypeII ext.

PowerSystems.set_internal! — Method

set_internal!(value::TGTypeII, val::Any) -> Any

Set TGTypeII internal.

PowerSystems.set_n_states! — Method

set_n_states!(value::TGTypeII, val::Any) -> Any

Set TGTypeII n_states.

PowerSystems.set_states! — Method

set_states!(value::TGTypeII, val::Any) -> Any

Set TGTypeII states.

PowerSystems.set_τ_max! — Method

set_τ_max!(value::TGTypeII, val::Any) -> Any

Set TGTypeII τ_max.

PowerSystems.set_τ_min! — Method

set_τ_min!(value::TGTypeII, val::Any) -> Any

Set TGTypeII τ_min.