VariableReserveNonSpinning

PowerSystems.VariableReserveNonSpinning — Type

mutable struct VariableReserveNonSpinning <: ReserveNonSpinning
    name::String
    available::Bool
    time_frame::Float64
    requirement::Float64
    sustained_time::Float64
    max_output_fraction::Float64
    max_participation_factor::Float64
    deployed_fraction::Float64
    ext::Dict{String, Any}
    time_series_container::InfrastructureSystems.TimeSeriesContainer
    internal::InfrastructureSystemsInternal
end

Data Structure for the procurement products for system simulations.

Arguments

name::String

available::Bool

time_frame::Float64: the saturation time_frame in minutes to provide reserve contribution, validation range: (0, nothing), action if invalid: error

requirement::Float64: the required quantity of the product should be scaled by a TimeSeriesData

sustained_time::Float64: the time in secounds reserve contribution must sustained at a specified level, validation range: (0, nothing), action if invalid: error

max_output_fraction::Float64: the maximum fraction of the device's output that can be assigned to the service, validation range: (0, 1), action if invalid: error

max_participation_factor::Float64: the maximum limit of reserve contribution per device, validation range: (0, 1), action if invalid: error

deployed_fraction::Float64: Fraction of ancillary services participation deployed from the assignment, validation range: (0, 1), action if invalid: error

ext::Dict{String, Any}

time_series_container::InfrastructureSystems.TimeSeriesContainer: internal time_series storage

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

InfrastructureSystems.get_name — Method

get_name(value::VariableReserveNonSpinning) -> String

Get VariableReserveNonSpinning name.

InfrastructureSystems.get_time_series_container — Method

get_time_series_container(
    value::VariableReserveNonSpinning
) -> InfrastructureSystems.TimeSeriesContainer

Get VariableReserveNonSpinning time_series_container.

PowerSystems.get_available — Method

get_available(value::VariableReserveNonSpinning) -> Bool

Get VariableReserveNonSpinning available.

PowerSystems.get_deployed_fraction — Method

get_deployed_fraction(
    value::VariableReserveNonSpinning
) -> Float64

Get VariableReserveNonSpinning deployed_fraction.

PowerSystems.get_ext — Method

get_ext(
    value::VariableReserveNonSpinning
) -> Dict{String, Any}

Get VariableReserveNonSpinning ext.

PowerSystems.get_max_output_fraction — Method

get_max_output_fraction(
    value::VariableReserveNonSpinning
) -> Float64

Get VariableReserveNonSpinning max_output_fraction.

PowerSystems.get_max_participation_factor — Method

get_max_participation_factor(
    value::VariableReserveNonSpinning
) -> Float64

Get VariableReserveNonSpinning max_participation_factor.

PowerSystems.get_requirement — Method

get_requirement(value::VariableReserveNonSpinning) -> Any

Get VariableReserveNonSpinning requirement.

PowerSystems.get_sustained_time — Method

get_sustained_time(
    value::VariableReserveNonSpinning
) -> Float64

Get VariableReserveNonSpinning sustained_time.

PowerSystems.get_time_frame — Method

get_time_frame(value::VariableReserveNonSpinning) -> Float64

Get VariableReserveNonSpinning time_frame.

PowerSystems.set_available! — Method

set_available!(
    value::VariableReserveNonSpinning,
    val
) -> Any

Set VariableReserveNonSpinning available.

PowerSystems.set_deployed_fraction! — Method

set_deployed_fraction!(
    value::VariableReserveNonSpinning,
    val
) -> Any

Set VariableReserveNonSpinning deployed_fraction.

PowerSystems.set_ext! — Method

set_ext!(value::VariableReserveNonSpinning, val) -> Any

Set VariableReserveNonSpinning ext.

PowerSystems.set_max_output_fraction! — Method

set_max_output_fraction!(
    value::VariableReserveNonSpinning,
    val
) -> Any

Set VariableReserveNonSpinning max_output_fraction.

PowerSystems.set_max_participation_factor! — Method

set_max_participation_factor!(
    value::VariableReserveNonSpinning,
    val
) -> Any

Set VariableReserveNonSpinning max_participation_factor.

PowerSystems.set_requirement! — Method

set_requirement!(
    value::VariableReserveNonSpinning,
    val
) -> Any

Set VariableReserveNonSpinning requirement.

PowerSystems.set_sustained_time! — Method

set_sustained_time!(
    value::VariableReserveNonSpinning,
    val
) -> Any

Set VariableReserveNonSpinning sustained_time.

PowerSystems.set_time_frame! — Method

set_time_frame!(
    value::VariableReserveNonSpinning,
    val
) -> Any

Set VariableReserveNonSpinning time_frame.

PowerSystems.set_time_series_container! — Method

set_time_series_container!(
    value::VariableReserveNonSpinning,
    val
) -> Any

Set VariableReserveNonSpinning time_series_container.