VariableReserveNonSpinning

PowerSystems.VariableReserveNonSpinning — Type

mutable struct VariableReserveNonSpinning <: ReserveNonSpinning
    name::String
    available::Bool
    time_frame::Float64
    requirement::Float64
    ext::Dict{String, Any}
    time_series_container::InfrastructureSystems.TimeSeriesContainer
    internal::InfrastructureSystemsInternal
end

Data Structure for the procurement products for system simulations.

Arguments

name::String

available::Bool

time_frame::Float64: the relative saturation time_frame, validation range: (0, nothing), action if invalid: error

requirement::Float64: the required quantity of the product should be scaled by a TimeSeriesData

ext::Dict{String, Any}

time_series_container::InfrastructureSystems.TimeSeriesContainer: internal time_series storage

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_available — Method

get_available(value::VariableReserveNonSpinning) -> Bool

Get VariableReserveNonSpinning available.

PowerSystems.get_ext — Method

get_ext(value::VariableReserveNonSpinning) -> Dict{String,Any}

Get VariableReserveNonSpinning ext.

PowerSystems.get_requirement — Method

get_requirement(value::VariableReserveNonSpinning) -> Float64

Get VariableReserveNonSpinning requirement.

PowerSystems.get_time_frame — Method

get_time_frame(value::VariableReserveNonSpinning) -> Float64

Get VariableReserveNonSpinning time_frame.

PowerSystems.set_available! — Method

set_available!(value::VariableReserveNonSpinning, val::Any) -> Any

Set VariableReserveNonSpinning available.

PowerSystems.set_ext! — Method

set_ext!(value::VariableReserveNonSpinning, val::Any) -> Any

Set VariableReserveNonSpinning ext.

PowerSystems.set_requirement! — Method

set_requirement!(value::VariableReserveNonSpinning, val::Any) -> Any

Set VariableReserveNonSpinning requirement.

PowerSystems.set_time_frame! — Method

set_time_frame!(value::VariableReserveNonSpinning, val::Any) -> Any

Set VariableReserveNonSpinning time_frame.