Reserves

Static Reserve

PowerSystems.StaticReserve — Type

mutable struct StaticReserve{T <: ReserveDirection} <: Reserve{T}
    name::String
    available::Bool
    time_frame::Float64
    requirement::Float64
    ext::Dict{String, Any}
    internal::InfrastructureSystemsInternal
end

Data Structure for a proportional reserve product for system simulations.

Arguments

name::String

available::Bool

time_frame::Float64: the relative saturation time_frame, validation range: (0, nothing), action if invalid: error

requirement::Float64: the static value of required reserves in system p.u., validation range: (0, nothing), action if invalid: error

ext::Dict{String, Any}

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

InfrastructureSystems.get_name — Method

get_name(value::StaticReserve) -> String

Get StaticReserve name.

PowerSystems.get_available — Method

get_available(value::StaticReserve) -> Bool

Get StaticReserve available.

PowerSystems.get_ext — Method

get_ext(value::StaticReserve) -> Dict{String, Any}

Get StaticReserve ext.

PowerSystems.get_requirement — Method

get_requirement(value::StaticReserve) -> Float64

Get StaticReserve requirement.

PowerSystems.get_time_frame — Method

get_time_frame(value::StaticReserve) -> Float64

Get StaticReserve time_frame.

PowerSystems.set_available! — Method

set_available!(value::StaticReserve, val) -> Any

Set StaticReserve available.

PowerSystems.set_ext! — Method

set_ext!(value::StaticReserve, val) -> Any

Set StaticReserve ext.

PowerSystems.set_requirement! — Method

set_requirement!(value::StaticReserve, val) -> Any

Set StaticReserve requirement.

PowerSystems.set_time_frame! — Method

set_time_frame!(value::StaticReserve, val) -> Any

Set StaticReserve time_frame.

Static Reserve Group

PowerSystems.StaticReserveGroup — Type

mutable struct StaticReserveGroup{T <: ReserveDirection} <: Service
    name::String
    available::Bool
    requirement::Float64
    ext::Dict{String, Any}
    contributing_services::Vector{Service}
    internal::InfrastructureSystemsInternal
end

Data Structure for a group reserve product for system simulations.

Arguments

name::String

available::Bool

requirement::Float64: the static value of required reserves in system p.u., validation range: (0, nothing), action if invalid: error

ext::Dict{String, Any}

contributing_services::Vector{Service}: Services that contribute for this requirement constraint

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

InfrastructureSystems.get_name — Method

get_name(value::StaticReserveGroup) -> String

Get StaticReserveGroup name.

PowerSystems.get_available — Method

get_available(value::StaticReserveGroup) -> Bool

Get StaticReserveGroup available.

PowerSystems.get_contributing_services — Method

get_contributing_services(value::StaticReserveGroup) -> Vector{Service}

Get StaticReserveGroup contributing_services.

PowerSystems.get_ext — Method

get_ext(value::StaticReserveGroup) -> Dict{String, Any}

Get StaticReserveGroup ext.

PowerSystems.get_requirement — Method

get_requirement(value::StaticReserveGroup) -> Float64

Get StaticReserveGroup requirement.

PowerSystems.set_available! — Method

set_available!(value::StaticReserveGroup, val) -> Any

Set StaticReserveGroup available.

PowerSystems.set_ext! — Method

set_ext!(value::StaticReserveGroup, val) -> Any

Set StaticReserveGroup ext.

PowerSystems.set_requirement! — Method

set_requirement!(value::StaticReserveGroup, val) -> Any

Set StaticReserveGroup requirement.

Variable Reserve

PowerSystems.VariableReserve — Type

mutable struct VariableReserve{T <: ReserveDirection} <: Reserve{T}
    name::String
    available::Bool
    time_frame::Float64
    requirement::Float64
    ext::Dict{String, Any}
    time_series_container::InfrastructureSystems.TimeSeriesContainer
    internal::InfrastructureSystemsInternal
end

Data Structure for the procurement products for system simulations.

Arguments

name::String

available::Bool

time_frame::Float64: the relative saturation time_frame, validation range: (0, nothing), action if invalid: error

requirement::Float64: the required quantity of the product should be scaled by a TimeSeriesData

ext::Dict{String, Any}

time_series_container::InfrastructureSystems.TimeSeriesContainer: internal time_series storage

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

InfrastructureSystems.get_name — Method

get_name(value::VariableReserve) -> String

Get VariableReserve name.

InfrastructureSystems.get_time_series_container — Method

get_time_series_container(value::VariableReserve) -> InfrastructureSystems.TimeSeriesContainer

Get VariableReserve time_series_container.

PowerSystems.get_available — Method

get_available(value::VariableReserve) -> Bool

Get VariableReserve available.

PowerSystems.get_ext — Method

get_ext(value::VariableReserve) -> Dict{String, Any}

Get VariableReserve ext.

PowerSystems.get_requirement — Method

get_requirement(value::VariableReserve) -> Float64

Get VariableReserve requirement.

PowerSystems.get_time_frame — Method

get_time_frame(value::VariableReserve) -> Float64

Get VariableReserve time_frame.

PowerSystems.set_available! — Method

set_available!(value::VariableReserve, val) -> Any

Set VariableReserve available.

PowerSystems.set_ext! — Method

set_ext!(value::VariableReserve, val) -> Any

Set VariableReserve ext.

PowerSystems.set_requirement! — Method

set_requirement!(value::VariableReserve, val) -> Any

Set VariableReserve requirement.

PowerSystems.set_time_frame! — Method

set_time_frame!(value::VariableReserve, val) -> Any

Set VariableReserve time_frame.

PowerSystems.set_time_series_container! — Method

set_time_series_container!(value::VariableReserve, val) -> Any

Set VariableReserve time_series_container.

Reserve Demand Curve

PowerSystems.ReserveDemandCurve — Type

mutable struct ReserveDemandCurve{T <: ReserveDirection} <: Reserve{T}
    variable::Union{Nothing, IS.TimeSeriesKey}
    name::String
    available::Bool
    time_frame::Float64
    ext::Dict{String, Any}
    time_series_container::InfrastructureSystems.TimeSeriesContainer
    internal::InfrastructureSystemsInternal
end

Data Structure for a operating reserve with demand curve product for system simulations.

Arguments

variable::Union{Nothing, IS.TimeSeriesKey}: Variable Cost TimeSeriesKey

name::String

available::Bool

time_frame::Float64: the relative saturation time_frame, validation range: (0, nothing), action if invalid: error

ext::Dict{String, Any}

time_series_container::InfrastructureSystems.TimeSeriesContainer: internal time_series storage

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

InfrastructureSystems.get_name — Method

get_name(value::ReserveDemandCurve) -> String

Get ReserveDemandCurve name.

InfrastructureSystems.get_time_series_container — Method

get_time_series_container(value::ReserveDemandCurve) -> InfrastructureSystems.TimeSeriesContainer

Get ReserveDemandCurve time_series_container.

PowerSystems.get_available — Method

get_available(value::ReserveDemandCurve) -> Bool

Get ReserveDemandCurve available.

PowerSystems.get_ext — Method

get_ext(value::ReserveDemandCurve) -> Dict{String, Any}

Get ReserveDemandCurve ext.

PowerSystems.get_time_frame — Method

get_time_frame(value::ReserveDemandCurve) -> Float64

Get ReserveDemandCurve time_frame.

PowerSystems.get_variable — Method

get_variable(value::ReserveDemandCurve) -> Union{Nothing, InfrastructureSystems.TimeSeriesKey}

Get ReserveDemandCurve variable.

PowerSystems.set_available! — Method

set_available!(value::ReserveDemandCurve, val) -> Any

Set ReserveDemandCurve available.

PowerSystems.set_ext! — Method

set_ext!(value::ReserveDemandCurve, val) -> Any

Set ReserveDemandCurve ext.

PowerSystems.set_time_frame! — Method

set_time_frame!(value::ReserveDemandCurve, val) -> Any

Set ReserveDemandCurve time_frame.

PowerSystems.set_time_series_container! — Method

set_time_series_container!(value::ReserveDemandCurve, val) -> Any

Set ReserveDemandCurve time_series_container.

PowerSystems.set_variable! — Method

set_variable!(value::ReserveDemandCurve, val) -> Any

Set ReserveDemandCurve variable.